% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotFunctions.R
\name{AddFacet}
\alias{AddFacet}
\title{AddFacet}
\usage{
AddFacet(p, fv1 = NULL, fv2 = NULL, Exclude = "None", Debug = FALSE)
}
\arguments{
\item{Debug}{= FALSE}

\item{data}{Source data.table}

\item{ShapColNames}{Names of the columns that contain shap values you want included}

\item{FacetVar1}{Column name}

\item{FacetVar2}{Column name}

\item{AggMethod}{A string for aggregating shapely values for importances. Choices include, 'mean', 'absmean', 'meanabs', 'sd', 'median', 'absmedian', 'medianabs'}

\item{TopN}{The number of variables to plot}
}
\description{
Add up to two facet variables for plots
}
\seealso{
Other Graphics: 
\code{\link{AppModelInsights}()},
\code{\link{AutoPlotter}()},
\code{\link{BarPlot}()},
\code{\link{BoxPlot}()},
\code{\link{ChartTheme}()},
\code{\link{HistPlot}()},
\code{\link{ViolinPlot}()},
\code{\link{multiplot}()}
}
\author{
Adrian Antico
}
\concept{Graphics}
