% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimpleApps.R
\name{App_BoxPlotsOverTime}
\alias{App_BoxPlotsOverTime}
\title{App_BoxPlotsOverTime}
\usage{
App_BoxPlotsOverTime(
  data,
  FeatureNames = NULL,
  GroupVariables = NULL,
  FilterVariable = NULL,
  DateName = NULL,
  DateBreaks = "1 day",
  DateLabels = "\%b \%d",
  Title = NULL,
  SubTitle = NULL,
  Debug = FALSE
)
}
\arguments{
\item{data}{Source data.table}

\item{FeatureNames}{Character vector of feature column names}

\item{GroupVariables}{Character vector of group column names}

\item{FilterVariable}{Variable to use to filter data before plotting}

\item{DateName}{Character scalar of the date column name}

\item{Title}{Title for plot}

\item{SubTitle}{SubTitle for plot}

\item{Debug}{FALSE}
}
\description{
Simple shiny app for viewing boxplots over time. You can use up to 3 categorical variables to filter by and one additional variable to filter as a bonus.
}
\examples{
\dontrun{
# Pull Data
data <- data.table::fread(system.file('tests/QA_DataSets/ThreeGroup-FC-Walmart-XREG3.csv', package = "RemixAutoML"))
data[, Date := as.Date(Date)]

# Run App
RemixAutoML::App_BoxPlotsOverTime(
  data,
  FeatureNames = names(data)[5L:ncol(data)],
  GroupVariables = names(data)[seq_len(3L)],
  FilterVariable = 'XREG1',
  DateName = 'Date',
  DateBreaks = '14 day',
  DateLabels = '\%Y \%b \%d',
  Title = 'Time Series Distribution',
  SubTitle = 'Dist. by Day',
  Debug = TRUE)

# FeatureNames = names(data)[5L:ncol(data)]
# GroupVariables = names(data)[seq_len(3L)]
# FilterVariable = 'XREG1'
# DateName = 'Date'
# DateBreaks = '1 day'
# DateLabels = '\%Y \%b \%d'
# Title = 'Time Series Distribution'
# SubTitle = 'Dist. by Day'
# Debug = TRUE
}

}
\seealso{
Other Shiny: 
\code{\link{ArgNullCheck2}()},
\code{\link{ArgNullCheck}()},
\code{\link{BlankRow}()},
\code{\link{DateInput}()},
\code{\link{GenerateEvaluationMetrics}()},
\code{\link{NumericInput}()},
\code{\link{PickerInput_GetLevels}()},
\code{\link{PickerInput}()},
\code{\link{PreparePlotData}()},
\code{\link{ReactiveLoadCSV}()},
\code{\link{ReturnParam}()},
\code{\link{StoreArgs}()},
\code{\link{TextInput}()},
\code{\link{observeEventLoad}()}
}
\author{
Adrian Antico
}
\concept{Shiny}
