% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimpleApps.R
\name{App_TimeSeriesPlots}
\alias{App_TimeSeriesPlots}
\title{App_TimeSeriesPlots}
\usage{
App_TimeSeriesPlots(
  data,
  FeatureNames = NULL,
  GroupVariables = NULL,
  FilterVariable = NULL,
  DateName = NULL,
  AppTitle = "Time Series Plotting",
  AppWidth = 12L,
  LogoWidth = "1300px",
  LogoHeight = "100px",
  Box1Color = "navy",
  Box2Color = "purple",
  Box3Color = "aqua",
  CreatePlotButtonColor = "default",
  UpdatePlotButtonColor = "default",
  ResetPlotButtonColor = "default",
  Debug = FALSE
)
}
\arguments{
\item{data}{Source data.table}

\item{FeatureNames}{Character vector of feature column names}

\item{GroupVariables}{Character vector of group column names}

\item{FilterVariable}{Variable to use to filter data before plotting}

\item{DateName}{Character scalar of the date column name}

\item{AppTitle}{Defaults to 'Time Series Plotting'}

\item{AppWidth}{Width of boxes}

\item{Box1Color}{Choose from 'red', 'yellow', 'aqua', 'blue', 'light-blue', 'green', 'navy', 'teal', 'olive', 'lime', 'orange', 'fuchsia', 'purple', 'maroon', 'black'}

\item{Box2Color}{Choose from 'red', 'yellow', 'aqua', 'blue', 'light-blue', 'green', 'navy', 'teal', 'olive', 'lime', 'orange', 'fuchsia', 'purple', 'maroon', 'black'}

\item{Box3Color}{Choose from 'red', 'yellow', 'aqua', 'blue', 'light-blue', 'green', 'navy', 'teal', 'olive', 'lime', 'orange', 'fuchsia', 'purple', 'maroon', 'black'}

\item{CreatePlotButtonColor}{Choose from 'default', 'primary', 'warning', 'danger', 'success', 'royal'}

\item{UpdatePlotButtonColor}{Choose from 'default', 'primary', 'warning', 'danger', 'success', 'royal'}

\item{ResetPlotButtonColor}{Choose from 'default', 'primary', 'warning', 'danger', 'success', 'royal'}

\item{Debug}{FALSE}
}
\description{
Simple shiny app for viewing time series plots with box plots or line plots. You can use up to 3 categorical variables to filter by and one additional variable to filter as a bonus. You'll need to have shiny, shinyWidgets, htmltools
}
\examples{
\dontrun{
# Pull Data
data <- data.table::fread(system.file('tests/QA_DataSets/ThreeGroup-FC-Walmart-XREG3.csv', package = "RemixAutoML"))
data[, Date := as.Date(Date)]

# Run App
RemixAutoML::App_TimeSeriesPlots(
  data,
  FeatureNames = names(data)[5L:ncol(data)],
  GroupVariables = names(data)[seq_len(3L)],
  FilterVariable = 'XREG1',
  DateName = 'Date',
  AppTitle = 'Time Series Plotting',
  AppWidth = 12L,
  LogoWidth = '1300px',
  LogoHeight = '100px',
  Box1Color = 'navy',
  Box2Color = 'purple',
  Box3Color = 'aqua',
  CreatePlotButtonColor = 'default',
  UpdatePlotButtonColor = 'default',
  ResetPlotButtonColor = 'default',
  Debug = TRUE)

# FeatureNames = names(data)[5L:ncol(data)]
# GroupVariables = names(data)[seq_len(3L)]
# FilterVariable = 'XREG1'
# DateName = 'Date'
# Debug = TRUE
# AppWidth = 12L,
# LogoWidth = '1300px',
# LogoHeight = '100px',
# Box1Color = 'navy',
# Box2Color = 'purple',
# Box3Color = 'aqua',
# CreatePlotButtonColor = 'default',
# UpdatePlotButtonColor = 'default',
# ResetPlotButtonColor = 'default',
# Debug = TRUE
}

}
\seealso{
Other Shiny: 
\code{\link{ArgNullCheck2}()},
\code{\link{ArgNullCheck}()},
\code{\link{BlankRow}()},
\code{\link{DateInput}()},
\code{\link{GenerateEvaluationMetrics}()},
\code{\link{NumericInput}()},
\code{\link{PickerInput_GetLevels}()},
\code{\link{PickerInput}()},
\code{\link{PreparePlotData}()},
\code{\link{ReactiveLoadCSV}()},
\code{\link{ReturnParam}()},
\code{\link{StoreArgs}()},
\code{\link{TextInput}()},
\code{\link{observeEventLoad}()}
}
\author{
Adrian Antico
}
\concept{Shiny}
