% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RemixAutoML.R
\name{Apply_Asin}
\alias{Apply_Asin}
\title{Inverse Asin Transformation}
\usage{
Apply_Asin(x)
}
\arguments{
\item{x}{The data in numerical vector form}
}
\value{
Asin results
}
\description{
Inverse Asin Transformation
}
\seealso{
Other Feature Engineering: \code{\link{Apply_Asinh}},
  \code{\link{Apply_BoxCox}}, \code{\link{Apply_Logit}},
  \code{\link{Apply_YeoJohnson}},
  \code{\link{AutoDataPartition}},
  \code{\link{AutoTransformationCreate}},
  \code{\link{AutoTransformationScore}},
  \code{\link{AutoWord2VecModeler}},
  \code{\link{CreateCalendarVariables}},
  \code{\link{DT_GDL_Feature_Engineering}},
  \code{\link{DummifyDT}},
  \code{\link{FAST_GDL_Feature_Engineering}},
  \code{\link{GDL_Feature_Engineering}},
  \code{\link{InvApply_Asinh}},
  \code{\link{InvApply_Asin}},
  \code{\link{InvApply_BoxCox}},
  \code{\link{InvApply_Logit}},
  \code{\link{InvApply_YeoJohnson}},
  \code{\link{ModelDataPrep}},
  \code{\link{Scoring_GDL_Feature_Engineering}}
}
\author{
Adrian Antico
}
\concept{Feature Engineering}
