% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotFunctions.R
\name{AppModelInsights}
\alias{AppModelInsights}
\title{AppModelInsights}
\usage{
AppModelInsights(
  dt = NULL,
  PlotType = NULL,
  ModelOutputList = NULL,
  TargetVar = NULL,
  PredictVar = NULL,
  PDPVar = NULL,
  DateVar = NULL,
  GamFit = FALSE,
  Buckets = 20,
  Subsetcheck = FALSE,
  Check2 = FALSE,
  Debug = FALSE
)
}
\arguments{
\item{dt}{= NULL,}

\item{PlotType}{= NULL,}

\item{ModelOutputList}{= NULL,}

\item{TargetVar}{= isolate(YVar()),}

\item{PredictVar}{= isolate(ScoreVar()),}

\item{PDPVar}{= NULL,}

\item{DateVar}{= isolate(DateVar()),}

\item{GamFit}{= FALSE,}

\item{Buckets}{= 20,}

\item{Subsetcheck}{= FALSE,}

\item{Check2}{= FALSE,}

\item{Debug}{= FALSE}
}
\description{
Plot model insights in app
}
\seealso{
Other Graphics: 
\code{\link{AutoPlotter}()},
\code{\link{ChartTheme}()},
\code{\link{multiplot}()}
}
\author{
Adrian Antico
}
\concept{Graphics}
