% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal_intelligence.R
\name{aq_adjustment_guidance}
\alias{aq_adjustment_guidance}
\title{Produce Question-Relative Adjustment Guidance}
\usage{
aq_adjustment_guidance(causal_context)
}
\arguments{
\item{causal_context}{An `aq_causal_context`.}
}
\description{
Produce Question-Relative Adjustment Guidance
}
