% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decision_valuation.R
\name{aq_alternative_cash_flows}
\alias{aq_alternative_cash_flows}
\title{Create Alternative Cash-Flow Records}
\usage{
aq_alternative_cash_flows(cash_flows)
}
\description{
Create Alternative Cash-Flow Records
}
