% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decision_valuation.R
\name{aq_alternative_dominance}
\alias{aq_alternative_dominance}
\title{Identify Dominated and Non-Dominated Alternatives}
\usage{
aq_alternative_dominance(economics, criteria = NULL, risks = NULL)
}
\description{
Identify Dominated and Non-Dominated Alternatives
}
