% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decision_valuation.R
\name{aq_assess_alternative_economics}
\alias{aq_assess_alternative_economics}
\title{Assess Alternative Economics}
\usage{
aq_assess_alternative_economics(
  context,
  cash_flows = NULL,
  impact_mappings = NULL,
  scenarios = NULL
)
}
\description{
Assess Alternative Economics
}
