% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal_experiment_design.R
\name{aq_assess_assignment_balance}
\alias{aq_assess_assignment_balance}
\title{Assess Assignment Balance}
\usage{
aq_assess_assignment_balance(
  assignment_plan,
  baseline_data = NULL,
  unit_id_col = "unit_id",
  baseline_cols = character()
)
}
\description{
Assess Assignment Balance
}
