% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal_intelligence.R
\name{aq_assess_causal_designs}
\alias{aq_assess_causal_designs}
\title{Assess Causal Design Eligibility}
\usage{
aq_assess_causal_designs(causal_context)
}
\arguments{
\item{causal_context}{An `aq_causal_context`.}
}
\description{
Assess Causal Design Eligibility
}
