% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decision_valuation.R
\name{aq_assess_decision_effort}
\alias{aq_assess_decision_effort}
\title{Assess Effort and Capacity Requirements}
\usage{
aq_assess_decision_effort(effort = NULL)
}
\description{
Assess Effort and Capacity Requirements
}
