% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decision_valuation.R
\name{aq_assess_decision_information_value}
\alias{aq_assess_decision_information_value}
\title{Assess Evidence-to-Decision Information Value}
\usage{
aq_assess_decision_information_value(
  uncertainties = NULL,
  sensitivity = NULL,
  experiment_cost = NA_real_,
  reversibility = NA
)
}
\description{
Assess Evidence-to-Decision Information Value
}
