% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decision_valuation.R
\name{aq_assess_decision_option_value}
\alias{aq_assess_decision_option_value}
\title{Assess Decision Optionality for Valuation}
\usage{
aq_assess_decision_option_value(optionality = NULL)
}
\description{
Assess Decision Optionality for Valuation
}
