% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal_completed_experiment.R
\name{aq_assess_estimand_preservation}
\alias{aq_assess_estimand_preservation}
\title{Assess Estimand Preservation}
\usage{
aq_assess_estimand_preservation(
  completed_experiment,
  reconciliation,
  integrity,
  fidelity,
  missingness,
  interference,
  outcomes = NULL,
  exclusions = NULL
)
}
\description{
Assess Estimand Preservation
}
