% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vnext_forecasting.R
\name{aq_assess_forecast}
\alias{aq_assess_forecast}
\title{Assess vNext Forecast Evidence}
\usage{
aq_assess_forecast(forecast)
}
\arguments{
\item{forecast}{An `aq_forecast_result`.}
}
\value{
An `aq_forecast_assessment_result`.
}
\description{
Assess vNext Forecast Evidence
}
