% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal_randomized_itt.R
\name{aq_assess_itt_materiality}
\alias{aq_assess_itt_materiality}
\title{Assess ITT Effect Materiality}
\usage{
aq_assess_itt_materiality(
  primary_estimate,
  minimum_meaningful_effect = NA_real_
)
}
\description{
Assess ITT Effect Materiality
}
