% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vnext_multitarget_forecasting.R
\name{aq_assess_multitarget_forecast}
\alias{aq_assess_multitarget_forecast}
\title{Assess Multi-Target Forecast Evidence}
\usage{
aq_assess_multitarget_forecast(forecast)
}
\arguments{
\item{forecast}{An `aq_multitarget_forecast_result`.}
}
\value{
An `aq_multitarget_forecast_assessment_result`.
}
\description{
Assess Multi-Target Forecast Evidence
}
