% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal_observational_planning.R
\name{aq_assess_observational_balance}
\alias{aq_assess_observational_balance}
\title{Assess Baseline Covariate Balance}
\usage{
aq_assess_observational_balance(data, treatment, covariates)
}
\description{
Assess Baseline Covariate Balance
}
