% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal_observational_planning.R
\name{aq_assess_observational_estimation_readiness}
\alias{aq_assess_observational_estimation_readiness}
\title{Assess Observational Estimation Readiness}
\usage{
aq_assess_observational_estimation_readiness(
  plan,
  overlap = NULL,
  adjustment_spec = NULL,
  selection_missingness = NULL,
  unmeasured_risk = NULL,
  falsification_plan = NULL
)
}
\description{
Assess Observational Estimation Readiness
}
