% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal_observational_planning.R
\name{aq_assess_observational_temporal_eligibility}
\alias{aq_assess_observational_temporal_eligibility}
\title{Assess Adjustment Variable Timing Eligibility}
\usage{
aq_assess_observational_temporal_eligibility(
  adjustment_spec,
  variable_timing = NULL
)
}
\description{
Assess Adjustment Variable Timing Eligibility
}
