% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal_completed_experiment.R
\name{aq_assess_randomization_integrity}
\alias{aq_assess_randomization_integrity}
\title{Assess Randomization Integrity}
\usage{
aq_assess_randomization_integrity(assignment_evidence, expected_arms = NULL)
}
\description{
Assess Randomization Integrity
}
