% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal_completed_experiment.R
\name{aq_assess_realized_balance}
\alias{aq_assess_realized_balance}
\title{Assess Realized Balance}
\usage{
aq_assess_realized_balance(
  assignment_evidence,
  baseline_data = NULL,
  unit_id_col = "unit_id",
  baseline_cols = character()
)
}
\description{
Assess Realized Balance
}
