% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal_completed_experiment.R
\name{aq_assess_treatment_fidelity}
\alias{aq_assess_treatment_fidelity}
\title{Assess Treatment Fidelity}
\usage{
aq_assess_treatment_fidelity(
  assignment_evidence,
  delivery_evidence = NULL,
  exposure_evidence = NULL,
  compliance_evidence = NULL
)
}
\description{
Assess Treatment Fidelity
}
