% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal_completed_experiment.R
\name{aq_assignment_evidence}
\alias{aq_assignment_evidence}
\title{Ingest Assignment Evidence}
\usage{
aq_assignment_evidence(
  data,
  unit_id_col = "unit_id",
  planned_arm_col = "planned_arm",
  realized_arm_col = "realized_assigned_arm",
  assignment_time_col = NULL,
  eligible_col = NULL,
  exclusion_col = NULL,
  override_col = NULL
)
}
\description{
Ingest Assignment Evidence
}
