% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal_intelligence.R
\name{aq_causal_variable_roles}
\alias{aq_causal_variable_roles}
\title{Create Question-Relative Causal Variable Roles}
\usage{
aq_causal_variable_roles(roles)
}
\arguments{
\item{roles}{Role records.}
}
\value{
A data.table.
}
\description{
Create Question-Relative Causal Variable Roles
}
