% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal_completed_experiment.R
\name{aq_completed_experiment}
\alias{aq_completed_experiment}
\title{Create a Completed Experiment Record}
\usage{
aq_completed_experiment(
  record,
  experiment_plan_artifact = NULL,
  completed_experiment_id = NULL
)
}
\description{
Create a Completed Experiment Record
}
