% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal_observational_estimation.R
\name{aq_construct_observational_matches}
\alias{aq_construct_observational_matches}
\title{Construct Governed Observational Matching Sets}
\usage{
aq_construct_observational_matches(data, spec, propensity_result)
}
\description{
Construct Governed Observational Matching Sets
}
