% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal_observational_estimation.R
\name{aq_construct_observational_weights}
\alias{aq_construct_observational_weights}
\title{Construct Governed Observational Weights}
\usage{
aq_construct_observational_weights(data, spec, propensity_result)
}
\description{
Construct Governed Observational Weights
}
