% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decision_valuation.R
\name{aq_decision_action_thresholds}
\alias{aq_decision_action_thresholds}
\title{Create Decision Action Thresholds}
\usage{
aq_decision_action_thresholds(thresholds)
}
\description{
Create Decision Action Thresholds
}
