% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decision_workflow.R
\name{aq_decision_implementation_plan}
\alias{aq_decision_implementation_plan}
\title{Create Decision Implementation Plans}
\usage{
aq_decision_implementation_plan(plans)
}
\description{
Create Decision Implementation Plans
}
