% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decision_workflow.R
\name{aq_decision_monitoring_plan}
\alias{aq_decision_monitoring_plan}
\title{Create Decision Monitoring Plans}
\usage{
aq_decision_monitoring_plan(monitors)
}
\description{
Create Decision Monitoring Plans
}
