% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decision_workflow.R
\name{aq_decision_review_request}
\alias{aq_decision_review_request}
\title{Create Decision Review Requests}
\usage{
aq_decision_review_request(requests)
}
\description{
Create Decision Review Requests
}
