% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decision_valuation.R
\name{aq_decision_valuation_campaign_seeds}
\alias{aq_decision_valuation_campaign_seeds}
\title{Create Decision Valuation Campaign Seeds}
\usage{
aq_decision_valuation_campaign_seeds(recommendation)
}
\description{
Create Decision Valuation Campaign Seeds
}
