% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decision_valuation.R
\name{aq_decision_valuation_scenarios}
\alias{aq_decision_valuation_scenarios}
\title{Create Explicit Decision Scenarios}
\usage{
aq_decision_valuation_scenarios(scenarios)
}
\description{
Create Explicit Decision Scenarios
}
