% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decision_workflow.R
\name{aq_decision_workflow_staleness}
\alias{aq_decision_workflow_staleness}
\title{Assess Workflow Staleness}
\usage{
aq_decision_workflow_staleness(change_type, material = TRUE)
}
\description{
Assess Workflow Staleness
}
