% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal_observational_did.R
\name{aq_did_composition_stability}
\alias{aq_did_composition_stability}
\title{Assess Difference-in-Differences Composition Stability}
\usage{
aq_did_composition_stability(data, spec)
}
\description{
Assess Difference-in-Differences Composition Stability
}
