% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal_observational_did.R
\name{aq_did_effect_report}
\alias{aq_did_effect_report}
\title{Create a Difference-in-Differences Review Report Object}
\usage{
aq_did_effect_report(result)
}
\description{
Create a Difference-in-Differences Review Report Object
}
