% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal_observational_did.R
\name{aq_did_pre_period_diagnostics}
\alias{aq_did_pre_period_diagnostics}
\title{Compute Difference-in-Differences Pre-Period Diagnostics}
\usage{
aq_did_pre_period_diagnostics(data, spec)
}
\description{
Compute Difference-in-Differences Pre-Period Diagnostics
}
