% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal_observational_did.R
\name{aq_did_readiness}
\alias{aq_did_readiness}
\title{Assess Difference-in-Differences Readiness}
\usage{
aq_did_readiness(data, spec)
}
\description{
Assess Difference-in-Differences Readiness
}
