% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epistemic_integrity.R
\name{aq_epistemic_adjudication}
\alias{aq_epistemic_adjudication}
\title{Create Epistemic Adjudication Records}
\usage{
aq_epistemic_adjudication(adjudications)
}
\description{
Create Epistemic Adjudication Records
}
