% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epistemic_integrity.R
\name{aq_epistemic_intervention_event}
\alias{aq_epistemic_intervention_event}
\title{Create Epistemic Intervention Events}
\usage{
aq_epistemic_intervention_event(events)
}
\description{
Create Epistemic Intervention Events
}
