% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epistemic_integrity.R
\name{aq_epistemic_quality_gates}
\alias{aq_epistemic_quality_gates}
\title{Create Epistemic Quality Gates}
\usage{
aq_epistemic_quality_gates(findings = NULL, claim_assessment = NULL)
}
\description{
Create Epistemic Quality Gates
}
