% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal_observational_did.R
\name{aq_estimate_did_effect}
\alias{aq_estimate_did_effect}
\title{Estimate a Governed Classic Difference-in-Differences Effect}
\usage{
aq_estimate_did_effect(data, spec)
}
\description{
Estimate a Governed Classic Difference-in-Differences Effect
}
