% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal_observational_estimation.R
\name{aq_estimate_observational_aipw}
\alias{aq_estimate_observational_aipw}
\title{Estimate a Governed Observational AIPW Effect}
\usage{
aq_estimate_observational_aipw(
  spec,
  propensity_result,
  outcome_result,
  balance_result = NULL
)
}
\description{
Estimate a Governed Observational AIPW Effect
}
