% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal_observational_estimation.R
\name{aq_estimate_observational_effect}
\alias{aq_estimate_observational_effect}
\title{Estimate a Governed Observational Effect}
\usage{
aq_estimate_observational_effect(data, spec)
}
\description{
Estimate a Governed Observational Effect
}
