% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal_randomized_itt.R
\name{aq_estimate_randomized_itt}
\alias{aq_estimate_randomized_itt}
\title{Estimate a Randomized ITT Effect}
\usage{
aq_estimate_randomized_itt(
  spec,
  completed_evidence,
  baseline_data = NULL,
  planned_analysis = NULL,
  readiness = NULL,
  authority_approved = TRUE,
  coverage_approved = TRUE
)
}
\description{
Estimate a Randomized ITT Effect
}
