% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decision_valuation.R
\name{aq_evidence_impact_mapping}
\alias{aq_evidence_impact_mapping}
\title{Create Evidence-to-Impact Mapping Records}
\usage{
aq_evidence_impact_mapping(mappings)
}
\description{
Create Evidence-to-Impact Mapping Records
}
