% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal_completed_experiment.R
\name{aq_exclusion_evidence}
\alias{aq_exclusion_evidence}
\title{Ingest Exclusion Evidence}
\usage{
aq_exclusion_evidence(
  data,
  unit_id_col = "unit_id",
  stage_col = "exclusion_stage",
  reason_col = "exclusion_reason",
  planned_col = NULL
)
}
\description{
Ingest Exclusion Evidence
}
