% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal_experiment_design.R
\name{aq_experiment_design_spec}
\alias{aq_experiment_design_spec}
\title{Create an Experimental Design Specification}
\usage{
aq_experiment_design_spec(
  experiment_question,
  design_type,
  assignment_unit,
  analysis_unit,
  treatment_delivery_unit = assignment_unit,
  cluster_unit = NA_character_,
  blocking_variables = character(),
  stratification_variables = character(),
  number_of_arms = 2L,
  allocation_ratios = c(control = 0.5, treatment = 0.5),
  rollout_schedule = NA_character_,
  washout_period = NA_character_,
  pre_period = NA_character_,
  treatment_period = NA_character_,
  follow_up_period = NA_character_,
  randomization_method = "deterministic_pseudo_random",
  seed = 20260713L,
  exposure_rules = NA_character_,
  compliance_expectations = NA_character_,
  contamination_risks = character(),
  interference_assumptions = "not yet assessed",
  planned_estimator_family = "future_intention_to_treat",
  supported_estimands = character(),
  validity_limitations = character(),
  design_id = NULL
)
}
\description{
Create an Experimental Design Specification
}
