% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal_experiment_design.R
\name{aq_experiment_gate_assessment}
\alias{aq_experiment_gate_assessment}
\title{Assess Experiment Approval Gates}
\usage{
aq_experiment_gate_assessment(
  experiment_question,
  design_spec,
  assignment_plan = NULL,
  power_plan = NULL,
  measurement_plan = NULL,
  validity_threats = NULL,
  authority_approved = FALSE,
  coverage_approved = FALSE,
  information_value = NULL
)
}
\description{
Assess Experiment Approval Gates
}
