% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal_experiment_design.R
\name{aq_experiment_information_value}
\alias{aq_experiment_information_value}
\title{Assess Experiment Information Value}
\usage{
aq_experiment_information_value(
  decision_sensitivity = "unknown",
  lever_importance = "unknown",
  experiment_cost = NA_real_,
  duration_days = NA_integer_,
  reversibility = NA,
  optionality_created = character(),
  opportunity_cost = NA_real_
)
}
\description{
Assess Experiment Information Value
}
