% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal_experiment_design.R
\name{aq_experiment_question}
\alias{aq_experiment_question}
\title{Create an Experiment Question}
\usage{
aq_experiment_question(
  question,
  causal_context = NULL,
  experiment_question_id = NULL
)
}
\arguments{
\item{question}{Experiment question record or list.}

\item{causal_context}{Optional `aq_causal_context`.}

\item{experiment_question_id}{Optional ID override.}
}
\value{
An `aq_experiment_question` object.
}
\description{
Create an Experiment Question
}
