% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal_completed_experiment.R
\name{aq_exposure_evidence}
\alias{aq_exposure_evidence}
\title{Ingest Exposure Evidence}
\usage{
aq_exposure_evidence(
  data,
  unit_id_col = "unit_id",
  exposure_col = "exposure",
  intensity_col = NULL,
  timing_col = NULL,
  quality_col = NULL
)
}
\description{
Ingest Exposure Evidence
}
