% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vnext_forecasting.R
\name{aq_fit_forecast}
\alias{aq_fit_forecast}
\title{Execute a vNext Forecast}
\usage{
aq_fit_forecast(spec, data, origin = NULL, future_data = NULL)
}
\arguments{
\item{spec}{An `aq_forecast_spec`.}

\item{data}{Time-series data.}

\item{origin}{Optional forecast origin override.}

\item{future_data}{Optional future data containing known-future regressors.}
}
\value{
An `aq_forecast_result`.
}
\description{
Execute a vNext Forecast
}
