% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vnext_funnel_forecasting.R
\name{aq_fit_funnel_forecast}
\alias{aq_fit_funnel_forecast}
\title{Fit a vNext Funnel Forecast}
\usage{
aq_fit_funnel_forecast(spec, data, origin = NULL)
}
\arguments{
\item{spec}{An `aq_funnel_forecast_spec`.}

\item{data}{Long-form funnel data.}

\item{origin}{Optional forecast origin.}
}
\value{
An `aq_funnel_forecast_result`.
}
\description{
Fit a vNext Funnel Forecast
}
