% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vnext_hurdle_forecasting.R
\name{aq_fit_hurdle_forecast}
\alias{aq_fit_hurdle_forecast}
\title{Fit a vNext Hurdle Forecast}
\usage{
aq_fit_hurdle_forecast(spec, data, origin = NULL, future_data = NULL)
}
\arguments{
\item{spec}{An `aq_hurdle_forecast_spec`.}

\item{data}{Intermittent-demand data.}

\item{origin}{Optional forecast origin.}

\item{future_data}{Optional future-known data.}
}
\value{
An `aq_hurdle_forecast_result`.
}
\description{
Fit a vNext Hurdle Forecast
}
