% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vnext_catboost_regression.R
\name{aq_fit_model}
\alias{aq_fit_model}
\title{Fit a vNext CatBoost Model}
\usage{
aq_fit_model(spec, data, validation_data = NULL)
}
\arguments{
\item{spec}{An `aq_model_spec`.}

\item{data}{Training/source data.}

\item{validation_data}{Optional explicit validation data.}
}
\value{
An `aq_fit_result`.
}
\description{
Fit a vNext CatBoost Model
}
