% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vnext_multitarget_forecasting.R
\name{aq_fit_multitarget_forecast}
\alias{aq_fit_multitarget_forecast}
\title{Fit a vNext Multi-Target Forecast}
\usage{
aq_fit_multitarget_forecast(spec, data, origin = NULL, future_data = NULL)
}
\arguments{
\item{spec}{An `aq_multitarget_forecast_spec`.}

\item{data}{Wide multi-target time-series data.}

\item{origin}{Optional forecast origin override.}

\item{future_data}{Optional future known-variable data.}
}
\value{
An `aq_multitarget_forecast_result`.
}
\description{
Fit a vNext Multi-Target Forecast
}
