% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal_observational_estimation.R
\name{aq_fit_observational_propensity}
\alias{aq_fit_observational_propensity}
\title{Fit a Governed Observational Propensity Model}
\usage{
aq_fit_observational_propensity(data, spec)
}
\description{
Fit a Governed Observational Propensity Model
}
