% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vnext_panel_forecasting.R
\name{aq_fit_panel_forecast}
\alias{aq_fit_panel_forecast}
\title{Fit a Global Panel Forecast}
\usage{
aq_fit_panel_forecast(spec, data, origin = NULL, future_data = NULL)
}
\description{
Fit a Global Panel Forecast
}
