% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vnext_hurdle_forecasting.R
\name{aq_fit_sba_forecast}
\alias{aq_fit_sba_forecast}
\title{Fit an SBA Intermittent Demand Forecast}
\usage{
aq_fit_sba_forecast(spec, data, origin = NULL)
}
\arguments{
\item{spec}{An `aq_sba_forecast_spec`.}

\item{data}{Intermittent-demand data.}

\item{origin}{Optional forecast origin.}
}
\value{
An `aq_sba_forecast_result`.
}
\description{
Fit an SBA Intermittent Demand Forecast
}
