% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vnext_hurdle_forecasting.R
\name{aq_fit_tsb_forecast}
\alias{aq_fit_tsb_forecast}
\title{Fit a TSB Intermittent Demand Forecast}
\usage{
aq_fit_tsb_forecast(spec, data, origin = NULL)
}
\arguments{
\item{spec}{An `aq_tsb_forecast_spec`.}

\item{data}{Intermittent-demand data.}

\item{origin}{Optional forecast origin.}
}
\value{
An `aq_tsb_forecast_result`.
}
\description{
Fit a TSB Intermittent Demand Forecast
}
