% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vnext_forecasting.R
\name{aq_forecast_partition}
\alias{aq_forecast_partition}
\title{Create a vNext Forecast Partition}
\usage{
aq_forecast_partition(spec, data, origin = NULL)
}
\arguments{
\item{spec}{An `aq_forecast_spec`.}

\item{data}{Time-series data.}

\item{origin}{Optional forecast origin override.}
}
\value{
An `aq_forecast_partition` object.
}
\description{
Create a vNext Forecast Partition
}
