% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vnext_catboost_regression.R
\name{aq_model_operator_capabilities}
\alias{aq_model_operator_capabilities}
\title{List vNext Model Operator Capabilities}
\usage{
aq_model_operator_capabilities()
}
\value{
A `data.table` of implemented vNext model capabilities.
}
\description{
List vNext Model Operator Capabilities
}
