% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decision_valuation.R
\name{aq_multi_criteria_assessment}
\alias{aq_multi_criteria_assessment}
\title{Assess Multiple Decision Criteria}
\usage{
aq_multi_criteria_assessment(criteria_values)
}
\description{
Assess Multiple Decision Criteria
}
