% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal_observational_planning.R
\name{aq_observational_adjustment_spec}
\alias{aq_observational_adjustment_spec}
\title{Create an Observational Adjustment Set Contract}
\usage{
aq_observational_adjustment_spec(
  approved_confounders = character(),
  required_baseline_variables = character(),
  optional_precision_variables = character(),
  assignment_predictors = character(),
  effect_modifiers = character(),
  excluded_mediators = character(),
  excluded_colliders = character(),
  excluded_post_treatment = character(),
  instruments = character(),
  proxy_variables = character(),
  measurement_quality = list(),
  missingness = list(),
  evidence_basis = character(),
  graph_version = NA_character_,
  human_approval = FALSE,
  unresolved_conflicts = character(),
  adjustment_spec_id = NULL
)
}
\description{
Create an Observational Adjustment Set Contract
}
