% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal_observational_planning.R
\name{aq_observational_falsification_plan}
\alias{aq_observational_falsification_plan}
\title{Plan Negative Controls and Falsification Checks}
\usage{
aq_observational_falsification_plan(tests)
}
\description{
Plan Negative Controls and Falsification Checks
}
