% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal_observational_estimation.R
\name{aq_observational_sensitivity_summary}
\alias{aq_observational_sensitivity_summary}
\title{Summarize Observational Sensitivity Obligations}
\usage{
aq_observational_sensitivity_summary(
  spec,
  propensity_result,
  weight_result,
  balance_result,
  aipw_result
)
}
\description{
Summarize Observational Sensitivity Obligations
}
