% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vnext_panel_forecasting.R
\name{aq_panel_forecast_partition}
\alias{aq_panel_forecast_partition}
\title{Create an Entity-Aware Panel Forecast Partition}
\usage{
aq_panel_forecast_partition(spec, data, origin = NULL)
}
\description{
Create an Entity-Aware Panel Forecast Partition
}
