% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal_intelligence.R
\name{aq_plan_causal_investigation}
\alias{aq_plan_causal_investigation}
\title{Plan the Smallest Useful Causal Investigation}
\usage{
aq_plan_causal_investigation(causal_context)
}
\arguments{
\item{causal_context}{An `aq_causal_context`.}
}
\description{
Plan the Smallest Useful Causal Investigation
}
