% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal_randomized_design_analysis.R
\name{aq_randomization_inference}
\alias{aq_randomization_inference}
\title{Run Randomization Inference for Supported Assignment Mechanisms}
\usage{
aq_randomization_inference(population, spec, design_spec = NULL)
}
\description{
Run Randomization Inference for Supported Assignment Mechanisms
}
