% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal_randomized_design_analysis.R
\name{aq_randomized_causal_effect_report}
\alias{aq_randomized_causal_effect_report}
\title{Create a Causal Effect Report Contract}
\usage{
aq_randomized_causal_effect_report(
  itt_result,
  design_result = NULL,
  report_id = NULL
)
}
\description{
Create a Causal Effect Report Contract
}
