% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal_randomized_design_analysis.R
\name{aq_randomized_cuped_adjustment}
\alias{aq_randomized_cuped_adjustment}
\title{Run CUPED-Style Variance Reduction}
\usage{
aq_randomized_cuped_adjustment(population, spec, pre_period_metric)
}
\description{
Run CUPED-Style Variance Reduction
}
