% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal_randomized_design_analysis.R
\name{aq_randomized_design_method_eligibility}
\alias{aq_randomized_design_method_eligibility}
\title{Determine Eligible Randomized Analysis Methods}
\usage{
aq_randomized_design_method_eligibility(spec, itt_result)
}
\description{
Determine Eligible Randomized Analysis Methods
}
