% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal_randomized_design_analysis.R
\name{aq_randomized_guardrail_decision}
\alias{aq_randomized_guardrail_decision}
\title{Interpret Guardrails for Randomized Decision Evidence}
\usage{
aq_randomized_guardrail_decision(
  primary_materiality,
  guardrails = NULL,
  maximum_acceptable_harm = NA_real_
)
}
\description{
Interpret Guardrails for Randomized Decision Evidence
}
