% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal_randomized_itt.R
\name{aq_randomized_itt_campaign_seeds}
\alias{aq_randomized_itt_campaign_seeds}
\title{Create Campaign Seeds from an ITT Result}
\usage{
aq_randomized_itt_campaign_seeds(result)
}
\description{
Create Campaign Seeds from an ITT Result
}
