% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal_randomized_itt.R
\name{aq_randomized_itt_effect_artifact}
\alias{aq_randomized_itt_effect_artifact}
\title{Create a Randomized ITT Effect Artifact}
\usage{
aq_randomized_itt_effect_artifact(result, artifact_id = NULL)
}
\description{
Create a Randomized ITT Effect Artifact
}
