% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal_randomized_design_analysis.R
\name{aq_randomized_materiality_regions}
\alias{aq_randomized_materiality_regions}
\title{Classify Randomized Effect Materiality Regions}
\usage{
aq_randomized_materiality_regions(
  primary_estimate,
  material_benefit = NA_real_,
  material_harm = NA_real_
)
}
\description{
Classify Randomized Effect Materiality Regions
}
