% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal_randomized_design_analysis.R
\name{aq_randomized_robustness_matrix}
\alias{aq_randomized_robustness_matrix}
\title{Create a Randomized Robustness Matrix}
\usage{
aq_randomized_robustness_matrix(itt_result, design_depth = NULL)
}
\description{
Create a Randomized Robustness Matrix
}
