% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decision_workflow.R
\name{aq_record_decision_implementation}
\alias{aq_record_decision_implementation}
\title{Record Realized Decision Implementation}
\usage{
aq_record_decision_implementation(records)
}
\description{
Record Realized Decision Implementation
}
