% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal_randomized_itt.R
\name{aq_review_randomized_itt_result}
\alias{aq_review_randomized_itt_result}
\title{Review a Randomized ITT Result}
\usage{
aq_review_randomized_itt_result(
  result,
  reviewer = NA_character_,
  review_notes = NA_character_,
  approve = FALSE
)
}
\description{
Review a Randomized ITT Result
}
