% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vnext_catboost_regression.R
\name{aq_save_model_bundle}
\alias{aq_save_model_bundle}
\title{Save a vNext Model Bundle}
\usage{
aq_save_model_bundle(fit, path, overwrite = FALSE)
}
\arguments{
\item{fit}{An `aq_fit_result` or `aq_model_bundle`.}

\item{path}{Directory path or `.rds` bundle file path.}

\item{overwrite}{Whether to overwrite an existing bundle file.}
}
\value{
An `aq_model_bundle_manifest`.
}
\description{
Save a vNext Model Bundle
}
