% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artifact_schema.R
\name{aq_supported_actions}
\alias{aq_supported_actions}
\title{Return Supported Downstream Actions for an AutoQuant Artifact}
\usage{
aq_supported_actions(artifact)
}
\arguments{
\item{artifact}{An AutoQuant artifact or vNext result object.}
}
\value{
Character vector of supported actions.
}
\description{
Return Supported Downstream Actions for an AutoQuant Artifact
}
