% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal_observational_planning.R
\name{aq_target_trial_spec}
\alias{aq_target_trial_spec}
\title{Create a Target-Trial Emulation Specification}
\usage{
aq_target_trial_spec(
  study,
  eligibility_criteria,
  treatment_strategies,
  assignment_time,
  follow_up,
  outcome,
  causal_contrast,
  analysis_plan = "observational_design_readiness_only",
  censoring = "not_assessed",
  adherence = "not_assessed",
  estimand = NA_character_,
  target_trial_id = NULL
)
}
\description{
Create a Target-Trial Emulation Specification
}
