% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal_observational_planning.R
\name{aq_unmeasured_confounding_risk_register}
\alias{aq_unmeasured_confounding_risk_register}
\title{Create an Unmeasured-Confounding Risk Register}
\usage{
aq_unmeasured_confounding_risk_register(risks)
}
\description{
Create an Unmeasured-Confounding Risk Register
}
