% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal_intelligence.R
\name{aq_validate_causal_graph}
\alias{aq_validate_causal_graph}
\title{Validate a Causal Graph Structurally}
\usage{
aq_validate_causal_graph(causal_context)
}
\arguments{
\item{causal_context}{An `aq_causal_context`.}
}
\description{
Validate a Causal Graph Structurally
}
