% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal_intelligence.R
\name{aq_validate_causal_question}
\alias{aq_validate_causal_question}
\title{Validate a Causal Question}
\usage{
aq_validate_causal_question(causal_question)
}
\arguments{
\item{causal_question}{An `aq_causal_question`.}
}
\value{
A data.table of deterministic diagnostics.
}
\description{
Validate a Causal Question
}
