% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal_completed_experiment.R
\name{aq_validate_completed_experiment}
\alias{aq_validate_completed_experiment}
\title{Validate a Completed Experiment Record}
\usage{
aq_validate_completed_experiment(completed_experiment)
}
\description{
Validate a Completed Experiment Record
}
