% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal_observational_did.R
\name{aq_validate_did_analysis_spec}
\alias{aq_validate_did_analysis_spec}
\title{Validate a Difference-in-Differences Analysis Specification}
\usage{
aq_validate_did_analysis_spec(spec)
}
\description{
Validate a Difference-in-Differences Analysis Specification
}
