% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal_intelligence.R
\name{aq_validate_estimand}
\alias{aq_validate_estimand}
\title{Validate an Estimand}
\usage{
aq_validate_estimand(estimand)
}
\arguments{
\item{estimand}{An `aq_estimand`.}
}
\description{
Validate an Estimand
}
