% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decision_valuation.R
\name{aq_validate_evidence_impact_mapping}
\alias{aq_validate_evidence_impact_mapping}
\title{Validate Evidence-to-Impact Mappings}
\usage{
aq_validate_evidence_impact_mapping(mapping)
}
\description{
Validate Evidence-to-Impact Mappings
}
