% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal_experiment_design.R
\name{aq_validate_experiment_design_spec}
\alias{aq_validate_experiment_design_spec}
\title{Validate an Experimental Design Specification}
\usage{
aq_validate_experiment_design_spec(design_spec, experiment_question = NULL)
}
\description{
Validate an Experimental Design Specification
}
