% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal_experiment_design.R
\name{aq_validate_experiment_question}
\alias{aq_validate_experiment_question}
\title{Validate an Experiment Question}
\usage{
aq_validate_experiment_question(experiment_question)
}
\description{
Validate an Experiment Question
}
