% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vnext_hierarchical_forecasting.R
\name{aq_validate_hierarchy_spec}
\alias{aq_validate_hierarchy_spec}
\title{Validate a vNext Hierarchy Specification}
\usage{
aq_validate_hierarchy_spec(spec)
}
\arguments{
\item{spec}{An `aq_hierarchy_spec`.}
}
\value{
A `data.table` of deterministic hierarchy diagnostics.
}
\description{
Validate a vNext Hierarchy Specification
}
