% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal_observational_estimation.R
\name{aq_validate_observational_analysis_spec}
\alias{aq_validate_observational_analysis_spec}
\title{Validate a Governed Observational Analysis Specification}
\usage{
aq_validate_observational_analysis_spec(spec)
}
\description{
Validate a Governed Observational Analysis Specification
}
