% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal_observational_estimation.R
\name{aq_validate_observational_balance}
\alias{aq_validate_observational_balance}
\title{Validate Observational Covariate Balance}
\usage{
aq_validate_observational_balance(data, spec, weight_result = NULL)
}
\description{
Validate Observational Covariate Balance
}
