% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal_observational_planning.R
\name{aq_validate_observational_study}
\alias{aq_validate_observational_study}
\title{Validate an Observational Causal Study Contract}
\usage{
aq_validate_observational_study(study)
}
\description{
Validate an Observational Causal Study Contract
}
