% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal_completed_experiment.R
\name{aq_validate_outcome_evidence}
\alias{aq_validate_outcome_evidence}
\title{Validate Outcome Evidence}
\usage{
aq_validate_outcome_evidence(outcome_evidence, primary_outcome = NULL)
}
\description{
Validate Outcome Evidence
}
