% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal_randomized_design_analysis.R
\name{aq_validate_randomized_design_analysis_spec}
\alias{aq_validate_randomized_design_analysis_spec}
\title{Validate a Design-Aware Randomized Analysis Specification}
\usage{
aq_validate_randomized_design_analysis_spec(spec)
}
\description{
Validate a Design-Aware Randomized Analysis Specification
}
