% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal_randomized_itt.R
\name{aq_validate_randomized_itt_readiness}
\alias{aq_validate_randomized_itt_readiness}
\title{Validate Randomized ITT Readiness}
\usage{
aq_validate_randomized_itt_readiness(
  spec,
  completed_evidence,
  readiness = NULL,
  planned_analysis = NULL,
  baseline_data = NULL,
  stale_readiness = FALSE,
  stale_planned_analysis = FALSE,
  authority_approved = TRUE,
  coverage_approved = TRUE
)
}
\description{
Validate Randomized ITT Readiness
}
