% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal_randomized_itt.R
\name{aq_validate_randomized_itt_spec}
\alias{aq_validate_randomized_itt_spec}
\title{Validate a Randomized ITT Specification}
\usage{
aq_validate_randomized_itt_spec(
  spec,
  planned_analysis = NULL,
  covariate_roles = NULL
)
}
\description{
Validate a Randomized ITT Specification
}
