% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal_observational_planning.R
\name{aq_validate_target_trial_spec}
\alias{aq_validate_target_trial_spec}
\title{Validate a Target-Trial Emulation Specification}
\usage{
aq_validate_target_trial_spec(spec)
}
\description{
Validate a Target-Trial Emulation Specification
}
