% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_semantics.R
\name{aq_validate_variable_semantics}
\alias{aq_validate_variable_semantics}
\title{Validate Canonical Variable Semantics}
\usage{
aq_validate_variable_semantics(semantics)
}
\arguments{
\item{semantics}{An \code{aq_variable_semantics} object.}
}
\value{
A data.table of validation diagnostics.
}
\description{
Validate Canonical Variable Semantics
}
