% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal_experiment_design.R
\name{aq_validity_threat_register}
\alias{aq_validity_threat_register}
\title{Build a Validity-Threat Register}
\usage{
aq_validity_threat_register(
  threats = NULL,
  design_spec = NULL,
  measurement_plan = NULL
)
}
\description{
Build a Validity-Threat Register
}
