% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyFunctions.R
\name{ArgNullCheck}
\alias{ArgNullCheck}
\title{ArgNullCheck}
\usage{
ArgNullCheck(Input, InputID, Default)
}
\arguments{
\item{Input}{input}

\item{InputID}{inputId}

\item{Default}{Default value}
}
\description{
ArgNullCheck
}
\examples{
\dontrun{
ArgValue <<- RemixAutoML::ArgNullCheck(Input = input, InputID = "TS_Value", Default = 2)
}
}
\seealso{
Other Shiny: 
\code{\link{App_BoxPlotsOverTime}()},
\code{\link{ArgNullCheck2}()},
\code{\link{BlankRow}()},
\code{\link{DateInput}()},
\code{\link{GenerateEvaluationMetrics}()},
\code{\link{NumericInput}()},
\code{\link{PickerInput_GetLevels}()},
\code{\link{PickerInput}()},
\code{\link{PreparePlotData}()},
\code{\link{ReactiveLoadCSV}()},
\code{\link{ReturnParam}()},
\code{\link{StoreArgs}()},
\code{\link{TextInput}()},
\code{\link{observeEventLoad}()}
}
\author{
Adrian Antico
}
\concept{Shiny}
