% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyFunctions.R
\name{ArgNullCheck2}
\alias{ArgNullCheck2}
\title{ArgNullCheck2}
\usage{
ArgNullCheck2(Input, InputID, Default, Type = "numeric")
}
\arguments{
\item{Input}{input}

\item{InputID}{inputId}

\item{Default}{Default value}

\item{Type}{numeric character logical}
}
\description{
ArgNullCheck2
}
\examples{
\dontrun{
ArgValue <<- RemixAutoML::ArgNullCheck2(Input = input, InputID = "TS_Value", Default = 2, Type = "numeric")
}
}
\seealso{
Other Shiny: 
\code{\link{ArgNullCheck}()},
\code{\link{BlankRow}()},
\code{\link{DateInput}()},
\code{\link{GenerateEvaluationMetrics}()},
\code{\link{NumericInput}()},
\code{\link{PickerInput_GetLevels2}()},
\code{\link{PickerInput_GetLevels}()},
\code{\link{PickerInput}()},
\code{\link{PreparePlotData}()},
\code{\link{ReactiveLoadCSV}()},
\code{\link{ReturnParam}()},
\code{\link{StoreArgs}()},
\code{\link{TextInput}()},
\code{\link{observeEventLoad}()}
}
\author{
Adrian Antico
}
\concept{Shiny}
