% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyFunctions.R
\name{AssignData}
\alias{AssignData}
\title{Assign a data.table by name from an environment}
\usage{
AssignData(data, env = globalenv())
}
\arguments{
\item{env}{an environment}

\item{df}{character, name of the object}
}
\value{
the object
}
\description{
Assign a data.table by name from an environment
}
