% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RemixAutoML.R
\name{AutoCatBoostClassifier}
\alias{AutoCatBoostClassifier}
\title{AutoCatBoostClassifier is an automated catboost model grid-tuning classifier and evaluation system}
\usage{
AutoCatBoostClassifier(data, TestData = NULL,
  TargetColumnName = "Doug", FeatureColNames = names(data)[2:18],
  TrainSplitRatio = 0.5, task_type = "CPU", eval_metric = "AUC",
  Trees = 50, GridTune = TRUE, grid_eval_metric = "f",
  MaxModelsInGrid = 10, model_path = getwd(), ModelID = "FirstModel",
  NumOfParDepPlots = 15, ReturnModelObjects = TRUE,
  SaveModelObjects = TRUE)
}
\arguments{
\item{data}{This is your data set for training and testing your model}

\item{TestData}{If you want to supply your own data for testing (column names and column ordering must be the same as data)}

\item{FeatureColNames}{Either supply the feature column names OR the column number where the target is located (but not mixed types)}

\item{TrainSplitRatio}{A decimal between 0.01 and 0.99 that tells the function how much data to keep for training and validation. 0.90 means 90% to train and 10% to validation.}

\item{task_type}{= "GPU" Set to "GPU" to utilize your GPU for training. Default is "CPU".}

\item{eval_metric}{This is the metric used inside catboost to measure performance on validation data during a grid-tune. "AUC" is the default, but other options include: "Logloss", "CrossEntropy", "Precision", "Recall", "F1", "BalancedAccuracy", "BalancedErrorRate", "MCC", "Accuracy", "CtrFactor", "AUC", "BrierScore", "HingeLoss", "HammingLoss", "ZeroOneLoss", "Kappa", "WKappa", "LogLikelihoodOfPrediction"}

\item{Trees}{The maximum number of trees you want in your models}

\item{GridTune}{Set to TRUE to run a grid tuning procedure. Set a number in MaxModelsInGrid to tell the procedure how many models you want to test.}

\item{grid_eval_metric}{This is the metric used to find the threshold "f" # c("auc","tpr","fnr","fpr","tnr","prbe","f","odds")}

\item{MaxModelsInGrid}{Number of models to test from grid options (1080 total possible options)}

\item{model_path}{A character string of your path file to where you want your output saved}

\item{ModelID}{A character string to name your model and output}

\item{NumOfParDepPlots}{Tell the function the number of partial dependence calibration plots you want to create. Calibration boxplots will only be created for numerical features (not dummy variables)}

\item{ReturnModelObjects}{Set to TRUE to output all modeling objects (E.g. plots and evaluation metrics)}

\item{SaveModelObjects}{Set to TRUE to return all modeling objects to your environment}

\item{TargetColName}{Either supply the target column name OR the column number where the target is located (but not mixed types)}
}
\value{
Saves to file: <ModelID>_VariableImportance.csv, <ModelID> (the model), <ModelID>_ValidationData.csv, <ModelID>_ROC_Plot.png, <ModelID>_EvalutionPlot.png, <ModelID>_EvaluationMetrics.csv, <ModelID>_ParDepPlots.R (named list of features with partial dependence calibration plots), and <ModelID>_ParDepBoxPlots.R (named list of non-dummy features with partial dependence calibration box plots)
}
\description{
AutoCatBoostClassifier is an automated modeling function that runs a variety of steps. First, the function will run a random grid tune over N number of models and find which model is the best (a default model is always included in that set). Once the model is identified and built, several other outputs are generated: validation data with predictions, ROC plot, evaluation plot, evaluation metrics, variable importance, partial dependence calibration plots, partial dependence calibration box plots, and column names used in model fitting.
}
\examples{
\dontrun{
Correl <- 0.85
N <- 1000
data <- data.table::data.table(Target = runif(N))
data[, x1 := qnorm(Target)]
data[, x2 := runif(N)]
data[, Independent_Variable1 := log(pnorm(Correl * x1 +
                                            sqrt(1-Correl^2) * qnorm(x2)))]
data[, Independent_Variable2 := (pnorm(Correl * x1 +
                                         sqrt(1-Correl^2) * qnorm(x2)))]
data[, Independent_Variable3 := exp(pnorm(Correl * x1 +
                                            sqrt(1-Correl^2) * qnorm(x2)))]
data[, Independent_Variable4 := exp(exp(pnorm(Correl * x1 +
                                                sqrt(1-Correl^2) * qnorm(x2))))]
data[, Independent_Variable5 := sqrt(pnorm(Correl * x1 +
                                             sqrt(1-Correl^2) * qnorm(x2)))]
data[, Independent_Variable6 := (pnorm(Correl * x1 +
                                         sqrt(1-Correl^2) * qnorm(x2)))^0.10]
data[, Independent_Variable7 := (pnorm(Correl * x1 +
                                         sqrt(1-Correl^2) * qnorm(x2)))^0.25]
data[, Independent_Variable8 := (pnorm(Correl * x1 +
                                         sqrt(1-Correl^2) * qnorm(x2)))^0.75]
data[, Independent_Variable9 := (pnorm(Correl * x1 +
                                         sqrt(1-Correl^2) * qnorm(x2)))^2]
data[, Independent_Variable10 := (pnorm(Correl * x1 +
                                          sqrt(1-Correl^2) * qnorm(x2)))^4]

data[, Independent_Variable11 := as.factor(ifelse(Independent_Variable2 < 0.20, "A",
                                        ifelse(Independent_Variable2 < 0.40, "B",
                                               ifelse(Independent_Variable2 < 0.6, "C",
                                                      ifelse(Independent_Variable2 < 0.8, "D", "E")))))]

# Dummify Categorical Variable
data <- RemixAutoML::DummifyDT(data = data,
                               cols = "Independent_Variable11",
                               KeepFactorCols = FALSE,
                               OneHot = TRUE,
                               ClustScore = FALSE)
data[, Predict := (pnorm(Correl * x1 +
                           sqrt(1-Correl^2) * qnorm(x2)))]
data[, ':=' (x1 = NULL, x2 = NULL)]
data[, Target := ifelse(Target < 0.5, 1, 0)]
TestModel <- AutoCatBoostClassifier(data,
                                    TestData = NULL,
                                    TargetColumnName = "Target",
                                    FeatureColNames = c(2:18),
                                    MaxModelsInGrid = 3,
                                    TrainSplitRatio = 0.50,
                                    task_type = "GPU",
                                    eval_metric = "AUC",
                                    grid_eval_metric = "auc",
                                    Trees = 50,
                                    GridTune = TRUE,
                                    model_path = getwd(),
                                    ModelID = "ModelTest",
                                    NumOfParDepPlots = 15,
                                    ReturnModelObjects = TRUE,
                                    SaveModelObjects = TRUE)
}
}
\seealso{
Other Supervised Learning: \code{\link{AutoH2OModeler}},
  \code{\link{AutoH2OScoring}}, \code{\link{AutoNLS}},
  \code{\link{AutoRecommenderScoring}},
  \code{\link{AutoRecommender}}, \code{\link{AutoTS}}
}
\author{
Adrian Antico
}
\concept{Supervised Learning}
