% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoCatBoostHurdleModel.R, R/RemixAutoML.R
\name{AutoCatBoostdHurdleModel}
\alias{AutoCatBoostdHurdleModel}
\title{AutoCatBoostdHurdleModel is a Retrain Function for the Regression Models for the Subsetted Data in P6}
\usage{
AutoCatBoostdHurdleModel(data, ValidationData = NULL, TestData = NULL,
  Buckets = c(1, 5, 10, 20), TargetColumnName = "Target",
  FeatureColNames = 4:ncol(data), PrimaryDateColumn = NULL,
  IDcols = NULL, ClassWeights = NULL, SplitRatios = c(0.7, 0.2, 0.1),
  task_type = "GPU", ModelID = "ModelTest", Paths = NULL,
  SaveModelObjects = TRUE, Trees = 15000, GridTune = TRUE,
  MaxModelsInGrid = 1, NumOfParDepPlots = 10, PassInGrid = NULL)

AutoCatBoostdHurdleModel(data, ValidationData = NULL, TestData = NULL,
  Buckets = c(1, 5, 10, 20), TargetColumnName = "Target",
  FeatureColNames = 4:ncol(data), PrimaryDateColumn = NULL,
  IDcols = NULL, ClassWeights = NULL, SplitRatios = c(0.7, 0.2, 0.1),
  task_type = "GPU", ModelID = "ModelTest", Paths = NULL,
  SaveModelObjects = TRUE, Trees = 15000, GridTune = TRUE,
  MaxModelsInGrid = 1, NumOfParDepPlots = 10, PassInGrid = NULL)
}
\arguments{
\item{data}{Source training data. Do not include a column that has the class labels for the buckets as they are created internally.}

\item{ValidationData}{Source validation data. Do not include a column that has the class labels for the buckets as they are created internally.}

\item{TestData}{Souce test data. Do not include a column that has the class labels for the buckets as they are created internally.}

\item{Buckets}{A numeric vector of the buckets used for subsetting the data. NOTE: the final Bucket value will first create a subset of data that is less than the value and a second one thereafter for data greater than the bucket value.}

\item{TargetColumnName}{Supply the column name or number for the target variable}

\item{FeatureColNames}{Supply the column names or number of the features (not included the PrimaryDateColumn)}

\item{PrimaryDateColumn}{Supply a date column if the data is functionally related to it}

\item{IDcols}{Includes PrimaryDateColumn and any other columns you want returned in the validation data with predictions}

\item{ClassWeights}{Utilize these for the classifier model}

\item{SplitRatios}{Supply vector of partition ratios. For example, c(0.70,0.20,0,10).}

\item{task_type}{Set to "GPU" or "CPU"}

\item{ModelID}{Define a character name for your models}

\item{Paths}{A character vector of the path file strings. EITHER SUPPLY 1 file path or N file paths for N models}

\item{SaveModelObjects}{Set to TRUE to save the model objects to file in the folders listed in Paths}

\item{Trees}{Default 15000}

\item{GridTune}{Set to TRUE if you want to grid tune the models}

\item{NumOfParDepPlots}{Set to pull back N number of partial dependence calibration plots.}

\item{PassInGrid}{Pass in a grid for changing up the parameter settings for catboost}

\item{RegressionModels}{Set to the model of choice. Currently only catboost is available.}

\item{ClassificationModels}{Set to the model of choice. Currently, only catboost is available.}

\item{NumberModelsInGrid}{Set to a numeric value for the number of models to try in grid tune}

\item{data}{Source training data. Do not include a column that has the class labels for the buckets as they are created internally.}

\item{ValidationData}{Source validation data. Do not include a column that has the class labels for the buckets as they are created internally.}

\item{TestData}{Souce test data. Do not include a column that has the class labels for the buckets as they are created internally.}

\item{Buckets}{A numeric vector of the buckets used for subsetting the data. NOTE: the final Bucket value will first create a subset of data that is less than the value and a second one thereafter for data greater than the bucket value.}

\item{TargetColumnName}{Supply the column name or number for the target variable}

\item{FeatureColNames}{Supply the column names or number of the features (not included the PrimaryDateColumn)}

\item{PrimaryDateColumn}{Supply a date column if the data is functionally related to it}

\item{IDcols}{Includes PrimaryDateColumn and any other columns you want returned in the validation data with predictions}

\item{ClassWeights}{Utilize these for the classifier model}

\item{SplitRatios}{Supply vector of partition ratios. For example, c(0.70,0.20,0,10).}

\item{task_type}{Set to "GPU" or "CPU"}

\item{ModelID}{Define a character name for your models}

\item{Paths}{A character vector of the path file strings. EITHER SUPPLY 1 file path or N file paths for N models}

\item{SaveModelObjects}{Set to TRUE to save the model objects to file in the folders listed in Paths}

\item{Trees}{Default 15000}

\item{GridTune}{Set to TRUE if you want to grid tune the models}

\item{NumberModelsInGrid}{Set to a numeric value for the number of models to try in grid tune}

\item{NumOfParDepPlots}{Set to pull back N number of partial dependence calibration plots.}

\item{PassInGrid}{Pass in a grid for changing up the parameter settings for catboost}
}
\value{
Returns AutoCatBoostRegression() model objects: VariableImportance.csv, Model, ValidationData.csv, EvalutionPlot.png, EvalutionBoxPlot.png, EvaluationMetrics.csv, ParDepPlots.R a named list of features with partial dependence calibration plots, ParDepBoxPlots.R, GridCollect, and catboostgrid

Returns AutoCatBoostRegression() model objects: VariableImportance.csv, Model, ValidationData.csv, EvalutionPlot.png, EvalutionBoxPlot.png, EvaluationMetrics.csv, ParDepPlots.R a named list of features with partial dependence calibration plots, ParDepBoxPlots.R, GridCollect, and catboostgrid
}
\description{
AutoCatBoostdHurdleModel is a Retrain Function for the Regression Models for the Subsetted Data in P6

AutoCatBoostdHurdleModel is a Retrain Function for the Regression Models for the Subsetted Data in P6
}
\examples{
\donttest{
Output <- RemixAutoML::AutoCatBoostdHurdleModel(
  data,
  ValidationData = NULL,
  TestData = NULL,
  Buckets = c(1, 5, 10, 20),
  TargetColumnName = "PLND_LABOR_UNITS",
  FeatureColNames = 4:ncol(data),
  PrimaryDateColumn = "PLND_STRT_DT",
  IDcols = c(1,3),
  ClassWeights = NULL,
  SplitRatios = c(0.7, 0.2, 0.1),
  task_type = "GPU",
  ModelID = "P6",
  Paths = c(paste0(getwd(),"/P6_Buckets")),
  SaveModelObjects = TRUE,
  Trees = 5000,
  GridTune = FALSE,
  MaxModelsInGrid = 1,
  NumOfParDepPlots = 10,
  PassInGrid = grid)
}
}
\seealso{
Other Supervised Learning: \code{\link{AutoCatBoostCARMA}},
  \code{\link{AutoCatBoostClassifier}},
  \code{\link{AutoCatBoostMultiClass}},
  \code{\link{AutoCatBoostRegression}},
  \code{\link{AutoCatBoostScoring}},
  \code{\link{AutoH2OMLScoring}},
  \code{\link{AutoH2OModeler}},
  \code{\link{AutoH2OScoring}},
  \code{\link{AutoH2oDRFClassifier}},
  \code{\link{AutoH2oDRFMultiClass}},
  \code{\link{AutoH2oDRFRegression}},
  \code{\link{AutoH2oGBMClassifier}},
  \code{\link{AutoH2oGBMMultiClass}},
  \code{\link{AutoH2oGBMRegression}},
  \code{\link{AutoNLS}},
  \code{\link{AutoRecommenderScoring}},
  \code{\link{AutoRecommender}}, \code{\link{AutoTS}},
  \code{\link{AutoXGBoostClassifier}},
  \code{\link{AutoXGBoostMultiClass}},
  \code{\link{AutoXGBoostRegression}},
  \code{\link{AutoXGBoostScoring}}
}
\concept{Awesome}
\concept{Supervised Learning}
