% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoFreqSizeDistScoringFunctions.R
\name{AutoCatBoostFreqSizeScoring}
\alias{AutoCatBoostFreqSizeScoring}
\title{AutoCatBoostFreqSizeScoring is for scoring the models build with AutoCatBoostSizeFreqDist()}
\usage{
AutoCatBoostFreqSizeScoring(
  ScoringData,
  TargetColumnNames = NULL,
  FeatureColumnNames = NULL,
  IDcols = NULL,
  CountQuantiles = seq(0.1, 0.9, 0.1),
  SizeQuantiles = seq(0.1, 0.9, 0.1),
  ModelPath = NULL,
  ModelIDs = c("CountModel", "SizeModel"),
  KeepFeatures = TRUE
)
}
\arguments{
\item{ScoringData}{The scoring data returned from IntermittentDemandScoringDataGenerator()}

\item{TargetColumnNames}{A character or numeric vector of the target names. E.g. c("Counts","TARGET_qty")}

\item{FeatureColumnNames}{A character vector of column names or column numbers}

\item{IDcols}{ID columns you want returned with the data that is not a model feature}

\item{CountQuantiles}{A numerical vector of the quantiles used in model building}

\item{SizeQuantiles}{A numerical vector of the quantiles used in model building}

\item{ModelPath}{The path file to where you models were saved}

\item{ModelIDs}{The ID's used in model building}

\item{KeepFeatures}{Set to TRUE to return the features with the predicted values}
}
\value{
Returns a list of CountData scores, SizeData scores, along with count and size prediction column names
}
\description{
AutoCatBoostFreqSizeScoring is for scoring the models build with AutoCatBoostSizeFreqDist(). It will return the predicted values for every quantile model for both distributions for 1 to the max forecast periods you provided to build the scoring data.
}
\examples{
\dontrun{
FinalData <- AutoCatBoostFreqSizeScoring(
  ScoringData,
  TargetColumnNames = c("Counts","TARGET_qty"),
  FeatureColumnNames = 1:ncol(ScoringData),
  IDcols = NULL,
  CountQuantiles = seq(0.10,0.90,0.10),
  SizeQuantiles = seq(0.10,0.90,0.10),
  ModelPath = getwd(),
  ModelIDs = c("CountModel","SizeModel"),
  KeepFeatures = TRUE)
}
}
\seealso{
Other Automated Time Series: 
\code{\link{AutoCatBoostCARMA}()},
\code{\link{AutoCatBoostSizeFreqDist}()},
\code{\link{AutoH2oDRFCARMA}()},
\code{\link{AutoH2oGBMCARMA}()},
\code{\link{AutoH2oGBMFreqSizeScoring}()},
\code{\link{AutoH2oGBMSizeFreqDist}()},
\code{\link{AutoH2oGLMCARMA}()},
\code{\link{AutoH2oMLCARMA}()},
\code{\link{AutoTS}()},
\code{\link{AutoXGBoostCARMA}()},
\code{\link{CLTrainer}()},
\code{\link{IntermittentDemandScoringDataGenerator}()},
\code{\link{PredictArima}()}
}
\author{
Adrian Antico
}
\concept{Automated Time Series}
