% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReinforcementLearningFunctions.R
\name{AutoCatBoostGridSet}
\alias{AutoCatBoostGridSet}
\title{AutoCatBoostGridSet}
\usage{
AutoCatBoostGridSet(NumberPartitions = 10,
  BootstrapType = c("Bayesian", "Bernoulli", "Poisson", "MVS", "No"),
  BaggingTemperature = seq(1, 3, 1), MinDataInLeaf = c(1, 5, 10),
  MaxLeaves = c(15, 31, 45), SubSample = c(seq(0.6, 1, 0.1)),
  RandomStrength = c(0.9, 1, 1.1), SamplingFrequency = c("PerTree",
  "PreTreeLevel"), RandomSubspaceMethod = c(0.65, 0.8, 1), ODWait = 50,
  L2LeafReg = c(seq(2, 4, 1)), LearningRate = c(seq(0.01, 0.05, 0.01)),
  Depth = c(seq(4, 16, 1)))
}
\arguments{
\item{NumberPartitions}{= 10}

\item{BootstrapType}{= c("Bayesian","Bernoulli","Poisson","MVS","No")}

\item{BaggingTemperature}{= seq(1,3,1)}

\item{MinDataInLeaf}{= c(1,5,10)}

\item{MaxLeaves}{= c(15,31,45)}

\item{SubSample}{= c(seq(0.6,1,0.10))}

\item{RandomStrength}{= c(0.90,1,1.10)}

\item{SamplingFrequency}{= c("PerTree","PreTreeLevel") CPU Only}

\item{RandomSubspaceMethod}{= c(0.65,0.80,1) CPU Only}

\item{ODWait}{= 50}

\item{L2LeafReg}{= c(seq(2.0,4.0,1.0))}

\item{LearningRate}{= c(seq(0.01,0.05,0.01))}

\item{Depth}{= c(seq(4,16,1))}
}
\description{
AutoCatBoostGridSet lets the user decide on grid options
}
\seealso{
Other Supervised Learning: \code{\link{AutoH2OScoring}}
}
\author{
Adrian Antico
}
\concept{Supervised Learning}
