% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoClustering.R
\name{AutoClustering}
\alias{AutoClustering}
\title{AutoClustering}
\usage{
AutoClustering(
  data,
  FeatureColumns = NULL,
  ModelID = "TestModel",
  SavePath = NULL,
  NThreads = 8,
  MaxMemory = "28G",
  MaxClusters = 50,
  ClusterMetric = "totss",
  RunDimReduction = TRUE,
  ShrinkRate = (sqrt(5) - 1)/2,
  Epochs = 5L,
  L2_Reg = 0.1,
  ElasticAveraging = TRUE,
  ElasticAveragingMovingRate = 0.9,
  ElasticAveragingRegularization = 0.001
)
}
\arguments{
\item{data}{is the source time series data.table}

\item{FeatureColumns}{Independent variables}

\item{ModelID}{For naming the files to save}

\item{SavePath}{Directory path for saving models}

\item{NThreads}{set based on number of threads your machine has available}

\item{MaxMemory}{set based on the amount of memory your machine has available}

\item{MaxClusters}{number of factors to test out in k-means to find the optimal number}

\item{ClusterMetric}{pick the metric to identify top model in grid tune c("totss","betweenss","withinss")}

\item{RunDimReduction}{If TRUE, an autoencoder will be built to reduce the feature space. Otherwise, all features in FeatureColumns will be used for clustering}

\item{ShrinkRate}{Node shrink rate for H2OAutoencoder. See that function for details.}

\item{Epochs}{For the autoencoder}

\item{L2_Reg}{For the autoencoder}

\item{ElasticAveraging}{For the autoencoder}

\item{ElasticAveragingMovingRate}{For the autoencoder}

\item{ElasticAveragingRegularization}{For the autoencoder}
}
\value{
Original data.table with added column with cluster number identifier
}
\description{
AutoClustering adds a column to your original data with a cluster number identifier. You can run request an autoencoder to be built to reduce the dimensionality of your data before running the clusering algo.
}
\examples{
\dontrun{
#########################
# Training Setup
#########################

# Create fake data
data <- RemixAutoML::FakeDataGenerator(
  Correlation = 0.85,
  N = 1000,
  ID = 2,
  ZIP = 0,
  AddDate = TRUE,
  Classification = FALSE,
  MultiClass = FALSE)

# Run function
data <- RemixAutoML::AutoClustering(
  data,
  FeatureColumns = names(data)[2:(ncol(data)-1)],
  ModelID = "TestModel",
  SavePath = getwd(),
  NThreads = 8,
  MaxMemory = "28G",
  MaxClusters = 50,
  ClusterMetric = "totss",
  RunDimReduction = TRUE,
  ShrinkRate = (sqrt(5) - 1) / 2,
  Epochs = 5L,
  L2_Reg = 0.10,
  ElasticAveraging = TRUE,
  ElasticAveragingMovingRate = 0.90,
  ElasticAveragingRegularization = 0.001)

#########################
# Scoring Setup
#########################

Sys.sleep(10)

# Create fake data
data <- RemixAutoML::FakeDataGenerator(
  Correlation = 0.85,
  N = 1000,
  ID = 2,
  ZIP = 0,
  AddDate = TRUE,
  Classification = FALSE,
  MultiClass = FALSE)

# Run function
data <- RemixAutoML::AutoClusteringScoring(
  data,
  FeatureColumns = names(data)[2:(ncol(data)-1)],
  ModelID = "TestModel",
  SavePath = getwd(),
  NThreads = 8,
  MaxMemory = "28G",
  DimReduction = TRUE)
}

}
\seealso{
Other Unsupervised Learning: 
\code{\link{AutoClusteringScoring}()},
\code{\link{GenTSAnomVars}()},
\code{\link{H2OIsolationForestScoring}()},
\code{\link{H2OIsolationForest}()},
\code{\link{ResidualOutliers}()}
}
\author{
Adrian Antico
}
\concept{Unsupervised Learning}
