% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoClustering.R
\name{AutoClusteringScoring}
\alias{AutoClusteringScoring}
\title{AutoClusteringScoring}
\usage{
AutoClusteringScoring(
  data,
  FeatureColumns = NULL,
  ModelID = "TestModel",
  SavePath = NULL,
  NThreads = 8,
  MaxMemory = "28G",
  DimReduction = TRUE
)
}
\arguments{
\item{data}{is the source time series data.table}

\item{FeatureColumns}{Independent variables}

\item{ModelID}{This is returned from the training run in the output list with element named 'model_name'. It's not identical to the ModelID used in training due to the grid tuning.}

\item{SavePath}{Directory path for saving models}

\item{NThreads}{set based on number of threads your machine has available}

\item{MaxMemory}{set based on the amount of memory your machine has available}

\item{DimReduction}{Set to TRUE if you set RunDimReduction in the training version of this function}
}
\value{
Original data.table with added column with cluster number identifier
}
\description{
AutoClusteringScoring adds a column to your original data with a cluster number identifier. You can run request an autoencoder to be built to reduce the dimensionality of your data before running the clusering algo.
}
\seealso{
Other Unsupervised Learning: 
\code{\link{AutoClustering}()},
\code{\link{GenTSAnomVars}()},
\code{\link{H2OIsolationForestScoring}()},
\code{\link{H2OIsolationForest}()},
\code{\link{ResidualOutliers}()}
}
\author{
Adrian Antico
}
\concept{Unsupervised Learning}
