% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EDAFunctions.R
\name{AutoCorrAnalysis}
\alias{AutoCorrAnalysis}
\title{AutoCorrAnalysis}
\usage{
AutoCorrAnalysis(
  data = NULL,
  CorVars = NULL,
  SkipCorVars = NULL,
  ByGroupVars = NULL,
  DataSampleRate = 0.5,
  MinRows = 30,
  KeepSignificantVars = TRUE,
  PValAdjMethod = "holm",
  RobustCalc = TRUE,
  PartialCorr = FALSE,
  BayesianCorr = FALSE
)
}
\arguments{
\item{data}{data.table}

\item{CorVars}{Can leave NULL or supply column names you want to analyze}

\item{SkipCorVars}{Can leave NULL or supply column names you want to skip}

\item{ByGroupVars}{Categorical variables to run correlation analysis by}
}
\description{
Generate correlation analysis over a data set
}
\examples{
\dontrun{
data <- RemixAutoML::FakeDataGenerator(
  Correlation = 0.85,
  N = 10000L,
  ID = 2L,
  FactorCount = 2L,
  AddDate = TRUE,
  ZIP = 2L,
  TimeSeries = FALSE,
  ChainLadder = FALSE,
  Classification = TRUE,
  MultiClass = FALSE)

# Run Analysis
data <- RemixAutoML::AutoCorrAnalysis(
  data = data,
  CorVars = NULL,
  SkipCorVars = c("IDcol_1","IDcol_2","DateTime"),
  ByGroupVars = "Factor_1",
  DataSampleRate = 0.50,
  MinRows = 30,
  KeepSignificantVars = TRUE,
  PValAdjMethod = "holm",
  RobustCalc = TRUE,
  PartialCorr = FALSE,
  BayesianCorr = FALSE)
}
}
\seealso{
Other EDA: 
\code{\link{AutoWordFreq}()},
\code{\link{BNLearnArcStrength}()},
\code{\link{ProblematicFeatures}()}
}
\author{
Adrian Antico
}
\concept{EDA}
