% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RemixAutoML.R
\name{AutoDataPartition}
\alias{AutoDataPartition}
\title{The AutoRecomScoring function scores recommender models from AutoRecommender()}
\usage{
AutoDataPartition(data, NumDataSets = 3, Ratios = c(0.7, 0.2, 0.1),
  PartitionType = "random", StratifyColumnNames = NULL,
  TimeColumnName = NULL)
}
\arguments{
\item{data}{The binary ratings matrix from RecomDataCreate()}

\item{NumDataSets}{The winning model returned from AutoRecommender()}

\item{Ratios}{Typically your customer ID}

\item{PartitionType}{Set to either "random", "timeseries", or "time". With "random", your data will be paritioned randomly (with stratified sampling if column names are supplied). With "timeseries", you can partition by time with a stratify option (so long as you have an equal number of records for each strata). With "time" you will have data sets generated so that the training data contains the earliest records in time, validation data the second earliest, test data the third earliest, etc.}

\item{StratifyColumnNames}{Supply column names of categorical features to use in a stratified sampling procedure for partitioning the data. Partition type must be "random" to use this option}

\item{TimeColumnName}{Supply a date column name or a name of a column with an ID for sorting by time such that the smallest number is the earliest in time.}
}
\value{
Returns a list of data.tables
}
\description{
This function will take your ratings matrix and model and score your data in parallel.
}
\examples{
\donttest{
dataSets <- AutoDataPartition(data,
                              NumDataSets = 3,
                              Ratios = c(0.70,0.20,0.10),
                              PartitionType = "random",
                              StratifyColumnNames = NULL,
                              TimeColumnName = NULL)
}
}
\seealso{
Other Feature Engineering: \code{\link{AutoWord2VecModeler}},
  \code{\link{CreateCalendarVariables}},
  \code{\link{DT_GDL_Feature_Engineering}},
  \code{\link{DummifyDT}},
  \code{\link{FAST_GDL_Feature_Engineering}},
  \code{\link{GDL_Feature_Engineering}},
  \code{\link{ModelDataPrep}},
  \code{\link{Scoring_GDL_Feature_Engineering}}
}
\author{
Adrian Antico and Douglas Pestana
}
\concept{Feature Engineering}
