% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoH2oDRFHurdleModel.R
\name{AutoH2oDRFHurdleModel}
\alias{AutoH2oDRFHurdleModel}
\title{AutoH2oDRFHurdleModel is generalized hurdle modeling framework}
\usage{
AutoH2oDRFHurdleModel(data, ValidationData = NULL, TestData = NULL,
  Buckets = 0L, TargetColumnName = NULL, FeatureColNames = NULL,
  TransformNumericColumns = NULL, SplitRatios = c(0.7, 0.2, 0.1),
  ModelID = "ModelTest", Paths = NULL, MetaDataPaths = NULL,
  SaveModelObjects = TRUE, IfSaveModel = "mojo", MaxMem = "28G",
  NThreads = max(1L, parallel::detectCores() - 2L), Trees = 1000L,
  GridTune = TRUE, MaxModelsInGrid = 1L, NumOfParDepPlots = 10L,
  PassInGrid = NULL)
}
\arguments{
\item{data}{Source training data. Do not include a column that has the class labels for the buckets as they are created internally.}

\item{ValidationData}{Source validation data. Do not include a column that has the class labels for the buckets as they are created internally.}

\item{TestData}{Souce test data. Do not include a column that has the class labels for the buckets as they are created internally.}

\item{Buckets}{A numeric vector of the buckets used for subsetting the data. NOTE: the final Bucket value will first create a subset of data that is less than the value and a second one thereafter for data greater than the bucket value.}

\item{TargetColumnName}{Supply the column name or number for the target variable}

\item{FeatureColNames}{Supply the column names or number of the features (not included the PrimaryDateColumn)}

\item{TransformNumericColumns}{Transform numeric column inside the AutoCatBoostRegression() function}

\item{SplitRatios}{Supply vector of partition ratios. For example, c(0.70,0.20,0,10).}

\item{ModelID}{Define a character name for your models}

\item{Paths}{The path to your folder where you want your model information saved}

\item{MetaDataPaths}{A character string of your path file to where you want your model evaluation output saved. If left NULL, all output will be saved to Paths.}

\item{SaveModelObjects}{Set to TRUE to save the model objects to file in the folders listed in Paths}

\item{IfSaveModel}{Save as "mojo" or "standard"}

\item{MaxMem}{Set the maximum memory your system can provide}

\item{NThreads}{Set the number of threads you want to dedicate to the model building}

\item{Trees}{Default 1000}

\item{GridTune}{Set to TRUE if you want to grid tune the models}

\item{MaxModelsInGrid}{Set to a numeric value for the number of models to try in grid tune}

\item{NumOfParDepPlots}{Set to pull back N number of partial dependence calibration plots.}

\item{PassInGrid}{Pass in a grid for changing up the parameter settings for catboost}
}
\value{
Returns AutoXGBoostRegression() model objects: VariableImportance.csv, Model, ValidationData.csv, EvalutionPlot.png, EvalutionBoxPlot.png, EvaluationMetrics.csv, ParDepPlots.R a named list of features with partial dependence calibration plots, ParDepBoxPlots.R, GridCollect, and the grid used
}
\description{
AutoH2oDRFHurdleModel is generalized hurdle modeling framework
}
\examples{
\donttest{
Output <- AutoH2oDRFHurdleModel(
  data,
  ValidationData = NULL,
  TestData = NULL,
  Buckets = 1L,
  TargetColumnName = "Target_Variable",
  FeatureColNames = 4:ncol(data),
  TransformNumericColumns = NULL,
  SplitRatios = c(0.7, 0.2, 0.1),
  NThreads = max(1L, parallel::detectCores()-2L),
  ModelID = "ModelID",
  Paths = NULL,
  MetaDataPaths = NULL,
  SaveModelObjects = TRUE,
  IfSaveModel = "mojo",
  Trees = 1000L,
  GridTune = FALSE,
  MaxModelsInGrid = 1L,
  NumOfParDepPlots = 10L,
  PassInGrid = NULL)
}
}
\seealso{
Other Automated Regression: \code{\link{AutoCatBoostHurdleModel}},
  \code{\link{AutoCatBoostRegression}},
  \code{\link{AutoH2oDRFRegression}},
  \code{\link{AutoH2oGBMHurdleModel}},
  \code{\link{AutoH2oGBMRegression}},
  \code{\link{AutoH2oGLMRegression}},
  \code{\link{AutoH2oMLRegression}}, \code{\link{AutoNLS}},
  \code{\link{AutoXGBoostHurdleModel}},
  \code{\link{AutoXGBoostRegression}}
}
\concept{Automated Regression}
