% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoFreqSizeDistScoringFunctions.R
\name{AutoH2oGBMFreqSizeScoring}
\alias{AutoH2oGBMFreqSizeScoring}
\title{AutoH2oGBMFreqSizeScoring is for scoring the models build with AutoH2oGBMSizeFreqDist()}
\usage{
AutoH2oGBMFreqSizeScoring(
  ScoringData,
  TargetColumnNames = NULL,
  CountQuantiles = seq(0.1, 0.9, 0.1),
  SizeQuantiles = seq(0.1, 0.9, 0.1),
  ModelPath = NULL,
  ModelIDs = c("CountModel", "SizeModel"),
  JavaOptions = "-Xmx1g -XX:ReservedCodeCacheSize=256m",
  KeepFeatures = TRUE
)
}
\arguments{
\item{ScoringData}{The scoring data returned from IntermittentDemandScoringDataGenerator()}

\item{TargetColumnNames}{A character or numeric vector of the target names. E.g. c("Counts","TARGET_qty")}

\item{CountQuantiles}{A numerical vector of the quantiles used in model building}

\item{SizeQuantiles}{A numerical vector of the quantiles used in model building}

\item{ModelPath}{The path file to where you models were saved}

\item{ModelIDs}{The ID's used in model building}

\item{JavaOptions}{For mojo scoring '-Xmx1g -XX:ReservedCodeCacheSize=256m',}

\item{KeepFeatures}{Set to TRUE to return the features with the predicted values}
}
\value{
Returns a list of CountData scores, SizeData scores, along with count and size prediction column names
}
\description{
AutoH2oGBMFreqSizeScoring is for scoring the models build with AutoH2oGBMSizeFreqDist(). It will return the predicted values for every quantile model for both distributions for 1 to the max forecast periods you provided to build the scoring data.
}
\examples{
\dontrun{
FinalData <- AutoH2oGBMFreqSizeScoring(
  ScoringData,
  TargetColumnNames = c("Counts","TARGET_qty"),
  CountQuantiles = seq(0.10,0.90,0.10),
  SizeQuantiles = seq(0.10,0.90,0.10),
  ModelPath = getwd(),
  ModelIDs = c("CountModel","SizeModel"),
  JavaOptions = '-Xmx1g -XX:ReservedCodeCacheSize=256m',
  KeepFeatures = TRUE)
}
}
\seealso{
Other Automated Time Series: 
\code{\link{AutoBanditNNet}()},
\code{\link{AutoBanditSarima}()},
\code{\link{AutoCatBoostFreqSizeScoring}()},
\code{\link{AutoTBATS}()},
\code{\link{AutoTS}()},
\code{\link{IntermittentDemandScoringDataGenerator}()}
}
\author{
Adrian Antico
}
\concept{Automated Time Series}
