% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RemixAutoML.R
\name{AutoH2OMLScoring}
\alias{AutoH2OMLScoring}
\title{AutoH2OMLScoring is an automated scoring function that compliments the AutoH2o model training functions.}
\usage{
AutoH2OMLScoring(ScoringData = NULL, FeatureColNames = NULL,
  ModelType = "mojo", H2OShutdown = TRUE, MaxMem = "28G",
  JavaOptions = "-Xmx1g -XX:ReservedCodeCacheSize=256m",
  ModelPath = NULL, ModelID = NULL, ReturnFeatures = TRUE,
  MDP_Impute = TRUE, MDP_CharToFactor = TRUE, MDP_RemoveDates = TRUE,
  MDP_MissFactor = "0", MDP_MissNum = -1)
}
\arguments{
\item{ScoringData}{This is your data.table of features for scoring. Can be a single row or batch.}

\item{ModelType}{Set to either "mojo" or "standard" depending on which version you saved}

\item{H2OShutdown}{Set to TRUE is you are scoring a "standard" model file and you aren't planning on continuing to score.}

\item{MaxMem}{Set to you dedicated amount of memory. E.g. "28G"}

\item{JavaOptions}{Change the default to your machines specification if needed. Default is '-Xmx1g -XX:ReservedCodeCacheSize=256m',}

\item{ModelPath}{Supply your path file used in the AutoH2o__() function}

\item{ModelID}{Supply the model ID used in the AutoH2o__() function}

\item{ReturnFeatures}{Set to TRUE to return your features with the predicted values.}

\item{MDP_Impute}{Set to TRUE if you did so for modeling and didn't do so before supplying ScoringData in this function}

\item{MDP_CharToFactor}{Set to TRUE to turn your character columns to factors if you didn't do so to your ScoringData that you are supplying to this function}

\item{MDP_RemoveDates}{Set to TRUE if you have date of timestamp columns in your ScoringData}

\item{MDP_MissFactor}{If you set MDP_Impute to TRUE, supply the character values to replace missing values with}

\item{MDP_MissNum}{If you set MDP_Impute to TRUE, supply a numeric value to replace missing values with}

\item{FeatureColumnNames}{Supply either column names or column numbers used in the AutoH2o__() function}
}
\value{
A data.table of predicted values with the option to return model features as well.
}
\description{
AutoH2OMLScoring is an automated scoring function that compliments the AutoH2oGBM__() and AutoH2oDRF__() models training functions. This function requires you to supply features for scoring. It will run ModelDataPrep()to prepare your features for H2O data conversion and scoring.
}
\examples{
\donttest{
Preds <- AutoH2OMLScoring(ScoringData = data,
                          FeatureColNames = 2:12,
                          ModelType = "mojo",
                          H2OShutdown = TRUE,
                          MaxMem = "28G",
                          JavaOptions = '-Xmx1g -XX:ReservedCodeCacheSize=256m',
                          ModelPath = NULL,
                          ModelID = "ModelTest",
                          ReturnFeatures = TRUE,
                          MDP_Impute = TRUE,
                          MDP_CharToFactor = TRUE,
                          MDP_RemoveDates = TRUE,
                          MDP_MissFactor = "0",
                          MDP_MissNum = -1)
}
}
\seealso{
Other Supervised Learning: \code{\link{AutoCatBoostClassifier}},
  \code{\link{AutoCatBoostMultiClass}},
  \code{\link{AutoCatBoostRegression}},
  \code{\link{AutoCatBoostScoring}},
  \code{\link{AutoH2OModeler}},
  \code{\link{AutoH2OScoring}},
  \code{\link{AutoH2oDRFClassifier}},
  \code{\link{AutoH2oDRFMultiClass}},
  \code{\link{AutoH2oDRFRegression}},
  \code{\link{AutoH2oGBMClassifier}},
  \code{\link{AutoH2oGBMMultiClass}},
  \code{\link{AutoH2oGBMRegression}},
  \code{\link{AutoMLTS}}, \code{\link{AutoNLS}},
  \code{\link{AutoRecommenderScoring}},
  \code{\link{AutoRecommender}}, \code{\link{AutoTS}},
  \code{\link{AutoXGBoostClassifier}},
  \code{\link{AutoXGBoostMultiClass}},
  \code{\link{AutoXGBoostRegression}},
  \code{\link{AutoXGBoostScoring}}
}
\concept{Supervised Learning}
