% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RemixAutoML.R
\name{AutoH2OTextPrepScoring}
\alias{AutoH2OTextPrepScoring}
\title{AutoH2OTextPrepScoring is for NLP scoring}
\usage{
AutoH2OTextPrepScoring(data, string, MaxMem, NThreads)
}
\arguments{
\item{data}{The text data}

\item{string}{The name of the string column to prepare}

\item{MaxMem}{Amount of memory you want to let H2O utilize}

\item{NThreads}{The number of threads you want to let H2O utilize}
}
\description{
This function returns prepared tokenized data for H2O Word2VecModeler scoring
}
\examples{
\donttest{
data <- AutoH2OTextPrepScoring(data = x,
                               string = "text_column",
                               MaxMem = "28G",
                               NThreads = 8)
}
}
\seealso{
Other Misc: \code{\link{ChartTheme}},
  \code{\link{PrintObjectsSize}}, \code{\link{RemixTheme}},
  \code{\link{SimpleCap}}, \code{\link{multiplot}},
  \code{\link{percRank}}, \code{\link{tempDatesFun}},
  \code{\link{tokenizeH2O}}
}
\author{
Adrian Antico
}
\concept{Misc}
