% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EconometricsFunctions.R
\name{AutoHierarchicalFourier}
\alias{AutoHierarchicalFourier}
\title{AutoHierarchicalFourier}
\usage{
AutoHierarchicalFourier(
  datax = data,
  xRegs = names(XREGS),
  FourierTermS = FourierTerms,
  TimeUniT = TimeUnit,
  FC_PeriodS = FC_Periods,
  TargetColumN = TargetColumn,
  DateColumN = DateColumnName,
  HierarchGroups = NULL,
  IndependentGroups = NULL
)
}
\arguments{
\item{datax}{data}

\item{xRegs}{The XREGS}

\item{FourierTermS}{Number of fourier pairs}

\item{TimeUniT}{Time unit}

\item{FC_PeriodS}{Number of forecast periods}

\item{TargetColumN}{Target column name}

\item{DateColumN}{Date column name}

\item{HierarchGroups}{Character vector of categorical columns to fully interact}

\item{IndependentGroups}{Character vector of categorical columns to run independently}
}
\description{
AutoHierarchicalFourier reverses the difference
}
\seealso{
Other Feature Engineering: 
\code{\link{AutoDataPartition}()},
\code{\link{AutoDiffLagN}()},
\code{\link{AutoInteraction}()},
\code{\link{AutoLagRollStatsScoring}()},
\code{\link{AutoLagRollStats}()},
\code{\link{AutoTransformationCreate}()},
\code{\link{AutoTransformationScore}()},
\code{\link{AutoWord2VecModeler}()},
\code{\link{AutoWord2VecScoring}()},
\code{\link{CategoricalEncoding}()},
\code{\link{CreateCalendarVariables}()},
\code{\link{CreateHolidayVariables}()},
\code{\link{DummifyDT}()},
\code{\link{H2OAutoencoderScoring}()},
\code{\link{H2OAutoencoder}()},
\code{\link{ModelDataPrep}()},
\code{\link{TimeSeriesFill}()}
}
\author{
Adrian Antico
}
\concept{Feature Engineering}
