% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoHurdleScoring.R
\name{AutoHurdleScoring}
\alias{AutoHurdleScoring}
\title{AutoHurdleScoring()}
\usage{
AutoHurdleScoring(
  TestData = NULL,
  ModelClass = "catboost",
  ArgList = Output$ArgsList,
  ModelList = Output$ModelList
)
}
\arguments{
\item{TestData}{scoring data.table}

\item{ModelClass}{Name of model type. "catboost" is currently the only available option}

\item{ArgList}{Output from the hurdle model}

\item{ModelList}{Output from the hurdle model}
}
\description{
AutoHurdleScoring()
}
\seealso{
Other Automated Model Scoring: 
\code{\link{AutoCatBoostScoring}()},
\code{\link{AutoH2OMLScoring}()},
\code{\link{AutoH2OModeler}()},
\code{\link{AutoXGBoostScoring}()}
}
\author{
Adrian Antico
}
\concept{Automated Model Scoring}
