% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataWranglingFunctions.R
\name{AutoInteraction}
\alias{AutoInteraction}
\title{AutoNumericInteraction}
\usage{
AutoInteraction(
  data = NULL,
  NumericVars = NULL,
  InteractionDepth = NULL,
  Center = TRUE,
  Scale = TRUE,
  SkipCols = NULL
)
}
\arguments{
\item{data}{Source data.table}

\item{InteractionDepth}{The max K in N choose K. If NULL, K will loop through 1 to length(NumVars)}

\item{Center}{TRUE to center the data}

\item{Scale}{TRUE to scale the data}

\item{SkipCols}{Use this to exclude features from being created. An example could be, you build a model with all variables and then use the varaible importance list to determine which features aren't necessary and pass that set of features into this argument as a character vector.}

\item{NumVars}{Names of numeric columns (if NULL, all numeric and integer columns will be used)}
}
\description{
AutoNumericInteraction creates interaction variables from your numerical features in your data. Supply a set of column names to utilize and set the interaction level. Supply a character vector of columns to exclude and the function will ignore those features.
}
\examples{
\dontrun{

# Create fake data
data <- RemixAutoML::FakeDataGenerator(
  Correlation = 0.70,
  N = 1000,
  ID = 2L,
  FactorCount = 2L,
  AddDate = TRUE,
  ZIP = 0L,
  TimeSeries = FALSE,
  ChainLadderData = FALSE,
  Classification = FALSE,
  MultiClass = FALSE)

# Store names of numeric and integer cols
Cols <-names(data)[c(which(unlist(lapply(data, is.numeric))),
                     which(unlist(lapply(data, is.integer))))]

# Create features
data <- RemixAutoML::AutoInteraction(
  data = data,
  NumericVars = Cols,
  InteractionDepth = 4,
  Center = TRUE,
  Scale = TRUE,
  SkipCols = NULL)
}

}
\seealso{
Other Feature Engineering: 
\code{\link{AutoDataPartition}()},
\code{\link{AutoHierarchicalFourier}()},
\code{\link{AutoLagRollStatsScoring}()},
\code{\link{AutoLagRollStats}()},
\code{\link{AutoTransformationCreate}()},
\code{\link{AutoTransformationScore}()},
\code{\link{AutoWord2VecModeler}()},
\code{\link{ContinuousTimeDataGenerator}()},
\code{\link{CreateCalendarVariables}()},
\code{\link{CreateHolidayVariables}()},
\code{\link{DT_GDL_Feature_Engineering}()},
\code{\link{DummifyDT}()},
\code{\link{H2oAutoencoder}()},
\code{\link{ModelDataPrep}()},
\code{\link{Partial_DT_GDL_Feature_Engineering}()},
\code{\link{TimeSeriesFill}()}
}
\author{
Adrian Antico
}
\concept{Feature Engineering}
