% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoMarketBasketModel.R
\name{AutoMarketBasketModel}
\alias{AutoMarketBasketModel}
\title{AutoMarketBasketModel function runs a market basket analysis automatically}
\usage{
AutoMarketBasketModel(data, OrderIDColumnName, ItemIDColumnName,
  LHS_Delimeter = ",", Support = 0.001, Confidence = 0.1,
  MaxLength = 2, MinLength = 2, MaxTime = 5)
}
\arguments{
\item{data}{This is your transactions data set}

\item{OrderIDColumnName}{Supply your column name for the Order ID Values}

\item{ItemIDColumnName}{Supply your column name for the Item ID Values}

\item{LHS_Delimeter}{Default delimeter for separating multiple ItemID's is a comma.}

\item{Support}{Threshold for inclusion using support}

\item{Confidence}{Threshold for inclusion using confidence}

\item{MaxLength}{Maximum combinations of Item ID (number of items in basket to consider)}

\item{MinLength}{Minimum length of combinations of ItemID (number of items in basket to consider)}

\item{MaxTime}{Max run time per iteration (default is 5 seconds)}
}
\description{
AutoMarketBasketModel function runs a market basket analysis automatically. It will convert your data, run the algorithm, and add on additional significance values not orginally contained within.
}
\examples{
\donttest{
rules_data <- AutoMarketBasketModel(
  data,
  OrderIDColumnName = "OrderNumber",
  ItemIDColumnName = "ItemNumber",
  LHS_Delimeter = ",",
  Support = 0.001,
  Confidence = 0.1,
  MaxLength = 2,
  MinLength = 2,
  MaxTime = 5)
}
}
\seealso{
Chi-sq statistics and p-values based on this paper: http://www.cs.bc.edu/~alvarez/ChiSquare/chi2tr.pdf

Other Recommender Systems: \code{\link{AutoRecomDataCreate}},
  \code{\link{AutoRecommenderScoring}},
  \code{\link{AutoRecommender}}
}
\author{
Adrian Antico and Douglas Pestana
}
\concept{Recommender Systems}
