% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RemixAutoML.R
\name{AutoNLS}
\alias{AutoNLS}
\title{AutoNLS is a function for automatically building nls models}
\usage{
AutoNLS(data, y, x, monotonic = TRUE)
}
\arguments{
\item{data}{Data is the data table you are building the modeling on}

\item{y}{Y is the target variable name in quotes}

\item{x}{X is the independent variable name in quotes}

\item{monotonic}{This is a TRUE/FALSE indicator - choose TRUE if you want monotonic regression over polynomial regression}
}
\value{
A list containing 1: A data table with your original column replaced by the nls model predictions; 2: The model name; 3: The winning model to later use.
}
\description{
This function will build models for 9 different nls models, along with a non-parametric monotonic regression and a polynomial regression. The models are evaluated, a winner is picked, and the predicted values are stored in your data table.
}
\examples{
data <- data.table::data.table(Variable = seq(1,500,1), Target = rep(1, 500))
for (i in as.integer(1:500)) {
  if(i == 1) {
    var <- data[i, "Variable"][[1]]
    data.table::set(data, i = i, j = 2L, value = var * (1 + runif(1)/100))
} else {
    var = data[i-1, "Target"][[1]]
    data.table::set(data, i = i, j = 2L, value = var * (1 + runif(1)/100))
  }
}

# To keep original values
data1 <- data.table::copy(data)

# Merge and Model data
data2 <- merge(data1,
               AutoNLS(data = data, y = "Target", x = "Variable", monotonic = FALSE),
               by = "Variable",
               all = TRUE)

# Plot graphs of predicted vs actual
p <- ggplot2::ggplot(data2, ggplot2::aes(x = Variable)) +
  ggplot2::geom_line(ggplot2::aes(y = data2[["Target.x"]], color = "blue")) +
  ggplot2::geom_line(ggplot2::aes(y = data2[["Target.y"]], color = "red")) +
  ChartTheme(Size = 12) + ggplot2::ggtitle("Growth Models") +
  ggplot2::ylab("Target Variable") + ggplot2::xlab("Independent Variable")
}
\seealso{
Other Supervised Learning: \code{\link{AutoH20Modeler}},
  \code{\link{AutoH20Scoring}}, \code{\link{AutoTS}}
}
\author{
Adrian Antico
}
\concept{Supervised Learning}
