% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotFunctions.R
\name{AutoPlotter}
\alias{AutoPlotter}
\title{AutoPlotter}
\usage{
AutoPlotter(
  dt = NULL,
  PlotType = "Scatter",
  YVar = NULL,
  XVar = NULL,
  YMin = NULL,
  YMax = NULL,
  XMin = NULL,
  XMax = NULL,
  ColorVariables = "None",
  SizeVar1 = "None",
  FacetVar1 = "None",
  FacetVar2 = "None",
  YTicks = "Default",
  XTicks = "Default",
  OutlierSize = 0.1,
  OutlierColor = "blue",
  BoxPlotFill = "gray",
  GamFitScatter = FALSE,
  TextSize = 12,
  AngleX = 90,
  AngleY = 0,
  ChartColor = "lightsteelblue1",
  BorderColor = "darkblue",
  TextColor = "darkblue",
  GridColor = "white",
  BackGroundColor = "gray95",
  SubTitleColor = "blue",
  LegendPosition = "bottom",
  LegendBorderSize = 0.5,
  LegendLineType = "solid",
  Debug = FALSE
)
}
\arguments{
\item{dt}{= NULL}

\item{PlotType}{= input[['PlotType']]}

\item{YVar}{= shiny::isolate(YVar())}

\item{XVar}{= shiny::isolate(DateVar())}

\item{YMin}{= input[['YMin']]}

\item{YMax}{= input[['YMax']]}

\item{XMin}{= input[['XMin']]}

\item{XMax}{= input[['XMax']]}

\item{ColorVariables}{= shiny::isolate(SelectedGroups())}

\item{SizeVar1}{= input[['SizeVar1']]}

\item{FacetVar1}{= shiny::isolate(input[['FacetVar1']])}

\item{FacetVar2}{= shiny::isolate(input[['FacetVar2']])}

\item{YTicks}{= input[['YTicks']]}

\item{XTicks}{= input[['XTicks']]}

\item{OutlierSize}{= input[['OutlierSize']]}

\item{OutlierColor}{= input[['OutlierColor']]}

\item{BoxPlotFill}{= input[['BoxPlotFill']]}

\item{GamFitScatter}{= input[['GamFitScatter']]}

\item{TextSize}{= input[['TextSize']]}

\item{AngleX}{= input[['AngleX']]}

\item{AngleY}{= input[['AngleY']]}

\item{ChartColor}{= input[['ChartColor']]}

\item{BorderColor}{= input[['BorderColor']]}

\item{TextColor}{= input[['TextColor']]}

\item{GridColor}{= input[['GridColor']]}

\item{BackGroundColor}{= input[['BackGroundColor']]}

\item{SubTitleColor}{= input[['SubTitleColor']]}

\item{LegendPosition}{= input[['LegendPosition']]}

\item{LegendBorderSize}{= as.numeric(input[['LegendBorderSize']])}

\item{LegendLineType}{= input[['LegendLineType']]}

\item{Debug}{= FALSE}
}
\description{
Create plots
}
\seealso{
Other Graphics: 
\code{\link{AppModelInsights}()},
\code{\link{ChartTheme}()},
\code{\link{multiplot}()}
}
\author{
Adrian Antico
}
\concept{Graphics}
