% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoRecomDataCreate.R
\name{AutoRecomDataCreate}
\alias{AutoRecomDataCreate}
\title{Convert transactional data.table to a binary ratings matrix}
\usage{
AutoRecomDataCreate(data, EntityColName = "CustomerID",
  ProductColName = "StockCode", MetricColName = "TotalSales",
  ReturnMatrix = FALSE)
}
\arguments{
\item{data}{This is your transactional data.table. Must include an Entity (typically customer), ProductCode (such as SKU), and a sales metric (such as total sales).}

\item{EntityColName}{This is the column name in quotes that represents the column name for the Entity, such as customer}

\item{ProductColName}{This is the column name in quotes that represents the column name for the product, such as SKU}

\item{MetricColName}{This is the column name in quotes that represents the column name for the metric, such as total sales}

\item{ReturnMatrix}{Set to FALSE to coerce the object (desired route) or TRUE to return a matrix}
}
\value{
A BinaryRatingsMatrix
}
\description{
Convert transactional data.table to a binary ratings matrix
}
\examples{
data <- data.table::data.table(CustomerID = c(1,1,2,2,3,3),
                               StockCode = c("A","B","A","A","B","A"),
                               TotalSales = c(2,3,4,5,1,2))
RatingsMatrix <- AutoRecomDataCreate(data,
                                     EntityColName = "CustomerID",
                                     ProductColName = "StockCode",
                                     MetricColName = "TotalSales",
                                     ReturnMatrix = TRUE)
}
\seealso{
Other Marketing Modeling: \code{\link{AutoMarketBasketModel}},
  \code{\link{AutoRecommenderScoring}},
  \code{\link{AutoRecommender}}
}
\author{
Adrian Antico and Douglas Pestana
}
\concept{Marketing Modeling}
