% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RemixAutoML.R
\name{AutoRecommender}
\alias{AutoRecommender}
\title{Automatically build the best recommendere model.}
\usage{
AutoRecommender(data, Partition = "Split", KFolds = 2, Ratio = 0.75,
  RatingType = "TopN", RatingsKeep = 20,
  SkipModels = "AssociationRules")
}
\arguments{
\item{data}{This is your BinaryRatingsMatrix. See function RecomDataCreate}

\item{Partition}{Choose from "split", "cross-validation", "bootstrap". See evaluationScheme in recommenderlab for details.}

\item{KFolds}{Choose 2 for traditional train / test. Choose > 2 for the number of cross validations}

\item{Ratio}{The ratio for train and test. E.g. 0.75 for 75% data allocated to training}

\item{RatingsKeep}{The total ratings you wish to return. Default is 20.}

\item{SkipModels}{AssociationRules runs the slowest and may crash your system. Choose from: "AssociationRules","ItemBasedCF","UserBasedCF","PopularItems","RandomItems"}

\item{RatingsType}{Choose from “topNList”, “ratings”, “ratingMatrix”}
}
\description{
Automatically build the best recommendere model.
}
\examples{
\dontrun{
ModelResults <- AutoRecommender(BinaryRatingsMatrix,
                                Partition = "Split",
                                KFolds = 2,
                                Ratio = 0.75,
                                RatingType = "TopN",
                                RatingsKeep = 20,
                                SkipModels = "AssociationRules")
}
}
\author{
Adrian Antico and Douglas Pestana
}
