% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoRecommender.R
\name{AutoRecommenderDataCreate}
\alias{AutoRecommenderDataCreate}
\title{AutoRecommenderDataCreate}
\usage{
AutoRecommenderDataCreate(
  data,
  EntityColName = "CustomerID",
  ProductColName = "StockCode",
  MetricColName = "TotalSales",
  ReturnMatrix = FALSE
)
}
\arguments{
\item{data}{This is your transactional data.table. Must include an Entity (typically customer), ProductCode (such as SKU), and a sales metric (such as total sales).}

\item{EntityColName}{This is the column name in quotes that represents the column name for the Entity, such as customer}

\item{ProductColName}{This is the column name in quotes that represents the column name for the product, such as SKU}

\item{MetricColName}{This is the column name in quotes that represents the column name for the metric, such as total sales}

\item{ReturnMatrix}{Set to FALSE to coerce the object (desired route) or TRUE to return a matrix}
}
\value{
A BinaryRatingsMatrix
}
\description{
AutoRecommenderDataCreate to create data that is prepared for modeling
}
\examples{
\dontrun{
RatingsMatrix <- RemixAutoML::AutoRecommenderDataCreate(
  data,
  EntityColName = "CustomerID",
  ProductColName = "StockCode",
  MetricColName = "TotalSales",
  ReturnMatrix = TRUE)
}
}
\seealso{
Other Recommenders: 
\code{\link{AutoMarketBasketModel}()},
\code{\link{AutoRecommenderScore}()},
\code{\link{AutoRecommenderTrain}()}
}
\author{
Adrian Antico and Douglas Pestana
}
\concept{Recommenders}
