% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoRecommender.R
\name{AutoRecommenderScore}
\alias{AutoRecommenderScore}
\title{AutoRecommenderScore}
\usage{
AutoRecommenderScore(
  data,
  WinningModel,
  EntityColName = "CustomerID",
  ProductColName = "StockCode",
  NumItemsReturn = 1
)
}
\arguments{
\item{data}{The binary ratings matrix from RecomDataCreate()}

\item{WinningModel}{The winning model returned from AutoRecommender()}

\item{EntityColName}{Typically your customer ID}

\item{ProductColName}{Something like "StockCode"}

\item{NumItemsReturn}{Number of items to return on scoring}
}
\value{
Returns the prediction data
}
\description{
This function will take your ratings matrix and model and score your data in parallel.
}
\examples{
\dontrun{
Results <- RemixAutoML::AutoRecommenderScore(
  data = AutoRecomDataCreate(
      data,
      EntityColName = "CustomerID",
      ProductColName = "StockCode",
      MetricColName = "TotalSales"),
  WinningModel = AutoRecommender(
      AutoRecomDataCreate(
        data,
        EntityColName = "CustomerID",
        ProductColName = "StockCode",
        MetricColName = "TotalSales"),
      Partition = "Split",
      KFolds = 2,
      Ratio = 0.75,
      RatingType = "TopN",
      RatingsKeep = 20,
      SkipModels = "AssociationRules",
      ModelMetric = "TPR"),
  EntityColName = "CustomerID",
  ProductColName = "StockCode")
}
}
\seealso{
Other Recommenders: 
\code{\link{AutoMarketBasketModel}()},
\code{\link{AutoRecommenderDataCreate}()},
\code{\link{AutoRecommenderTrain}()}
}
\author{
Adrian Antico and Douglas Pestana
}
\concept{Recommenders}
