% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RemixAutoML.R
\name{AutoRecommenderScoring}
\alias{AutoRecommenderScoring}
\title{The AutoRecomScoring function scores recommender models from AutoRecommender()}
\usage{
AutoRecommenderScoring(RatingMatrix, WinningModel,
  EntityColName = "CustomerID", ProductColName = "StockCode",
  MetricColName = "TotalSales")
}
\arguments{
\item{WinningModel}{The winning model returned from AutoRecommender()}

\item{EntityColName}{Typically your customer ID}

\item{ProductColName}{Something like "StockCode"}

\item{MetricColName}{Something like "TotalSales"}

\item{BinaryRatingMatrix}{The binary ratings matrix from RecomDataCreate()}
}
\value{
Returns the prediction data
}
\description{
This function will take your ratings matrix and model and score your data in parallel.
}
\seealso{
Other Supervised Learning: \code{\link{AutoH20Modeler}},
  \code{\link{AutoH20Scoring}}, \code{\link{AutoNLS}},
  \code{\link{AutoRecommender}}, \code{\link{AutoTS}}
}
\author{
Adrian Antico and Douglas Pestana
}
\concept{Supervised Learning}
