% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataWranglingFunctions.R
\name{AutoShapeShap}
\alias{AutoShapeShap}
\title{AutoShapeShap}
\usage{
AutoShapeShap(
  ScoringData = NULL,
  Threads = max(1L, parallel::detectCores() - 2L),
  DateColumnName = "Date",
  ByVariableName = "GroupVariable"
)
}
\arguments{
\item{ScoringData}{Scoring data from AutoCatBoostScoring with classification or regression}

\item{Threads}{Number of threads to use for the parellel routine}

\item{DateColumnName}{Name of the date column in scoring data}

\item{ByVariableName}{Name of your base entity column name}
}
\description{
AutoShapeShap will convert your scored shap values from CatBoost
}
\seealso{
Other Model Interpretation: 
\code{\link{SingleRowShapeShap}()}
}
\author{
Adrian Antico
}
\concept{Model Interpretation}
