% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RemixAutoML.R
\name{AutoTS}
\alias{AutoTS}
\title{AutoTS is an automated time series modeling function}
\usage{
AutoTS(data, TargetName = "Target", DateName = "DateTime",
  FCPeriods = 30, HoldOutPeriods = 30, TimeUnit = "day", Lags = 25,
  SLags = 2, NumCores = 4, SkipModels = NULL, StepWise = TRUE,
  TSClean = TRUE)
}
\arguments{
\item{data}{is the source time series data as a data.table - or a data structure that can be converted to a data.table}

\item{TargetName}{is the name of the dependent variable in your data.table}

\item{DateName}{is the name of the date column in your data.table}

\item{FCPeriods}{is the number of periods into the future you wish to forecast}

\item{HoldOutPeriods}{is the number of periods to use for validation testing}

\item{TimeUnit}{is the level of aggregation your dataset comes in}

\item{Lags}{is the number of lags you wish to test in various models (same as moving averages)}

\item{SLags}{is the number of seasonal lags you wish to test in various models (same as moving averages)}

\item{NumCores}{is the number of cores available on your computer}

\item{SkipModels}{Don't run specified models - e.g. exclude all models "DSHW" "ARFIMA" "ARIMA" "ETS" "NNET" "TBATS" "TSLM"}

\item{StepWise}{Set to TRUE to have ARIMA and ARFIMA run a stepwise selection process. Otherwise, all models will be generated in parallel execution, but still run much slower.}

\item{TSClean}{Set to TRUE to have missing values interpolated and outliers replaced with interpolated values: creates separate models for a larger comparison set}
}
\value{
Returns a list containing 1: A data.table object with a date column and the forecasted values; 2: The model evaluation results; 3: The champion model for later use if desired; 4: The name of the champion model.
}
\description{
Step 1 is to build all the models and evaluate them on the number of HoldOutPeriods periods you specify. Step 2 is to pick the winner and rebuild the winning model on the full data set. Step 3 is to generate forecasts with the final model for FCPeriods that you specify.
AutoTS builds the best time series models for each type, using optimized box-cox transformations and using a user-supplied frequency for the ts data conversion along with a model-based frequency for the ts data conversion, compares all types, selects the winner, and generates a forecast. Models include:
}
\details{
DSHW: Double Seasonal Holt Winters

ARFIMA: Auto Regressive Fractional Integrated Moving Average

ARIMIA: Stepwise Auto Regressive Integrated Moving Average with specified max lags, seasonal lags, moving averages, and seasonal moving averages

ETS: Additive and Multiplicitive Exponential Smoothing and Holt Winters

NNetar: Auto Regressive Neural Network models automatically compares models with 1 lag or 1 seasonal lag compared to models with up to N lags and N seasonal lags

TBATS: Exponential smoothing state space model with Box-Cox transformation, ARMA errors, Trend and Seasonal components

TSLM: Time Series Linear Model - builds a linear model with trend and season components extracted from the data

Facebook Prophet: Additive model where non-linear trends are fit with yearly, weekly, and daily seasonality, plus holiday effects. Uses Facebook's open source Prophet model.
}
\seealso{
Other Supervised Learning: \code{\link{AutoH2OModeler}},
  \code{\link{AutoH2OScoring}}, \code{\link{AutoNLS}},
  \code{\link{AutoRecommenderScoring}},
  \code{\link{AutoRecommender}}
}
\author{
Adrian Antico and Douglas Pestana
}
\concept{Supervised Learning}
