% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RemixAutoML.R
\name{AutoWord2VecModeler}
\alias{AutoWord2VecModeler}
\title{Automated word2vec data generation via H20}
\usage{
AutoWord2VecModeler(data, stringCol = c("Text_Col1", "Text_Col2"),
  KeepStringCol = FALSE, model_path = getwd(),
  ModelID = c("Text_Col1", "Text_Col2"), vects = 100,
  SaveStopWords = FALSE, MinWords = 1, WindowSize = 12,
  Epochs = 25, StopWords = NULL, SaveModel = "standard",
  Threads = 6, MaxMemory = "28G")
}
\arguments{
\item{data}{Source data table to merge vects onto}

\item{stringCol}{A string name for the column to convert via word2vec}

\item{KeepStringCol}{Set to TRUE if you want to keep the original string column that you convert via word2vec}

\item{model_path}{A string path to the location where you want the model and metadata stored}

\item{vects}{The number of vectors to retain from the word2vec model}

\item{SaveStopWords}{Set to TRUE to save the stop words used}

\item{MinWords}{For H20 word2vec model}

\item{WindowSize}{For H20 word2vec model}

\item{Epochs}{For H20 word2vec model}

\item{StopWords}{For H20 word2vec model}

\item{SaveModel}{Set to "standard" to save normally; set to "mojo" to save as mojo. NOTE: while you can save a mojo, I haven't figured out how to score it in the AutoH20Scoring function.}

\item{Threads}{Number of available threads you want to dedicate to model building}

\item{MaxMemory}{Amount of memory you want to dedicate to model building}
}
\description{
This function allows you to automatically build a word2vec model and merge the data onto your supplied dataset
}
\examples{
\dontrun{
data <- Word2VecModel(data,
                      stringCol     = c("Text_Col1",
                      "Text_Col2"),
                      KeepStringCol = FALSE,
                      model_path    = getwd(),
                      vects         = 5,
                      SaveStopWords = FALSE,
                      MinWords      = 1,
                      WindowSize    = 1,
                      Epochs        = 25,
                      StopWords     = NULL,
                      SaveModel     = "standard",
                      Threads       = 6,
                      MaxMemory     = "28G")
}
}
\seealso{
Other Feature Engineering: \code{\link{DT_GDL_Feature_Engineering}},
  \code{\link{DummifyDT}},
  \code{\link{FAST_GDL_Feature_Engineering}},
  \code{\link{GDL_Feature_Engineering}},
  \code{\link{ModelDataPrep}},
  \code{\link{Scoring_GDL_Feature_Engineering}}
}
\author{
Adrian Antico
}
\concept{Feature Engineering}
