% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RemixAutoML.R
\name{AutoWordFreq}
\alias{AutoWordFreq}
\title{Automated Word Frequency and Word Cloud Creation}
\usage{
AutoWordFreq(data, TextColName = "DESCR",
  GroupColName = "ClusterAllNoTarget", GroupLevel = 0,
  RemoveEnglishStopwords = TRUE, Stemming = TRUE,
  StopWords = c("bla", "bla2"))
}
\arguments{
\item{data}{Source data table}

\item{TextColName}{A string name for the column}

\item{GroupColName}{Set to NULL to ignore, otherwise set to Cluster column name (or factor column name)}

\item{GroupLevel}{Must be set if GroupColName is defined. Set to cluster ID (or factor level)}

\item{RemoveEnglishStopwords}{Set to TRUE to remove English stop words, FALSE to ignore}

\item{Stemming}{Set to TRUE to run stemming on your text data}

\item{StopWords}{Add your own stopwords, in vector format}
}
\description{
This function builds a word frequency table and a word cloud. It prepares data, cleans text, and generates output.
}
\examples{
\dontrun{
data <- WordFreq(data,
                 TextColName = "DESCR",
                 GroupColName = "ClusterAllNoTarget",
                 GroupLevel = 0,
                 RemoveEnglishStopwords = TRUE,
                 Stemming = TRUE,
                 StopWords = c("bla1", "bla2")
}
}
\seealso{
Other Misc: \code{\link{AutoH2OTextPrepScoring}},
  \code{\link{ChartTheme}}, \code{\link{PrintObjectsSize}},
  \code{\link{RecomDataCreate}}, \code{\link{RemixTheme}},
  \code{\link{SimpleCap}}, \code{\link{multiplot}},
  \code{\link{percRank}}, \code{\link{tempDatesFun}},
  \code{\link{tokenizeH2O}}
}
\author{
Adrian Antico
}
\concept{Misc}
