% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoXGBoostHurdleModelScoring.R
\name{AutoXGBoostHurdleModelScoring}
\alias{AutoXGBoostHurdleModelScoring}
\title{AutoXGBoostHurdleModelScoring}
\usage{
AutoXGBoostHurdleModelScoring(
  TestData = NULL,
  Path = NULL,
  ModelID = NULL,
  ArgsList = NULL,
  ModelList = NULL,
  Threshold = NULL,
  CARMA = FALSE
)
}
\arguments{
\item{TestData}{scoring data.table}

\item{Path}{Supply if ArgsList is NULL or ModelList is null.}

\item{ModelID}{Supply if ArgsList is NULL or ModelList is null. Same as used in model training.}

\item{ArgsList}{Output from the hurdle model}

\item{ModelList}{Output from the hurdle model}

\item{Threshold}{NULL to use raw probabilities to predict. Otherwise, supply a threshold}

\item{CARMA}{Keep FALSE. Used for CARMA functions internals}
}
\value{
A data.table with the final predicted value, the intermediate model predictions, and your source data
}
\description{
AutoXGBoostHurdleModelScoring can score AutoXGBoostHurdleModel() models
}
\examples{
\dontrun{

# Define file path
Path <- getwd()

# Create hurdle data with correlated features
data <- RemixAutoML::FakeDataGenerator(
  Correlation = 0.70,
  N = 25000,
  ID = 3,
  FactorCount = 2L,
  AddDate = TRUE,
  ZIP = 1,
  Classification = FALSE,
  MultiClass = FALSE)

# Define features
Features <- names(data)[!names(data) \%in\%
  c("Adrian","IDcol_1","IDcol_2","IDcol_3","DateTime")]

Output <- RemixAutoML::AutoXGBoostHurdleModel(

   # Operationalization args
   TrainOnFull = FALSE,
   PassInGrid = NULL,

   # Metadata args
   NThreads = max(1L, parallel::detectCores()-2L),
   ModelID = "ModelTest",
   Paths = Path,
   MetaDataPaths = NULL,

   # data args
   data,
   ValidationData = NULL,
   TestData = NULL,
   Buckets = 0L,
   TargetColumnName = NULL,
   FeatureColNames = NULL,
   PrimaryDateColumn = NULL,
   WeightsColumnName = NULL,
   IDcols = NULL,
   ClassWeights = c(1,1),
   DebugMode = FALSE,

   # options
   EncodingMethod = "credibility",
   TransformNumericColumns = NULL,
   Methods = c('Asinh','Asin','Log','LogPlus1','Sqrt','Logit'),
   SplitRatios = c(0.70, 0.20, 0.10),
   ReturnModelObjects = TRUE,
   SaveModelObjects = FALSE,
   NumOfParDepPlots = 10L,

   # grid tuning args
   GridTune = FALSE,
   grid_eval_metric = "accuracy",
   MaxModelsInGrid = 1L,
   BaselineComparison = "default",
   MaxRunsWithoutNewWinner = 10L,
   MaxRunMinutes = 60L,

   # XGBoost parameters
   TreeMethod = "hist",
   Trees = list("classifier" = 1000, "regression" = 1000),
   eta = list("classifier" = 0.05, "regression" = 0.05),
   max_depth = list("classifier" = 4L, "regression" = 4L),
   min_child_weight = list("classifier" = 1.0, "regression" = 1.0),
   subsample = list("classifier" = 0.55, "regression" = 0.55),
   colsample_bytree = list("classifier" = 0.55, "regression" = 0.55))

# Score XGBoost Hurdle Model
HurdleScores <- RemixAutoML::AutoXGBoostHurdleModelScoring(
  TestData = data,
  Path = Path,
  ModelID = "ModelTest",
  ModelList = NULL,
  ArgsList = NULL,
  Threshold = NULL)
}
}
\seealso{
Other Automated Model Hurdle Modeling: 
\code{\link{AutoCatBoostHurdleModelScoring}()},
\code{\link{AutoLightGBMHurdleModelScoring}()}
}
\author{
Adrian Antico
}
\concept{Automated Model Hurdle Modeling}
