% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoXGBoostScoring.R
\name{AutoXGBoostScoring}
\alias{AutoXGBoostScoring}
\title{AutoXGBoostScoring is an automated scoring function that compliments the AutoCatBoost model training functions.}
\usage{
AutoXGBoostScoring(
  TargetType = NULL,
  ScoringData = NULL,
  FeatureColumnNames = NULL,
  IDcols = NULL,
  FactorLevelsList = NULL,
  TargetLevels = NULL,
  Objective = "multi:softmax",
  OneHot = FALSE,
  ModelObject = NULL,
  ModelPath = NULL,
  ModelID = NULL,
  ReturnFeatures = TRUE,
  TransformNumeric = FALSE,
  BackTransNumeric = FALSE,
  TargetColumnName = NULL,
  TransformationObject = NULL,
  TransID = NULL,
  TransPath = NULL,
  MDP_Impute = TRUE,
  MDP_CharToFactor = TRUE,
  MDP_RemoveDates = TRUE,
  MDP_MissFactor = "0",
  MDP_MissNum = -1
)
}
\arguments{
\item{TargetType}{Set this value to "regression", "classification", or "multiclass" to score models built using AutoCatBoostRegression(), AutoCatBoostClassify() or AutoCatBoostMultiClass().}

\item{ScoringData}{This is your data.table of features for scoring. Can be a single row or batch.}

\item{FeatureColumnNames}{Supply either column names or column numbers used in the AutoXGBoost__() function}

\item{IDcols}{Supply ID column numbers for any metadata you want returned with your predicted values}

\item{FactorLevelsList}{Supply the factor variables' list from DummifyDT()}

\item{TargetLevels}{Supply the target levels output from AutoXGBoostMultiClass() or the scoring function will go looking for it in the file path you supply.}

\item{Objective}{Set to 'multi:softprobs' if you did so in training. Default is softmax}

\item{OneHot}{Set to TRUE to have one-hot-encoding run. Otherwise, N columns will be made for N levels of a factor variable}

\item{ModelObject}{Supply a model for scoring, otherwise it will have to search for it in the file path you specify}

\item{ModelPath}{Supply your path file used in the AutoXGBoost__() function}

\item{ModelID}{Supply the model ID used in the AutoXGBoost__() function}

\item{ReturnFeatures}{Set to TRUE to return your features with the predicted values.}

\item{TransformNumeric}{Set to TRUE if you have features that were transformed automatically from an Auto__Regression() model AND you haven't already transformed them.}

\item{BackTransNumeric}{Set to TRUE to generate back-transformed predicted values. Also, if you return features, those will also be back-transformed.}

\item{TargetColumnName}{Input your target column name used in training if you are utilizing the transformation service}

\item{TransformationObject}{Set to NULL if you didn't use transformations or if you want the function to pull from the file output from the Auto__Regression() function. You can also supply the transformation data.table object with the transformation details versus having it pulled from file.}

\item{TransID}{Set to the ID used for saving the transformation data.table object or set it to the ModelID if you are pulling from file from a build with Auto__Regression().}

\item{TransPath}{Set the path file to the folder where your transformation data.table detail object is stored. If you used the Auto__Regression() to build, set it to the same path as ModelPath.}

\item{MDP_Impute}{Set to TRUE if you did so for modeling and didn't do so before supplying ScoringData in this function}

\item{MDP_CharToFactor}{Set to TRUE to turn your character columns to factors if you didn't do so to your ScoringData that you are supplying to this function}

\item{MDP_RemoveDates}{Set to TRUE if you have date of timestamp columns in your ScoringData}

\item{MDP_MissFactor}{If you set MDP_Impute to TRUE, supply the character values to replace missing values with}

\item{MDP_MissNum}{If you set MDP_Impute to TRUE, supply a numeric value to replace missing values with}
}
\value{
A data.table of predicted values with the option to return model features as well.
}
\description{
AutoXGBoostScoring is an automated scoring function that compliments the AutoCatBoost model training functions. This function requires you to supply features for scoring. It will run ModelDataPrep() and the DummifyDT() function to prepare your features for xgboost data conversion and scoring.
}
\examples{
\donttest{
Preds <- AutoXGBoostScoring(TargetType = "regression",
                            ScoringData = data,
                            FeatureColumnNames = 2:12,
                            IDcols = NULL,
                            FactorLevelsList = NULL,
                            TargetLevels = NULL,
                            Objective = "multi:softmax",
                            OneHot = FALSE,
                            ModelObject = NULL,
                            ModelPath = "home",
                            ModelID = "ModelTest",
                            ReturnFeatures = TRUE,
                            TransformNumeric = FALSE,
                            BackTransNumeric = FALSE,
                            TargetColumnName = NULL,
                            TransformationObject = NULL,
                            TransID = NULL,
                            TransPath = NULL,
                            MDP_Impute = TRUE,
                            MDP_CharToFactor = TRUE,
                            MDP_RemoveDates = TRUE,
                            MDP_MissFactor = "0",
                            MDP_MissNum = -1)
}
}
\seealso{
Other Automated Model Scoring: 
\code{\link{AutoCatBoostScoring}()},
\code{\link{AutoH2OMLScoring}()},
\code{\link{AutoH2OModeler}()},
\code{\link{AutoHurdleScoring}()},
\code{\link{CLForecast}()}
}
\author{
Adrian Antico
}
\concept{Automated Model Scoring}
