% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FakeDataGenerator.R
\name{BenchmarkData}
\alias{BenchmarkData}
\title{BenchmarkData}
\usage{
BenchmarkData(
  NRows = 1e+07,
  Levels = 1e+06,
  NAs = -1L,
  FixedEffects = c(5, 10, 15),
  CharVars = TRUE,
  IntVars = TRUE,
  Sort = TRUE
)
}
\arguments{
\item{NAs}{= -1L}

\item{FixedEffects}{c(5,10,15), number of levels for each variable}

\item{CharVars}{FALSE}

\item{IntVars}{TRUE}

\item{Sort}{= TRUE}

\item{N}{= 10000000,}

\item{RandomLevels}{= 1000000}

\item{RandomEffects}{c(3)}
}
\description{
Modified version of h2o datatable benchmark data
}
