% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelMetrics.R
\name{BinaryMetrics}
\alias{BinaryMetrics}
\title{BinaryMetrics}
\usage{
BinaryMetrics(
  MLModels = "catboost",
  ClassWeights. = ClassWeights,
  SaveModelObjects. = SaveModelObjects,
  ValidationData. = ValidationData,
  TrainOnFull. = TrainOnFull,
  TargetColumnName. = TargetColumnName,
  ModelID. = ModelID,
  model_path. = model_path,
  metadata_path. = metadata_path
)
}
\arguments{
\item{MLModels}{= "catboost"}

\item{ClassWeights.}{= ClassWeights}

\item{SaveModelObjects.}{= SaveModelObjects}

\item{ValidationData.}{= ValidationData}

\item{TrainOnFull.}{= TrainOnFull}

\item{TargetColumnName.}{= TargetColumnName}

\item{ModelID.}{= ModelID}

\item{model_path.}{= model_path}

\item{metadata_path.}{= metadata_path}
}
\description{
Compute binary metrics and save them to file
}
\seealso{
Other Model Evaluation: 
\code{\link{ClassificationMetrics}()},
\code{\link{DT_BinaryConfusionMatrix}()},
\code{\link{MultiClassMetrics}()},
\code{\link{RegressionMetrics}()},
\code{\link{RemixClassificationMetrics}()}
}
\author{
Adrian Antico
}
\concept{Model Evaluation}
