% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyFunctions.R
\name{BlankRow}
\alias{BlankRow}
\title{BlankRow}
\usage{
BlankRow(W)
}
\arguments{
\item{W}{width of column}
}
\value{
Adds a row to your UI of width W
}
\description{
BlankRow add blank row with width w
}
\examples{
\dontrun{
RemixAutoML::BlankRow(12)
}
}
\seealso{
Other Shiny: 
\code{\link{ArgNullCheck2}()},
\code{\link{ArgNullCheck}()},
\code{\link{DateInput}()},
\code{\link{GenerateEvaluationMetrics}()},
\code{\link{NumericInput}()},
\code{\link{PickerInput_GetLevels2}()},
\code{\link{PickerInput_GetLevels}()},
\code{\link{PickerInput}()},
\code{\link{PreparePlotData}()},
\code{\link{ReactiveLoadCSV}()},
\code{\link{ReturnParam}()},
\code{\link{StoreArgs}()},
\code{\link{TextInput}()},
\code{\link{observeEventLoad}()}
}
\author{
Adrian Antico
}
\concept{Shiny}
