% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EDAFunctions.R
\name{BNLearnArcStrength}
\alias{BNLearnArcStrength}
\title{BNLearnArcStrength}
\usage{
BNLearnArcStrength(
  data = NULL,
  NetworkVars = NULL,
  DataSampleRate = 0.5,
  ByGroupVars = NULL,
  MinRows = 30
)
}
\arguments{
\item{data}{data.table}

\item{NetworkVars}{Names of the columns to utilize in the analysis}

\item{DataSampleRate}{Sample your data to reduce runtime}

\item{ByGroupVars}{Group variables that you want to have the analysis done by}

\item{MinRows}{Minimum number of rows to utilize in the ByGroupVars analysis}
}
\description{
Utilize bnlearn to create a bayesian network and return the arc strengths for features and their edges
}
\seealso{
Other EDA: 
\code{\link{AutoCorrAnalysis}()},
\code{\link{AutoWordFreq}()}
}
\author{
Adrian Antico
}
\concept{EDA}
