% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CARMA-HelperFunctions.R
\name{CARMA_Define_Args}
\alias{CARMA_Define_Args}
\title{CARMA_Define_Args}
\usage{
CARMA_Define_Args(TimeUnit = NULL, TimeGroups = NULL,
  HierarchGroups = NULL, GroupVariables = NULL, FC_Periods = NULL,
  PartitionType = NULL, TrainOnFull = NULL, SplitRatios = NULL,
  SD_Periods = 0L, Skew_Periods = 0L, Kurt_Periods = 0L,
  Quantile_Periods = 0L)
}
\arguments{
\item{TimeUnit}{= TimeUnit}

\item{TimeGroups}{= TimeGroups}

\item{HierarchGroups}{= HierarchGroups}

\item{GroupVariables}{= GroupVariables}

\item{FC_Periods}{= FC_Periods}

\item{PartitionType}{= PartitionType}

\item{TrainOnFull}{= TrainOnFull}

\item{SplitRatios}{= SplitRatios}

\item{SD_Periods}{= 0L turns it off, otherwise values must be greater than 1 such as c(2L,5L,6L,25L)}

\item{Skew_Periods}{= 0L turns it off, otherwise values must be greater than 2 such as c(3L,5L,6L,25L)}

\item{Kurt_Periods}{= 0L turns it off, otherwise values must be greater than 3 such as c(4L,5L,6L,25L)}

\item{Quantile_Periods}{= 0L turns it off, otherwise values must be greater than 3 such as c(5L,6L,25L)}
}
\description{
CARMA_Define_Args is to help manage carma code
}
\seealso{
Other Carma Helper: \code{\link{CARMA_Get_IndepentVariablesPass}},
  \code{\link{CarmaCatBoostKeepVarsGDL}},
  \code{\link{CarmaH2OKeepVarsGDL}},
  \code{\link{CarmaXGBoostKeepVarsGDL}}
}
\author{
Adrian Antico
}
\concept{Carma Helper}
