% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CARMA-HelperFunctions.R
\name{CARMA_GroupHierarchyCheck}
\alias{CARMA_GroupHierarchyCheck}
\title{CARMA_GroupHierarchyCheck}
\usage{
CARMA_GroupHierarchyCheck(
  data = data,
  Group_Variables = GroupVariables,
  HierarchyGroups = HierarchGroups
)
}
\arguments{
\item{data}{data fed into function}

\item{Group_Variables}{Takes GroupVariables from caram function}

\item{HierarchyGroups}{Vector of group variables}
}
\description{
CARMA_GroupHierarchyCheck
}
\seealso{
Other Carma Helper: 
\code{\link{CARMA_Define_Args}()},
\code{\link{CARMA_Get_IndepentVariablesPass}()},
\code{\link{CarmaCatBoostKeepVarsGDL}()},
\code{\link{CarmaH2OKeepVarsGDL}()},
\code{\link{CarmaXGBoostKeepVarsGDL}()}
}
\author{
Adrian Antico
}
\concept{Carma Helper}
