% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CARMA-HelperFunctions.R
\name{CarmaH2OKeepVarsGDL}
\alias{CarmaH2OKeepVarsGDL}
\title{CarmaH2OKeepVarsGDL}
\usage{
CarmaH2OKeepVarsGDL(
  data,
  IndepVarPassTRUE = "GroupVar",
  UpdateData,
  CalendarFeatures,
  XREGS,
  Difference,
  HierarchGroups,
  GroupVariables,
  GroupVarVector,
  CalendarVariables = NULL,
  HolidayVariable = NULL,
  TargetColumnName,
  DateColumnName
)
}
\arguments{
\item{data}{Supply data}

\item{IndepVarPassTRUE}{Name of the column used as a single grouping variable.}

\item{UpdateData}{Supply UpdateData}

\item{CalendarFeatures}{Supply CalendarFeatures}

\item{XREGS}{Supply XREGS}

\item{Difference}{Supply Difference}

\item{HierarchGroups}{Supply HierarchGroups}

\item{GroupVariables}{Supply GroupVariables}

\item{GroupVarVector}{Supply GroupVarVector}

\item{CalendarVariables}{Supply CalendarVariables}

\item{HolidayVariable}{Supply HolidayVariable}

\item{TargetColumnName}{Supply TargetColumnName}

\item{DateColumnName}{Supply DateColumnName}
}
\description{
CarmaH2OKeepVarsGDL is to help manage carma code
}
\seealso{
Other Carma Helper: 
\code{\link{CARMA_Define_Args}()},
\code{\link{CARMA_Get_IndepentVariablesPass}()},
\code{\link{CarmaCatBoostKeepVarsGDL}()},
\code{\link{CarmaXGBoostKeepVarsGDL}()}
}
\author{
Adrian Antico
}
\concept{Carma Helper}
