% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelMetrics.R
\name{CarmaHoldoutMetrics}
\alias{CarmaHoldoutMetrics}
\title{CarmaHoldoutMetrics}
\usage{
CarmaHoldoutMetrics(
  DATA = TestDataEval,
  TARGETCOLUMNNAME = TargetColumnName,
  GROUPVARIABLES = GroupingVariables
)
}
\arguments{
\item{DATA}{TestDataEval}

\item{TARGETCOLUMNNAME}{TargetColumnName}

\item{GROUPVARIABLES}{GroupVariables}
}
\description{
CarmaHoldoutMetrics
}
\seealso{
Other Carma Helper: 
\code{\link{CARMA_Define_Args}()},
\code{\link{CARMA_Get_IndepentVariablesPass}()},
\code{\link{CARMA_GroupHierarchyCheck}()},
\code{\link{CarmaCatBoostKeepVarsGDL}()},
\code{\link{CarmaH2OKeepVarsGDL}()},
\code{\link{CarmaXGBoostKeepVarsGDL}()}
}
\author{
Adrian Antico
}
\concept{Carma Helper}
