% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CARMA-HelperFunctions.R
\name{CarmaXGBoostKeepVarsGDL}
\alias{CarmaXGBoostKeepVarsGDL}
\title{CarmaXGBoostKeepVarsGDL}
\usage{
CarmaXGBoostKeepVarsGDL(data, IndepVarPassTRUE = "GroupVar", UpdateData,
  CalendarFeatures, XREGS, Difference, HierarchGroups, GroupVariables,
  GroupVarVector, CalendarVariables, HolidayVariable, TargetColumnName,
  DateColumnName)
}
\arguments{
\item{data}{Supply data}

\item{IndepVarPassTRUE}{Name of the column used as a single grouping variable.}

\item{UpdateData}{Supply UpdateData}

\item{CalendarFeatures}{Supply CalendarFeatures}

\item{XREGS}{Supply XREGS}

\item{Difference}{Supply Difference}

\item{HierarchGroups}{Supply HierarchGroups}

\item{GroupVariables}{Supply GroupVariables}

\item{GroupVarVector}{Supply GroupVarVector}

\item{CalendarVariables}{Supply CalendarVariables}

\item{HolidayVariable}{Supply HolidayVariable}

\item{TargetColumnName}{Supply TargetColumnName}

\item{DateColumnName}{Supply DateColumnName}
}
\description{
CarmaXGBoostKeepVarsGDL is to help manage carma code
}
\seealso{
Other Carma Helper: \code{\link{CARMA_Define_Args}},
  \code{\link{CARMA_Get_IndepentVariablesPass}},
  \code{\link{CarmaCatBoostKeepVarsGDL}},
  \code{\link{CarmaH2OKeepVarsGDL}}
}
\author{
Adrian Antico
}
\concept{Carma Helper}
