% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CatBoostHelpers.R
\name{CatBoostArgsCheck}
\alias{CatBoostArgsCheck}
\title{CatBoostArgsCheck}
\usage{
CatBoostArgsCheck(
  ModelType = "regression",
  DummifyCols. = DummifyCols,
  data. = data,
  FeatureColNames. = FeatureColNames,
  PrimaryDateColumn. = PrimaryDateColumn,
  GridTune. = GridTune,
  model_path. = model_path,
  metadata_path. = metadata_path,
  ClassWeights. = ClassWeights,
  LossFunction. = LossFunction,
  loss_function. = loss_function,
  loss_function_value. = loss_function_value,
  eval_metric. = eval_metric,
  eval_metric_value. = eval_metric_value,
  task_type. = task_type,
  NumGPUs. = NumGPUs,
  MaxModelsInGrid. = MaxModelsInGrid,
  NumOfParDepPlots. = NumOfParDepPlots,
  ReturnModelObjects. = ReturnModelObjects,
  SaveModelObjects. = SaveModelObjects,
  PassInGrid. = PassInGrid,
  MetricPeriods. = MetricPeriods,
  langevin. = langevin,
  diffusion_temperature. = diffusion_temperature,
  Trees. = Trees,
  Depth. = Depth,
  LearningRate. = LearningRate,
  L2_Leaf_Reg. = L2_Leaf_Reg,
  RandomStrength. = RandomStrength,
  BorderCount. = BorderCount,
  RSM. = RSM,
  BootStrapType. = BootStrapType,
  GrowPolicy. = GrowPolicy,
  model_size_reg. = model_size_reg,
  feature_border_type. = feature_border_type,
  sampling_unit. = sampling_unit,
  subsample. = subsample,
  score_function. = score_function,
  min_data_in_leaf. = min_data_in_leaf
)
}
\arguments{
\item{ModelType}{Passthrough}

\item{DummifyCols.}{Passthrough}

\item{data.}{Passthrough}

\item{FeatureColNames.}{Passthrough}

\item{PrimaryDateColumn.}{Passthrough}

\item{GridTune.}{Passthrough}

\item{model_path.}{Passthrough}

\item{metadata_path.}{Passthrough}

\item{ClassWeights.}{Passthrough}

\item{LossFunction.}{Passthrough}

\item{loss_function.}{Passthrough regression}

\item{loss_function_value.}{Passthrough regression}

\item{eval_metric.}{Passthrough regression}

\item{eval_metric_value.}{Passthrough regression}

\item{task_type.}{Passthrough}

\item{NumGPUs.}{Passthrough}

\item{MaxModelsInGrid.}{Passthrough}

\item{NumOfParDepPlots.}{Passthrough}

\item{SaveModelObjects.}{Passthrough}

\item{PassInGrid.}{Passthrough}

\item{MetricPeriods.}{Passthrough}

\item{langevin.}{Passthrough}

\item{diffusion_temperature.}{Passthrough}

\item{Trees.}{Passthrough}

\item{Depth.}{Passthrough}

\item{LearningRate.}{Passthrough}

\item{L2_Leaf_Reg.}{Passthrough}

\item{RandomStrength.}{Passthrough}

\item{BorderCount.}{Passthrough}

\item{RSM.}{Passthrough}

\item{BootStrapType.}{Passthrough}

\item{GrowPolicy.}{Passthrough}

\item{model_size_reg.}{Passthrough}

\item{feature_border_type.}{Passthrough}

\item{sampling_unit.}{Passthrough}

\item{subsample.}{Passthrough}

\item{score_function.}{Passthrough}

\item{min_data_in_leaf.}{Passthrough}

\item{ReturnModelObject.}{Passthrough}
}
\description{
Ensure arguments are defined correctly
}
\seealso{
Other CatBoost Helpers: 
\code{\link{CatBoostDataConversion}()},
\code{\link{CatBoostDataPrep}()},
\code{\link{CatBoostEvalPlots}()},
\code{\link{CatBoostFinalParams}()},
\code{\link{CatBoostGridParams}()},
\code{\link{CatBoostImportances}()},
\code{\link{CatBoostPDF}()},
\code{\link{CatBoostParameterGrids}()},
\code{\link{CatBoostRemoveFiles}()},
\code{\link{CatBoostValidationData}()}
}
\author{
Adrian
}
\concept{CatBoost Helpers}
