% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CatBoostHelpers.R
\name{CatBoostDataConversion}
\alias{CatBoostDataConversion}
\title{CatBoostDataConversion}
\usage{
CatBoostDataConversion(
  CatFeatures. = CatFeatures,
  dataTrain. = dataTrain,
  dataTest. = dataTest,
  TestData. = TestData,
  TrainTarget. = TrainTarget,
  TestTarget. = TestTarget,
  FinalTestTarget. = FinalTestTarget,
  TrainOnFull. = TrainOnFull
)
}
\arguments{
\item{CatFeatures.}{Passthrough}

\item{dataTrain.}{Passthrough}

\item{dataTest.}{Passthrough}

\item{TestData.}{Passthrough}

\item{TrainTarget.}{Passthrough}

\item{TestTarget.}{Passthrough}

\item{FinalTestTarget.}{Passthrough}

\item{TrainOnFull.}{Passthrough}
}
\description{
Convert data to catboost format
}
\seealso{
Other CatBoost Helpers: 
\code{\link{CatBoostArgsCheck}()},
\code{\link{CatBoostDataPrep}()},
\code{\link{CatBoostEvalPlots}()},
\code{\link{CatBoostFinalParams}()},
\code{\link{CatBoostGridParams}()},
\code{\link{CatBoostImportances}()},
\code{\link{CatBoostPDF}()},
\code{\link{CatBoostParameterGrids}()},
\code{\link{CatBoostRemoveFiles}()},
\code{\link{CatBoostValidationData}()}
}
\author{
Adrian
}
\concept{CatBoost Helpers}
