% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CatBoostHelpers.R
\name{CatBoostDataPrep}
\alias{CatBoostDataPrep}
\title{CatBoostDataPrep}
\usage{
CatBoostDataPrep(
  ModelType = "regression",
  data. = data,
  ValidationData. = ValidationData,
  TestData. = TestData,
  TargetColumnName. = TargetColumnName,
  FeatureColNames. = FeatureColNames,
  PrimaryDateColumn. = PrimaryDateColumn,
  IDcols. = IDcols,
  TransformNumericColumns. = TransformNumericColumns,
  Methods. = Methods,
  ModelID. = ModelID,
  model_path. = model_path,
  DummifyCols. = DummifyCols,
  LossFunction. = LossFunction,
  EvalMetric. = EvalMetric,
  TrainOnFull. = TrainOnFull,
  SaveModelObjects. = SaveModelObjects
)
}
\arguments{
\item{ModelType}{'regression', 'vector', 'classification', or 'multiclass'}

\item{data.}{Passthrough}

\item{ValidationData.}{Passthrough}

\item{TestData.}{Passthrough}

\item{TargetColumnName.}{Passthrough}

\item{FeatureColNames.}{Passthrough}

\item{PrimaryDateColumn.}{Passthrough}

\item{IDcols.}{Passthrough}

\item{TransformNumericColumns.}{Passthrough regression}

\item{Methods.}{Passthrough regression}

\item{ModelID.}{Passthrough regression}

\item{model_path.}{Passthrough regression}

\item{DummifyCols.}{Passthrough regression}

\item{LossFunction.}{Passthrough regression}

\item{EvalMetric.}{Passthrough regression}

\item{TrainOnFull.}{Passthrough}

\item{SaveModelObjects.}{Passthrough}
}
\description{
Prepare data for loading into catboost format
}
\seealso{
Other CatBoost Helpers: 
\code{\link{CatBoostArgsCheck}()},
\code{\link{CatBoostDataConversion}()},
\code{\link{CatBoostEvalPlots}()},
\code{\link{CatBoostFinalParams}()},
\code{\link{CatBoostGridParams}()},
\code{\link{CatBoostImportances}()},
\code{\link{CatBoostPDF}()},
\code{\link{CatBoostParameterGrids}()},
\code{\link{CatBoostRemoveFiles}()},
\code{\link{CatBoostValidationData}()}
}
\author{
Adrian Antico
}
\concept{CatBoost Helpers}
