% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CatBoostHelpers.R
\name{CatBoostEvalPlots}
\alias{CatBoostEvalPlots}
\title{CatBoostEvalPlots}
\usage{
CatBoostEvalPlots(
  ModelType = "classification",
  TrainOnFull. = TrainOnFull,
  LossFunction. = LossFunction,
  EvalMetric. = EvalMetric,
  EvaluationMetrics. = EvaluationMetrics,
  ValidationData. = ValidationData,
  NumOfParDepPlots. = NumOfParDepPlots,
  VariableImportance. = VariableImportance,
  TargetColumnName. = TargetColumnName,
  FeatureColNames. = FeatureColNames,
  SaveModelObjects. = SaveModelObjects,
  ModelID. = ModelID,
  metadata_path. = metadata_path,
  model_path. = model_path,
  predict. = predict
)
}
\arguments{
\item{ModelType}{'classification', 'multiclass', 'regression', or 'vector'}

\item{TrainOnFull.}{Passthrough}

\item{LossFunction.}{Passthrough regression}

\item{EvalMetric.}{Passthrough regression}

\item{EvaluationMetrics.}{Passthrough regression}

\item{ValidationData.}{Passthrough}

\item{NumOfParDepPlots.}{Passthrough}

\item{VariableImportance.}{Passthrough}

\item{TargetColumnName.}{Passthrough}

\item{FeatureColNames.}{Passthrough}

\item{SaveModelObjects.}{Passthrough}

\item{ModelID.}{Passthrough}

\item{metadata_path.}{Passthrough}

\item{model_path.}{Passthrough}
}
\description{
Generate evaluation plots
}
\seealso{
Other CatBoost Helpers: 
\code{\link{CatBoostArgsCheck}()},
\code{\link{CatBoostDataConversion}()},
\code{\link{CatBoostDataPrep}()},
\code{\link{CatBoostFinalParams}()},
\code{\link{CatBoostGridParams}()},
\code{\link{CatBoostImportances}()},
\code{\link{CatBoostPDF}()},
\code{\link{CatBoostParameterGrids}()},
\code{\link{CatBoostRemoveFiles}()},
\code{\link{CatBoostValidationData}()}
}
\author{
Adrian
}
\concept{CatBoost Helpers}
