% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CatBoostHelpers.R
\name{CatBoostFinalParams}
\alias{CatBoostFinalParams}
\title{CatBoostFinalParams}
\usage{
CatBoostFinalParams(
  ModelType = "classification",
  UseBestModel. = UseBestModel,
  ClassWeights. = ClassWeights,
  PassInGrid. = PassInGrid,
  ExperimentalGrid. = ExperimentalGrid,
  BestGrid. = BestGrid,
  GridTune. = GridTune,
  TrainOnFull. = TrainOnFull,
  MetricPeriods. = MetricPeriods,
  LossFunction. = LossFunction,
  EvalMetric. = EvalMetric,
  score_function. = score_function,
  HasTime. = HasTime,
  task_type. = task_type,
  NumGPUs. = NumGPUs,
  NTrees. = Trees,
  Depth. = Depth,
  LearningRate. = LearningRate,
  L2_Leaf_Reg. = L2_Leaf_Reg,
  langevin. = langevin,
  diffusion_temperature. = diffusion_temperature,
  sampling_unit. = sampling_unit,
  RandomStrength. = RandomStrength,
  BorderCount. = BorderCount,
  RSM. = RSM,
  GrowPolicy. = GrowPolicy,
  BootStrapType. = BootStrapType,
  model_size_reg. = model_size_reg,
  feature_border_type. = feature_border_type,
  subsample. = subsample,
  min_data_in_leaf. = min_data_in_leaf
)
}
\arguments{
\item{ModelType}{'regression', 'classification', 'multiclass', 'vector'}

\item{UseBestModel.}{Passthrough}

\item{ClassWeights.}{Passthrough}

\item{PassInGrid.}{Passthrough}

\item{ExperimentalGrid.}{Passthrough}

\item{BestGrid.}{Passthrough}

\item{GridTune.}{Passthrough}

\item{TrainOnFull.}{Passthrough}

\item{MetricPeriods.}{Passthrough}

\item{LossFunction.}{Passthrough}

\item{EvalMetric.}{Passthrough}

\item{score_function.}{Passthrough}

\item{HasTime.}{Passthrough}

\item{task_type.}{Passthrough}

\item{NumGPUs.}{Passthrough}

\item{NTrees.}{Passthrough}

\item{Depth.}{Passthrough}

\item{LearningRate.}{Passthrough}

\item{L2_Leaf_Reg.}{Passthrough}

\item{langevin.}{Passthrough}

\item{diffusion_temperature.}{Passthrough}

\item{sampling_unit.}{Passthrough}

\item{RandomStrength.}{Passthrough}

\item{BorderCount.}{Passthrough}

\item{RSM.}{Passthrough}

\item{GrowPolicy.}{Passthrough}

\item{BootStrapType.}{Passthrough}

\item{model_size_reg.}{Passthrough}

\item{feature_border_type.}{Passthrough}

\item{subsample.}{Passthrough}

\item{min_data_in_leaf.}{Passthrough}

\item{TargetColumnName.}{Passthrough}
}
\description{
Convert data to catboost format
}
\seealso{
Other CatBoost Helpers: 
\code{\link{CatBoostArgsCheck}()},
\code{\link{CatBoostDataConversion}()},
\code{\link{CatBoostDataPrep}()},
\code{\link{CatBoostGridParams}()},
\code{\link{CatBoostImportances}()},
\code{\link{CatBoostPDF}()},
\code{\link{CatBoostParameterGrids}()},
\code{\link{CatBoostRemoveFiles}()},
\code{\link{CatBoostValidationData}()}
}
\author{
Adrian
}
\concept{CatBoost Helpers}
