% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyFunctions.R
\name{CatBoostGridEvalMetricsOptions}
\alias{CatBoostGridEvalMetricsOptions}
\title{CatBoostGridEvalMetricsOptions}
\usage{
CatBoostGridEvalMetricsOptions(TT)
}
\arguments{
\item{TT}{Target Type: 'regression', 'classification', 'multiclass', all lower case}
}
\description{
CatBoostGridEvalMetricsOptions
}
\keyword{internal}
