% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CatBoostHelpers.R
\name{CatBoostGridParams}
\alias{CatBoostGridParams}
\title{CatBoostClassifierParams}
\usage{
CatBoostGridParams(
  N. = N,
  counter. = NULL,
  BanditArmsN. = NULL,
  HasTime. = NULL,
  MetricPeriods. = NULL,
  ClassWeights. = NULL,
  EvalMetric. = NULL,
  LossFunction. = NULL,
  task_type. = NULL,
  NumGPUs. = NULL,
  model_path. = NULL,
  NewGrid. = NULL,
  Grid. = NULL,
  ExperimentalGrid. = NULL,
  GridClusters. = NULL
)
}
\arguments{
\item{N.}{Iteration for specific grid in grid clusters}

\item{counter.}{Passthrough}

\item{BanditArmsN.}{Passthrough}

\item{HasTime.}{Passthrough}

\item{MetricPeriods.}{Passthrough}

\item{ClassWeights.}{Passthrough}

\item{EvalMetric.}{Passthrough}

\item{LossFunction.}{Passthrough}

\item{task_type.}{Passthrough}

\item{NumGPUs.}{Passthrough}

\item{model_path.}{Passthrough}

\item{NewGrid.}{Passthrough}

\item{Grid.}{Passthrough}

\item{ExperimentalGrid.}{Passthrough}

\item{GridClusters.}{Passthrough}
}
\description{
CatBoostClassifierParams
}
\seealso{
Other CatBoost Helpers: 
\code{\link{CatBoostArgsCheck}()},
\code{\link{CatBoostDataConversion}()},
\code{\link{CatBoostDataPrep}()},
\code{\link{CatBoostFinalParams}()},
\code{\link{CatBoostImportances}()},
\code{\link{CatBoostPDF}()},
\code{\link{CatBoostParameterGrids}()},
\code{\link{CatBoostRemoveFiles}()},
\code{\link{CatBoostValidationData}()}
}
\author{
Adrian Antico
}
\concept{CatBoost Helpers}
