% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CatBoostHelpers.R
\name{CatBoostImportances}
\alias{CatBoostImportances}
\title{CatBoostImportances}
\usage{
CatBoostImportances(
  ModelType = "regression",
  TargetColumnName. = TargetColumnName,
  BestGrid. = BestGrid,
  TrainOnFull. = TrainOnFull,
  TrainPool. = TrainPool,
  TestPool. = TestPool,
  FinalTestPool. = FinalTestPool,
  TestDataCheck = !is.null(TestData),
  ValidationData. = ValidationData,
  FeatureColNames. = FeatureColNames,
  GridTune. = GridTune,
  task_type. = task_type,
  SaveModelObjects. = SaveModelObjects,
  model. = model,
  ModelID. = ModelID,
  model_path. = model_path,
  metadata_path. = metadata_path,
  GrowPolicy. = GrowPolicy
)
}
\arguments{
\item{ModelType}{'regression', 'classification', or 'multiclass'}

\item{TargetColumnName.}{Passthrough}

\item{BestGrid.}{Passthrough}

\item{TrainOnFull.}{Passthrough}

\item{TrainPool.}{Passthrough}

\item{TestPool.}{Passthrough}

\item{FinalTestPool.}{Passthrough}

\item{TestDataCheck}{Check if TestData is not null}

\item{ValidationData.}{Passthrough}

\item{FeatureColNames.}{Passthrough}

\item{GridTune.}{Passthrough}

\item{task_type.}{Passthrough}

\item{SaveModelObjects.}{Passthrough}

\item{model.}{Passthrough}

\item{ModelID.}{Passthrough}

\item{model_path.}{Passthrough}

\item{metadata_path.}{Passthrough}

\item{GrowPolicy.}{= GrowPolicy}
}
\description{
Generate variable importance, interaction importance, and shap values
}
\seealso{
Other CatBoost Helpers: 
\code{\link{CatBoostArgsCheck}()},
\code{\link{CatBoostDataConversion}()},
\code{\link{CatBoostDataPrep}()},
\code{\link{CatBoostFinalParams}()},
\code{\link{CatBoostGridParams}()},
\code{\link{CatBoostPDF}()},
\code{\link{CatBoostParameterGrids}()},
\code{\link{CatBoostRemoveFiles}()},
\code{\link{CatBoostValidationData}()}
}
\author{
Adrian
}
\concept{CatBoost Helpers}
