% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyFunctions.R
\name{CatBoostLossFunctionOptions}
\alias{CatBoostLossFunctionOptions}
\title{CatBoostLossFunctionOptions}
\usage{
CatBoostLossFunctionOptions(TT)
}
\arguments{
\item{TT}{Target Type: 'regression', 'classification', 'multiclass', all lower case}
}
\description{
CatBoostLossFunctionOptions
}
\keyword{internal}
