% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CatBoostHelpers.R
\name{CatBoostPDF}
\alias{CatBoostPDF}
\title{CatBoostPDF}
\usage{
CatBoostPDF(
  ModelClass = "catboost",
  ModelType = "regression",
  TrainOnFull. = TrainOnFull,
  SaveInfoToPDF. = SaveInfoToPDF,
  EvaluationPlot. = EvaluationPlot,
  EvaluationBoxPlot. = EvaluationBoxPlot,
  ParDepPlots. = ParDepPlots,
  ParDepBoxPlots. = ParDepBoxPlots,
  EvalMetrics. = EvalMetrics,
  VariableImportance. = VariableImportance,
  Interaction. = Interaction,
  model_path. = model_path,
  metadata_path. = metadata_path
)
}
\arguments{
\item{ModelClass}{'catboost', 'xgboost', 'h2o'}

\item{ModelType}{'regression', 'classification', 'multiclass', or 'vector'}

\item{TrainOnFull.}{Passthrough}

\item{SaveInfoToPDF.}{Passthrough}

\item{EvaluationPlot.}{Passthrough}

\item{EvaluationBoxPlot.}{Passthrough}

\item{ParDepPlots.}{Passthrough}

\item{ParDepBoxPlots.}{Passthrough}

\item{EvalMetrics.}{Passthrough}

\item{VariableImportance.}{Passthrough}

\item{Interaction.}{Passthrough}

\item{model_path.}{Passthrough}

\item{metadata_path.}{Passthrough}
}
\description{
Send model insights to pdf
}
\seealso{
Other CatBoost Helpers: 
\code{\link{CatBoostArgsCheck}()},
\code{\link{CatBoostClassifierParams}()},
\code{\link{CatBoostDataConversion}()},
\code{\link{CatBoostDataPrep}()},
\code{\link{CatBoostEvalPlots}()},
\code{\link{CatBoostFinalParams}()},
\code{\link{CatBoostImportances}()},
\code{\link{CatBoostMultiClassParams}()},
\code{\link{CatBoostParameterGrids}()},
\code{\link{CatBoostRegressionParams}()},
\code{\link{CatBoostRemoveFiles}()},
\code{\link{CatBoostValidationData}()}
}
\author{
Adrian
}
\concept{CatBoost Helpers}
