% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReinforcementLearningFunctions.R
\name{CatBoostRegressionParams}
\alias{CatBoostRegressionParams}
\title{CatBoostRegressionParams}
\usage{
CatBoostRegressionParams(counter, BanditArmsN, HasTime, MetricPeriods,
  eval_metric, task_type, model_path, NewGrid, Grid, ExperimentalGrid,
  GridClusters)
}
\arguments{
\item{counter}{Passthrough}

\item{BanditArmsN}{Passthrough}

\item{HasTime}{Passthrough}

\item{MetricPeriods}{Passthrough}

\item{eval_metric}{Passthrough}

\item{task_type}{Passthrough}

\item{model_path}{Passthrough}

\item{NewGrid}{Passthrough}

\item{Grid}{Passthrough}

\item{ExperimentalGrid}{Passthrough}

\item{GridClusters}{Passthrough}
}
\description{
CatBoostRegressionParams
}
\seealso{
Other Supervised Learning: \code{\link{AutoH2OScoring}},
  \code{\link{CatBoostClassifierParams}},
  \code{\link{CatBoostParameterGrids}}
}
\author{
Adrian Antico
}
\concept{Supervised Learning}
