% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CatBoostHelpers.R
\name{CatBoostRemoveFiles}
\alias{CatBoostRemoveFiles}
\title{CatBoostRemoveFiles}
\usage{
CatBoostRemoveFiles(GridTune. = GridTune)
}
\arguments{
\item{GridTune.}{Passthrough}
}
\description{
Remove temp files generated by catboost
}
\seealso{
Other CatBoost Helpers: 
\code{\link{CatBoostArgsCheck}()},
\code{\link{CatBoostClassifierParams}()},
\code{\link{CatBoostDataConversion}()},
\code{\link{CatBoostDataPrep}()},
\code{\link{CatBoostEvalPlots}()},
\code{\link{CatBoostFinalParams}()},
\code{\link{CatBoostImportances}()},
\code{\link{CatBoostMultiClassParams}()},
\code{\link{CatBoostPDF}()},
\code{\link{CatBoostParameterGrids}()},
\code{\link{CatBoostRegressionParams}()},
\code{\link{CatBoostValidationData}()}
}
\author{
Adrian
}
\concept{CatBoost Helpers}
