% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CatBoostHelpers.R
\name{CatBoostValidationData}
\alias{CatBoostValidationData}
\title{CatBoostValidationData}
\usage{
CatBoostValidationData(
  ModelType = "classification",
  TrainOnFull. = TrainOnFull,
  TestDataCheck = !is.null(TestData),
  FinalTestTarget. = FinalTestTarget,
  TestTarget. = TestTarget,
  TrainTarget. = TrainTarget,
  TestMerge. = TestMerge,
  dataTest. = dataTest,
  data. = data,
  predict. = predict,
  TargetColumnName. = TargetColumnName,
  SaveModelObjects. = SaveModelObjects,
  metadata_path. = metadata_path,
  model_path. = model_path,
  ModelID. = ModelID,
  LossFunction. = LossFunction,
  TransformNumericColumns. = TransformNumericColumns,
  GridTune. = GridTune,
  TransformationResults. = TransformationResults,
  TargetLevels. = TargetLevels
)
}
\arguments{
\item{ModelType}{= "classification",}

\item{TrainOnFull.}{Passthrough}

\item{TestDataCheck}{= !is.null(TestData),}

\item{FinalTestTarget.}{Passthrough}

\item{TestTarget.}{Passthrough}

\item{TrainTarget.}{Passthrough}

\item{TestMerge.}{Passthrough}

\item{dataTest.}{Passthrough}

\item{data.}{Passthrough}

\item{predict.}{Passthrough}

\item{TargetColumnName.}{Passthrough}

\item{SaveModelObjects.}{Passthrough}

\item{metadata_path.}{Passthrough}

\item{model_path.}{Passthrough}

\item{ModelID.}{Passthrough}

\item{LossFunction.}{Passthrough regression}

\item{TransformNumericColumns.}{Passthrough regression}

\item{GridTune.}{Passthrough regression}

\item{TransformationResults.}{Passthrough regression}

\item{TargetLevels.}{Passthrough multiclass}
}
\description{
Return validation data with predictions and save to file if requested
}
\seealso{
Other CatBoost Helpers: 
\code{\link{CatBoostArgsCheck}()},
\code{\link{CatBoostDataConversion}()},
\code{\link{CatBoostDataPrep}()},
\code{\link{CatBoostFinalParams}()},
\code{\link{CatBoostGridParams}()},
\code{\link{CatBoostImportances}()},
\code{\link{CatBoostPDF}()},
\code{\link{CatBoostParameterGrids}()},
\code{\link{CatBoostRemoveFiles}()}
}
\author{
Adrian Antico
}
\concept{CatBoost Helpers}
