% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CausalMediation.R
\name{CausalMediation}
\alias{CausalMediation}
\title{CausalMediation}
\usage{
CausalMediation(
  data,
  OutcomeTargetVariable = NULL,
  TreatmentVariable = NULL,
  MediatorVariable = NULL,
  Covariates = NULL,
  MM_TreatmentCovariates = NULL,
  OM_TreatmentCovariates = NULL,
  OM_MediatorCovariates = NULL,
  SurvivalEventVariable = NULL,
  UnTreated_ReferenceIndicator = NULL,
  Treated_ReferenceIndicator = NULL,
  Mediator_ControlDirectEffectLevel = NULL,
  Covariate_NaturalDirectIndirect = 0,
  MediatorTargetType = "linear",
  OutcomeTargetType = "linear",
  TreatmentMediatorInteraction = TRUE,
  CaseControlSourceData = FALSE,
  RemoveNA = FALSE
)
}
\arguments{
\item{data}{Data frame containing the following relevant variables.}

\item{OutcomeTargetVariable}{yvar in underlying model. A character vector of length 1. Outcome variable name. It should be the time variable for the survival outcome.}

\item{TreatmentVariable}{avar in underlying model. A character vector of length 1. Treatment variable name.}

\item{MediatorVariable}{mvar in underlying model. A character vector of length 1. Mediator variable name.}

\item{Covariates}{For main model}

\item{MM_TreatmentCovariates}{emm_ac_mreg in underlying model. A character vector of length > 0. Effect modifiers names. The covariate vector in treatment-covariate product term in the mediator model.}

\item{OM_TreatmentCovariates}{emm_ac_yreg in underlying model. A character vector of length > 0. Effect modifiers names. The covariate vector in treatment-covariate product term in the outcome model.}

\item{OM_MediatorCovariates}{emm_mc_yreg in underlying model. A character vector of length > 0. Effect modifiers names. The covariate vector in mediator-covariate product term in outcome model.}

\item{SurvivalEventVariable}{eventvar in underlying model. An character vector of length 1. Only required for survival outcome regression models. Note that the coding is 1 for event and 0 for censoring, following the R survival package convention.}

\item{UnTreated_ReferenceIndicator}{a0 in underlying model. A numeric vector of length 1. The reference level of treatment variable that is considered "untreated" or "unexposed".}

\item{Treated_ReferenceIndicator}{a1 in underlying model. A numeric vector of length 1.}

\item{Mediator_ControlDirectEffectLevel}{m_cde in underlying model. A numeric vector of length 1. Mediator level at which controlled direct effect is evaluated at.}

\item{Covariate_NaturalDirectIndirect}{c_cond in underlying model. A numeric vector of the same length as \code{cvar}. Covariate levels at which natural direct and indirect effects are evaluated at.}

\item{MediatorTargetType}{mreg in underlying model. A character vector of length 1. Mediator regression type: \code{"linear"} or \code{"logistic"}.}

\item{OutcomeTargetType}{yreg in underlying model. A character vector of length 1. Outcome regression type: \code{"linear"}, \code{"logistic"}, \code{"loglinear"}, \code{"poisson"}, \code{"negbin"}, \code{"survCox"}, \code{"survAFT_exp"}, or \code{"survAFT_weibull"}.}

\item{TreatmentMediatorInteraction}{interaction in underlying model. A logical vector of length 1. The presence of treatment-mediator interaction in the outcome model. Default to TRUE.}

\item{CaseControlSourceData}{casecontrol in underlying model. A logical vector of length 1. Default to FALSE. Whether data comes from a case-control study.}

\item{RemoveNA}{na_omit in underlying model. A logical vector of length 1. Default to FALSE. Whether to remove NAs in the columns of interest before fitting the models.}

\item{ConfoundingVariables}{cvar in underlying model. A character vector of length > 0. Covariate names. Use \code{NULL} if there is no covariate. However, this is a highly suspicious situation. Even if \code{avar} is randomized, \code{mvar} is not. Thus, there are usually some confounder(s) to account for the common cause structure (confounding) between \code{mvar} and \code{yvar}.}
}
\value{
list with model output object, summary output, effects output, and an effects plot
}
\description{
CausalMediation utilizes models from regmedint package
}
\examples{
\dontrun{
library(regmedint) # to load vv2015
data(vv2015)
Output <- RemixAutoML::CausalMediation(
  data = vv2015,
  OutcomeTargetVariable = 'y',            # yvar char length = 0
  TreatmentVariable = "x",                # avar char length = 0 (binary)
  MediatorVariable = "m",                 # mvar char length = 0 (binary)
  Covariates = "c",                       # cvar char length > 0
  MM_TreatmentCovariates = NULL,          # emm_ac_mreg = NULL char length > 0
  OM_TreatmentCovariates = NULL,          # emm_ac_yreg = NULL char length > 0
  OM_MediatorCovariates = NULL,           # emm_mc_yreg = NULL char length > 0
  SurvivalEventVariable = "event",        # eventvar char length = 0
  UnTreated_ReferenceIndicator = 0,       # ao num length = 1
  Treated_ReferenceIndicator = 1,         # a1 num length = 1
  Mediator_ControlDirectEffectLevel = 1,  # m_cde num length = 1
  Covariate_NaturalDirectIndirect = 3,    # c_cond; same length as Covariates num length = length(Covariates)
  MediatorTargetType = 'logistic',        # mreg "linear" or "logistic",
  OutcomeTargetType = 'survAFT_weibull',  # yreg "linear", "logistic", "loglinear", "poisson", "negbin", "survCox", "survAFT_exp", or "survAFT_weibull"
  TreatmentMediatorInteraction = TRUE,    # interaction = TRUE,
  CaseControlSourceData = FALSE,          # casecontrol = FALSE,
  RemoveNA = FALSE)

# data = vv2015
# OutcomeTargetVariable = 'y'
# TreatmentVariable = "x"
# MediatorVariable = "m"
# Covariates = "c"
# MM_TreatmentCovariates = NULL
# OM_TreatmentCovariates = NULL
# OM_MediatorCovariates = NULL
# SurvivalEventVariable = "event"
# UnTreated_ReferenceIndicator = 0
# Treated_ReferenceIndicator = 1
# Mediator_ControlDirectEffectLevel = 1
# Covariate_NaturalDirectIndirect = 3
# MediatorTargetType = 'logistic'
# OutcomeTargetType = 'survAFT_weibull'
# TreatmentMediatorInteraction = TRUE
# CaseControlSourceData = FALSE
# RemoveNA = FALSE
}

}
\author{
Adrian Antico
}
\concept{Statistical Modeling}
