% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Adrian_Antico_Modeling_Tools_Library.R
\name{ChartTheme}
\alias{ChartTheme}
\title{ChartTheme function is a ggplot theme generator for ggplots}
\usage{
ChartTheme(BackGround = "lightsteelblue1", OtherColor = "navyblue",
  Size = 0.25)
}
\arguments{
\item{OtherColor}{This is the color for borders and text}

\item{Background}{This is the color of the background on the chart}
}
\value{
An object to pass along to ggplot objects following the "+" sign
}
\description{
This function helps your ggplots look professional with the choice of the two main colors that will dominate the theme
}
\examples{
p <- ggplot(data, aes(x = weeks, y = quantity)) + geom_line()
p <- p + ChartTheme(BackGround = "aquamarine", OtherColor = "purple4")
p
}
\author{
Adrian Antico at RemixInstitute.com
}
