% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyFunctions.R
\name{ChoicesByType}
\alias{ChoicesByType}
\title{ChoicesByType}
\usage{
ChoicesByType(data)
}
\arguments{
\item{data}{data.table}
}
\value{
a list of columns names by data type
}
\description{
ChoicesByType
}
\keyword{internal}
