% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelMetrics.R
\name{ClassificationMetrics}
\alias{ClassificationMetrics}
\title{ClassificationMetrics}
\usage{
ClassificationMetrics(
  TestData,
  Thresholds,
  Target,
  PredictColumnName,
  PositiveOutcome,
  NegativeOutcome,
  CostMatrix = c(0, 1, 1, 0)
)
}
\arguments{
\item{TestData}{Test data from your modeling}

\item{Thresholds}{Value}

\item{Target}{Name of your target variable}

\item{PredictColumnName}{Name of your predicted value variable}

\item{PositiveOutcome}{The value of the positive outcome level}

\item{NegativeOutcome}{The value of the negative outcome level}

\item{CostMatrix}{c(True Positive Cost, False Negative Cost, False Positive Cost, True Negative Cost)}
}
\description{
ClassificationMetrics
}
\seealso{
Other Model Evaluation: 
\code{\link{BinaryMetrics}()},
\code{\link{DT_BinaryConfusionMatrix}()},
\code{\link{MultiClassMetrics}()},
\code{\link{RegressionMetrics}()},
\code{\link{RemixClassificationMetrics}()}
}
\author{
Adrian Antico
}
\concept{Model Evaluation}
