% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataWranglingFunctions.R
\name{ColumnSubsetDataTable}
\alias{ColumnSubsetDataTable}
\title{ColumnSubsetDataTable}
\usage{
ColumnSubsetDataTable(
  data,
  TargetColumnName = NULL,
  DateColumnName = NULL,
  GroupVars = NULL
)
}
\arguments{
\item{data}{data.table}

\item{TargetColumnName}{Target variable}

\item{DateColumnName}{Date variable}

\item{GroupVars}{Group variables}
}
\description{
ColumnSubsetDataTable will subset data tables by column
}
\seealso{
Other Data Wrangling: 
\code{\link{DataDisplayMeta}()},
\code{\link{FakeDataGenerator}()},
\code{\link{FullFactorialCatFeatures}()},
\code{\link{TimeSeriesMelt}()}
}
\author{
Adrian Antico
}
\concept{Data Wrangling}
