% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotFunctions.R
\name{CorrMatrixPlot}
\alias{CorrMatrixPlot}
\title{CorrMatrixPlot}
\usage{
CorrMatrixPlot(data = NULL, CorrVars = NULL, Method = "spearman")
}
\arguments{
\item{data}{Source data.table}

\item{CorrVars}{Column names of variables you want included in the correlation matrix}

\item{Method}{'spearman' default, 'pearson' otherwise}
}
\description{
Build a violin plot by simply passing arguments to a single function. It will sample your data using SampleSize number of rows. Sampled data is randomized.
}
\examples{
\dontrun{
data <- data.table::fread(file.choose())
CorrVars <- c('Weekly_Sales', 'XREG1', 'XREG2', 'XREG3')
p <- cor(data[, .SD, .SDcols = c(CorrVars)])
p1 <- heatmaply::heatmaply_cor(
  p,
  colors = c('red', 'white', 'blue'),
  xlab = "Features",
  ylab = "Features",
  k_col = 2,
  k_row = 2)
}

}
\seealso{
Other Graphics: 
\code{\link{AddFacet}()},
\code{\link{BarPlot}()},
\code{\link{BoxPlot}()},
\code{\link{ChartTheme}()},
\code{\link{HeatMapPlot}()},
\code{\link{HistPlot}()},
\code{\link{PlotlyConversion}()},
\code{\link{StockData}()},
\code{\link{StockPlot}()},
\code{\link{ViolinPlot}()},
\code{\link{multiplot}()}
}
\author{
Adrian Antico
}
\concept{Graphics}
