% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Adrian_Antico_Modeling_Tools_Library.R
\name{CountSingleDigits}
\alias{CountSingleDigits}
\title{CountSingleDigits counts the number of digits in a string}
\usage{
CountSingleDigits(data, col)
}
\arguments{
\item{data}{the source data.table}

\item{col}{the column number with the string to evaluate}
}
\value{
The original data.table with the added columns merged in
}
\description{
CountSingleDigits counts the number of digits in a string
}
\examples{
VIMP[, variable2 := ifelse(CountSingleDigits(VIMP,1) == 2, substr(variable, 1,7), substr(variable, 1,6))]
}
\author{
Adrian Antico at RemixInstitute.com
}
