% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateCalendarVariables.R
\name{CreateCalendarVariables}
\alias{CreateCalendarVariables}
\title{CreateCalendarVariables Create Caledar Variables}
\usage{
CreateCalendarVariables(
  data,
  DateCols = NULL,
  AsFactor = FALSE,
  TimeUnits = "wday"
)
}
\arguments{
\item{data}{This is your data}

\item{DateCols}{Supply either column names or column numbers of your date columns you want to use for creating calendar variables}

\item{AsFactor}{Set to TRUE if you want factor type columns returned; otherwise integer type columns will be returned}

\item{TimeUnits}{Supply a character vector of time units for creating calendar variables. Options include: "second", "minute", "hour", "wday", "mday", "yday", "week", "isoweek", "month", "quarter", "year"}
}
\value{
Returns your data.table with the added calendar variables at the end
}
\description{
CreateCalendarVariables Rapidly creates calendar variables based on the date column you provide
}
\examples{
data <- data.table::data.table(Date = "2018-01-01 00:00:00")
data <- CreateCalendarVariables(
   data,
   DateCols = "Date",
   AsFactor = FALSE,
   TimeUnits = c("wday", "month", "year"))
}
\seealso{
Other Feature Engineering: 
\code{\link{AutoDataPartition}()},
\code{\link{AutoFourierFeatures}()},
\code{\link{AutoLagRollStatsScoring}()},
\code{\link{AutoLagRollStats}()},
\code{\link{AutoTransformationCreate}()},
\code{\link{AutoTransformationScore}()},
\code{\link{AutoWord2VecModeler}()},
\code{\link{ContinuousTimeDataGenerator}()},
\code{\link{CreateHolidayVariables}()},
\code{\link{DT_GDL_Feature_Engineering}()},
\code{\link{DummifyDT}()},
\code{\link{H2oAutoencoder}()},
\code{\link{ModelDataPrep}()},
\code{\link{Partial_DT_GDL_Feature_Engineering}()},
\code{\link{TimeSeriesFill}()}
}
\author{
Adrian Antico
}
\concept{Feature Engineering}
