% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateHolidayVariables.R
\name{CreateHolidayVariables}
\alias{CreateHolidayVariables}
\title{CreateHolidayVariables Create Holiday Count Columns}
\usage{
CreateHolidayVariables(data, DateCols = "DateTime", TimeUnit = "day",
  HolidayGroups = c("USPublicHolidays", "EasterGroup", "ChristmasGroup",
  "OtherEcclesticalFeasts"), Holidays = NULL)
}
\arguments{
\item{data}{This is your data}

\item{DateCols}{Supply either column names or column numbers of your date columns you want to use for creating calendar variables}

\item{TimeUnit}{Supply a character value for the time unit your date columns are in, such as 'day'}

\item{AsFactor}{Set to TRUE if you want factor type columns returned; otherwise integer type columns will be returned}
}
\value{
Returns your data.table with the added holiday indicator variable
}
\description{
CreateHolidayVariables Rapidly creates holiday count variables based on the date columns you provide
}
\examples{
data <- data.table::data.table(Date = "2018-01-01 00:00:00")
data <- CreateHolidayVariables(data,
                               DateCols = "DateTime",
                               TimeUnit = "day",
                               HolidayGroups = c("USPublicHolidays","EasterGroup",
                                                 "ChristmasGroup","OtherEcclesticalFeasts"),
                               Holidays = NULL)
}
\seealso{
Other Feature Engineering: \code{\link{AutoDataPartition}},
  \code{\link{AutoTransformationCreate}},
  \code{\link{AutoTransformationScore}},
  \code{\link{AutoWord2VecModeler}},
  \code{\link{CreateCalendarVariables}},
  \code{\link{DT_GDL_Feature_Engineering}},
  \code{\link{DummifyDT}},
  \code{\link{GDL_Feature_Engineering}},
  \code{\link{ModelDataPrep}},
  \code{\link{Partial_DT_GDL_Feature_Engineering2}},
  \code{\link{Partial_DT_GDL_Feature_Engineering}},
  \code{\link{Scoring_GDL_Feature_Engineering}}
}
\author{
Adrian Antico
}
\concept{Feature Engineering}
