% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelEvaluationPlots.R
\name{CumGainsChart}
\alias{CumGainsChart}
\title{CumGainsChart}
\usage{
CumGainsChart(
  data = NULL,
  PredictedColumnName = "p1",
  TargetColumnName = NULL,
  NumBins = 20,
  SavePlot = FALSE,
  Name = NULL,
  metapath = NULL,
  modelpath = NULL
)
}
\arguments{
\item{data}{Test data with predictions. data.table}

\item{TargetColumnName}{Name of your target variable column}

\item{NumBins}{Number of percentile bins to plot}

\item{SavePlot}{FALSE by default}

\item{Name}{File name for saving}

\item{metapath}{Path to directory}

\item{modelpath}{Path to directory}

\item{PredictionColumnName}{Name of column that is the model score}
}
\description{
Create a cumulative gains chart
}
\seealso{
Other Model Evaluation and Interpretation: 
\code{\link{EvalPlot}()},
\code{\link{ParDepCalPlots}()},
\code{\link{ROCPlot}()},
\code{\link{RedYellowGreen}()},
\code{\link{threshOptim}()}
}
\author{
Adrian Antico
}
\concept{Model Evaluation and Interpretation}
