% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyFunctions.R
\name{DateInput}
\alias{DateInput}
\title{DateInput}
\usage{
DateInput(
  InputID = "TS_CARMA_HolidayMovingAverages",
  Label = "Import Data Creation Date",
  Value = Sys.Date(),
  Min = "1970-01-01",
  Max = "2100-01-01",
  Format = "yyyy-mm-dd"
)
}
\arguments{
\item{InputID}{Feeds ProjectList and inputId. Argument saved in ProjectList}

\item{Label}{Feeds label}

\item{Value}{Default}

\item{Min}{Min value}

\item{Max}{Max value}

\item{Format}{Date format}
}
\value{
PickerInput object for server.R to go into renderUI({PickerInput()})
}
\description{
DateInput automatically builds a date input with ProjectList argument usage if it exists
}
\examples{
\dontrun{
output$TS_Date <- renderUI({
  RemixAutoML::DateInput(InputID = "TS_CARMA_HolidayMovingAverages",
                         Label = "Import Data Creation Date",
                         Value = Sys.Date(),
                         Min = "1970-01-01",
                         Max = "2100-01-01",
                         Format = "yyyy-mm-dd")})
}
}
\seealso{
Other Shiny: 
\code{\link{App_TimeSeriesPlots}()},
\code{\link{ArgNullCheck2}()},
\code{\link{ArgNullCheck}()},
\code{\link{BlankRow}()},
\code{\link{GenerateEvaluationMetrics}()},
\code{\link{NumericInput}()},
\code{\link{PickerInput_GetLevels}()},
\code{\link{PickerInput}()},
\code{\link{PreparePlotData}()},
\code{\link{ReactiveLoadCSV}()},
\code{\link{ReturnParam}()},
\code{\link{StoreArgs}()},
\code{\link{TextInput}()},
\code{\link{observeEventLoad}()}
}
\author{
Adrian Antico
}
\concept{Shiny}
