% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscFunctions.R
\name{DeleteFile}
\alias{DeleteFile}
\title{DeleteFile}
\usage{
DeleteFile(File = NULL)
}
\arguments{
\item{File}{If NULL a prompt will allow you to click on the file to have it removed. Otherwise, supply a path to the file including its name and extension}
}
\description{
DeleteFile will prompt you for a file to delete and then permanently delete a file. You won't have to go the the recycle bin to delete it a second time
}
\seealso{
Other System Functions: 
\code{\link{CreateProjectFolders}()}
}
\author{
Adrian Antico
}
\concept{System Functions}
