% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoFreqSizeDistSimFunctions.R
\name{Demand}
\alias{Demand}
\title{IntermittentDemandSimulator}
\usage{
Demand(sku, FC_Periods, nSims, CountList, SizeList)
}
\description{
IntermittentDemandSimulator
}
\seealso{
Other Automated Time Series: \code{\link{AutoCatBoostCARMA}},
  \code{\link{AutoCatBoostFreqSizeScoring}},
  \code{\link{AutoCatBoostSizeFreqDist}},
  \code{\link{AutoH2oDRFCARMA}},
  \code{\link{AutoH2oGBMCARMA}}, \code{\link{AutoTS}},
  \code{\link{AutoXGBoostCARMA}},
  \code{\link{IntermittentDemandScoringDataGenerator}}
}
\author{
Adrian Antico
}
\concept{Automated Time Series}
