% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotFunctions.R
\name{DensityPlot}
\alias{DensityPlot}
\title{DensityPlot}
\usage{
DensityPlot(data, GroupVariables, MeasureVars)
}
\arguments{
\item{data}{data.table}

\item{GroupVariables}{= NULL}

\item{MeasureVariables}{= NULL}
}
\description{
Density plots, by groups, with transparent continuous plots
}
\seealso{
Other Graphics: 
\code{\link{AddFacet}()},
\code{\link{BarPlot}()},
\code{\link{BoxPlot}()},
\code{\link{ChartTheme}()},
\code{\link{CorrMatrixPlot}()},
\code{\link{HeatMapPlot}()},
\code{\link{HistPlot}()},
\code{\link{PlotlyConversion}()},
\code{\link{StockData}()},
\code{\link{StockPlot}()},
\code{\link{ViolinPlot}()},
\code{\link{multiplot}()}
}
\concept{Graphics}
