% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SystemFunctions.R
\name{DownloadCSVFromStorageExplorer}
\alias{DownloadCSVFromStorageExplorer}
\title{DownloadCSVFromStorageExplorer}
\usage{
DownloadCSVFromStorageExplorer(
  UploadCSVObjectName = "TransfersRawData.csv",
  SaveCSVFilePath = file.path(Root),
  SaveCSVName = "RawData.csv",
  UploadLocation = "Analytics Sandbox/Machine Learning",
  DataStoreName = NULL
)
}
\arguments{
\item{UploadCSVObjectName}{Name of the file you uploaded to the Microsoft Azure Storage Explorer}

\item{SaveCSVFilePath}{Path file to where you want to save your csv in Azure}

\item{SaveCSVName}{The name you want to give the csv that will be saved}

\item{UploadLocation}{The location to where the data is saved in the Azure Storage Explorer}

\item{DataStoreName}{The name of the store in data factory where you uploaded your data}
}
\description{
DownloadCSVFromStorageExplorer
}
\author{
Adrian Antico
}
\concept{Azure}
