% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelMetrics.R
\name{DT_BinaryConfusionMatrix}
\alias{DT_BinaryConfusionMatrix}
\title{DT_BinaryConfusionMatrix}
\usage{
DT_BinaryConfusionMatrix(
  data = MetricsData,
  GroupVariables = "IntervalNum",
  Target = "ActiveAtInterval",
  Predicted = "p1"
)
}
\arguments{
\item{data}{Supply your model validation data with predictions}

\item{GroupVariables}{Supply grouping variables to generate statistics by groups}

\item{Target}{The name of your target variable column}

\item{Predicted}{The name of your predicted value column#'}
}
\description{
DT_BinaryConfusionMatrix is for computing all metrics related to binary modeling outcomes
}
\examples{
\dontrun{
AggMetricsByGroup <- DT_BinaryConfusionMatrix(
  data,
  GroupVariables = c("Store","Dept"),
  Target = "HitTarget",
  Predicted = "p1")
}
}
\seealso{
Other Model Evaluation: 
\code{\link{ClassificationMetrics}()},
\code{\link{RemixClassificationMetrics}()}
}
\author{
Adrian Antico
}
\concept{Model Evaluation}
