% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Adrian_Antico_Modeling_Tools_Library.R
\name{DummifyDT}
\alias{DummifyDT}
\title{DummifyDT creates dummy variables for the selected columns}
\usage{
DummifyDT(data, cols, KeepBaseCols = TRUE)
}
\arguments{
\item{data}{the data set to run the micro auc on}

\item{cols}{a vector with the names of the columns you wish to dichotomize}

\item{KeepBaseCols}{set to TRUE to keep the original columns used in the dichotomization process}
}
\value{
data table with new dummy variables columns and optionally removes base columns
}
\description{
DummifyDT creates dummy variables for the selected columns
}
\examples{
cols <- c("ColumnA")
data <- DummifyDT(data, cols)
}
\author{
Adrian Antico
}
