% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EDAFunctions.R
\name{EDA_Histograms}
\alias{EDA_Histograms}
\title{EDA_Histograms}
\usage{
EDA_Histograms(
  data = NULL,
  PlotColumns = NULL,
  SampleCount = 1e+05,
  SavePath = NULL,
  FactorCountPerPlot = 10,
  PrintOutput = FALSE,
  Size = 12,
  AngleX = 35,
  AngleY = 0,
  ChartColor = "lightsteelblue1",
  BorderColor = "darkblue",
  TextColor = "darkblue",
  GridColor = "white",
  BackGroundColor = "gray95",
  LegendPosition = "bottom"
)
}
\arguments{
\item{data}{Input data.table}

\item{PlotColumns}{Default NULL. If NULL, all columns will be plotted (except date cols). Otherwise, supply a character vector of columns names to plot}

\item{SampleCount}{Number of random samples to use from data. data is first shuffled and then random samples taken}

\item{SavePath}{Output file path to where you can optionally save pdf}

\item{FactorCountPerPlot}{Default 10}

\item{PrintOutput}{Default FALSE. TRUE will print results upon running function}

\item{Size}{Default 12}

\item{AngleX}{Default 35}

\item{AngleY}{Default 0}

\item{ChartColor}{Default "lightsteelblue1"}

\item{BorderColor}{Default "darkblue"}

\item{TextColor}{Default "darkblue"}

\item{GridColor}{Default "white"}

\item{BackGroundColor}{Default "gray95"}

\item{LegendPosition}{Default "bottom"}
}
\description{
Creates histograms
}
\seealso{
Other EDA: 
\code{\link{AutoWordFreq}()},
\code{\link{ScatterCopula}()}
}
\author{
Adrian Antico
}
\concept{EDA}
