% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyFunctions.R
\name{FilterLogicData}
\alias{FilterLogicData}
\title{FilterLogicData}
\usage{
FilterLogicData(
  data1,
  FilterLogic = input[["FilterLogic"]],
  FilterVariable = input[["FilterVariable_1"]],
  FilterValue = input[["FilterValue_1a"]],
  FilterValue2 = input[["FilterValue_1b"]],
  Debug = FALSE
)
}
\arguments{
\item{data1}{data.table}

\item{FilterLogic}{passthrough}

\item{FilterVariable}{passthrough}

\item{FilterValue}{passthrough}

\item{FilterValue2}{passthrough}
}
\description{
FilterLogicData
}
