% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EconometricsFunctions.R
\name{FinalBuildArfima}
\alias{FinalBuildArfima}
\title{FinalBuildArfima}
\usage{
FinalBuildArfima(ModelOutputGrid = NULL,
  TimeSeriesPrepareOutput = NULL, FCPeriods = 1,
  MetricSelection = "MAE", NumberModelsScore = 1, ByDataType = TRUE)
}
\arguments{
\item{ModelOutputGrid}{Pass along the grid output from ParallelOptimzeArima()}

\item{TimeSeriesPrepareOutput}{Output from TimeSeriesPrepare()}

\item{FCPeriods}{The number of periods ahead to forecast}

\item{MetricSelection}{The value returned from TimeSeriesPrepare()}

\item{NumberModelsScore}{The value returned from TimeSeriesPrepare()}

\item{ByDataType}{Set to TRUE if you want to have models represented from all data sets utilized in training}
}
\value{
Time series data sets to pass onto auto modeling functions
}
\description{
FinalBuildArfima to generate forecasts and ensemble data
}
\examples{
FinalBuildArfima(
  Output = NULL, 
  TimeSeriesPrepareOutput = NULL,
  MaxFourierTerms = 0,
  TrainValidateShare = c(0.50,0.50),
  MaxNumberModels = 5,
  MaxRunMinutes = 5)
}
\seealso{
Other Time Series: \code{\link{AutoBanditNNet}},
  \code{\link{AutoBanditSarima}}, \code{\link{AutoTBATS}},
  \code{\link{CarmaHoldoutMetrics}},
  \code{\link{DifferenceDataReverse}},
  \code{\link{DifferenceData}},
  \code{\link{FinalBuildArima}},
  \code{\link{FinalBuildETS}},
  \code{\link{FinalBuildNNET}},
  \code{\link{FinalBuildTBATS}},
  \code{\link{FinalBuildTSLM}},
  \code{\link{GenerateParameterGrids}},
  \code{\link{OptimizeArfima}},
  \code{\link{OptimizeArima}}, \code{\link{OptimizeETS}},
  \code{\link{OptimizeNNET}}, \code{\link{OptimizeTBATS}},
  \code{\link{OptimizeTSLM}},
  \code{\link{ParallelAutoARIMA}},
  \code{\link{ParallelAutoArfima}},
  \code{\link{ParallelAutoETS}},
  \code{\link{ParallelAutoNNET}},
  \code{\link{ParallelAutoTBATS}},
  \code{\link{ParallelAutoTSLM}},
  \code{\link{RL_Performance}},
  \code{\link{Regular_Performance}},
  \code{\link{StackedTimeSeriesEnsembleForecast}},
  \code{\link{TimeSeriesDataPrepare}},
  \code{\link{WideTimeSeriesEnsembleForecast}}
}
\author{
Adrian Antico
}
\concept{Time Series}
