% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EconometricsFunctions.R
\name{FinalBuildTBATS}
\alias{FinalBuildTBATS}
\title{FinalBuildTBATS}
\usage{
FinalBuildTBATS(
  ModelOutputGrid = NULL,
  SavePath = NULL,
  TimeSeriesPrepareOutput = NULL,
  FCPeriods = 1,
  MetricSelection = "MAE",
  NumberModelsScore = 1,
  ByDataType = FALSE,
  DebugMode = FALSE
)
}
\arguments{
\item{ModelOutputGrid}{Pass along the grid output from ParallelOptimzeArima()}

\item{SavePath}{NULL returns nothing. Provide a path to save model object}

\item{TimeSeriesPrepareOutput}{Output from TimeSeriesPrepare()}

\item{FCPeriods}{The number of periods ahead to forecast}

\item{MetricSelection}{The value returned from TimeSeriesPrepare()}

\item{NumberModelsScore}{The value returned from TimeSeriesPrepare()}

\item{ByDataType}{Set to TRUE if you want to have models represented from all data sets utilized in training}

\item{DebugMode}{Set to TRUE to print steps}
}
\value{
Time series data sets to pass onto auto modeling functions
}
\description{
FinalBuildTBATS to generate forecasts and ensemble data
}
\examples{
\dontrun{
FinalBuildTBATS(
  Output = NULL,
  SavePath = NULL,
  TimeSeriesPrepareOutput = NULL,
  MaxFourierTerms = 0,
  TrainValidateShare = c(0.50,0.50),
  MaxNumberModels = 5,
  MaxRunMinutes = 5,
  ByDataType = FALSE,
  DebugMode = FALSE)
}
}
\seealso{
Other Time Series Helper: 
\code{\link{FinalBuildArfima}()},
\code{\link{FinalBuildArima}()},
\code{\link{FinalBuildETS}()},
\code{\link{FinalBuildNNET}()},
\code{\link{FinalBuildTSLM}()},
\code{\link{GenerateParameterGrids}()},
\code{\link{OptimizeArfima}()},
\code{\link{OptimizeArima}()},
\code{\link{OptimizeETS}()},
\code{\link{OptimizeNNET}()},
\code{\link{OptimizeTBATS}()},
\code{\link{OptimizeTSLM}()},
\code{\link{ParallelAutoARIMA}()},
\code{\link{ParallelAutoArfima}()},
\code{\link{ParallelAutoETS}()},
\code{\link{ParallelAutoNNET}()},
\code{\link{ParallelAutoTBATS}()},
\code{\link{ParallelAutoTSLM}()},
\code{\link{PredictArima}()},
\code{\link{RL_Performance}()},
\code{\link{Regular_Performance}()},
\code{\link{StackedTimeSeriesEnsembleForecast}()},
\code{\link{TimeSeriesDataPrepare}()},
\code{\link{WideTimeSeriesEnsembleForecast}()}
}
\author{
Adrian Antico
}
\concept{Time Series Helper}
