% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyFunctions.R
\name{GenerateEvaluationMetrics}
\alias{GenerateEvaluationMetrics}
\title{GenerateEvaluationMetrics}
\usage{
GenerateEvaluationMetrics(
  EvalData = NULL,
  TargetName = NULL,
  DateName = NULL,
  GroupNames = NULL
)
}
\arguments{
\item{EvalData}{Source data in shiny app}

\item{TargetName}{Target variable name}

\item{DateName}{Date variable name}

\item{GroupNames}{Group variable names}
}
\value{
PreparePlotData object for server.R to
}
\description{
GenerateEvaluationMetrics calculates evaluation metrics for out of sample forecast and evaluation data
}
\examples{
\dontrun{
  PlotData <- RemixAutoML::PreparePlotData(input, TargetVariable = "TargetVariables", DateVariable = "DateVariables", GroupVariables = GroupVariables, G1Levels = "TS_Group1Levels", G2Levels = "TS_Group2Levels", G3Levels = "TS_Group3Levels")
}
}
\seealso{
Other Shiny: 
\code{\link{App_TimeSeriesPlots}()},
\code{\link{ArgNullCheck2}()},
\code{\link{ArgNullCheck}()},
\code{\link{BlankRow}()},
\code{\link{DateInput}()},
\code{\link{NumericInput}()},
\code{\link{PickerInput_GetLevels}()},
\code{\link{PickerInput}()},
\code{\link{PreparePlotData}()},
\code{\link{ReactiveLoadCSV}()},
\code{\link{ReturnParam}()},
\code{\link{StoreArgs}()},
\code{\link{TextInput}()},
\code{\link{observeEventLoad}()}
}
\author{
Adrian Antico
}
\concept{Shiny}
