% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EconometricsFunctions.R
\name{GenerateParameterGrids}
\alias{GenerateParameterGrids}
\title{GenerateParameterGrids}
\usage{
GenerateParameterGrids(
  Model = NULL,
  test = NULL,
  MinVal = NULL,
  DataSetName = NULL,
  SeasonalDifferences = NULL,
  SeasonalMovingAverages = NULL,
  SeasonalLags = NULL,
  MaxFourierTerms = NULL,
  Differences = NULL,
  MovingAverages = NULL,
  Lags = NULL
)
}
\arguments{
\item{Model}{'arima', 'ets', 'tbats', 'nnet', 'arfima', 'tslm', 'dshw'}

\item{test}{validation data}

\item{MinVal}{Minimum value of time series}

\item{DataSetName}{Passthrough}

\item{SeasonalDifferences}{Passthrough}

\item{SeasonalMovingAverages}{Passthrough}

\item{SeasonalLags}{Passthrough}

\item{MaxFourierTerms}{Passthrough}

\item{Differences}{Passthrough}

\item{MovingAverages}{Passthrough}

\item{Lags}{Passthrough}
}
\description{
GenerateParameterGrids creates and stores model results in Experiment Grid
}
\seealso{
Other Time Series Helper: 
\code{\link{FinalBuildArfima}()},
\code{\link{FinalBuildArima}()},
\code{\link{FinalBuildETS}()},
\code{\link{FinalBuildNNET}()},
\code{\link{FinalBuildTBATS}()},
\code{\link{FinalBuildTSLM}()},
\code{\link{OptimizeArfima}()},
\code{\link{OptimizeArima}()},
\code{\link{OptimizeETS}()},
\code{\link{OptimizeNNET}()},
\code{\link{OptimizeTBATS}()},
\code{\link{OptimizeTSLM}()},
\code{\link{ParallelAutoARIMA}()},
\code{\link{ParallelAutoArfima}()},
\code{\link{ParallelAutoETS}()},
\code{\link{ParallelAutoNNET}()},
\code{\link{ParallelAutoTBATS}()},
\code{\link{ParallelAutoTSLM}()},
\code{\link{PredictArima}()},
\code{\link{RL_Performance}()},
\code{\link{Regular_Performance}()},
\code{\link{StackedTimeSeriesEnsembleForecast}()},
\code{\link{TimeSeriesDataPrepare}()},
\code{\link{WideTimeSeriesEnsembleForecast}()}
}
\author{
Adrian Antico
}
\concept{Time Series Helper}
