% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Adrian_Antico_Modeling_Tools_Library.R
\name{GenTSAnomVars}
\alias{GenTSAnomVars}
\title{GenTSAnomVars is an automated z-score anomaly detection via GLM-like procedure}
\usage{
GenTSAnomVars(data, GroupVar1 = "SKU", GroupVar2 = NULL,
  DateVar = "DATE", High = 1.96, Low = -1.96, KeepAllCols = FALSE)
}
\arguments{
\item{data}{the source residuals data.table}

\item{GroupVar1}{this is a group by variable}

\item{GroupVar2}{this is another group by variable}

\item{DateVar}{this is a time variable for grouping}

\item{High}{this is the threshold on the high end}

\item{Low}{this is the threshold on the low end}

\item{KeepAllCols}{set to TRUE to remove the intermediate features}
}
\value{
The original data.table with the added columns merged in
}
\description{
GenTSAnomVars is an automated z-score anomaly detection via GLM-like procedure. Data is z-scaled and grouped by factors and time periods to determine which points are above and below the control limits in a cumulative time fashion. Then a cumulative rate is created as the final variable. Set KeepAllCols to FALSE to utilize the intermediate features to create rolling stats from them.
}
\examples{
data <- data.table(a = seq(0,10000,1), predicted = sde::GBM(N=10000)*1000)[-1,]
data <- data.table(a = seq(1,10000,1), predicted = sde::rcCIR(n=10000, Dt=0.1, x0=1, theta=c(6,2,2)))
data <- data.table(a = seq(1,10000,1), predicted = sde::rsOU(n=10000, theta=c(0,2,1)))
stuff    <- GenTSAnomVars(data,
                          GroupVar1   = "a",
                          GroupVar2   = NULL,
                          DateVar     = NULL,
                          High        = 1.96,
                          Low         = -1.96,
                          KeepAllCols = FALSE)
}
\author{
Adrian Antico at RemixInstitute.com
}
