% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyFunctions.R
\name{Get}
\alias{Get}
\title{Get}
\usage{
Get(CharRef, max_env_pos = 10L)
}
\arguments{
\item{x}{code being collected}

\item{y}{collection list to append to}

\item{Debug}{passthrough}
}
\description{
Get
}
\examples{
\dontrun{
microbenchmark::microbenchmark(
  times = 100,
  Get('data1', max_env_pos = 10L),
  Get('data1', max_env_pos = 9),
  Get('data1', max_env_pos = 8),
  Get('data1', max_env_pos = 7),
  Get('data1', max_env_pos = 6),
  Get('data1', max_env_pos = 5),
  Get('data1', max_env_pos = 4),
  Get('data1', max_env_pos = 3),
  Get('data1', max_env_pos = 2),
  Get('data1', max_env_pos = 1))

# Output
# Unit: microseconds
# expr     min       lq      mean   median       uq       max neval
# Get("data1", max_env_pos = 10L) 486.301 501.1010 748.82397 527.5510 727.6010 13876.300   100
# Get("data1", max_env_pos = 9) 446.901 458.0010 544.01307 481.7510 582.3510  1000.300   100
# Get("data1", max_env_pos = 8) 393.501 410.6510 501.97888 431.5010 552.6010  1033.002   100
# Get("data1", max_env_pos = 7) 338.802 351.0015 421.11700 367.7010 451.5505   881.001   100
# Get("data1", max_env_pos = 6) 288.201 296.4505 359.87810 303.2010 372.8010   827.301   100
# Get("data1", max_env_pos = 5) 236.001 244.0010 299.16408 253.4510 316.3505   567.500   100
# Get("data1", max_env_pos = 4) 184.101 191.4015 247.81099 203.1010 291.0510   586.801   100
# Get("data1", max_env_pos = 3) 134.201 138.8020 172.01197 144.1505 182.9505   364.900   100
# Get("data1", max_env_pos = 2)  82.201  85.6010 103.08305  88.3510 112.5510   220.701   100
# Get("data1", max_env_pos = 1)  30.500  32.3505  42.08696  34.4010  46.3005   134.602   100
}

}
\keyword{internal}
