% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyFunctions.R
\name{GetFilterValueLabel}
\alias{GetFilterValueLabel}
\title{GetFilterValueLabel}
\usage{
GetFilterValueLabel(data, VarName = NULL, type = 1)
}
\arguments{
\item{data}{data.table}

\item{VarName}{Variable name}

\item{type}{1 for min, 2 for max}
}
\description{
GetFilterValueLabel
}
