% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Adrian_Antico_Modeling_Tools_Library.R
\name{GLRM_KMeans_Col}
\alias{GLRM_KMeans_Col}
\title{GLRM_KMeans_Col Automated row clustering for mixed column types}
\usage{
GLRM_KMeans_Col(data, GridTuneGLRM = TRUE, nthreads = 4,
  MaxMem = "14G", glrmCols = 1:ncol(data), IgnoreConstCols = TRUE,
  glrmFactors = 5, Loss = "Quadratic", Gamma_x = 0.5, Gamma_y = 0,
  glrmMaxIters = 1000, SVDMethod = "Randomized",
  MaxRunTimeSecs = 3600, KMeansK = 50)
}
\arguments{
\item{data}{is the source time series data.table}

\item{GridTuneGLRM}{If you want to grid tune the glrm model}

\item{nthreads}{set based on number of threads your machine has available}

\item{MaxMem}{set based on the amount of memory your machine has available}

\item{glrmCols}{the column numbers for the glrm}

\item{IgnoreConstCols}{tell H20 to ignore any columns that have zero variance}

\item{glrmFactors}{similar to the number of factors to return from PCA}

\item{Loss}{set to one of "Quadratic", "Absolute", "Huber", "Poisson", "Hinge", "Logistic", "Periodic"}

\item{Gamma_x}{regularization weight on X matrix}

\item{Gamma_y}{regularization weight on y matrix}

\item{glrmMaxIters}{max number of iterations}

\item{SVDMethod}{choose from "Randomized","GramSVD","Power"}

\item{MaxRunTimeSecs}{set the timeout for max run time}

\item{KMeansK}{number of factors to test out in k-means to find the optimal number}
}
\value{
Original data.table with added column with cluster number identifier
}
\description{
GLRM_KMeans_Col adds a column to your original data with a cluster number identifier. Uses glrm (grid tune-able) and then k-means to find optimal k.
}
\author{
Adrian Antico at RemixInstitute.com
}
