% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GridTuning.R
\name{GridTuner}
\alias{GridTuner}
\title{GridTuner}
\usage{
GridTuner(
  AlgoType = "catboost",
  ModelType = "classification",
  TrainOnFull. = TrainOnFull,
  TargetColumnName. = TargetColumnName,
  DebugMode. = DebugMode.,
  HasTime = HasTime,
  task_type. = task_type,
  Trees. = Trees,
  Depth. = Depth,
  LearningRate. = LearningRate,
  L2_Leaf_Reg. = L2_Leaf_Reg,
  BorderCount. = BorderCount,
  RandomStrength. = RandomStrength,
  RSM. = RSM,
  BootStrapType. = BootStrapType,
  GrowPolicy. = GrowPolicy,
  NumGPUs = NumGPUs,
  LossFunction = LossFunction,
  EvalMetric = EvalMetric,
  MetricPeriods = MetricPeriods,
  ClassWeights = ClassWeights,
  CostMatrixWeights = CostMatrixWeights,
  data = data,
  TrainPool. = TrainPool,
  TestPool. = TestPool,
  FinalTestTarget. = FinalTestTarget,
  TestTarget. = TestTarget,
  FinalTestPool. = FinalTestPool,
  TestData. = TestData,
  TestMerge. = TestMerge,
  TargetLevels. = TargetLevels,
  MaxRunsWithoutNewWinner = MaxRunsWithoutNewWinner,
  MaxModelsInGrid = MaxModelsInGrid,
  MaxRunMinutes = MaxRunMinutes,
  BaselineComparison. = BaselineComparison,
  SaveModelObjects = SaveModelObjects,
  metadata_path = metadata_path,
  model_path = model_path,
  ModelID = ModelID,
  grid_eval_metric. = grid_eval_metric
)
}
\arguments{
\item{AlgoType}{"catboost"}

\item{ModelType}{"classification"}

\item{TrainOnFull.}{TrainOnFull}

\item{TargetColumnName.}{TargetColumnName}

\item{DebugMode.}{DebugMode}

\item{task_type.}{task_type}

\item{Trees.}{Trees}

\item{Depth.}{Depth}

\item{LearningRate.}{LearningRate}

\item{L2_Leaf_Reg.}{L2_Leaf_Reg}

\item{BorderCount.}{BorderCount}

\item{RandomStrength.}{RandomStrength}

\item{RSM.}{RSM}

\item{BootStrapType.}{BootStrapType}

\item{GrowPolicy.}{GrowPolicy}

\item{NumGPUs}{NumGPUs}

\item{LossFunction}{LossFunction}

\item{EvalMetric}{EvalMetric}

\item{MetricPeriods}{MetricPeriods}

\item{ClassWeights}{ClassWeights}

\item{CostMatrixWeights}{CostMatrixWeights}

\item{data}{data}

\item{TrainPool.}{TrainPool}

\item{TestPool.}{TestPool}

\item{TestTarget.}{TestTarget}

\item{FinalTestPool.}{FinalTestPool}

\item{TestData.}{TestData}

\item{TestMerge.}{TestMerge}

\item{TargetLevels.}{TargetLevels}

\item{MaxRunsWithoutNewWinner}{MaxRunsWithoutNewWinner}

\item{MaxModelsInGrid}{MaxModelsInGrid}

\item{MaxRunMinutes}{MaxRunMinutes}

\item{BaselineComparison.}{BaselineComparison}

\item{SaveModelObjects}{SaveModelObjects}

\item{metadata_path}{metadata_path}

\item{model_path}{model_path}

\item{ModelID}{ModelID}

\item{grid_eval_metric.}{MultiClass}
}
\description{
Execute grid tuning for catboost or xgboost
}
\seealso{
Other Reinforcement Learning: 
\code{\link{RL_Initialize}()},
\code{\link{RL_ML_Update}()},
\code{\link{RL_Update}()},
\code{\link{RPM_Binomial_Bandit}()}
}
\author{
Adrian Antico
}
\concept{Reinforcement Learning}
