% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Adrian_Antico_Modeling_Tools_Library.R
\name{H20MultinomialAUC}
\alias{H20MultinomialAUC}
\title{H20MultinomialAUC computes the micro auc from a multinomial model}
\usage{
H20MultinomialAUC(validate, best_model, targetColNum = 1,
  targetName = "bla")
}
\arguments{
\item{validate}{the data set to run the micro auc on}

\item{best_model}{the model object you wish to test}
}
\value{
Micro AUC
}
\description{
H20MultinomialAUC computes the micro auc from a multinomial model
}
\examples{
auc_val <- H20MultinomialAUC(validate, best_model)
}
\author{
Adrian Antico
}
