% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/H2OIsolationForestScoring.R
\name{H2OIsolationForestScoring}
\alias{H2OIsolationForestScoring}
\title{H2OIsolationForestScoring}
\usage{
H2OIsolationForestScoring(
  data,
  Features = NULL,
  IDcols = NULL,
  H2OStart = TRUE,
  H2OShutdown = TRUE,
  ModelID = "TestModel",
  SavePath = NULL,
  Threshold = 0.975,
  MaxMem = "28G",
  NThreads = -1,
  Debug = FALSE
)
}
\arguments{
\item{data}{The data.table with the columns you wish to have analyzed}

\item{Features}{A character vector with the column names to utilize in the isolation forest}

\item{IDcols}{A character vector with the column names to not utilize in the isolation forest but have returned with the data output. Otherwise those columns will be removed}

\item{H2OStart}{TRUE to have H2O started inside function}

\item{H2OShutdown}{TRUE to shutdown H2O inside function}

\item{ModelID}{Name for model that gets saved to file if SavePath is supplied and valid}

\item{SavePath}{Path directory to store saved model}

\item{Threshold}{Quantile value to find the cutoff value for classifying outliers}

\item{MaxMem}{Specify the amount of memory to allocate to H2O. E.g. "28G"}

\item{NThreads}{Specify the number of threads (E.g. cores * 2)}

\item{Debug}{Debugging}
}
\value{
Source data.table with predictions. Note that any columns not listed in Features nor IDcols will not be returned with data. If you want columns returned but not modeled, supply them as IDcols
}
\description{
H2OIsolationForestScoring for dimensionality reduction and / or anomaly detection scoring on new data
}
\examples{
\dontrun{
# Create simulated data
data <- RemixAutoML::FakeDataGenerator(
  Correlation = 0.70,
  N = 50000,
  ID = 2L,
  FactorCount = 2L,
  AddDate = TRUE,
  ZIP = 0L,
  TimeSeries = FALSE,
  ChainLadderData = FALSE,
  Classification = FALSE,
  MultiClass = FALSE)

# Run algo
data <- RemixAutoML::H2OIsolationForest(
  data,
  Features = names(data)[2L:ncol(data)],
  IDcols = c("Adrian", "IDcol_1", "IDcol_2"),
  ModelID = "Adrian",
  SavePath = getwd(),
  Threshold = 0.95,
  MaxMem = "28G",
  NThreads = -1,
  NTrees = 100,
  SampleRate = (sqrt(5)-1)/2,
  MaxDepth = 8,
  MinRows = 1,
  ColSampleRate = 1,
  ColSampleRatePerLevel = 1,
  ColSampleRatePerTree = 1,
  CategoricalEncoding = c("AUTO"),
  Debug = TRUE)

# Remove output from data and then score
data[, eval(names(data)[17:ncol(data)]) := NULL]

# Run algo
Outliers <- RemixAutoML::H2OIsolationForestScoring(
  data,
  Features = names(data)[2:ncol(data)],
  IDcols = c("Adrian", "IDcol_1", "IDcol_2"),
  H2OStart = TRUE,
  H2OShutdown = TRUE,
  ModelID = "TestModel",
  SavePath = getwd(),
  Threshold = 0.95,
  MaxMem = "28G",
  NThreads = -1,
  Debug = FALSE)
}
}
\seealso{
Other Unsupervised Learning: 
\code{\link{AutoKMeans}()},
\code{\link{GenTSAnomVars}()},
\code{\link{H2OIsolationForest}()},
\code{\link{ResidualOutliers}()}
}
\author{
Adrian Antico
}
\concept{Unsupervised Learning}
