% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotFunctions.R
\name{HeatMapPlot}
\alias{HeatMapPlot}
\title{HeatMapPlot}
\usage{
HeatMapPlot(
  dt,
  x = NULL,
  y = NULL,
  z = NULL,
  AggMethod = "mean",
  PercentileBuckets_X = 0.1,
  PercentileBuckets_Y = 0.1,
  NLevels_X = 33,
  NLevels_Y = 33
)
}
\arguments{
\item{dt}{Source data.table}

\item{x}{X-Axis variable}

\item{y}{Y-Axis variable}

\item{z}{Z-Axis variable}

\item{AggMethod}{'mean', 'median', 'sum', 'sd', 'count'}

\item{PercentileBuckets_X}{= 0.10}

\item{PercentileBuckets_Y}{= 0.10}

\item{NLevels_X}{= 20}

\item{NLevels_Y}{= 20}

\item{RankLevels_X}{= 'mean'}
}
\description{
Create heat maps with numeric or categorical dt
}
