% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Adrian_Antico_Modeling_Tools_Library.R
\name{HolidayCounts}
\alias{HolidayCounts}
\title{HolidayCounts is a vectorized function to count up the number of holidays in a range of dates}
\usage{
HolidayCounts(a, b)
}
\arguments{
\item{a}{The start date}

\item{b}{The end date}
}
\value{
The counts either as a scalar or a column in your table
}
\description{
HolidayCounts is a vectorized function to count up the number of holidays in a range of dates
}
\examples{
data[, holidays := HolidayCounts(Date1, Date2)]
}
\author{
Adrian Antico
}
