% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ContinuousTimeDataGenerator.R
\name{ID_BuildTrainDataSets}
\alias{ID_BuildTrainDataSets}
\title{ID_BuildTrainDataSets for assembling data}
\usage{
ID_BuildTrainDataSets(
  MetaData,
  data,
  Case = 2L,
  TargetVariableName = NULL,
  DateVariableName = NULL,
  GroupingVariables = NULL,
  FC_Periods,
  TimeUnit = "week",
  PowerRate = 0.5,
  SampleRate = 5L,
  TargetWindowSamples = 5L
)
}
\arguments{
\item{MetaData}{This is the metadata returned from the ID_MetadataGenerator() function}

\item{data}{This is your transactional data}

\item{Case}{Indicate which data constructor method to use}

\item{TargetVariableName}{Your target variable names}

\item{DateVariableName}{Your date variable names}

\item{GroupingVariables}{Your grouping variables}

\item{FC_Periods}{The number of periods to forecast}

\item{TimeUnit}{The time period unit, such as "day", "week", or "month"}

\item{PowerRate}{The calculated for determining the total samples is number of records to the power of PowerRate. Then that values is multiplied by the SampleRate. This ensures that a more representative sample is generated across the data set.}

\item{SampleRate}{The value used to sample from each level of the grouping variables}

\item{TargetWindowSamples}{The number of different targets to utilize for a single random start date}
}
\value{
Returns the count modeling data and the size modeling data
}
\description{
ID_BuildTrainDataSets for assembling data for the IntermittentDemandBootStrapper() function.
}
\seealso{
Other Feature Engineering Helper: 
\code{\link{AutoFourierFeatures}()},
\code{\link{ID_MetadataGenerator}()},
\code{\link{ID_TrainingDataGenerator2}()},
\code{\link{ID_TrainingDataGenerator}()}
}
\concept{Feature Engineering Helper}
