% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoFreqSizeDistSimFunctions.R
\name{ID_Forecast}
\alias{ID_Forecast}
\title{ID_Forecast for forecasting intermittent demand data}
\usage{
ID_Forecast(CountData = FinalData$CountData,
  SizeData = FinalData$SizeData,
  CountDataNames = FinalData$CountPredNames,
  SizeDataNames = FinalData$SizePredNames, GroupVar = NULL,
  FC_Periods = 26, NumSims = 1000, PredictionIntervals = c(0.05, 0.2,
  0.8, 0.95))
}
\arguments{
\item{CountData}{This is the count data returned from AutoCatBoostFreqSizeScoring() or AutoH2oGBMFreqSizeScoring()}

\item{SizeData}{This is the size data returned from AutoCatBoostFreqSizeScoring() or AutoH2oGBMFreqSizeScoring()}

\item{CountDataNames}{This is the count data names returned from AutoCatBoostFreqSizeScoring() or AutoH2oGBMFreqSizeScoring()}

\item{SizeDataNames}{This is the size data returned from AutoCatBoostFreqSizeScoring() or AutoH2oGBMFreqSizeScoring()}

\item{GroupVar}{This is your grouping variable. E.g. sku as being the name of the column that contains all skus}

\item{FC_Periods}{The max period of your forecast. E.g. if you want 52 weeks of forecasts, set to 52}

\item{NumSims}{Set the number of collapsed gibbs simulations to run for each time unit forecast. E.g. if you want 52 weeks of forecasts, the simulations will run NumSims for each period up to 52.}

\item{PredictionIntervals}{Set the prediction intervals you want returned. E.g. c(seq(0.05,0.95,0.05))}
}
\value{
Returns your entire set of groupvar forecasts in a single data.table
}
\description{
ID_Forecast for forecasting intermittent demand data. It runs collapsed gibbs sampler simulations, using the count and size quantile regression models, for every period for all group levels in your data.
}
\examples{
\donttest{
Results <- ID_Forecast(
  CountData, 
  SizeData, 
  GroupVar = "sku", 
  FC_Periods = 26, 
  NumSims = 1000, 
  PredictionIntervals = c(0.05,0.20,0.50,0.80,0.95))
}
}
\seealso{
Other Automated Time Series: \code{\link{AutoCatBoostCARMA}},
  \code{\link{AutoCatBoostFreqSizeScoring}},
  \code{\link{AutoCatBoostSizeFreqDist}},
  \code{\link{AutoH2oDRFCARMA}},
  \code{\link{AutoH2oGBMCARMA}},
  \code{\link{AutoH2oGBMFreqSizeScoring}},
  \code{\link{AutoH2oGBMSizeFreqDist}},
  \code{\link{AutoTS}}, \code{\link{AutoXGBoostCARMA}},
  \code{\link{ID_SingleLevelGibbsSampler}},
  \code{\link{IntermittentDemandScoringDataGenerator}}
}
\author{
Adrian Antico
}
\concept{Automated Time Series}
