% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ContinuousTimeDataGenerator.R
\name{ID_MetadataGenerator}
\alias{ID_MetadataGenerator}
\title{ID_MetadataGenerator}
\usage{
ID_MetadataGenerator(
  data,
  RestrictDateRange = TRUE,
  DateVariableName = NULL,
  GroupingVariables = NULL,
  MinTimeWindow = 1L,
  MinTxnRecords = 2L,
  DateInterval = "day"
)
}
\arguments{
\item{data}{This is your transactional level data}

\item{RestrictDateRange}{= TRUE}

\item{DateVariableName}{Bla}

\item{GroupingVariables}{Bla}

\item{MinTimeWindow}{The number of time periods you would like to omit for training. Default is 1 so that at a minimum, there is at least one period of values to forecast. You can set it up to a larger value if you do not want more possible target windows for the lower target window values.}

\item{MinTxnRecords}{I typically set this to 2 so that there is at least one other instance of demand so that the forecasted values are not complete nonsense.}

\item{DateInterval}{This is the time unit for determining date calculations}
}
\value{
Returns a data.table with summary information for the IntermittentDemandBootStrapper() function.
}
\description{
ID_MetadataGenerator for summary metadata for transactional data. The data returned from this function feeds into the IntermittentDemandBootStrapper() function.
}
\examples{
\dontrun{
# Generate Metadata----
MetaData <- ID_MetadataGenerator(
  data = data,
  RestrictDateRange = TRUE,
  DateVariableName = DateVariableName,
  GroupingVariables = GroupingVariables,
  MinTimeWindow = MinTimeWindow,
  MinTxnRecords = MinTxnRecords,
  DateInterval = TimeUnit,
  TimeUnit = TimeUnit
)
}
}
\seealso{
Other Feature Engineering Helper: 
\code{\link{AutoFourierFeatures}()},
\code{\link{ID_BuildTrainDataSets}()},
\code{\link{ID_TrainingDataGenerator2}()},
\code{\link{ID_TrainingDataGenerator}()}
}
\author{
Adrian Antico
}
\concept{Feature Engineering Helper}
