% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoFreqSizeDistSimFunctions.R
\name{ID_SingleLevelGibbsSampler}
\alias{ID_SingleLevelGibbsSampler}
\title{ID_SingleLevelGibbsSampler for collapsed gibbs sampler}
\usage{
ID_SingleLevelGibbsSampler(CountDataLevel, SizeDataLevel,
  FC_Periods = NULL, nSims = 5000, CountList = NULL,
  SizeList = NULL, CountVectorSize = length(CountList),
  SizeVectorSize = length(SizeVectorList))
}
\arguments{
\item{CountDataLevel}{Single record of predicted values from Count quantile regressions in a numeric vector form}

\item{SizeDataLevel}{Single record of predicted values from Size quantile regressions in a numeric vector form}

\item{FC_Periods}{The number of periods you set up to forecast}

\item{nSims}{The number of simulations you want to have run for each period in 1 to FC_Periods}

\item{CountList}{A numeric vector of the quantiles used for Count modeling}

\item{SizeList}{A numeric vector of the quantiles used for Size modeling}

\item{CountVectorSize}{length of CountList}

\item{SizeVectorSize}{length of SizeList}
}
\value{
The posterior predicted distribution of simulated values
}
\description{
ID_SingleLevelGibbsSampler for collapsed gibbs sampler from quantile regressions
}
\seealso{
Other Automated Time Series: \code{\link{AutoCatBoostCARMA}},
  \code{\link{AutoCatBoostFreqSizeScoring}},
  \code{\link{AutoCatBoostSizeFreqDist}},
  \code{\link{AutoH2oDRFCARMA}},
  \code{\link{AutoH2oGBMCARMA}},
  \code{\link{AutoH2oGBMFreqSizeScoring}},
  \code{\link{AutoH2oGBMSizeFreqDist}},
  \code{\link{AutoTS}}, \code{\link{AutoXGBoostCARMA}},
  \code{\link{ID_Forecast}},
  \code{\link{IntermittentDemandScoringDataGenerator}}
}
\author{
Adrian Antico
}
\concept{Automated Time Series}
