% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ContinuousTimeDataGenerator.R
\name{ID_TrainingDataGenerator}
\alias{ID_TrainingDataGenerator}
\title{ID_TrainingDataGenerator for subsetting data}
\usage{
ID_TrainingDataGenerator(
  data,
  Type = "timetoevent1",
  TargetVariableName = NULL,
  Level = NULL,
  DateVariableName = NULL,
  GroupingVariables = NULL,
  RandomStartDate = NULL,
  TimeUnit = NULL,
  TargetWindow = NULL
)
}
\arguments{
\item{data}{Source data}

\item{Type}{"timetoevent1", "eventinwindow1"}

\item{TargetVariableName}{Name of the variables to run feature engineering on. List the actual target variable name first.}

\item{Level}{The individual level of your group variable}

\item{DateVariableName}{Name of your date variable}

\item{GroupingVariables}{Your grouping variables}

\item{RandomStartDate}{The date to partition the data}

\item{TimeUnit}{This is the TimeUnit you selected for aggregation}

\item{TargetWindow}{The length of the target window sampled}
}
\value{
Returns two data sets for the IntermittentDemandBootStrapper() function based on a single level from the grouping variables.
}
\description{
ID_TrainingDataGenerator for subsetting data for the IntermittentDemandBootStrapper() function.
}
\seealso{
Other Feature Engineering Helper: 
\code{\link{AutoFourierFeatures}()},
\code{\link{ID_BuildTrainDataSets}()},
\code{\link{ID_MetadataGenerator}()},
\code{\link{ID_TrainingDataGenerator2}()}
}
\concept{Feature Engineering Helper}
