% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ContinuousTimeDataGenerator.R
\name{ID_TrainingDataGenerator2}
\alias{ID_TrainingDataGenerator2}
\title{ID_TrainingDataGenerator2 for subsetting data}
\usage{
ID_TrainingDataGenerator2(
  data,
  TargetVariableName = NULL,
  Level = NULL,
  GroupingVariables = NULL,
  DateVariableName = NULL,
  RandomStartDate = NULL,
  TimeUnit = NULL,
  TargetWindow = NULL
)
}
\arguments{
\item{data}{Source data}

\item{TargetVariableName}{vector of variable names}

\item{Level}{The individual level of your group variable}

\item{GroupingVariables}{Your grouping variables}

\item{DateVariableName}{Name of your date variable}

\item{RandomStartDate}{The date to partition the data}

\item{TimeUnit}{This is the TimeUnit you selected for aggregation}

\item{TargetWindow}{The length of the target window sampled}
}
\value{
Returns two data sets for the IntermittentDemandBootStrapper() function based on a single level from the grouping variables.
}
\description{
ID_TrainingDataGenerator2 for subsetting data for the IntermittentDemandBootStrapper() function.
}
\seealso{
Other Feature Engineering Helper: 
\code{\link{AutoFourierFeatures}()},
\code{\link{ID_BuildTrainDataSets}()},
\code{\link{ID_MetadataGenerator}()},
\code{\link{ID_TrainingDataGenerator}()}
}
\concept{Feature Engineering Helper}
