% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Algorithms.R
\name{InsertSortedValue}
\alias{InsertSortedValue}
\title{InsertSortedValue}
\usage{
InsertSortedValue(Vec, Val, order = "left")
}
\arguments{
\item{Val}{value to insert}

\item{order}{'left' or 'right', insert value location}

\item{vec}{numeric vector}
}
\description{
Update a sorted vector with a new value that preserves sort order

Algorithms
}
\examples{
\dontrun{
RemixAutoML::InsertSortedValue(vec = seq(2, 2000, 2), Val = 741, order = "left")
}

}
\author{
Adrian Antico
}
