% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoFreqSizeDistScoringFunctions.R
\name{IntermittentDemandScoringDataGenerator}
\alias{IntermittentDemandScoringDataGenerator}
\title{IntermittentDemandScoringDataGenerator creates the scoring data for forecasting}
\usage{
IntermittentDemandScoringDataGenerator(data = NULL, FC_Periods = 52,
  SaveData = FALSE, FilePath = NULL, TargetVariableName = "qty",
  DateVariableName = "date", GroupingVariables = "sku", Lags = 1:7,
  MovingAverages = seq(7, 28, 7), TimeTrendVariable = TRUE,
  TimeUnit = "day", CurrentDate = NULL, CalendarVariables = c("wday",
  "mday", "yday", "week", "isoweek", "month", "quarter", "year"),
  HolidayGroups = "USPublicHolidays")
}
\arguments{
\item{data}{This is your source data}

\item{FC_Periods}{The number of periods you set up to forecast}

\item{SaveData}{Set to TRUE to save the output data to file}

\item{FilePath}{Set a path file have the data saved there}

\item{TargetVariableName}{Name or column number of your target variable}

\item{DateVariableName}{Name or column number of your date variable}

\item{GroupingVariables}{Name or column number of your group variables}

\item{Lags}{The number of lags used in building the modeling data sets}

\item{MovingAverages}{The number of moving averages used in building the modeling data sets}

\item{TimeTrendVariable}{Set to TRUE if you did so in model data creation}

\item{TimeUnit}{Set to the same time unit used in modeling data creation}

\item{CurrentDate}{Set this to the current date or a date that you want. It is user specified in case you want to score historical data.}

\item{CalendarVariables}{Set this to the same setting you used in modeling data creation}

\item{HolidayGroups}{Set this to the same setting you used in modeling data creation}
}
\value{
Returns the most recent records for every level of your grouping variables with all the feature used in model building
}
\description{
IntermittentDemandScoringDataGenerator creates the scoring data for forecasting. It will recreate the same features used for modeling, take the most recent record, and then duplicate those records for each forecast period specifed.
}
\examples{
\donttest{
 ScoringData <- IntermittentDemandScoringDataGenerator(
   data = data,
   SaveData = FALSE,
   FilePath = NULL,
   TargetVariableName = "qty",
   DateVariableName = "date",
   GroupingVariables = "sku",
   Lags = 1:7,
   MovingAverages = seq(7,28,7),
   TimeTrendVariable = TRUE,
   TimeUnit = "day",
   CurrentDate = NULL,
   CalendarVariables = c("wday",
                         "mday",
                         "yday",
                         "week",
                         "isoweek",
                         "month",
                         "quarter",
                         "year"),
   HolidayGroups = "USPublicHolidays")
}
}
\seealso{
Other Automated Time Series: \code{\link{AutoCatBoostCARMA}},
  \code{\link{AutoCatBoostFreqSizeScoring}},
  \code{\link{AutoCatBoostSizeFreqDist}},
  \code{\link{AutoH2oDRFCARMA}},
  \code{\link{AutoH2oGBMCARMA}},
  \code{\link{AutoH2oGBMFreqSizeScoring}},
  \code{\link{AutoH2oGBMSizeFreqDist}},
  \code{\link{AutoTS}}, \code{\link{AutoXGBoostCARMA}},
  \code{\link{ID_Forecast}},
  \code{\link{ID_SingleLevelGibbsSampler}}
}
\author{
Adrian Antico
}
\concept{Automated Time Series}
