% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyFunctions.R
\name{IntraSessionDefaults}
\alias{IntraSessionDefaults}
\title{IntraSessionDefaults}
\usage{
IntraSessionDefaults(
  List = NULL,
  InputName = "Plot1_SelectData",
  ArgName = "SelectedDefault",
  Default = names(DataList)[[1L]],
  Debug = Debug
)
}
\arguments{
\item{List}{named list of args where the args can be many}

\item{InputName}{'Plot1_SelectData' in input$Plot1_SelectData}

\item{ArgName}{'SelectedDefault' in List[[InputName]][[Default]]}

\item{Default}{default value to assign}

\item{Debug}{Logical}
}
\description{
IntraSessionDefaults
}
\keyword{internal}
