% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RemixAutoML.R
\name{InvApply_Asinh}
\alias{InvApply_Asinh}
\title{Inverse Asinh Transformation}
\usage{
InvApply_Asinh(x)
}
\arguments{
\item{x}{The data in numerical vector form}
}
\value{
Asinh results
}
\description{
Inverse Asinh Transformation
}
\seealso{
Other Feature Engineering: \code{\link{Apply_Asinh}},
  \code{\link{Apply_Asin}}, \code{\link{Apply_BoxCox}},
  \code{\link{Apply_Logit}},
  \code{\link{Apply_YeoJohnson}},
  \code{\link{AutoDataPartition}},
  \code{\link{AutoTransformationCreate}},
  \code{\link{AutoTransformationScore}},
  \code{\link{AutoWord2VecModeler}},
  \code{\link{CreateCalendarVariables}},
  \code{\link{DT_GDL_Feature_Engineering}},
  \code{\link{DummifyDT}},
  \code{\link{FAST_GDL_Feature_Engineering}},
  \code{\link{GDL_Feature_Engineering}},
  \code{\link{InvApply_Asin}},
  \code{\link{InvApply_BoxCox}},
  \code{\link{InvApply_Logit}},
  \code{\link{InvApply_YeoJohnson}},
  \code{\link{ModelDataPrep}},
  \code{\link{Scoring_GDL_Feature_Engineering}}
}
\author{
Adrian Antico
}
\concept{Feature Engineering}
