% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_export_helpers.R
\name{is_htmlwidget_object}
\alias{is_htmlwidget_object}
\title{Check Whether an Object Is an HTML Widget}
\usage{
is_htmlwidget_object(object)
}
\arguments{
\item{object}{Any R object.}
}
\value{
Logical scalar.
}
\description{
Internal helper used by the object export functions.
}
\keyword{internal}
