% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NLP.R
\name{Keras}
\alias{Keras}
\title{Keras}
\usage{
Keras(LayerFunc, PriorOutput = NULL, Name = NULL, Debug = FALSE, ...)
}
\arguments{
\item{LayerFunc}{Required, Positional. E.g. keras::layer_input without quotes}

\item{PriorOutput}{The resulting list contained in Output <- RemixAutoML::Keras(). If you pass this LayerList and ArgsList will be ignored. Same Output from ArgsList and LayerList.}

\item{Name}{Name for the LayerList and ArgsList elements}

\item{Debug}{FALSE}

\item{...}{Passthrough args for keras layer_input}
}
\description{
Keras functions simplified for tracking args
}
\examples{
\dontrun{
Output <- RemixAutoML::KerasInputs(keras::flatten, PriorOutput = Output, Name = 'flatten1', object = Output$LayerList$PreviousLayer)
}


}
\seealso{
Other DL: 
\code{\link{KerasTrainEval}()},
\code{\link{Keras_BinaryScoreEval}()}
}
\author{
Adrian Antico
}
\concept{DL}
