% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NLP.R
\name{Keras_BinaryScoreEval}
\alias{Keras_BinaryScoreEval}
\title{Keras_BinaryScoreEval}
\usage{
Keras_BinaryScoreEval(
  CompiledModel = NULL,
  YTest = NULL,
  Features = NULL,
  CostMatrixWeights = c(0, 1, 1, 0),
  ConfMatThresh = 0.5,
  ReturnMetrics = TRUE,
  ReturnFeatures = TRUE
)
}
\arguments{
\item{CompiledModel}{= model_word2vec,}

\item{YTest}{Vector of 0s and 1s}

\item{Features}{Input features as a matrix}

\item{CostMatrixWeights}{= c(0,1,1,0) TP, FN, FP, TN}

\item{ConfMatThresh}{'max' or 'min'. Tells function whether to grab the threshold associated with the max or min of the chose CutOffMetric}

\item{ReturnMetrics}{= TRUE. FALSE and only the score data will be returned}

\item{ReturnFeatures}{= TRUE. FALSE and only the predicted values will be returned}
}
\description{

}
\seealso{
Other DL: 
\code{\link{KerasTrainEval}()},
\code{\link{Keras}()}
}
\author{
Adrian Antico
}
\concept{DL}
