% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NLP.R
\name{Keras_t2s_pad}
\alias{Keras_t2s_pad}
\title{Keras_text2seq_pad}
\usage{
Keras_t2s_pad(
  RunText2Seq = TRUE,
  RunPadding = FALSE,
  T2S_tokenizer = NULL,
  T2S_text = NULL,
  PadSeq_sequences = NULL,
  PadSeq_padding = "pre",
  PadSeq_maxlen = 150,
  PadSeq_value = 0L,
  PadSeq_truncating = "pre",
  PadSeq_dtype = "int32"
)
}
\arguments{
\item{RunText2Seq}{= TRUE,}

\item{RunPadding}{= FALSE,}

\item{T2S_tokenizer}{= NULL, keras::text_tokenizer keras::fit_text_tokenizer}

\item{T2S_text}{text character vector or a list() of text character vectors. If the list is named, those names will be used for the names in the output list that contain the matching data sets}

\item{PadSeq_sequences}{NULL. Either a result from keras::texts_to_sequences() or a list() of results. If the list is named, those names will be used for the names in the output list that contain the matching data sets}

\item{PadSeq_padding}{= "pre",}

\item{PadSeq_maxlen}{= 150,}

\item{PadSeq_value}{= 0L,}

\item{PadSeq_truncating}{= "pre",}

\item{PadSeq_dtype}{= "int32"}
}
\description{
return padded text to sequence keras object
}
\examples{
\dontrun{

}

}
\seealso{
Other NLP: 
\code{\link{KerasTokenizerFit}()}
}
\author{
Adrian Antico
}
\concept{NLP}
