% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NLP.R
\name{KerasTokenizerFit}
\alias{KerasTokenizerFit}
\title{KerasTokenizerFit}
\usage{
KerasTokenizerFit(
  Fit_Object = NULL,
  Fit_Text = NULL,
  Tok_num_words = NULL,
  Tok_filters = "!\\"#$\%&()*+,-./:;<=>?@[\\\\]^_`{|}~\\t\\n",
  Tok_lower = TRUE,
  Tok_split = " ",
  Tok_char_level = FALSE,
  Tok_oov_token = NULL,
  RunTokenizer = TRUE,
  RunFitTextTokenizer = TRUE
)
}
\arguments{
\item{Fit_Object}{Object to be passed into layer}

\item{Fit_Text}{= x_train_text, character vector of text. e.g. data[['TextColumn']]}

\item{Tok_num_words}{= NULL,}

\item{Tok_filters}{see example below}

\item{Tok_lower}{= TRUE,}

\item{Tok_split}{= " ",}

\item{Tok_char_level}{= FALSE,}

\item{Tok_oov_token}{= NULL,}

\item{RunTokenizer}{= TRUE,}

\item{RunFitTextTokenizer}{= TRUE}
}
\value{
List with two elements: VectorizedText and Tokenizer.
}
\description{
Tokenize and Fit with the one function. Convert column of long text within rows to a long vector of words
}
\examples{
\dontrun{
Tokenize_Object <- RemixAutoML:::KerasTokenizerFit(
  Fit_Text = x_train_text,
  Fit_Object = NULL,
  Tok_num_words = NULL,
  Tok_filters = "!\"#$\%&()*+,-./:;<=>?@[\\\\]^_`{|}~\t\n",
  Tok_lower = TRUE,
  Tok_split = " ",
  Tok_char_level = FALSE,
  Tok_oov_token = NULL,
  RunTokenizer = TRUE,
  RunFitTextTokenizer = TRUE)
}

}
\seealso{
Other NLP: 
\code{\link{Keras_t2s_pad}()}
}
\author{
Adrian Antico
}
\concept{NLP}
