% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LimeFunctions.R
\name{LimeModel}
\alias{LimeModel}
\title{LimeModel to build a lime model}
\usage{
LimeModel(
  data,
  Model = NULL,
  Bins = 10,
  ModelType = "xgboost",
  NThreads = parallel::detectCores(),
  MaxMem = "32G",
  ModelPath = NULL,
  ModelID = NULL
)
}
\arguments{
\item{data}{Supply a training data set. This data set should be the data right before it gets converted to an h2o, catboost, or xgboost data object.}

\item{Model}{Supply the model returned from training with the Auto__() functions.}

\item{Bins}{Number of bins for discretizing numeric features}

\item{ModelType}{Select from xgboost, h2o, and catboost}

\item{NThreads}{Number of CPU threads}

\item{MaxMem}{For use with H2O models. E.g. set to "28G"}

\item{ModelPath}{Set to the path where your ML model is saved}

\item{ModelID}{ID used to identify your ML model}
}
\value{
Model for utilizing lime
}
\description{
LimeModel to build a lime model for prediction explanations in this package#'
}
\seealso{
Other Model Evaluation and Interpretation: 
\code{\link{AutoLimeAid}()},
\code{\link{EvalPlot}()},
\code{\link{ParDepCalPlots}()},
\code{\link{ROCPlot}()},
\code{\link{RedYellowGreen}()},
\code{\link{VI_Plot}()},
\code{\link{threshOptim}()}
}
\author{
Adrian Antico
}
\concept{Model Evaluation and Interpretation}
