% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscFunctions.R
\name{Logger}
\alias{Logger}
\title{Logger}
\usage{
Logger(x)
}
\arguments{
\item{x}{Function to call repeatedly}
}
\description{
Logging errors and warnings from repeated calls to a function
}
\examples{
\dontrun{
Output <- lapply(1:10, FUN = Logger(PrintToPDF))
}

}
\seealso{
Other Misc: 
\code{\link{AutoH2OTextPrepScoring}()},
\code{\link{DeleteFile}()},
\code{\link{PrintToPDF}()},
\code{\link{tokenizeH2O}()}
}
\author{
Adrian Antico
}
\concept{Misc}
