% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelEvaluationPlots.R
\name{ML_EvalPlots}
\alias{ML_EvalPlots}
\title{ML_EvalPlots}
\usage{
ML_EvalPlots(
  ModelType = "classification",
  TrainOnFull. = TrainOnFull,
  LossFunction. = LossFunction,
  EvalMetric. = EvalMetric,
  EvaluationMetrics. = EvaluationMetrics,
  ValidationData. = ValidationData,
  NumOfParDepPlots. = NumOfParDepPlots,
  VariableImportance. = VariableImportance,
  TargetColumnName. = TargetColumnName,
  FeatureColNames. = FeatureColNames,
  SaveModelObjects. = SaveModelObjects,
  ModelID. = ModelID,
  metadata_path. = metadata_path,
  model_path. = model_path,
  predict. = predict
)
}
\arguments{
\item{ModelType}{'classification', 'multiclass', 'regression', or 'vector'}

\item{TrainOnFull.}{Passthrough}

\item{LossFunction.}{Passthrough regression}

\item{EvalMetric.}{Passthrough regression}

\item{EvaluationMetrics.}{Passthrough regression}

\item{ValidationData.}{Passthrough}

\item{NumOfParDepPlots.}{Passthrough}

\item{VariableImportance.}{Passthrough}

\item{TargetColumnName.}{Passthrough}

\item{FeatureColNames.}{Passthrough}

\item{SaveModelObjects.}{Passthrough}

\item{ModelID.}{Passthrough}

\item{metadata_path.}{Passthrough}

\item{model_path.}{Passthrough}
}
\description{
Generate evaluation plots
}
\seealso{
Other Model Evaluation and Interpretation: 
\code{\link{AutoLimeAid}()},
\code{\link{EvalPlot}()},
\code{\link{LimeModel}()},
\code{\link{ParDepCalPlots}()},
\code{\link{ROCPlot}()},
\code{\link{RedYellowGreen}()},
\code{\link{VI_Plot}()},
\code{\link{threshOptim}()}
}
\author{
Adrian
}
\concept{Model Evaluation and Interpretation}
