% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyFunctions.R
\name{ModelDataObjects}
\alias{ModelDataObjects}
\title{ModelDataObjects}
\usage{
ModelDataObjects(ModelOutput, TT = "catboost")
}
\arguments{
\item{ModelOutput}{Output from RemixAutoML:: supervised learning functions}
}
\description{
ModelDataObjects
}
\keyword{internal}
