% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelInsights.R
\name{ModelInsightsReport}
\alias{ModelInsightsReport}
\title{ModelInsightsReport}
\usage{
ModelInsightsReport(
  TrainDataInclude = FALSE,
  FeatureColumnNames = NULL,
  SampleSize = 1e+05,
  ModelObject = NULL,
  ModelID = "ModelTest",
  SourcePath = NULL,
  OutputPath = NULL,
  KeepOutput = NULL,
  GlobalVars = ls()
)
}
\arguments{
\item{TrainDataInclude}{Default FALSE. If FALSE, no derived visuals or metrics for TrainData}

\item{FeatureColumnNames}{NULL. Feature column names as character vector.}

\item{SampleSize}{Default 100000}

\item{ModelObject}{Returned output from regression, classificaiton, and multiclass Remix Auto_() models. Currenly supports CatBoost, XGBoost, and LightGBM models}

\item{ModelID}{ModelID used in the AutoQuant supervised learning function}

\item{SourcePath}{Path to directory with AutoQuant Model Output}

\item{OutputPath}{Path to directory where the html will be saved}

\item{KeepOutput}{NULL A list of output names to select. Pass in as a character vector. E.g. c('Test_VariableImportance', 'Train_VariableImportance')}

\item{GlobalVars}{ls() don't use}
}
\description{
ModelInsightsReport is an Rmarkdown report for viewing the model insights generated by AutoQuant supervised learning functions
}
\author{
Adrian Antico
}
\concept{Reports}
