% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelMetrics.R
\name{MultiClassMetrics}
\alias{MultiClassMetrics}
\title{MultiClassMetrics}
\usage{
MultiClassMetrics(
  SaveModelObjects. = SaveModelObjects,
  ValidationData. = ValidationData,
  PredictData. = predict,
  TrainOnFull. = TrainOnFull,
  TargetColumnName. = TargetColumnName,
  TargetLevels. = TargetLevels,
  ModelID. = ModelID,
  model_path. = model_path,
  metadata_path. = metadata_path
)
}
\arguments{
\item{SaveModelObjects.}{= SaveModelObjects}

\item{ValidationData.}{= ValidationData}

\item{PredictData.}{= predict}

\item{TrainOnFull.}{= TrainOnFull}

\item{TargetColumnName.}{= TargetColumnName}

\item{TargetLevels.}{= TargetLevels}

\item{ModelID.}{= ModelID}

\item{model_path.}{= model_path}

\item{metadata_path.}{= metadata_path}
}
\description{
Compute regression metrics and save them to file
}
\seealso{
Other Model Evaluation: 
\code{\link{BinaryMetrics}()},
\code{\link{ClassificationMetrics}()},
\code{\link{DT_BinaryConfusionMatrix}()},
\code{\link{RegressionMetrics}()},
\code{\link{RemixClassificationMetrics}()}
}
\author{
Adrian Antico
}
\concept{Model Evaluation}
