% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Adrian_Antico_Modeling_Tools_Library.R
\name{multiplot}
\alias{multiplot}
\title{Multiplot is a function for combining multiple plots}
\usage{
multiplot(..., plotlist = NULL, file, cols = 1, layout = NULL)
}
\arguments{
\item{plotlist}{This is the list of your charts}

\item{cols}{This is the number of columns in your multiplot}

\item{rows}{This is the number of rows in your multiplot}
}
\value{
An object to pass along to ggplot objects following the "+" sign
}
\description{
Sick of copying this one into your code? Well, not anymore.
}
\examples{
p1 <- ggplot(data1, aes(x = weeks, y = quantity)) + geom_line()
p1 <- p1 + ChartTheme(Size = 12)

p2 <- ggplot(data2, aes(x = weeks, y = quantity)) + geom_line()
p2 <- p2 + ChartTheme(Size = 12)

p3 <- ggplot(data3, aes(x = weeks, y = quantity)) + geom_line()
p3 <- p3 + ChartTheme(Size = 12)

p4 <- ggplot(data4, aes(x = weeks, y = quantity)) + geom_line()
p4 <- p4 + ChartTheme(Size = 12)

multiplot(plotlist=list(p1,p2,p3,p4), cols = 2)
}
\author{
Adrian Antico
}
