% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Adrian_Antico_Modeling_Tools_Library.R
\name{nlsModelFit}
\alias{nlsModelFit}
\title{nlsModelFit is a function for automatically building nls models}
\usage{
nlsModelFit(data, y, x, monotonic = TRUE)
}
\arguments{
\item{data}{Data is the data table you are building the modeling on}

\item{y}{Y is the target variable name in quotes}

\item{x}{X is the independent variable name in quotes}

\item{monotonic}{This is a TRUE/FALSE indicator - choose TRUE if you want monotonic regression over polynomial regression}
}
\value{
A data table with your original column replaced by the nls model predictions
}
\description{
This function will build models for 9 different nls models, along with a non-parametric monotonic regression and a polynomial regression. The models are evaluated, a winner is picked, and the predicted values are stored in your data table.
}
\examples{
data <- data[, value := nlsModelFit(y = "Target", x = "Variable")]
}
\author{
Adrian Antico at RemixInstitute.com
}
