% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyFunctions.R
\name{NumericInput}
\alias{NumericInput}
\title{NumericInput}
\usage{
NumericInput(
  InputID = "TS_CARMA_HolidayMovingAverages",
  Label = "Select Holiday Count MA's",
  Step = 10,
  Value = 1,
  Min = 1,
  Max = 10
)
}
\arguments{
\item{InputID}{Feeds ProjectList and inputId. Argument saved in ProjectList}

\item{Label}{Feeds label}

\item{Step}{Feeds size in the options list}

\item{Value}{Default}

\item{Min}{Min value}

\item{Max}{Max value}
}
\value{
PickerInput object for server.R to go into renderUI({PickerInput()})
}
\description{
NumericInput automatically builds a numeric input with tryCatch's and ProjectList argument usage if it exists
}
\examples{
\dontrun{
output$TS_CARMA_HolidayMovingAverages <- renderUI({
  RemixAutoML::NumericInput(InputID = "TS_CARMA_HolidayMovingAverages",
                            Label = "Select Holiday Count MA's",
                            Min = 1,
                            Max = 1,
                            Value = 1,
                            Step = 1)})
}
}
\seealso{
Other Shiny: 
\code{\link{ArgNullCheck2}()},
\code{\link{ArgNullCheck}()},
\code{\link{BlankRow}()},
\code{\link{DateInput}()},
\code{\link{GenerateEvaluationMetrics}()},
\code{\link{PickerInput_GetLevels}()},
\code{\link{PickerInput}()},
\code{\link{PreparePlotData}()},
\code{\link{ReactiveLoadCSV}()},
\code{\link{ReturnParam}()},
\code{\link{StoreArgs}()},
\code{\link{TextInput}()},
\code{\link{observeEventLoad}()}
}
\author{
Adrian Antico
}
\concept{Shiny}
