% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyFunctions.R
\name{observeEventLoad}
\alias{observeEventLoad}
\title{observeEventLoad}
\usage{
observeEventLoad(input, InputVal = NULL, ObjectName = NULL)
}
\arguments{
\item{input}{Passthrough}

\item{InputVal}{The values that goes after input$}

\item{ObjectName}{The name of the object to assign the load output to}
}
\description{
Used to load .Rdata files in Shiny and assign the object a name
}
\seealso{
Other Shiny: 
\code{\link{ArgNullCheck2}()},
\code{\link{ArgNullCheck}()},
\code{\link{BlankRow}()},
\code{\link{DateInput}()},
\code{\link{GenerateEvaluationMetrics}()},
\code{\link{NumericInput}()},
\code{\link{PickerInput_GetLevels2}()},
\code{\link{PickerInput_GetLevels}()},
\code{\link{PickerInput}()},
\code{\link{PreparePlotData}()},
\code{\link{ReactiveLoadCSV}()},
\code{\link{ReturnParam}()},
\code{\link{StoreArgs}()},
\code{\link{TextInput}()}
}
\author{
Adrian Antico
}
\concept{Shiny}
