% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EconometricsFunctions.R
\name{OptimizeNNET}
\alias{OptimizeNNET}
\title{OptimizeNNET is a function that takes raw data and returns time series data}
\usage{
OptimizeNNET(
  Output,
  MetricSelection = "MAE",
  DataSetName = NULL,
  train = NULL,
  test = NULL,
  FullData = NULL,
  HoldOutPeriods = NULL,
  MinVal = NULL,
  TargetName = NULL,
  DateName = NULL,
  Lags = NULL,
  SeasonalLags = NULL,
  MaxFourierTerms = NULL,
  TrainValidateShare = NULL,
  MaxRunsWithoutNewWinner = 20,
  MaxNumberModels = NULL,
  MaxRunMinutes = NULL,
  FinalGrid = NULL
)
}
\arguments{
\item{Output}{This is passed through as output from TimeSeriesDataPrepare() and passed through ParallelArima()}

\item{MetricSelection}{Select from "MSE", "MAE", or "MAPE"}

\item{DataSetName}{This is the name of the data set passed through in parallel loop}

\item{train}{Training data returned from TimeSeriesDataPrepare()}

\item{test}{Test data returned from TimeSeriesDataPrepare()}

\item{FullData}{Full series data for scoring and ensemble}

\item{HoldOutPeriods}{Holdout periods returned from TimeSeriesDataPrepare()}

\item{MinVal}{Minimum value of target variable returned from TimeSeriesDataPrepare()}

\item{TargetName}{Target variable name returned from TimeSeriesDataPrepare()}

\item{DateName}{Date variable name returned from TimeSeriesDataPrepare()}

\item{Lags}{Max value of lag returned from TimeSeriesDataPrepare()}

\item{SeasonalLags}{Max value of seasonal lags returned from TimeSeriesDataPrepare()}

\item{MaxFourierTerms}{Max value of fourier pairs}

\item{TrainValidateShare}{A two-element numeric vector. The first element is the weight applied to the training performance and the remainder is applied to the validation performance.}

\item{MaxRunsWithoutNewWinner}{The number of runs without a new winner which if passed tells the function to stop}

\item{MaxNumberModels}{The number of models you want to test.}

\item{FinalGrid}{If NULL, regular train optimization occurs. If the grid is supplied, final builds are conducted.}
}
\value{
Time series data sets to pass onto auto modeling functions
}
\description{
OptimizeNNET is a function that takes raw data and returns the necessary time series data and objects for model building. It also fills any time gaps with zeros. Use this before you run any time series model functions.
}
\examples{
Results <- OptimizeNNET(
  Output,
  MetricSelection = "MAE",
  DataSetName = NULL,
  train = NULL,
  test = NULL,
  FullData = NULL,
  HoldOutPeriods = NULL,
  MinVal = NULL,
  TargetName = NULL,
  DateName = NULL,
  Lags = NULL,
  SeasonalLags = NULL,
  MaxFourierTerms = NULL,
  TrainValidateShare = NULL,
  MaxRunsWithoutNewWinner = 20,
  MaxNumberModels = 5,
  MaxRunMinutes = NULL,
  FinalGrid = NULL)
}
\seealso{
Other Time Series: 
\code{\link{AutoBanditNNet}()},
\code{\link{AutoBanditSarima}()},
\code{\link{AutoTBATS}()},
\code{\link{CarmaHoldoutMetrics}()},
\code{\link{DifferenceDataReverse}()},
\code{\link{DifferenceData}()},
\code{\link{FinalBuildArfima}()},
\code{\link{FinalBuildArima}()},
\code{\link{FinalBuildETS}()},
\code{\link{FinalBuildNNET}()},
\code{\link{FinalBuildTBATS}()},
\code{\link{FinalBuildTSLM}()},
\code{\link{GenerateParameterGrids}()},
\code{\link{OptimizeArfima}()},
\code{\link{OptimizeArima}()},
\code{\link{OptimizeETS}()},
\code{\link{OptimizeTBATS}()},
\code{\link{OptimizeTSLM}()},
\code{\link{ParallelAutoARIMA}()},
\code{\link{ParallelAutoArfima}()},
\code{\link{ParallelAutoETS}()},
\code{\link{ParallelAutoNNET}()},
\code{\link{ParallelAutoTBATS}()},
\code{\link{ParallelAutoTSLM}()},
\code{\link{RL_Performance}()},
\code{\link{Regular_Performance}()},
\code{\link{StackedTimeSeriesEnsembleForecast}()},
\code{\link{TimeSeriesDataPrepare}()},
\code{\link{WideTimeSeriesEnsembleForecast}()}
}
\author{
Adrian Antico
}
\concept{Time Series}
