% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EconometricsFunctions.R
\name{OptimizeTSLM}
\alias{OptimizeTSLM}
\title{OptimizeTSLM is a function that takes raw data and returns time series data}
\usage{
OptimizeTSLM(
  Output,
  MetricSelection = "MAE",
  DataSetName = NULL,
  train = NULL,
  test = NULL,
  FullData = NULL,
  HoldOutPeriods = NULL,
  MinVal = NULL,
  TargetName = NULL,
  DateName = NULL,
  TrainValidateShare = NULL,
  FinalGrid = NULL
)
}
\arguments{
\item{Output}{This is passed through as output from TimeSeriesDataPrepare() and passed through ParallelArima()}

\item{MetricSelection}{Select from "MSE", "MAE", or "MAPE"}

\item{DataSetName}{This is the name of the data set passed through in parallel loop}

\item{train}{Training data returned from TimeSeriesDataPrepare()}

\item{test}{Test data returned from TimeSeriesDataPrepare()}

\item{FullData}{Full series data for scoring and ensemble}

\item{HoldOutPeriods}{Holdout periods returned from TimeSeriesDataPrepare()}

\item{MinVal}{Minimum value of target variable returned from TimeSeriesDataPrepare()}

\item{TargetName}{Target variable name returned from TimeSeriesDataPrepare()}

\item{DateName}{Date variable name returned from TimeSeriesDataPrepare()}

\item{TrainValidateShare}{A two-element numeric vector. The first element is the weight applied to the training performance and the remainder is applied to the validation performance.}

\item{FinalGrid}{Grid for forecasting models}
}
\value{
Time series data sets to pass onto auto modeling functions
}
\description{
OptimizeTSLM is a function that takes raw data and returns the necessary time series data and objects for model building. It also fills any time gaps with zeros. Use this before you run any time series model functions.
}
\examples{
\dontrun{
Results <- OptimizeTSLM(
  Output,
  MetricSelection = "MAE",
  DataSetName = NULL,
  train = NULL,
  test = NULL,
  FullData = NULL,
  HoldOutPeriods = NULL,
  MinVal = NULL,
  TargetName = NULL,
  DateName = NULL,
  TrainValidateShare = NULL,
  FinalGrid = NULL)
}
}
\seealso{
Other Time Series Helper: 
\code{\link{FinalBuildArfima}()},
\code{\link{FinalBuildArima}()},
\code{\link{FinalBuildETS}()},
\code{\link{FinalBuildNNET}()},
\code{\link{FinalBuildTBATS}()},
\code{\link{FinalBuildTSLM}()},
\code{\link{GenerateParameterGrids}()},
\code{\link{OptimizeArfima}()},
\code{\link{OptimizeArima}()},
\code{\link{OptimizeETS}()},
\code{\link{OptimizeNNET}()},
\code{\link{OptimizeTBATS}()},
\code{\link{ParallelAutoARIMA}()},
\code{\link{ParallelAutoArfima}()},
\code{\link{ParallelAutoETS}()},
\code{\link{ParallelAutoNNET}()},
\code{\link{ParallelAutoTBATS}()},
\code{\link{ParallelAutoTSLM}()},
\code{\link{PredictArima}()},
\code{\link{RL_Performance}()},
\code{\link{Regular_Performance}()},
\code{\link{StackedTimeSeriesEnsembleForecast}()},
\code{\link{TimeSeriesDataPrepare}()},
\code{\link{WideTimeSeriesEnsembleForecast}()}
}
\author{
Adrian Antico
}
\concept{Time Series Helper}
