% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EconometricsFunctions.R
\name{ParallelAutoNNET}
\alias{ParallelAutoNNET}
\title{ParallelAutoNNET to run the 4 data sets at once}
\usage{
ParallelAutoNNET(
  Output,
  MetricSelection = "MAE",
  MaxFourierTerms = 1,
  TrainValidateShare = c(0.5, 0.5),
  MaxNumberModels = 20,
  MaxRunMinutes = 5,
  MaxRunsWithoutNewWinner = 12,
  NumCores = max(1L, min(4L, parallel::detectCores() - 2L))
)
}
\arguments{
\item{Output}{The output returned from TimeSeriesDataPrepare()}

\item{MetricSelection}{Choose from MAE, MSE, and MAPE}

\item{MaxFourierTerms}{Fourier pairs}

\item{TrainValidateShare}{c(0.50,0.50)}

\item{MaxNumberModels}{20}

\item{MaxRunMinutes}{5}

\item{MaxRunsWithoutNewWinner}{12}

\item{NumCores}{Default of max(1L, min(4L, parallel::detectCores())). Up to 4 cores can be utilized.}
}
\value{
Time series data sets to pass onto auto modeling functions
}
\description{
ParallelAutoNNET to run the 4 data sets at once
}
\examples{
\dontrun{
ParallelAutoNNET(
  MetricSelection = "MAE",
  Output = NULL,
  MaxRunsWithoutNewWinner = 20,
  TrainValidateShare = c(0.50,0.50),
  MaxNumberModels = 5,
  MaxRunMinutes = 5,
  NumCores = max(1L, min(4L, parallel::detectCores()-2L)))
}
}
\seealso{
Other Time Series Helper: 
\code{\link{FinalBuildArfima}()},
\code{\link{FinalBuildArima}()},
\code{\link{FinalBuildETS}()},
\code{\link{FinalBuildNNET}()},
\code{\link{FinalBuildTBATS}()},
\code{\link{FinalBuildTSLM}()},
\code{\link{GenerateParameterGrids}()},
\code{\link{OptimizeArfima}()},
\code{\link{OptimizeArima}()},
\code{\link{OptimizeETS}()},
\code{\link{OptimizeNNET}()},
\code{\link{OptimizeTBATS}()},
\code{\link{OptimizeTSLM}()},
\code{\link{ParallelAutoARIMA}()},
\code{\link{ParallelAutoArfima}()},
\code{\link{ParallelAutoETS}()},
\code{\link{ParallelAutoTBATS}()},
\code{\link{ParallelAutoTSLM}()},
\code{\link{PredictArima}()},
\code{\link{RL_Performance}()},
\code{\link{Regular_Performance}()},
\code{\link{StackedTimeSeriesEnsembleForecast}()},
\code{\link{TimeSeriesDataPrepare}()},
\code{\link{WideTimeSeriesEnsembleForecast}()}
}
\author{
Adrian Antico
}
\concept{Time Series Helper}
