% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SystemFunctions.R
\name{PathNormalizer}
\alias{PathNormalizer}
\title{PathNormalizer Converts path files to proper path files}
\usage{
PathNormalizer(Path)
}
\arguments{
\item{Path}{This is the path file you are supplying#'}
}
\value{
Returns a proper path file string
}
\description{
PathNormalizer Converts path files to proper path files
}
\seealso{
Other System Functions: 
\code{\link{CreateProjectFolders}()}
}
\author{
Adrian Antico
}
\concept{System Functions}
