% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureEngineering_NumericTypes.R
\name{PercRank}
\alias{PercRank}
\title{PercRank}
\usage{
PercRank(data, ColNames, GroupVars = NULL, Granularity = 0.001)
}
\arguments{
\item{data}{Source data.table}

\item{ColNames}{Character vector of column names}

\item{GroupVars}{Character vector of column names to have percent ranks by the group levels}

\item{Granularity}{Provide a value such that data.table::frank(Variable) * (1 / Granularity) / .N * Granularity. Default is 0.001}
}
\description{
Generate percent ranks for multiple variables, by groups if provided, and with a selected granularity
}
\author{
Adrian Antico
}
\concept{Utility}
