% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyFunctions.R
\name{PickerInput_GetLevels2}
\alias{PickerInput_GetLevels2}
\title{PickerInput_GetLevels2}
\usage{
PickerInput_GetLevels2(
  DataExist = TRUE,
  NumGroupVar = 3,
  InputID = "TS_CARMA_HolidayMovingAverages",
  InputID2 = "timeSeriesGroupVars",
  Choices = as.character(0:50),
  SelectedDefault = as.character(c(1, 2)),
  Multiple = TRUE
)
}
\arguments{
\item{DataExist}{Logical}

\item{NumGroupVar}{Which group var to select}

\item{InputID}{Feeds ProjectList and inputId. Argument saved in ProjectList}

\item{InputID2}{Values from input2. In first version the input is referenced inside function}

\item{Choices}{Feeds choices}

\item{SelectedDefault}{Feeds selected for cases where ProjectList has a null element}

\item{Multiple}{Feeds multiple for enabling selecting more than one element from list}
}
\value{
PickerInput object for server.R to go into renderUI({PickerInput()})
}
\description{
PickerInput_GetLevels2 automatically builds a picker input with tryCatch's and ProjectList argument usage if it exists
}
\examples{
\dontrun{
output$TS_CARMA_HolidayMovingAverages <- renderUI({
  RemixAutoML::PickerInput_GetLevels2(
    InputID = "TS_CARMA_HolidayMovingAverages", Label = "Select Holiday Count MA's", Choices = as.character(0:50),
    SelectedDefault = as.character(c(1,2)), Multiple = TRUE)})
}
}
\seealso{
Other Shiny: 
\code{\link{ArgNullCheck2}()},
\code{\link{ArgNullCheck}()},
\code{\link{BlankRow}()},
\code{\link{DateInput}()},
\code{\link{GenerateEvaluationMetrics}()},
\code{\link{NumericInput}()},
\code{\link{PickerInput_GetLevels}()},
\code{\link{PickerInput}()},
\code{\link{PreparePlotData}()},
\code{\link{ReactiveLoadCSV}()},
\code{\link{ReturnParam}()},
\code{\link{SelectizeInput}()},
\code{\link{StoreArgs}()},
\code{\link{TextInput}()},
\code{\link{observeEventLoad}()}
}
\author{
Adrian Antico
}
\concept{Shiny}
