% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyFunctions.R
\name{PlotLimits}
\alias{PlotLimits}
\title{PlotLimits}
\usage{
PlotLimits(p, YMin, YMax, XMin, XMax, Debug = FALSE)
}
\arguments{
\item{p}{data.table}

\item{YMin}{Y Min Value}

\item{YMax}{Y Max Value}

\item{XMin}{X Min Value}

\item{XMax}{X Max Value}

\item{Debug}{= FALSE}
}
\description{
PlotLimits
}
