% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotFunctions.R
\name{PlotlyConversion}
\alias{PlotlyConversion}
\title{PlotlyConversion}
\usage{
PlotlyConversion(p1)
}
\arguments{
\item{p1}{ggplot2 object}
}
\description{
PlotlyConversion
}
\seealso{
Other Graphics: 
\code{\link{AddFacet}()},
\code{\link{AppModelInsights}()},
\code{\link{AutoPlotter}()},
\code{\link{BarPlot}()},
\code{\link{BoxPlot}()},
\code{\link{ChartTheme}()},
\code{\link{CorrMatrixPlot}()},
\code{\link{HeatMapPlot}()},
\code{\link{HistPlot}()},
\code{\link{StockPlot}()},
\code{\link{ViolinPlot}()},
\code{\link{multiplot}()}
}
\author{
Adrian Antico
}
\concept{Graphics}
