% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SQL_Functions.R
\name{PosteGRE_DropDB}
\alias{PosteGRE_DropDB}
\title{PosteGRE_DropDB}
\usage{
PosteGRE_DropDB(
  DBName = NULL,
  Host = "localhost",
  Port = 5432,
  User = "postgres",
  Password = "",
  Connection = NULL,
  CloseConnection = TRUE
)
}
\arguments{
\item{DBName}{name of db}

\item{Host}{See args from related functions}

\item{Port}{See args from related functions}

\item{User}{See args from related functions}

\item{Password}{See args from related functions}

\item{Connection}{See args from related functions}

\item{CloseConnection}{See args from related functions}
}
\description{
PosteGRE_DropDB Drop selected database if it exists
}
\seealso{
Other Database: 
\code{\link{AutoDataDictionaries}()},
\code{\link{PostGRE_AppendData}()},
\code{\link{PostGRE_CreateTable}()},
\code{\link{PostGRE_GetTableNames}()},
\code{\link{PostGRE_ListTables}()},
\code{\link{PostGRE_Query}()},
\code{\link{PostGRE_RemoveCreateAppend}()},
\code{\link{PostGRE_RemoveTable}()},
\code{\link{PosteGRE_CreateDatabase}()},
\code{\link{PosteGRE_ListDatabases}()},
\code{\link{SQL_ClearTable}()},
\code{\link{SQL_DropTable}()},
\code{\link{SQL_Query_Push}()},
\code{\link{SQL_Query}()},
\code{\link{SQL_SaveTable}()},
\code{\link{SQL_Server_DBConnection}()}
}
\author{
Adrian Antico
}
\concept{Database}
