% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SQL_Functions.R
\name{PostGRE_AppendData}
\alias{PostGRE_AppendData}
\title{PostGRE_AppendData}
\usage{
PostGRE_AppendData(
  data = NULL,
  TableName = NULL,
  Append = FALSE,
  Connection = NULL,
  CloseConnection = FALSE,
  Host = NULL,
  DBName = NULL,
  User = NULL,
  Port = NULL,
  Password = NULL
)
}
\arguments{
\item{data}{Source data.table}

\item{Append}{Set to TRUE to append data, FALSE to overwrite data}

\item{Connection}{db connection}

\item{CloseConnection}{= FALSE}

\item{Host}{If Connection is NULL then this must be supplied. host}

\item{DBName}{If Connection is NULL then this must be supplied. dbname}

\item{User}{If Connection is NULL then this must be supplied. user}

\item{Port}{If Connection is NULL then this must be supplied. port}

\item{Password}{If Connection is NULL then this must be supplied. password}
}
\description{
PostGRE_AppendData get data from a database table
}
\examples{
\dontrun{
RemixAutoML::PostGRE_AppendData(
  data = data,
  TableName = 'somename',
  Append = FALSE,
  CloseConnection = FALSE,
  Host = 'localhost',
  DBName = 'RemixAutoML',
  User = 'postgres',
  Port = 5432,
  Password = 'Aa...')

# data = data
# CloseConnection = FALSE,
# Host = 'localhost'
# DBName = 'Testing'
# User = 'postgres'
# Port = 5432
# Password = 'Aa...'
}

}
\seealso{
Other Database: 
\code{\link{AutoDataDictionaries}()},
\code{\link{PostGRE_CreateTable}()},
\code{\link{PostGRE_GetTableNames}()},
\code{\link{PostGRE_ListTables}()},
\code{\link{PostGRE_Query}()},
\code{\link{PostGRE_RemoveCreateAppend}()},
\code{\link{PostGRE_RemoveTable}()},
\code{\link{PosteGRE_CreateDatabase}()},
\code{\link{PosteGRE_DropDB}()},
\code{\link{PosteGRE_ListDatabases}()},
\code{\link{SQL_ClearTable}()},
\code{\link{SQL_DropTable}()},
\code{\link{SQL_Query_Push}()},
\code{\link{SQL_Query}()},
\code{\link{SQL_SaveTable}()},
\code{\link{SQL_Server_DBConnection}()}
}
\author{
Adrian Antico
}
\concept{Database}
