% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SQL_Functions.R
\name{PostGRE_GetTableNames}
\alias{PostGRE_GetTableNames}
\title{PostGRE_GetTableNames}
\usage{
PostGRE_GetTableNames(
  Host = NULL,
  CloseConnection = FALSE,
  DBName = NULL,
  TableName = NULL,
  User = NULL,
  Port = NULL,
  Password = NULL
)
}
\arguments{
\item{Host}{See args from related functions}

\item{CloseConnection}{See args from related functions}

\item{DBName}{See args from related functions}

\item{TableName}{Name of postgres table}

\item{User}{See args from related functions}

\item{Port}{See args from related functions}

\item{Password}{See args from related functions}
}
\value{
A character vector of names. Exactly like names R base function for a data.frame
}
\description{
PostGRE_GetTableNames will list all column names from a table
}
\seealso{
Other Database: 
\code{\link{AutoDataDictionaries}()},
\code{\link{PostGRE_AppendData}()},
\code{\link{PostGRE_CreateTable}()},
\code{\link{PostGRE_ListTables}()},
\code{\link{PostGRE_Query}()},
\code{\link{PostGRE_RemoveCreateAppend}()},
\code{\link{PostGRE_RemoveTable}()},
\code{\link{PosteGRE_CreateDatabase}()},
\code{\link{PosteGRE_DropDB}()},
\code{\link{PosteGRE_ListDatabases}()},
\code{\link{SQL_ClearTable}()},
\code{\link{SQL_DropTable}()},
\code{\link{SQL_Query}()},
\code{\link{SQL_Query_Push}()},
\code{\link{SQL_SaveTable}()},
\code{\link{SQL_Server_DBConnection}()}
}
\author{
Adrian Antico
}
\concept{Database}
