% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SQL_Functions.R
\name{PostGRE_ListTables}
\alias{PostGRE_ListTables}
\title{PostGRE_ListTables}
\usage{
PostGRE_ListTables(
  DBName = NULL,
  Connection = NULL,
  CloseConnection = TRUE,
  Host = NULL,
  Port = NULL,
  User = NULL,
  Password = NULL
)
}
\arguments{
\item{DBName}{See args from related functions}

\item{Connection}{See args from related functions}

\item{CloseConnection}{See args from related functions}

\item{Host}{See args from related functions}

\item{Port}{See args from related functions}

\item{User}{See args from related functions}

\item{Password}{See args from related functions}

\item{Temporary}{See args from related functions}
}
\description{
PostGRE_ListTables will list all tables with an associated db
}
\seealso{
Other Database: 
\code{\link{AutoDataDictionaries}()},
\code{\link{PostGRE_AppendData}()},
\code{\link{PostGRE_CreateTable}()},
\code{\link{PostGRE_GetTableNames}()},
\code{\link{PostGRE_Query}()},
\code{\link{PostGRE_RemoveCreateAppend}()},
\code{\link{PostGRE_RemoveTable}()},
\code{\link{PosteGRE_CreateDatabase}()},
\code{\link{PosteGRE_DropDB}()},
\code{\link{PosteGRE_ListDatabases}()},
\code{\link{SQL_ClearTable}()},
\code{\link{SQL_DropTable}()},
\code{\link{SQL_Query_Push}()},
\code{\link{SQL_Query}()},
\code{\link{SQL_SaveTable}()},
\code{\link{SQL_Server_DBConnection}()}
}
\author{
Adrian Antico
}
\concept{Database}
