% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SQL_Functions.R
\name{PostGRE_Query}
\alias{PostGRE_Query}
\title{PostGRE_Query}
\usage{
PostGRE_Query(
  Query = NULL,
  Connection = NULL,
  CloseConnection = FALSE,
  Host = "localhost",
  DBName = "RemixAutoML",
  User = "postgres",
  Port = 5432,
  Password = ""
)
}
\arguments{
\item{Query}{SQL Statement in quotes}

\item{Connection}{db connection}

\item{CloseConnection}{= FALSE}

\item{Host}{If Connection is NULL then this must be supplied. host}

\item{DBName}{If Connection is NULL then this must be supplied. dbname}

\item{User}{If Connection is NULL then this must be supplied. user}

\item{Port}{If Connection is NULL then this must be supplied. port}

\item{Password}{If Connection is NULL then this must be supplied. password}
}
\description{
PostGRE_Query get data from a database table
}
\examples{
\dontrun{

# Query data from table with Uppercase name
data <- RemixAutoML::PostGRE_Query(
  Query = paste0("SELECT * FROM ", shQuote('Devices')),
  Host = 'localhost',
  CloseConnection = FALSE,
  DBName = 'RemixAutoML',
  User = 'postgres',
  Port = 5432,
  Password = 'Aa...')

# Query = 'Select * from static_data'
# Host = 'localhost'
# DBName = 'RemixAutoML'
# CloseConnection = FALSE,
# User = 'postgres'
# Port = 5432
# Password = 'Aa...'

# Create Schema
query <- "CREATE SCHEMA remixautoml AUTHORIZATION postgres;"
RemixAutoML::PostGRE_Query(
  Query = query,
  Host = 'localhost',
  CloseConnection = FALSE,
  DBName = 'RemixAutoML',
  User = 'postgres',
  Port = 5432,
  Password = 'Aa...')
}

}
\seealso{
Other Database: 
\code{\link{AutoDataDictionaries}()},
\code{\link{PostGRE_AppendData}()},
\code{\link{PostGRE_CreateTable}()},
\code{\link{PostGRE_RemoveCreateAppend}()},
\code{\link{PostGRE_RemoveTable}()},
\code{\link{SQL_ClearTable}()},
\code{\link{SQL_DropTable}()},
\code{\link{SQL_Query_Push}()},
\code{\link{SQL_Query}()},
\code{\link{SQL_SaveTable}()},
\code{\link{SQL_Server_DBConnection}()}
}
\author{
Adrian Antico
}
\concept{Database}
