% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EconometricsFunctions.R
\name{PredictArima}
\alias{PredictArima}
\title{PredictArima to forecast an Arima() model from the stats package}
\usage{
PredictArima(
  object = Results,
  n.ahead = FCPeriods,
  newxreg = NULL,
  se.fit = TRUE
)
}
\arguments{
\item{object}{Object that stores the output from Arima()}

\item{n.ahead}{Number of forecast periods to forecast}

\item{newxreg}{NULL by default. Forward looking independent variables as matrix type}

\item{se.fit}{Set to FALSE to not return prediction intervals with the forecast}
}
\description{
PredictArima is a function to overwrite the s3 generic <code>getS3method('predict','Arima')</code>
}
\seealso{
Other Automated Time Series: 
\code{\link{AutoCatBoostCARMA}()},
\code{\link{AutoCatBoostFreqSizeScoring}()},
\code{\link{AutoCatBoostSizeFreqDist}()},
\code{\link{AutoH2oDRFCARMA}()},
\code{\link{AutoH2oGBMCARMA}()},
\code{\link{AutoH2oGBMFreqSizeScoring}()},
\code{\link{AutoH2oGBMSizeFreqDist}()},
\code{\link{AutoH2oGLMCARMA}()},
\code{\link{AutoH2oMLCARMA}()},
\code{\link{AutoMLChainLadder}()},
\code{\link{AutoTS}()},
\code{\link{AutoXGBoostCARMA}()},
\code{\link{IntermittentDemandScoringDataGenerator}()}
}
\author{
Adrian Antico
}
\concept{Automated Time Series}
