% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyFunctions.R
\name{PreparePlotData}
\alias{PreparePlotData}
\title{PreparePlotData}
\usage{
PreparePlotData(
  input,
  PlotDataForecast,
  SubsetOnly = FALSE,
  Aggregate = NULL,
  TargetVariable = NULL,
  DateVariable = NULL,
  GroupVariables = NULL,
  G1Levels = NULL,
  G2Levels = NULL,
  G3Levels = NULL
)
}
\arguments{
\item{input}{input object within shiny context}

\item{PlotDataForecast}{Source data in shiny app}

\item{SubsetOnly}{Set to TRUE to only subset data}

\item{Aggregate}{Session object indicating whether to use mean or sum}

\item{TargetVariable}{Target variable name}

\item{DateVariable}{Date variable name}

\item{GroupVariables}{Group variable names}

\item{G1Levels}{Name of group 1 levels list element}

\item{G2Levels}{Name of group 2 levels list element}

\item{G3Levels}{Name of group 3 levels list element}
}
\value{
PreparePlotData object for server.R to
}
\description{
PreparePlotData automatically builds a picker input with tryCatch's and ProjectList argument usage if it exists
}
\examples{
\dontrun{
  PlotData <- RemixAutoML::PreparePlotData(
    input,
    PlotDataForecast,
    SubsetOnly = FALSE,
    Aggregate = "mean",
    TargetVariable = "TargetVariables",
    DateVariable = "DateVariables",
    GroupVariables = GroupVariables,
    G1Levels = "TS_Group1Levels",
    G2Levels = "TS_Group2Levels",
    G3Levels = "TS_Group3Levels")
}
}
\seealso{
Other Shiny: 
\code{\link{AppsPlotting}()},
\code{\link{ArgNullCheck2}()},
\code{\link{ArgNullCheck}()},
\code{\link{BlankRow}()},
\code{\link{DateInput}()},
\code{\link{GenerateEvaluationMetrics}()},
\code{\link{NumericInput}()},
\code{\link{PickerInput_GetLevels}()},
\code{\link{PickerInput}()},
\code{\link{ReactiveLoadCSV}()},
\code{\link{ReturnParam}()},
\code{\link{StoreArgs}()},
\code{\link{TextInput}()},
\code{\link{observeEventLoad}()}
}
\author{
Adrian Antico
}
\concept{Shiny}
