% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_export_helpers.R
\name{print.llm_image_artifact}
\alias{print.llm_image_artifact}
\title{Print an LLM Image Artifact}
\usage{
\method{print}{llm_image_artifact}(x, ...)
}
\arguments{
\item{x}{An object created by [ObjectToLLMArtifact()].}

\item{...}{Unused.}
}
\value{
Invisibly returns `x`.
}
\description{
Print an LLM Image Artifact
}
