% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Adrian_Antico_Modeling_Tools_Library.R
\name{PrintObjectsSize}
\alias{PrintObjectsSize}
\title{PrintObjectsSize prints out the top N objects and their associated sizes, sorted by size}
\usage{
PrintObjectsSize(N = 10)
}
\arguments{
\item{N}{The number of objects to display}
}
\value{
The objects in your environment and their sizes
}
\description{
PrintObjectsSize prints out the top N objects and their associated sizes, sorted by size
}
\examples{
PrintObjectsSize(N = 10)
}
\author{
Adrian Antico
}
