% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscFunctions.R
\name{PrintToPDF}
\alias{PrintToPDF}
\title{PrintToPDF}
\usage{
PrintToPDF(
  Path,
  OutputName,
  ObjectList = NULL,
  Knitr = FALSE,
  Title = "Model Output",
  Width = 7,
  Height = 7,
  Paper = "USr",
  BackgroundColor = "transparent",
  ForegroundColor = "black"
)
}
\arguments{
\item{Path}{Path file to the location where you want your pdf saved}

\item{OutputName}{Supply a name for the file you want saved}

\item{ObjectList}{List of objects to print to pdf}

\item{Knitr}{TRUE for data tables, FALSE for plots}

\item{Title}{The title of the pdf}

\item{Width}{Default is 7}

\item{Height}{Default is 7}

\item{Paper}{'USr' for landscape. 'special' means that Width and Height are used to determine page size}

\item{BackgroundColor}{Default is 'transparent'}

\item{ForegroundColor}{Default is 'black'}
}
\description{
PrintToPDF
}
\seealso{
Other Misc: 
\code{\link{AutoH2OTextPrepScoring}()},
\code{\link{ChartTheme}()},
\code{\link{PrintObjectsSize}()},
\code{\link{RPM_Binomial_Bandit}()},
\code{\link{SimpleCap}()},
\code{\link{tempDatesFun}()},
\code{\link{tokenizeH2O}()}
}
\author{
Adrian Antico
}
\concept{Misc}
