% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epistemic_integrity.R
\name{qa_epistemic_integrity_contracts}
\alias{qa_epistemic_integrity_contracts}
\title{QA for Epistemic Integrity Contracts}
\usage{
qa_epistemic_integrity_contracts()
}
\description{
QA for Epistemic Integrity Contracts
}
