% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/generate_binary_classification_model_insights_artifacts.R
\name{qa_generate_binary_classification_model_insights_artifacts}
\alias{qa_generate_binary_classification_model_insights_artifacts}
\title{QA for Binary Classification Model Insights Artifacts}
\usage{
qa_generate_binary_classification_model_insights_artifacts()
}
\value{
A compact data.table of QA checks.
}
\description{
QA for Binary Classification Model Insights Artifacts
}
\seealso{
Other Reports: 
\code{\link{BinaryClassificationModelInsightsReport}()},
\code{\link{EDAReport}()},
\code{\link{ModelInsightsReport}()},
\code{\link{RegressionModelInsightsReport}()},
\code{\link{TargetAnalysisReport}()},
\code{\link{generate_binary_classification_model_insights_artifacts}()},
\code{\link{generate_regression_model_insights_artifacts}()}
}
\concept{Reports}
