% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_catboost_builder_artifacts.R
\name{qa_generate_catboost_builder_artifacts_binary}
\alias{qa_generate_catboost_builder_artifacts_binary}
\title{QA for CatBoost Builder Binary Classification Artifacts}
\usage{
qa_generate_catboost_builder_artifacts_binary()
}
\value{
A compact data.table of QA checks.
}
\description{
QA for CatBoost Builder Binary Classification Artifacts
}
\seealso{
Other Reports: 
\code{\link{BinaryClassificationModelInsightsReport}()},
\code{\link{BinaryClassificationShapAnalysisReport}()},
\code{\link{EDAReport}()},
\code{\link{ModelInsightsReport}()},
\code{\link{RegressionModelInsightsReport}()},
\code{\link{RegressionShapAnalysisReport}()},
\code{\link{TargetAnalysisReport}()},
\code{\link{generate_binary_classification_model_insights_artifacts}()},
\code{\link{generate_catboost_builder_artifacts}()},
\code{\link{generate_regression_model_insights_artifacts}()},
\code{\link{qa_generate_binary_classification_model_insights_artifacts}()},
\code{\link{qa_generate_catboost_builder_artifacts}()},
\code{\link{qa_generate_catboost_builder_artifacts_regression}()}
}
\concept{Reports}
