% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal_observational_did.R
\name{qa_observational_did_estimation}
\alias{qa_observational_did_estimation}
\title{Deterministic QA for Governed Difference-in-Differences}
\usage{
qa_observational_did_estimation()
}
\description{
Deterministic QA for Governed Difference-in-Differences
}
