% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shap_autonls_effect_curves.R
\name{qa_shap_autonls_backend}
\alias{qa_shap_autonls_backend}
\title{QA for optional AutoNLS-backed SHAP effect curve artifacts}
\usage{
qa_shap_autonls_backend()
}
\value{
A data.table with QA checks and statuses.
}
\description{
QA for optional AutoNLS-backed SHAP effect curve artifacts
}
