% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supervised_reference_contract.R
\name{qa_supervised_reference_contract}
\alias{qa_supervised_reference_contract}
\title{QA Supervised Reference Contract and CatBoost Conformance}
\usage{
qa_supervised_reference_contract()
}
\value{
A validation table.
}
\description{
QA Supervised Reference Contract and CatBoost Conformance
}
