% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_semantics.R
\name{qa_variable_semantics_framework}
\alias{qa_variable_semantics_framework}
\title{QA for Canonical Variable Semantics}
\usage{
qa_variable_semantics_framework()
}
\value{
A data.table of deterministic QA checks.
}
\description{
QA for Canonical Variable Semantics
}
