% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vnext_catboost_regression.R
\name{qa_vnext_catboost_binary}
\alias{qa_vnext_catboost_binary}
\title{QA for vNext CatBoost Binary Classification}
\usage{
qa_vnext_catboost_binary()
}
\value{
A `data.table` of deterministic QA checks.
}
\description{
QA for vNext CatBoost Binary Classification
}
