% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vnext_catboost_regression.R
\name{qa_vnext_catboost_regression}
\alias{qa_vnext_catboost_regression}
\title{QA for vNext CatBoost Regression}
\usage{
qa_vnext_catboost_regression()
}
\value{
A `data.table` of deterministic QA checks.
}
\description{
QA for vNext CatBoost Regression
}
