% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vnext_forecasting_experiments.R
\name{qa_vnext_forecasting_experiment_campaigns}
\alias{qa_vnext_forecasting_experiment_campaigns}
\title{QA for vNext Forecasting Experiment Campaigns}
\usage{
qa_vnext_forecasting_experiment_campaigns()
}
\value{
A data.table of deterministic QA checks.
}
\description{
QA for vNext Forecasting Experiment Campaigns
}
