% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vnext_funnel_forecasting.R
\name{qa_vnext_funnel_forecasting_foundation}
\alias{qa_vnext_funnel_forecasting_foundation}
\title{QA for vNext Funnel Forecasting Foundation}
\usage{
qa_vnext_funnel_forecasting_foundation()
}
\value{
A data.table of deterministic QA checks.
}
\description{
QA for vNext Funnel Forecasting Foundation
}
