% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vnext_hierarchical_forecasting.R
\name{qa_vnext_hierarchical_forecasting_foundation}
\alias{qa_vnext_hierarchical_forecasting_foundation}
\title{QA for vNext Hierarchical Forecasting Foundation}
\usage{
qa_vnext_hierarchical_forecasting_foundation()
}
\value{
A QA `data.table`.
}
\description{
QA for vNext Hierarchical Forecasting Foundation
}
