% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vnext_hurdle_forecasting.R
\name{qa_vnext_hurdle_forecasting_foundation}
\alias{qa_vnext_hurdle_forecasting_foundation}
\title{QA for vNext Hurdle Forecasting Foundation}
\usage{
qa_vnext_hurdle_forecasting_foundation()
}
\value{
A QA `data.table`.
}
\description{
QA for vNext Hurdle Forecasting Foundation
}
