% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vnext_panel_strategy.R
\name{qa_vnext_panel_strategy_selection}
\alias{qa_vnext_panel_strategy_selection}
\title{QA for vNext Panel Strategy Selection}
\usage{
qa_vnext_panel_strategy_selection()
}
\value{
A QA `data.table`.
}
\description{
QA for vNext Panel Strategy Selection
}
