% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoCARMA_QA_FUNCTIONS.R
\name{QA_WALMARTDATAGENERATOR}
\alias{QA_WALMARTDATAGENERATOR}
\title{QA_WALMARTDATAGENERATOR}
\usage{
QA_WALMARTDATAGENERATOR(data, Groups = 1L, TimeUnit__ = "WEEK")
}
\arguments{
\item{data}{supply walmart data for either a single group or two group case. For no group, use XX}

\item{Groups}{Supply either 0L, 1L, or 2L to indicate the number of group variables to have tested}

\item{TimeUnit__}{= TimeUnit_}
}
\description{
QA_WALMARTDATAGENERATOR
}
\author{
Adrian Antico
}
\concept{QA}
