% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyFunctions.R
\name{ReactiveLoadCSV}
\alias{ReactiveLoadCSV}
\title{ReactiveLoadCSV}
\usage{
ReactiveLoadCSV(
  Infile = input[[eval(InputVal)]],
  ProjectList = NULL,
  DateUpdateName = NULL,
  RemoveObjects = NULL,
  Debug = FALSE
)
}
\arguments{
\item{ProjectList}{Supply the project list if available. NULL otherwise}

\item{DateUpdateName}{Supply the name for the ProjectList to store the import time}

\item{RemoveObjects}{List of objects to remove}

\item{Debug}{FALSE}

\item{input}{Passthrough}

\item{InputVal}{Values that follows input$}
}
\description{
Use this function to import csv's, track the time it was imported, and remove other objects
}
\seealso{
Other Shiny: 
\code{\link{ArgNullCheck2}()},
\code{\link{ArgNullCheck}()},
\code{\link{BlankRow}()},
\code{\link{DateInput}()},
\code{\link{GenerateEvaluationMetrics}()},
\code{\link{NumericInput}()},
\code{\link{PickerInput_GetLevels2}()},
\code{\link{PickerInput_GetLevels}()},
\code{\link{PickerInput}()},
\code{\link{PreparePlotData}()},
\code{\link{ReturnParam}()},
\code{\link{SelectizeInput}()},
\code{\link{StoreArgs}()},
\code{\link{TextInput}()},
\code{\link{observeEventLoad}()}
}
\author{
Adrian Antico
}
\concept{Shiny}
