% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Adrian_Antico_Modeling_Tools_Library.R
\name{RedYellowGreen}
\alias{RedYellowGreen}
\title{RedYellowGreen is for determining the optimal thresholds for binary classification when do-nothing is an option}
\usage{
RedYellowGreen(calibEval, PredictColNumber = 1, ActualColNumber = 767,
  TruePositiveCost = 0, TrueNegativeCost = 0,
  FalsePositiveCost = -10, FalseNegativeCost = -50, MidTierCost = -2,
  Cores = 8)
}
\arguments{
\item{calibEval}{data is the data table with your predicted and actual values from a classification model}

\item{PredictColNumber}{The column number where the actual target variable is located (in binary form)}

\item{ActualColNumber}{The column number where the predicted values are located}

\item{TruePositiveCost}{This is the utility for generating a true positive prediction}

\item{FalsePositiveCost}{This is the cost of generating a false positive prediction}

\item{FalseNegativeCost}{This is the cost of generating a false negative prediction}

\item{MidTierCost}{This is the cost of doing nothing (or whatever it means to not classify in your case)}

\item{Cores}{Number of cores on your machine}

\item{TruePositiveCost}{This is the utility for generating a true negative prediction}
}
\value{
A data table with all evaluated strategies, parameters, and utilities
}
\description{
This function will find the optimial thresholds for applying the main label and for finding the optimial range for doing nothing when you can quantity the cost of doing nothing
}
\examples{
test <- data.table(actual = ifelse(runif(1000) > 0.5,1,0),target = runif(1000))
data <- RedYellowGreen(calibEval,
                       PredictColNumber  = 1,
                       ActualColNumber   = 767,
                       TruePositiveCost  = 0,
                       TrueNegativeCost  = 0,
                       FalsePositiveCost = -1,
                       FalseNegativeCost = -10,
                       MidTierCost       = -5)
}
\author{
Adrian Antico
}
