% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelMetrics.R
\name{RegressionMetrics}
\alias{RegressionMetrics}
\title{RegressionMetrics}
\usage{
RegressionMetrics(
  SaveModelObjects. = SaveModelObjects,
  data. = data,
  ValidationData. = ValidationData,
  TrainOnFull. = TrainOnFull,
  LossFunction. = LossFunction,
  EvalMetric. = EvalMetric,
  TargetColumnName. = TargetColumnName,
  ModelID. = ModelID,
  model_path. = model_path,
  metadata_path. = metadata_path
)
}
\arguments{
\item{SaveModelObjects.}{= SaveModelObjects}

\item{data.}{= data}

\item{ValidationData.}{= ValidationData}

\item{TrainOnFull.}{= TrainOnFull}

\item{LossFunction.}{= LossFunction}

\item{EvalMetric.}{= EvalMetric}

\item{TargetColumnName.}{= TargetColumnName}

\item{ModelID.}{= ModelID}

\item{model_path.}{= model_path}

\item{metadata_path.}{= metadata_path}
}
\description{
Compute regression metrics and save them to file
}
\seealso{
Other Model Evaluation: 
\code{\link{BinaryMetrics}()},
\code{\link{ClassificationMetrics}()},
\code{\link{DT_BinaryConfusionMatrix}()},
\code{\link{MultiClassMetrics}()},
\code{\link{RemixClassificationMetrics}()}
}
\author{
Adrian Antico
}
\concept{Model Evaluation}
