% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EconometricsFunctions.R
\name{Regular_Performance}
\alias{Regular_Performance}
\title{Regular_Performance creates and stores model results in Experiment Grid}
\usage{
Regular_Performance(
  Model = NULL,
  Results = Results,
  GridList = GridList,
  TrainValidateShare = c(0.5, 0.5),
  ExperimentGrid = ExperimentGrid,
  run = run,
  train = train,
  ValidationData = ValidationData,
  HoldOutPeriods = HoldOutPeriods
)
}
\arguments{
\item{Model}{Set to ets, tbats, arfima, tslm, nnetar}

\item{Results}{This is a time series model}

\item{GridList}{List}

\item{TrainValidateShare}{The values used to blend training and validation performance}

\item{ExperimentGrid}{The results collection table}

\item{run}{Iterator}

\item{train}{Data set}

\item{ValidationData}{Data set}

\item{HoldOutPeriods}{Passthrough}
}
\description{
Regular_Performance creates and stores model results in Experiment Grid
}
\seealso{
Other Time Series Helper: 
\code{\link{FinalBuildArfima}()},
\code{\link{FinalBuildArima}()},
\code{\link{FinalBuildETS}()},
\code{\link{FinalBuildNNET}()},
\code{\link{FinalBuildTBATS}()},
\code{\link{FinalBuildTSLM}()},
\code{\link{GenerateParameterGrids}()},
\code{\link{OptimizeArfima}()},
\code{\link{OptimizeArima}()},
\code{\link{OptimizeETS}()},
\code{\link{OptimizeNNET}()},
\code{\link{OptimizeTBATS}()},
\code{\link{OptimizeTSLM}()},
\code{\link{ParallelAutoARIMA}()},
\code{\link{ParallelAutoArfima}()},
\code{\link{ParallelAutoETS}()},
\code{\link{ParallelAutoNNET}()},
\code{\link{ParallelAutoTBATS}()},
\code{\link{ParallelAutoTSLM}()},
\code{\link{PredictArima}()},
\code{\link{RL_Performance}()},
\code{\link{StackedTimeSeriesEnsembleForecast}()},
\code{\link{TimeSeriesDataPrepare}()},
\code{\link{WideTimeSeriesEnsembleForecast}()}
}
\author{
Adrian Antico
}
\concept{Time Series Helper}
