\name{RemixAutoML-package}
\alias{RemixAutoML-package}
\alias{RemixAutoML}
\docType{package}
\title{
  Automated Machine Learning Remixed
}
\description{
  Automated Machine Learning Remixed for real-world use-cases. The package utilizes data.table under the hood for all data wrangling like operations so it's super fast and memory efficient. All ML methods are available in R or Python. The forecasting functions are unique and state of the art. There are feature engineering functions in this package that you cannot find anywhere else.
}
\details{
  See the github README for details and examples www.github.com/AdrianAntico/RemixAutoML
}
\author{
Adrian Antico, adrianantico@gmail.com,
Douglas Pestana
}
