% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelMetrics.R
\name{RemixClassificationMetrics}
\alias{RemixClassificationMetrics}
\title{RemixClassificationMetrics}
\usage{
RemixClassificationMetrics(MLModels = c("catboost", "h2ogbm", "h2odrf",
  "xgboost"), TargetVariable = "Value",
  CatBoostTestData = CatModel$ValidationData,
  H2oGBMTestData = H2oGBMModel$ValidationData,
  H2oDRFTestData = H2oDRFModel$ValidationData,
  XGBoostTestData = XGBoostModel$ValidationData)
}
\arguments{
\item{MLModels}{A vector of model names from remixautoml}

\item{TargetVariable}{Name of your target variable}

\item{CatBoostTestData}{Test data returned from AutoCatBoostClassifier}

\item{H2oGBMTestData}{Test data returned from AutoCatBoostClassifier}

\item{H2oDRFTestData}{Test data returned from AutoCatBoostClassifier}

\item{XGBoostTestData}{Test data returned from AutoCatBoostClassifier}
}
\description{
RemixClassificationMetrics
}
\seealso{
Other Model Evaluation: \code{\link{ClassificationMetrics}},
  \code{\link{DT_BinaryConfusionMatrix}}
}
\author{
Adrian Antico
}
\concept{Model Evaluation}
