% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Adrian_Antico_Modeling_Tools_Library.R
\name{ResidualOutliers}
\alias{ResidualOutliers}
\title{ResidualOutliers is an automated time series outlier detection function}
\usage{
ResidualOutliers(data, maxN = 5, cvar = 4)
}
\arguments{
\item{data}{the source residuals data.table}

\item{maxN}{the largest lag or moving average (seasonal too) values for the arima fit}

\item{cvar}{the t-stat value for tsoutliers}
}
\value{
A data.table with outliers, the arima model, and residuals from the arima fit
}
\description{
ResidualOutliers is an automated time series outlier detection function that utilizes tsoutliers and auto.arima.
}
\examples{
data <- data.table(a = seq(0,10000,1),
                   predicted = sde::GBM(N=10000)*1000)[-1,]
data <- data.table(a = seq(1,10000,1),
                   predicted = sde::rcCIR(n=10000,
                                          Dt=0.1,
                                          x0=1,
                                          theta=c(6,2,2)))
data <- data.table(a = seq(1,10000,1),
                   predicted = sde::rsOU(n=10000,
                                         theta=c(0,2,1)))
stuff    <- ResidualOutliers(data = data, maxN = 5, cvar = 4)
data     <- stuff[[1]]
model    <- stuff[[2]]
resid    <- stuff[[3]]
outliers <- data[type != "<NA>"]
}
\author{
Adrian Antico
}
