% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelEvaluationPlots.R
\name{ResidualPlots}
\alias{ResidualPlots}
\title{ResidualPlots}
\usage{
ResidualPlots(
  TestData = NULL,
  Target = "Adrian",
  Predicted = "Independent_Variable1",
  DateColumnName = NULL,
  Gam_Fit = FALSE
)
}
\arguments{
\item{TestData}{= NULL,}

\item{Target}{= "Adrian",}

\item{Predicted}{= "Independent_Variable1",}

\item{DateColumnName}{"DateTime"}

\item{Gam_Fit}{= TRUE}
}
\description{
Residual plots for regression models
}
\examples{
\dontrun{
# Create fake data
test_data <- RemixAutoML::FakeDataGenerator(
  Correlation = 0.80,
  N = 250000,
  ID = 0,
  FactorCount = 0,
  AddDate = TRUE,
  AddComment = FALSE,
  AddWeightsColumn = FALSE,
  ZIP = 0)

# Build Plots
output <- RemixAutoML::ResidualPlots(
  TestData = test_data,
  Target = "Adrian",
  Predicted = "Independent_Variable1",
  DateColumnName = "DateTime",
  Gam_Fit = TRUE)
}

}
\seealso{
Other Model Evaluation and Interpretation: 
\code{\link{AutoShapeShap}()},
\code{\link{CumGainsChart}()},
\code{\link{EvalPlot}()},
\code{\link{ParDepCalPlots}()},
\code{\link{ROCPlot}()},
\code{\link{RedYellowGreen}()},
\code{\link{ShapPlot}()},
\code{\link{SingleRowShapeShap}()},
\code{\link{threshOptim}()}
}
\author{
Adrian Antico
}
\concept{Model Evaluation and Interpretation}
