% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReinforcementLearningFunctions.R
\name{RL_Initialize}
\alias{RL_Initialize}
\title{RL_Initialize}
\usage{
RL_Initialize(
  ParameterGridSet = NULL,
  Alpha = 1L,
  Beta = 1L,
  SubDivisions = 1000L
)
}
\arguments{
\item{ParameterGridSet}{This is a list of tuning grids}

\item{Alpha}{Prior successes}

\item{Beta}{Prior trials}

\item{SubDivisions}{Tolerance for integration}
}
\description{
RL_Initialize sets up the components necessary for RL
}
\examples{
RL_Start <- RL_Initialize(
  ParameterGridSet = GridClusters,
  Alpha = Alpha,
  Beta = Beta,
  SubDivisions = 1000L)
BanditArmsN <- RL_Start[["BanditArmsN"]]
Successes <- RL_Start[["Successes"]]
Trials <- RL_Start[["Trials"]]
GridIDs <- RL_Start[["GridIDs"]]
BanditProbs <- RL_Start[["BanditProbs"]]
}
\seealso{
Other Reinforcement Learning: 
\code{\link{RL_ML_Update}()},
\code{\link{RL_Update}()}
}
\author{
Adrian Antico
}
\concept{Reinforcement Learning}
