% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelEvaluationPlots.R
\name{ROCPlot}
\alias{ROCPlot}
\title{ROCPlot}
\usage{
ROCPlot(
  data = ValidationData,
  TargetName = TargetColumnName,
  SavePlot = SaveModelObjects,
  Name = ModelID,
  metapath = metadata_path,
  modelpath = model_path
)
}
\arguments{
\item{data}{validation data}

\item{TargetName}{Target variable name}

\item{SavePlot}{TRUE or FALSE}

\item{Name}{Name for saving}

\item{metapath}{Passthrough}

\item{modelpath}{Passthrough}
}
\value{
ROC Plot for classification models
}
\description{
Internal usage for classification methods. Returns an ROC plot
}
\seealso{
Other Model Evaluation and Interpretation: 
\code{\link{AutoShapeShap}()},
\code{\link{CumGainsChart}()},
\code{\link{EvalPlot}()},
\code{\link{ParDepCalPlots}()},
\code{\link{RedYellowGreen}()},
\code{\link{ResidualPlots}()},
\code{\link{SingleRowShapeShap}()},
\code{\link{threshOptim}()}
}
\author{
Adrian Antico
}
\concept{Model Evaluation and Interpretation}
