% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelInsights.R
\name{Run_ModelInsightsReport}
\alias{Run_ModelInsightsReport}
\title{ModelInsightsReport}
\usage{
Run_ModelInsightsReport(
  TrainData = NULL,
  ValidationData = NULL,
  TestData = NULL,
  TargetColumnName = NULL,
  PredictionColumnName = "Predict",
  FeatureColumnNames = NULL,
  DateColumnName = NULL,
  TargetType = "regression",
  ModelID = "ModelTest",
  Algo = "catboost",
  SourcePath = NULL,
  OutputPath = NULL,
  RemixOutput = NULL,
  Test_Importance_dt = NULL,
  Validation_Importance_dt = NULL,
  Train_Importance_dt = NULL,
  Test_Interaction_dt = NULL,
  Validation_Interaction_dt = NULL,
  Train_Interaction_dt = NULL
)
}
\arguments{
\item{TrainData}{data.table or something that converts to data.table via as.data.table}

\item{ValidationData}{= data.table or something that converts to data.table via as.data.table}

\item{TestData}{= data.table or something that converts to data.table via as.data.table}

\item{TargetColumnName}{For selecting columns}

\item{PredictionColumnName}{NULL}

\item{FeatureColumnNames}{NULL}

\item{DateColumnName}{NULL}

\item{TargetType}{'regression', 'classification', or 'multiclass'}

\item{ModelID}{ModelID used in the RemixAutoML supervised learning function}

\item{Algo}{'catboost', 'xgboost', 'lightgbm', 'h2o'}

\item{SourcePath}{Path to directory with RemixAutoML Model Output}

\item{OutputPath}{Path to directory where the html will be saved}

\item{RemixOutput}{Model Output passed from RemixAutoML::Auto_*() functions}

\item{Test_Importance_dt}{NULL.. Ignore if using RemixAutoML Models. Otherwise, supply a two column data.table with colnames 'Variable' and 'Importance'}

\item{Validation_Importance_dt}{NULL.. Ignore if using RemixAutoML Models. Otherwise, supply a two column data.table with colnames 'Variable' and 'Importance'}

\item{Train_Importance_dt}{NULL.. Ignore if using RemixAutoML Models. Otherwise, supply a two column data.table with colnames 'Variable' and 'Importance'}

\item{Test_Interaction_dt}{NULL.. Ignore if using RemixAutoML Models. Otherwise, supply a three column data.table with colnames 'Feature1', 'Feature2' and 'Importance'}

\item{Validation_Interaction_dt}{NULL.. Ignore if using RemixAutoML Models. Otherwise, supply a three column data.table with colnames 'Feature1', 'Feature2'  and 'Importance'}

\item{Train_Interaction_dt}{NULL.. Ignore if using RemixAutoML Models. Otherwise, supply a three column data.table with colnames 'Feature1', 'Feature2'  and 'Importance'}

\item{Path}{Path to Model Output if RemixOutput is left NULL}
}
\description{
ModelInsightsReport is an Rmarkdown report for viewing the model insights generated by RemixAutoML supervised learning functions
}
\seealso{
Other Model Insights: 
\code{\link{ModelInsightsReport}()}
}
\author{
Adrian Antico
}
\concept{Model Insights}
