% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimpleApps.R
\name{RunRemixAutoML}
\alias{RunRemixAutoML}
\title{RunRemixAutoML}
\usage{
RunRemixAutoML(
  data = NULL,
  ModelOutput = NULL,
  TargetName = NULL,
  PredictName = NULL,
  DateName = NULL,
  Debug = FALSE
)
}
\description{
GUI for running RemixAutoML functions. Time series evaluation and forecasting, machine learning, etc.
}
\seealso{
Other GUI: 
\code{\link{AppsPlotting}()}
}
\author{
Adrian Antico
}
\concept{GUI}
