% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotFunctions.R
\name{ShapImportancePlot}
\alias{ShapImportancePlot}
\title{ShapImportancePlot}
\usage{
ShapImportancePlot(
  data,
  ShapColNames = NULL,
  FacetVar1 = NULL,
  FacetVar2 = NULL,
  AggMethod = "mean",
  TopN = 25,
  Debug = FALSE
)
}
\arguments{
\item{data}{Source data.table}

\item{ShapColNames}{Names of the columns that contain shap values you want included}

\item{FacetVar1}{Column name}

\item{FacetVar2}{Column name}

\item{AggMethod}{A string for aggregating shapely values for importances. Choices include, 'mean', 'absmean', 'meanabs', 'sd', 'median', 'absmedian', 'medianabs'}

\item{TopN}{The number of variables to plot}

\item{Debug}{= FALSE}
}
\description{
Generate Variable Importance Plots using Shapely Values of given data set
}
\author{
Adrian Antico
}
\concept{Model Insights}
