% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyUtils.R
\name{Shiny.CodePrint.OrganizeCode}
\alias{Shiny.CodePrint.OrganizeCode}
\title{Shiny.CodePrint.OrganizeCode}
\usage{
Shiny.CodePrint.OrganizeCode(
  DM = NULL,
  DW = NULL,
  FE = NULL,
  ML = NULL,
  PL = NULL
)
}
\arguments{
\item{DM}{= NULL}

\item{DW}{= NULL}

\item{FE}{= NULL}

\item{ML}{= NULL}

\item{PL}{= NULL}
}
\description{
Shiny.CodePrint.OrganizeCode
}
\examples{
\dontrun{
# Step through
# #listname in c('DataMgtCode','DataWranglingCode','FeatureEngineeringCode','MachineLearningCode','PlottingCode')
# listname <- 'DataMgtCode'
# DataMgtCode <- list()
# DataMgtCode[[paste0(Sys.time())]] <- 'data.table::data.table(A = 1:12)'
# DataMgtCode[[paste0(Sys.time())]] <- 'data.table::data.table(B = 1:12)'
# DataWranglingCode <- list()
# DataWranglingCode[[paste0(Sys.time())]] <- 'data.table::data.table(C = 1:12)'
# DataWranglingCode[[paste0(Sys.time())]] <- 'data.table::data.table(D = 1:12)'
# MasterSet <- data.table::data.table(TimeStamp = Sys.time(), Code = 'Sys.time() # Lists intialization')
}

}
\keyword{internal}
