% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyFE.R
\name{Shiny.FE.Date.Holiday}
\alias{Shiny.FE.Date.Holiday}
\title{Shiny.FE.Date.Holiday}
\usage{
Shiny.FE.Date.Holiday(
  input,
  output,
  session,
  DataList,
  CodeList,
  CacheDir = NULL,
  CacheName = "data",
  Debug = FALSE
)
}
\arguments{
\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}

\item{DataList}{DataList contains the data sets in session}

\item{CodeList}{From app}

\item{CacheDir}{From app}

\item{CacheName}{'data' from app. Function will strip any '.csv' from CacheName if it exists and then place one there manually.}

\item{Debug}{Debug from app}
}
\description{
server.R observeEvent() for executing CreateHolidayVariables() for given inputs from app
}
\keyword{internal}
