% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyFunctions.R
\name{Shiny.ML.CatBoost.LossFunctionOptions}
\alias{Shiny.ML.CatBoost.LossFunctionOptions}
\title{Shiny.ML.CatBoost.LossFunctionOptions}
\usage{
Shiny.ML.CatBoost.LossFunctionOptions(TT)
}
\arguments{
\item{TT}{Target Type: 'regression', 'classification', 'multiclass', all lower case}
}
\description{
Shiny.ML.CatBoost.LossFunctionOptions
}
\keyword{internal}
