% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyML.R
\name{Shiny.ML.LightGBM.LossFunctionOptions}
\alias{Shiny.ML.LightGBM.LossFunctionOptions}
\title{LightGBMLossFunctionOptions}
\usage{
Shiny.ML.LightGBM.LossFunctionOptions(TT)
}
\arguments{
\item{TT}{Target Type: 'regression', 'classification', 'multiclass', all lower case}
}
\description{
LightGBMLossFunctionOptions
}
\keyword{internal}
