% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyFunctions.R
\name{Shiny.ML.ModelDataObjects}
\alias{Shiny.ML.ModelDataObjects}
\title{Shiny.ML.ModelDataObjects}
\usage{
Shiny.ML.ModelDataObjects(ModelOutput, TT = "catboost")
}
\arguments{
\item{ModelOutput}{Output from RemixAutoML:: supervised learning functions}
}
\description{
Shiny.ML.ModelDataObjects
}
\keyword{internal}
