% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyML.R
\name{Shiny.ML.Trainer}
\alias{Shiny.ML.Trainer}
\title{Shiny.ML.Trainer}
\usage{
Shiny.ML.Trainer(
  input,
  output,
  DataList,
  ArgsList,
  TT,
  ML_ExperimentTable,
  run,
  n,
  Debug,
  Algo = "CatBoost"
)
}
\arguments{
\item{input}{shiny input}

\item{output}{shiny output}

\item{DataList}{DataList stores data in app}

\item{ArgsList}{CatBoostArgsList, XGBoostArgsList}

\item{TT}{TargetType}

\item{ML_ExperimentTable}{Collection table}

\item{run}{Cross Validation Iteration Number in app}

\item{n}{Total Cross Validation Iterations in app}

\item{Debug}{Debug from app}

\item{Algo}{'CatBoost', 'XGBoost', 'LightGBM'}
}
\value{
a list of columns names by data type
}
\description{
Shiny.ML.Trainer
}
\keyword{internal}
