% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyML.R
\name{Shiny.ML.XGBoost_Params}
\alias{Shiny.ML.XGBoost_Params}
\title{Shiny.ML.XGBoost_Params}
\usage{
Shiny.ML.XGBoost_Params(l, input, Debug, TT)
}
\arguments{
\item{l}{ArgsList created inside server.R}

\item{input}{input from shiny app}

\item{Debug}{Debug. TRUE or FALSE}

\item{TT}{E.g. XGBoost_TargetType. Can be 'Regression', 'Binary Classification', or 'MultiClass'. Case sensitive}
}
\description{
Shiny.ML.XGBoost_Params
}
\keyword{internal}
