% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyDW.R
\name{Shiny.Utils.CachePath}
\alias{Shiny.Utils.CachePath}
\title{Shiny.Utils.CachePath}
\usage{
Shiny.Utils.CachePath(CacheName, CacheDir, Ext = ".csv")
}
\arguments{
\item{CacheName}{Name of file}

\item{CacheDir}{Path to folder}
}
\description{
Combines a directory path with a file name and extension
}
\keyword{internal}
