% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Adrian_Antico_Modeling_Tools_Library.R
\name{SimpleCap}
\alias{SimpleCap}
\title{SimpleCap function is for capitalizing the first letter of words}
\usage{
SimpleCap(x)
}
\arguments{
\item{x}{Column of interest}
}
\value{
An object to pass along to ggplot objects following the "+" sign
}
\description{
SimpleCap function is for capitalizing the first letter of words (need I say more?)
}
\examples{
x <- "adrian"
x <- SimpleCap(x)
}
\author{
Adrian Antico
}
