% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataWranglingFunctions.R
\name{SingleRowShapeShap}
\alias{SingleRowShapeShap}
\title{SingleRowShapeShap}
\usage{
SingleRowShapeShap(ShapData = NULL, EntityID = NULL, DateColumnName = NULL)
}
\arguments{
\item{ShapData}{Scoring data from AutoCatBoostScoring with classification or regression}
}
\description{
SingleRowShapeShap will convert a single row of your shap data into a table
}
\seealso{
Other Model Interpretation: 
\code{\link{AutoShapeShap}()}
}
\author{
Adrian Antico
}
\concept{Model Interpretation}
