% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SQL_Functions.R
\name{SQL_ClearTable}
\alias{SQL_ClearTable}
\title{SQL_ClearTable}
\usage{
SQL_ClearTable(
  DBConnection,
  SQLTableName = "",
  CloseChannel = TRUE,
  Errors = TRUE
)
}
\arguments{
\item{DBConnection}{AutoQuant::SQL_Server_DBConnection()}

\item{SQLTableName}{The SQL statement you want to run}

\item{CloseChannel}{TRUE to close when done, FALSE to leave the channel open}

\item{Errors}{Set to TRUE to halt, FALSE to return -1 in cases of errors}
}
\description{
SQL_ClearTable remove all rows from a database table
}
\seealso{
Other Database: 
\code{\link{AutoDataDictionaries}()},
\code{\link{PostGRE_AppendData}()},
\code{\link{PostGRE_CreateTable}()},
\code{\link{PostGRE_GetTableNames}()},
\code{\link{PostGRE_ListTables}()},
\code{\link{PostGRE_Query}()},
\code{\link{PostGRE_RemoveCreateAppend}()},
\code{\link{PostGRE_RemoveTable}()},
\code{\link{PosteGRE_CreateDatabase}()},
\code{\link{PosteGRE_DropDB}()},
\code{\link{PosteGRE_ListDatabases}()},
\code{\link{SQL_DropTable}()},
\code{\link{SQL_Query_Push}()},
\code{\link{SQL_Query}()},
\code{\link{SQL_SaveTable}()},
\code{\link{SQL_Server_DBConnection}()}
}
\author{
Adrian Antico
}
\concept{Database}
