% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SQL_Functions.R
\name{SQL_DropTable}
\alias{SQL_DropTable}
\title{SQL_DropTable}
\usage{
SQL_DropTable(
  DBConnection,
  SQLTableName = "",
  CloseChannel = TRUE,
  Errors = TRUE
)
}
\arguments{
\item{DBConnection}{RemixAutoML::SQL_Server_DBConnection()}

\item{SQLTableName}{The SQL statement you want to run}

\item{CloseChannel}{TRUE to close when done, FALSE to leave the channel open}

\item{Errors}{Set to TRUE to halt, FALSE to return -1 in cases of errors}
}
\description{
SQL_DropTable drop a database table
}
\seealso{
Other Database: 
\code{\link{AutoDataDictionaries}()},
\code{\link{SQL_ClearTable}()},
\code{\link{SQL_Query_Push}()},
\code{\link{SQL_Query}()},
\code{\link{SQL_SaveTable}()},
\code{\link{SQL_Server_BulkPull}()},
\code{\link{SQL_Server_BulkPush}()},
\code{\link{SQL_Server_DBConnection}()},
\code{\link{SQL_UpdateTable}()}
}
\author{
Adrian Antico
}
\concept{Database}
