% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataWranglingFunctions.R
\name{SQL_SaveTable}
\alias{SQL_SaveTable}
\title{SQL_SaveTable}
\usage{
SQL_SaveTable(
  DataToPush,
  DBConnection,
  SQLTableName = "",
  RowNames = NULL,
  ColNames = TRUE,
  VarTypes = NULL,
  CloseChannel = TRUE,
  AppendData = FALSE,
  AddPK = TRUE,
  Errors = TRUE,
  Safer = TRUE
)
}
\arguments{
\item{DataToPush}{data to be sent to warehouse}

\item{DBConnection}{RemixAutoML::SQL_Server_DBConnection()}

\item{SQLTableName}{The SQL statement you want to run}

\item{VarTypes}{= NULL,}

\item{CloseChannel}{TRUE to close when done, FALSE to leave the channel open}

\item{AppendData}{TRUE or FALSE}

\item{AddPK}{Add a PK column to table}

\item{Errors}{Set to TRUE to halt, FALSE to return -1 in cases of errors}

\item{rownames}{c("Segment","Date")}
}
\description{
SQL_SaveTable get data from a database
}
\seealso{
Other Data Wrangling: 
\code{\link{AutoDataDictionaries}()},
\code{\link{AutoHierarchicalFourier}()},
\code{\link{ColumnSubsetDataTable}()},
\code{\link{DataDisplayMeta}()},
\code{\link{FakeDataGenerator}()},
\code{\link{FullFactorialCatFeatures}()},
\code{\link{SQL_ClearTable}()},
\code{\link{SQL_DropTable}()},
\code{\link{SQL_Query}()},
\code{\link{SQL_Server_DBConnection}()},
\code{\link{SQL_UpdateTable}()},
\code{\link{TimeSeriesMelt}()}
}
\author{
Adrian Antico
}
\concept{Data Wrangling}
