% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SQL_Functions.R
\name{SQL_Server_BulkPull}
\alias{SQL_Server_BulkPull}
\title{SQL_Server_BulkPull}
\usage{
SQL_Server_BulkPull(
  Server = NULL,
  DBName = NULL,
  TableName = NULL,
  Query = NULL,
  FinalColumnNames = NULL,
  SavePath = NULL,
  SaveFileName = NULL,
  DeleteTextFile = TRUE
)
}
\arguments{
\item{Server}{Server name}

\item{DBName}{Name of the database}

\item{TableName}{Name of the table to pull}

\item{Query}{Leave NULL to pull entire talbe or supply a query}

\item{FinalColumnNames}{Supply this if you supply a query that isn't a select * query}

\item{SavePath}{Path file to where you want the text file saved}

\item{SaveFileName}{Name of the text file to create}

\item{DeleteTextFile}{Remove text file when done loading into R}
}
\description{
Pull data from a sql server warehouse using bulk copy process
}
\seealso{
Other Database: 
\code{\link{AutoDataDictionaries}()},
\code{\link{ExecuteSSIS}()},
\code{\link{SQL_ClearTable}()},
\code{\link{SQL_DropTable}()},
\code{\link{SQL_Query_Push}()},
\code{\link{SQL_Query}()},
\code{\link{SQL_SaveTable}()},
\code{\link{SQL_Server_BulkPush}()},
\code{\link{SQL_Server_DBConnection}()},
\code{\link{SQL_UpdateTable}()}
}
\author{
Adrian Antico
}
\concept{Database}
