% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SQL_Functions.R
\name{SQL_UpdateTable}
\alias{SQL_UpdateTable}
\title{SQL_UpdateTable}
\usage{
SQL_UpdateTable(
  DataToPush,
  DBConnection,
  SQLTableName = "",
  Index = NULL,
  CloseChannel = TRUE,
  Verbose = TRUE,
  Test = FALSE,
  NAString = "NA",
  Fast = TRUE
)
}
\arguments{
\item{DataToPush}{Update data table in warehouse with new values}

\item{DBConnection}{RemixAutoML::SQL_Server_DBConnection()}

\item{SQLTableName}{The SQL statement you want to run}

\item{Index}{Column name of index}

\item{CloseChannel}{TRUE to close when done, FALSE to leave the channel open}

\item{Verbose}{TRUE or FALSE}

\item{Test}{Set to TRUE to see if what you plan to do will work}

\item{NAString}{Supply character string to supply missing values}

\item{Fast}{Set to TRUE to update table in one shot versus row by row}
}
\description{
SQL_UpdateTable update a database table
}
\seealso{
Other Database: 
\code{\link{AutoDataDictionaries}()},
\code{\link{SQL_ClearTable}()},
\code{\link{SQL_DropTable}()},
\code{\link{SQL_Query_Push}()},
\code{\link{SQL_Query}()},
\code{\link{SQL_SaveTable}()},
\code{\link{SQL_Server_BulkPull}()},
\code{\link{SQL_Server_BulkPush}()},
\code{\link{SQL_Server_DBConnection}()}
}
\author{
Adrian Antico
}
\concept{Database}
