% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EconometricsFunctions.R
\name{StackedTimeSeriesEnsembleForecast}
\alias{StackedTimeSeriesEnsembleForecast}
\title{TimeSeriesEnsembleForecast}
\usage{
StackedTimeSeriesEnsembleForecast(TS_Models = c("arima", "tbats",
  "nnet"), ML_Methods = c("CatBoost", "XGBoost", "H2oGBM", "H2oDRF"),
  CalendarFeatures = TRUE, HolidayFeatures = TRUE,
  FourierFeatures = NULL, Path = "C:/Users/aantico/Documents/Package",
  TargetName = "Weekly_Sales", DateName = "Date", NTrees = 750,
  TaskType = "GPU", GridTune = FALSE, FCPeriods = 5,
  MaxNumberModels = 5)
}
\arguments{
\item{TS_Models}{Select which ts model forecasts to ensemble}

\item{ML_Methods}{Select which models to build for the ensemble}

\item{CalendarFeatures}{TRUE or FALSE}

\item{HolidayFeatures}{TRUE or FALSE}

\item{FourierFeatures}{Full set of fourier features for train and score}

\item{Path}{The path to the folder where the ts forecasts are stored}

\item{TargetName}{"Weekly_Sales"}

\item{DateName}{"Date"}

\item{NTrees}{Select the number of trees to utilize in ML models}

\item{GridTune}{Set to TRUE to grid tune the ML models}

\item{FCPeriods}{Number of periods to forecast}

\item{MaxNumberModels}{The number of models to try for each ML model}
}
\description{
TimeSeriesEnsembleForecast to generate forecasts and ensemble data
}
\seealso{
Other Time Series: \code{\link{AutoBanditNNet}},
  \code{\link{AutoBanditSarima}}, \code{\link{AutoTBATS}},
  \code{\link{CarmaHoldoutMetrics}},
  \code{\link{DifferenceDataReverse}},
  \code{\link{DifferenceData}},
  \code{\link{FinalBuildArfima}},
  \code{\link{FinalBuildArima}},
  \code{\link{FinalBuildETS}},
  \code{\link{FinalBuildNNET}},
  \code{\link{FinalBuildTBATS}},
  \code{\link{FinalBuildTSLM}},
  \code{\link{GenerateParameterGrids}},
  \code{\link{OptimizeArfima}},
  \code{\link{OptimizeArima}}, \code{\link{OptimizeETS}},
  \code{\link{OptimizeNNET}}, \code{\link{OptimizeTBATS}},
  \code{\link{OptimizeTSLM}},
  \code{\link{ParallelAutoARIMA}},
  \code{\link{ParallelAutoArfima}},
  \code{\link{ParallelAutoETS}},
  \code{\link{ParallelAutoNNET}},
  \code{\link{ParallelAutoTBATS}},
  \code{\link{ParallelAutoTSLM}},
  \code{\link{RL_Performance}},
  \code{\link{Regular_Performance}},
  \code{\link{TimeSeriesDataPrepare}},
  \code{\link{WideTimeSeriesEnsembleForecast}}
}
\author{
Adrian Antico
}
\concept{Time Series}
