% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureEngineering_NumericTypes.R
\name{StandardizeScoring}
\alias{StandardizeScoring}
\title{StandardizeScoring}
\usage{
StandardizeScoring(data, ScoreTable, Apply = "apply", GroupVars = NULL)
}
\arguments{
\item{data}{Source data.table}

\item{Apply}{'apply' or 'backtransform'}

\item{GroupVars}{Character vector of column names to have percent ranks by the group levels}

\item{ColNames}{Character vector of column names}

\item{Center}{TRUE}

\item{Scale}{TRUE}
}
\description{
Generate standardized values for multiple variables, by groups if provided, and with a selected granularity
}
\examples{
\dontrun{
x <- Standardize(data = data, ColNames = c('Weekly_Sales', 'XREG1'), GroupVars = c('Region','Store','Dept'), Center = TRUE, Scale = TRUE)
}

}
\seealso{
Other Feature Engineering: 
\code{\link{AutoDataPartition}()},
\code{\link{AutoDiffLagN}()},
\code{\link{AutoHierarchicalFourier}()},
\code{\link{AutoInteraction}()},
\code{\link{AutoLagRollMode}()},
\code{\link{AutoLagRollStatsScoring}()},
\code{\link{AutoLagRollStats}()},
\code{\link{AutoTransformationCreate}()},
\code{\link{AutoTransformationScore}()},
\code{\link{AutoWord2VecModeler}()},
\code{\link{AutoWord2VecScoring}()},
\code{\link{CategoricalEncoding}()},
\code{\link{CreateCalendarVariables}()},
\code{\link{CreateHolidayVariables}()},
\code{\link{DummifyDT}()},
\code{\link{H2OAutoencoderScoring}()},
\code{\link{H2OAutoencoder}()},
\code{\link{ModelDataPrep}()},
\code{\link{PercRankScoring}()},
\code{\link{PercRank}()},
\code{\link{Standardize}()},
\code{\link{TimeSeriesFillRoll}()},
\code{\link{TimeSeriesFill}()}
}
\author{
Adrian Antico
}
\concept{Feature Engineering}
