% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyOutputs.R
\name{StandardPlotsModal2}
\alias{StandardPlotsModal2}
\title{StandardPlotsModal2}
\usage{
StandardPlotsModal2(id, PlotNumber, AppWidth = 12L)
}
\arguments{
\item{id}{= 'PlotVariablesModals'}

\item{PlotNumber}{1}

\item{AppWidth}{= 12L}
}
\description{
Plot Variables Modals
}
\seealso{
Other Shiny: 
\code{\link{BoxPlotModals}()},
\code{\link{DataTable}()},
\code{\link{StandardPlotsModal1}()}
}
\author{
Adrian Antico
}
\concept{Shiny}
