% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotFunctions.R
\name{StockData}
\alias{StockData}
\title{StockData}
\usage{
StockData(
  PolyOut = NULL,
  Symbol = "TSLA",
  CompanyName = "Tesla Inc. Common Stock",
  Metric = "Stock Price",
  TimeAgg = "days",
  StartDate = "2022-01-01",
  EndDate = "2022-01-01",
  APIKey = NULL
)
}
\arguments{
\item{PolyOut}{NULL. If NULL, data is pulled. If supplied, data is not pulled.}

\item{Symbol}{ticker symbol string}

\item{CompanyName}{company name if you have it. ends up in title, that is all}

\item{Metric}{Stock Price, Percent Returns (use symbol for percent), Percent Log Returns (use symbol for percent), Index, Quadratic Variation}

\item{TimeAgg}{= 'days', 'weeks', 'months'}

\item{StartDate}{Supply a start date. E.g. '2022-01-01'}

\item{EndDate}{Supply an end date. E.g. `Sys.Date()`}

\item{APIKey}{Supply your polygon API key}

\item{Type}{'candlestick', 'ohlc'}
}
\description{
Create stock data for plotting using StockPlot()
}
\seealso{
Other Graphics: 
\code{\link{AddFacet}()},
\code{\link{BarPlot}()},
\code{\link{BoxPlot}()},
\code{\link{ChartTheme}()},
\code{\link{CorrMatrixPlot}()},
\code{\link{DensityPlot}()},
\code{\link{HeatMapPlot}()},
\code{\link{HistPlot}()},
\code{\link{PlotlyConversion}()},
\code{\link{StockPlot}()},
\code{\link{ViolinPlot}()},
\code{\link{multiplot}()}
}
\author{
Adrian Antico
}
\concept{Graphics}
