% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotFunctions.R
\name{StockPlot}
\alias{StockPlot}
\title{StockPlot}
\usage{
StockPlot(StockDataOutput, Type = "candlestick")
}
\arguments{
\item{StockDataOutput}{PolyOut returned from StockData()}

\item{Type}{'candlestick', 'ohlc'}
}
\description{
Create a candlestick plot for stocks. See https://plotly.com/r/figure-labels/
}
\seealso{
Other Graphics: 
\code{\link{AddFacet}()},
\code{\link{BarPlot}()},
\code{\link{BoxPlot}()},
\code{\link{ChartTheme}()},
\code{\link{CorrMatrixPlot}()},
\code{\link{HeatMapPlot}()},
\code{\link{HistPlot}()},
\code{\link{PlotlyConversion}()},
\code{\link{StockData}()},
\code{\link{ViolinPlot}()},
\code{\link{multiplot}()}
}
\author{
Adrian Antico
}
\concept{Graphics}
