% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyFunctions.R
\name{StoreArgs}
\alias{StoreArgs}
\title{Store Args values within a project}
\usage{
StoreArgs(input, ProjectList, VarName, Type, Default)
}
\arguments{
\item{input}{This is the input value within a Shiny context}

\item{ProjectList}{This is the VarNameument collection list}

\item{VarName}{The name of the VarNameument you want to store}

\item{Type}{"character" "numeric" "logical"}

\item{Default}{default value that gets saved}
}
\value{
Updates ProjectList inside function. Do not assign function to anything
}
\description{
Automatically save VarNameuments to project list
}
\examples{
\dontrun{
StoreVarNames(input, ProjectList, "NTrees", "numeric", 1000)
}

}
\seealso{
Other Shiny: 
\code{\link{ArgNullCheck2}()},
\code{\link{ArgNullCheck}()},
\code{\link{DateInput}()},
\code{\link{GenerateEvaluationMetrics}()},
\code{\link{NumericInput}()},
\code{\link{PickerInput_GetLevels}()},
\code{\link{PickerInput}()},
\code{\link{PreparePlotData}()},
\code{\link{ReturnParam}()},
\code{\link{TextInput}()}
}
\author{
Adrian Antico
}
\concept{Shiny}
