% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Adrian_Antico_Modeling_Tools_Library.R
\name{tempDatesFun}
\alias{tempDatesFun}
\title{tempDatesFun Convert Excel datetime char columns to Date columns}
\usage{
tempDatesFun(x)
}
\value{
An object to pass along to ggplot objects following the "+" sign
}
\description{
tempDatesFun takes the Excel datetime column, which imports as character, and converts it into a date type
}
\examples{
Cdata[, DAY_DATE := tempDatesFun(DAY_DATE)]
Cdata[, DAY_DATE := as.Date(DAY_DATE, "\%m/\%d/\%Y")]
}
\author{
Adrian Antico at RemixInstitute.com
}
