% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshOptim.R
\name{threshOptim}
\alias{threshOptim}
\title{threshOptim}
\usage{
threshOptim(
  data,
  actTar = "target",
  predTar = "p1",
  tpProfit = 0,
  tnProfit = 0,
  fpProfit = -1,
  fnProfit = -2,
  MinThresh = 0.001,
  MaxThresh = 0.999,
  ThresholdPrecision = 0.001
)
}
\arguments{
\item{data}{data is the data table you are building the modeling on}

\item{actTar}{The column name where the actual target variable is located (in binary form)}

\item{predTar}{The column name where the predicted values are located}

\item{tpProfit}{This is the utility for generating a true positive prediction}

\item{tnProfit}{This is the utility for generating a true negative prediction}

\item{fpProfit}{This is the cost of generating a false positive prediction}

\item{fnProfit}{This is the cost of generating a false negative prediction}

\item{MinThresh}{Minimum value to consider for model threshold}

\item{MaxThresh}{Maximum value to consider for model threshold}

\item{ThresholdPrecision}{Incrementing value in search}
}
\value{
Optimal threshold and corresponding utilities for the range of thresholds tested
}
\description{
threshOptim will return the utility maximizing threshold for future predictions along with the data generated to estimate the threshold
}
\examples{
\dontrun{
data <- data.table::data.table(Target = runif(10))
data[, x1 := qnorm(Target)]
data[, x2 := runif(10)]
data[, Predict := log(pnorm(0.85 * x1 + sqrt(1-0.85^2) * qnorm(x2)))]
data[, ':=' (x1 = NULL, x2 = NULL)]
data <- threshOptim(data     = data,
                    actTar   = "Target",
                    predTar  = "Predict",
                    tpProfit = 0,
                    tnProfit = 0,
                    fpProfit = -1,
                    fnProfit = -2,
                    MinThresh = 0.001,
                    MaxThresh = 0.999,
                    ThresholdPrecision = 0.001)
optimalThreshold <- data$Thresholds
allResults <- data$EvaluationTable
}
}
\seealso{
Other Model Evaluation and Interpretation: 
\code{\link{AutoLimeAid}()},
\code{\link{EvalPlot}()},
\code{\link{LimeModel}()},
\code{\link{ML_EvalPlots}()},
\code{\link{ParDepCalPlots}()},
\code{\link{ROCPlot}()},
\code{\link{RedYellowGreen}()},
\code{\link{VI_Plot}()}
}
\author{
Adrian Antico
}
\concept{Model Evaluation and Interpretation}
