% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EconometricsFunctions.R
\name{TimeSeriesDataPrepare}
\alias{TimeSeriesDataPrepare}
\title{TimeSeriesDataPrepare}
\usage{
TimeSeriesDataPrepare(
  data,
  TargetName,
  DateName,
  Lags,
  SeasonalLags,
  MovingAverages,
  SeasonalMovingAverages,
  TimeUnit,
  FCPeriods,
  HoldOutPeriods,
  TSClean = TRUE,
  ModelFreq = TRUE,
  FinalBuild = FALSE
)
}
\arguments{
\item{data}{Source data.table for forecasting}

\item{TargetName}{Name of your target variable}

\item{DateName}{Name of your date variable}

\item{Lags}{The max number of lags you want to test}

\item{SeasonalLags}{The max number of seasonal lags you want to test}

\item{MovingAverages}{The max number of moving average terms}

\item{SeasonalMovingAverages}{The max number of seasonal moving average terms}

\item{TimeUnit}{The level of aggregation your dataset comes in. Choices include: 1Min, 5Min, 10Min, 15Min, and 30Min, hour, day, week, month, quarter, year}

\item{FCPeriods}{The number of forecast periods you want to have forecasted}

\item{HoldOutPeriods}{The number of holdout samples to compare models against}

\item{TSClean}{TRUE or FALSE. TRUE will kick off a time series cleaning operation. Outliers will be smoothed and imputation will be conducted.}

\item{ModelFreq}{TRUE or FALSE. TRUE will enable a model-based time frequency calculation for an alternative frequency value to test models on.}

\item{FinalBuild}{Set to TRUE to create data sets with full data}
}
\value{
Time series data sets to pass onto auto modeling functions
}
\description{
TimeSeriesDataPrepare is a function that takes raw data and returns the necessary time series data and objects for model building. It also fills any time gaps with zeros. Use this before you run any time series model functions.
}
\examples{
\dontrun{
data <- data.table::fread(
  file.path(PathNormalizer(
    "C:\\\\Users\\\\aantico\\\\Documents\\\\Package\\\\data"),
    "tsdata.csv"))
TimeSeriesDataPrepare(
  data = data,
  TargetName = "Weekly_Sales",
  DateName = "Date",
  Lags = 5,
  MovingAverages,
  SeasonalMovingAverages,
  SeasonalLags = 1,
  TimeUnit = "week",
  FCPeriods = 10,
  HoldOutPeriods = 10,
  TSClean = TRUE,
  ModelFreq = TRUE,
  FinalBuild = FALSE)
}
}
\author{
Adrian Antico
}
\concept{Time Series Helper}
