% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureEngineering_DataSets.R
\name{TimeSeriesFillRoll}
\alias{TimeSeriesFillRoll}
\title{TimeSeriesFillRoll}
\usage{
TimeSeriesFillRoll(
  data = NULL,
  DateColumnName = NULL,
  RollVars = NULL,
  NonRollVars = NULL,
  GroupVariables = NULL,
  RollDirection = "backward",
  TimeUnit = "days",
  SimpleImpute = FALSE
)
}
\arguments{
\item{data}{Supply your full series data set here}

\item{DateColumnName}{Supply the name of your date column}

\item{RollVars}{= NULL,}

\item{NonRollVars}{= NULL,}

\item{GroupVariables}{Supply the column names of your group variables. E.g. "Group" or c("Group1","Group2")}

\item{RollDirection}{'backward' or 'forward'}

\item{TimeUnit}{Choose from "second", "minute", "hour", "day", "week", "month", "quarter", "year"}

\item{SimpleImpute}{Set to TRUE or FALSE. With TRUE numeric cols will fill NAs with a 0 and non-numeric cols with a "0"}
}
\value{
Returns a data table with missing time series records filled (currently just zeros)
}
\description{
TimeSeriesFillRoll For Completing Time Series Data For Single Series or Time Series by Group
}
\examples{
\dontrun{

# Pull in data
data <- data <- data.table::fread("https://www.dropbox.com/s/2str3ek4f4cheqi/walmart_train.csv?dl=1")

# Run function
data <- TimeSeriesFillRoll(
  data,
  RollVars = c('Net_Revenue', 'Units', 'SIZE_UNITS', 'Liters', 'Accum_Units'),
  NonRollVars = c('Diff_1_DATE_ISO','Net_Revenue_PerDay','Liters_PerDay','Units_PerDay'),
  DateColumnName = "Date",
  GroupVariables = c("Store","Dept"),
  RollDirection = 'backward',
  TimeUnit = "weeks",
  SimpleImpute = FALSE)
}
}
\seealso{
Other Feature Engineering: 
\code{\link{AutoDataPartition}()},
\code{\link{AutoDiffLagN}()},
\code{\link{AutoHierarchicalFourier}()},
\code{\link{AutoInteraction}()},
\code{\link{AutoLagRollMode}()},
\code{\link{AutoLagRollStatsScoring}()},
\code{\link{AutoLagRollStats}()},
\code{\link{AutoTransformationCreate}()},
\code{\link{AutoTransformationScore}()},
\code{\link{AutoWord2VecModeler}()},
\code{\link{AutoWord2VecScoring}()},
\code{\link{CategoricalEncoding}()},
\code{\link{CreateCalendarVariables}()},
\code{\link{CreateHolidayVariables}()},
\code{\link{DummifyDT}()},
\code{\link{H2OAutoencoderScoring}()},
\code{\link{H2OAutoencoder}()},
\code{\link{ModelDataPrep}()},
\code{\link{PercRankScoring}()},
\code{\link{PercRank}()},
\code{\link{StandardizeScoring}()},
\code{\link{Standardize}()},
\code{\link{TimeSeriesFill}()}
}
\author{
Adrian Antico
}
\concept{Feature Engineering}
