% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataWranglingFunctions.R
\name{TimeSeriesMelt}
\alias{TimeSeriesMelt}
\title{TimeSeriesMelt}
\usage{
TimeSeriesMelt(
  data,
  TargetVariable = NULL,
  DateVariable = NULL,
  GroupVariables = NULL
)
}
\arguments{
\item{data}{source data}

\item{TargetVariable}{vector of target variable names}

\item{DateVariable}{Name of date variable}

\item{GroupVariables}{Vector of group variable names}
}
\description{
TimeSeriesMelt
}
\seealso{
Other Data Wrangling: 
\code{\link{AutoDataDictionaries}()},
\code{\link{ColumnSubsetDataTable}()},
\code{\link{DataDisplayMeta}()},
\code{\link{FakeDataGenerator}()},
\code{\link{FullFactorialCatFeatures}()},
\code{\link{SQL_ClearTable}()},
\code{\link{SQL_DropTable}()},
\code{\link{SQL_Query_Push}()},
\code{\link{SQL_Query}()},
\code{\link{SQL_SaveTable}()},
\code{\link{SQL_Server_DBConnection}()},
\code{\link{SQL_UpdateTable}()}
}
\author{
Adrian Antico
}
\concept{Data Wrangling}
