% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotFunctions.R
\name{TimeSeriesPlotter}
\alias{TimeSeriesPlotter}
\title{TimeSeriesPlotter}
\usage{
TimeSeriesPlotter(
  data = data,
  TargetVariable = "TargetVariableName",
  DateVariable = "DateVariableName",
  GroupVariables = "GroupVariableName",
  EvaluationMode = FALSE,
  VLineDate = NULL,
  Aggregate = NULL,
  NumberGroupsDisplay = 5,
  LevelsToDisplay = NULL,
  OtherGroupLabel = "Other",
  DisplayOtherGroup = FALSE,
  TextSize = 12,
  LineWidth = 1,
  Color = "blue",
  XTickMarks = "1 year",
  AngleX = 35,
  AngleY = 0,
  ChartColor = "lightsteelblue1",
  BorderColor = "darkblue",
  TextColor = "darkblue",
  GridColor = "white",
  BackGroundColor = "gray95",
  LegendPosition = "bottom",
  LegendTextColor = "darkblue",
  LegendTextSize = 10,
  ForecastLineColor = "black",
  PredictionIntervals = FALSE,
  TS_ModelID = NULL,
  SSForecast = FALSE,
  PredictionIntervalColorInner = "aquamarine1",
  PredictionIntervalColorOuter = "peachpuff1"
)
}
\arguments{
\item{data}{Source data}

\item{TargetVariable}{Target variable}

\item{DateVariable}{Date variable}

\item{GroupVariables}{Group variables}

\item{EvaluationMode}{TRUE means two lines are displayed for Actual and Forecast}

\item{VLineDate}{Date of last actual target value}

\item{Aggregate}{Choose from 'sum' or 'mean'}

\item{NumberGroupsDisplay}{Number of lines to display}

\item{LevelsToDisplay}{Value}

\item{OtherGroupLabel}{Label to call all other group levels}

\item{DisplayOtherGroup}{If TRUE, a line will be shown with all levels that fall into 'other' otherwise no line will be shown}

\item{TextSize}{Default 12}

\item{LineWidth}{Numeric value. Default is 1}

\item{Color}{Set to "blue", "red", etc}

\item{XTickMarks}{Number of tick marks on x-axis. "1 minute","15 minutes","30 minutes","1 hour","3 hour","6 hour","12 hour","1 day","3 day","1 week","2 week","1 month","3 month","6 month","1 year","2 year","5 year","10 year"}

\item{AngleX}{Angle of text on x axis}

\item{AngleY}{Angle of text on y axis}

\item{ChartColor}{Color of chart background}

\item{BorderColor}{Color of border}

\item{TextColor}{Text color}

\item{GridColor}{Grid color}

\item{BackGroundColor}{Background color}

\item{LegendPosition}{Legend position}

\item{LegendTextColor}{Text color}

\item{LegendTextSize}{Text size}

\item{ForecastLineColor}{Forecast line color}

\item{PredictionIntervals}{Set to TRUE to plot prediction intervals}

\item{TS_ModelID}{Select a model from the list for forecasting viewer}

\item{SSForecast}{Default FALSE. Set to TRUE for single series models}

\item{PredictionIntervalColorInner}{Fills 20th to 80th percentiles}

\item{PredictionIntervalColorOuter}{Fills 5th to 20th and 80th to 95th percentiles}
}
\description{
TimeSeriesPlotter is a function to plot single or multiple lines on a single plot
}
\seealso{
Other Graphics: 
\code{\link{RemixTheme}()},
\code{\link{multiplot}()}
}
\author{
Adrian Antico
}
\concept{Graphics}
