% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Adrian_Antico_Modeling_Tools_Library.R
\name{tokenizeH20}
\alias{tokenizeH20}
\title{For NLP work}
\usage{
tokenizeH20(data3)
}
\arguments{
\item{data3}{The text data}

\item{sentences}{Source data table to merge vects onto}

\item{stop.words}{A string name for the column to convert via word2vec}
}
\description{
This function tokenizes data
}
\author{
Adrian Antico at RemixInstitute.com
}
