% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TextPrepFunctions.R
\name{tokenizeH2O}
\alias{tokenizeH2O}
\title{For NLP work}
\usage{
tokenizeH2O(data)
}
\arguments{
\item{data}{The text data}
}
\description{
This function tokenizes text data
}
\examples{
\dontrun{
data <- tokenizeH2O(data = data[["StringColumn"]])
}
}
\seealso{
Other Misc: 
\code{\link{AutoH2OTextPrepScoring}()},
\code{\link{DeleteFile}()},
\code{\link{LB}()},
\code{\link{Logger}()},
\code{\link{PrintToPDF}()}
}
\author{
Adrian Antico
}
\concept{Misc}
