% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyFunctions.R
\name{UniqueLevels}
\alias{UniqueLevels}
\title{UniqueLevels}
\usage{
UniqueLevels(input, data, n, GroupVars = NULL)
}
\arguments{
\item{input}{passthrough}

\item{data}{data.table}

\item{GroupVars}{passthrough}
}
\description{
UniqueLevels
}
