% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EDAFunctions.R
\name{UserBaseEvolution}
\alias{UserBaseEvolution}
\title{UserBaseEvolution}
\usage{
UserBaseEvolution(
  data,
  Entity = NULL,
  DateColumnName = NULL,
  TimeAgg = NULL,
  ChurnPeriods = 1
)
}
\arguments{
\item{data}{Source data.table}

\item{Entity}{Column name of the entity / user}

\item{DateColumnName}{Name of the date column used for inclusion of users in time periods}

\item{TimeAgg}{Choose from 'Month', 'Week', or 'Day'}

\item{ChurnPeriods}{Defaults to 1. This means for TimeAgg = 'Month' a one month churn period is used. For TimeAgg = 'Week' you will have a one week churn period. If you set ChurnPeriods to 2 then it will be a 2 month churn or a 2 week churn. Same logic applies for daily.}
}
\description{
This function creates a table of user counts over time for accumulated unique users, active unique users, new unique users, retained unique users, churned unique users, and reactivated unique users. You can run this with several specifications. You can request monthly, weekly, or daily counts and you can specify a churn window for the computations.
}
\seealso{
Other EDA: 
\code{\link{AutoWordFreq}()},
\code{\link{EDA_Histograms}()},
\code{\link{Mode}()},
\code{\link{PlotGUI}()},
\code{\link{ScatterCopula}()}
}
\author{
Adrian Antico
}
\concept{EDA}
