% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artifact_schema.R
\name{validate_artifact_collection}
\alias{validate_artifact_collection}
\title{Validate an AutoQuant Artifact Collection}
\usage{
validate_artifact_collection(artifacts)
}
\arguments{
\item{artifacts}{A list of artifacts.}
}
\value{
A `data.table` with artifact-level validation checks.
}
\description{
Validate an AutoQuant Artifact Collection
}
\seealso{
Other AutoQuant artifact validation: 
\code{\link{validate_artifact}()}
}
\concept{AutoQuant artifact validation}
