% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotFunctions.R
\name{VaporWaveTheme}
\alias{VaporWaveTheme}
\title{VaporWaveTheme}
\usage{
VaporWaveTheme(Size1 = 1, Size2 = 0.5, Size3 = 12, LegendPosition = "bottom")
}
\arguments{
\item{Size1}{Grid line size}

\item{Size2}{Axis line sizes}

\item{Size3}{Legend line sizes}

\item{LegendPosition}{"bottom" is default}
}
\description{
VaporWaveTheme
}
\seealso{
Other Misc: 
\code{\link{AutoH2OTextPrepScoring}()},
\code{\link{ChartTheme}()},
\code{\link{PrintToPDF}()},
\code{\link{tokenizeH2O}()}
}
\author{
Adrian Antico
}
\concept{Misc}
