% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelEvaluationPlots.R
\name{VI_Plot}
\alias{VI_Plot}
\title{VI_Plot}
\usage{
VI_Plot(Type = "catboost", VI_Data, ColorHigh = "darkblue", ColorLow = "white")
}
\arguments{
\item{Type}{'catboost', 'xgboost', 'h2o'}

\item{VI_Data}{Source data}
}
\value{
ROC Plot for classification models
}
\description{
Generate variable importance plots
}
\seealso{
Other Model Evaluation and Interpretation: 
\code{\link{AutoLimeAid}()},
\code{\link{EvalPlot}()},
\code{\link{LimeModel}()},
\code{\link{ParDepCalPlots}()},
\code{\link{ROCPlot}()},
\code{\link{RedYellowGreen}()},
\code{\link{threshOptim}()}
}
\author{
Adrian Antico
}
\concept{Model Evaluation and Interpretation}
