% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotFunctions.R
\name{ViolinPlot}
\alias{ViolinPlot}
\title{ViolinPlot}
\usage{
ViolinPlot(
  data = NULL,
  XVar = NULL,
  YVar = NULL,
  FacetVar1 = NULL,
  FacetVar2 = NULL,
  SampleSize = 1000000L,
  FillColor = "gray",
  YTicks = "Default",
  XTicks = "Default",
  TextSize = 12,
  AngleX = 90,
  AngleY = 0,
  ChartColor = "lightsteelblue1",
  BorderColor = "darkblue",
  TextColor = "darkblue",
  GridColor = "white",
  BackGroundColor = "gray95",
  SubTitleColor = "blue",
  LegendPosition = "bottom",
  LegendBorderSize = 0.5,
  LegendLineType = "solid",
  Debug = FALSE
)
}
\arguments{
\item{data}{Source data.table}

\item{XVar}{Column name of X-Axis variable. If NULL then ignored}

\item{YVar}{Column name of Y-Axis variable. If NULL then ignored}

\item{FacetVar1}{Column name of facet variable 1. If NULL then ignored}

\item{FacetVar2}{Column name of facet variable 2. If NULL then ignored}

\item{SampleSize}{An integer for the number of rows to use. Sampled data is randomized. If NULL then ignored}

\item{FillColor}{'gray'}

\item{YTicks}{Choose from 'Default', 'Percentiles', 'Every 5th percentile', 'Deciles', 'Quantiles', 'Quartiles'}

\item{XTicks}{Choose from 'Default', '1 year', '1 day', '3 day', '1 week', '2 week', '1 month', '3 month', '6 month', '2 year', '5 year', '10 year', '1 minute', '15 minutes', '30 minutes', '1 hour', '3 hour', '6 hour', '12 hour'}

\item{TextSize}{14}

\item{AngleX}{90}

\item{AngleY}{0}

\item{ChartColor}{'lightsteelblue'}

\item{BorderColor}{'darkblue'}

\item{TextColor}{'darkblue'}

\item{GridColor}{'white'}

\item{BackGroundColor}{'gray95'}

\item{SubTitleColor}{'darkblue'}

\item{LegendPosition}{'bottom'}

\item{LegendBorderSize}{0.50}

\item{LegendLineType}{'solid'}

\item{Debug}{FALSE}
}
\description{
Build a violin plot by simply passing arguments to a single function. It will sample your data using SampleSize number of rows. Sampled data is randomized.
}
\seealso{
Other EDA: 
\code{\link{AutoWordFreq}()},
\code{\link{BarPlot}()},
\code{\link{BoxPlot}()},
\code{\link{EDA_Histograms}()},
\code{\link{HistPlot}()},
\code{\link{Mode}()},
\code{\link{PlotGUI}()},
\code{\link{ScatterCopula}()},
\code{\link{UserBaseEvolution}()}
}
\author{
Adrian Antico
}
\concept{EDA}
