% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyFunctions.R
\name{withConsoleRedirect}
\alias{withConsoleRedirect}
\title{withConsoleRedirect}
\usage{
withConsoleRedirect(containerId, expr)
}
\arguments{
\item{containerId}{Passthrough}

\item{expr}{Code}
}
\description{
withConsoleRedirect
}
