% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Adrian_Antico_Modeling_Tools_Library.R
\name{Word2VecModel}
\alias{Word2VecModel}
\title{Automated word2vec data generation via H20}
\usage{
Word2VecModel(datax, stringCol = c("TITLE", "DESCR", "LI_NAME",
  "ACTION_TAKEN_TEXT", "SUG_DISP"), KeepStringCol = FALSE,
  model_path = "Query_Multilabel_Models", ModelID = c("TITLE", "DESCR",
  "LI_NAME", "ACTION_TAKEN_TEXT", "SUG_DISP"), vects = 25,
  SaveStopWords = FALSE, MinWords = 1, WindowSize = 1, Epochs = 25,
  StopWords = NULL)
}
\arguments{
\item{stringCol}{A string name for the column to convert via word2vec}

\item{KeepStringCol}{Set to TRUE if you want to keep the original string column that you convert via word2vec}

\item{model_path}{A string path to the location where you want the model and metadata stored}

\item{ModelID}{A vector of your model names}

\item{vects}{The number of vectors to retain from the word2vec model}

\item{SaveStopWords}{Set to TRUE to save the stop words used}

\item{MinWords}{For H20 word2vec model}

\item{WindowSize}{For H20 word2vec model}

\item{Epochs}{For H20 word2vec model}

\item{StopWords}{For H20 word2vec model}

\item{data}{Source data table to merge vects onto}
}
\description{
This function allows you to automatically build a word2vec model and merge the data onto your supplied dataset
}
\examples{
Word2VecModel(data,
             stringCol = "Comment",
             KeepStringCol = FALSE,
             model_path = getwd(),
             vects = 50,
             SaveStopWords = FALSE)
}
\author{
Adrian Antico
}
