% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Adrian_Antico_Modeling_Tools_Library.R
\name{WordFreq}
\alias{WordFreq}
\title{Automated Word Frequency and Word Cloud Creation}
\usage{
WordFreq(data, TextColName = "DESCR",
  ClusterCol = "ClusterAllNoTarget", ClusterID = 0,
  RemoveEnglishStopwords = TRUE, Stemming = TRUE,
  StopWords = c("blabla1", "blabla2"))
}
\arguments{
\item{data}{Source data table}

\item{TextColName}{A string name for the column}

\item{ClusterCol}{Set to NULL to ignore, otherwise set to Cluster column name (or factor column name)}

\item{ClusterID}{Must be set if ClusterCol is defined. Set to cluster ID (or factor level)}

\item{RemoveEnglishStopwords}{Set to TRUE to remove English stop words, FALSE to ignore}

\item{Stemming}{Set to TRUE to run stemming on your text data}

\item{StopWords}{Add your own stopwords, in vector format}
}
\description{
This function builds a word frequency table and a word cloud. It prepares data, cleans text, and generates output.
}
\examples{
Word2VecModel(data,
             stringCol     = "Comment",
             KeepStringCol = FALSE,
             model_path    = getwd(),
             ModelID       = "bla",
             vects         = 50,
             SaveStopWords = FALSE,
             MinWords      = 1,
             WindowSize    = 1,
             Epochs        = 25,
             StopWords     = NULL)
}
\author{
Adrian Antico
}
