% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XGBoostHelpers.R
\name{XGBoostArgsCheck}
\alias{XGBoostArgsCheck}
\title{XGBoostArgsCheck}
\usage{
XGBoostArgsCheck(
  GridTune. = GridTune,
  model_path. = model_path,
  metadata_path. = metadata_path,
  Trees. = NTrees,
  max_depth. = max_depth,
  eta. = eta,
  min_child_weight. = min_child_weight,
  subsample. = subsample,
  colsample_bytree. = colsample_bytree
)
}
\arguments{
\item{GridTune.}{Passthrough}

\item{model_path.}{Passthrough}

\item{metadata_path.}{Passthrough}

\item{Trees.}{Passthrough}

\item{max_depth.}{Passthrough}

\item{eta.}{Passthrough}

\item{min_child_weight.}{Passthrough}

\item{subsample.}{Passthrough}

\item{colsample_bytree.}{Passthrough}
}
\value{
A list containing data.table's with the parameters shuffled and ready to test in the bandit framework
}
\description{
XGBoostArgsCheck
}
\seealso{
Other XGBoost Helpers: 
\code{\link{XGBoostClassifierParams}()},
\code{\link{XGBoostDataPrep}()},
\code{\link{XGBoostFinalParams}()},
\code{\link{XGBoostMultiClassParams}()},
\code{\link{XGBoostRegressionMetrics}()},
\code{\link{XGBoostRegressionParams}()}
}
\author{
Adrian Antico
}
\concept{XGBoost Helpers}
