% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XGBoostHelpers.R
\name{XGBoostDataPrep}
\alias{XGBoostDataPrep}
\title{XGBoostDataPrep}
\usage{
XGBoostDataPrep(
  ModelType = "regression",
  data. = data,
  ValidationData. = ValidationData,
  TestData. = TestData,
  TargetColumnName. = TargetColumnName,
  FeatureColNames. = FeatureColNames,
  IDcols. = IDcols,
  TransformNumericColumns. = TransformNumericColumns,
  Methods. = Methods,
  ModelID. = ModelID,
  model_path. = model_path,
  TrainOnFull. = TrainOnFull,
  SaveModelObjects. = SaveModelObjects,
  ReturnFactorLevels. = ReturnFactorLevels
)
}
\arguments{
\item{ModelType}{'regression', 'vector', 'classification', or 'multiclass'}

\item{data.}{Passthrough}

\item{ValidationData.}{Passthrough}

\item{TestData.}{Passthrough}

\item{TargetColumnName.}{Passthrough}

\item{FeatureColNames.}{Passthrough}

\item{IDcols.}{Passthrough}

\item{TransformNumericColumns.}{Passthrough regression}

\item{Methods.}{Passthrough regression}

\item{ModelID.}{Passthrough regression}

\item{model_path.}{Passthrough regression}

\item{TrainOnFull.}{Passthrough}

\item{SaveModelObjects.}{Passthrough}

\item{ReturnFactorLevels.}{Passthrough}

\item{PrimaryDateColumn.}{Passthrough}

\item{DummifyCols.}{Passthrough regression}

\item{LossFunction.}{Passthrough regression}

\item{EvalMetric.}{Passthrough regression}
}
\description{
Prepare data for XGBoost modeling
}
\seealso{
Other XGBoost Helpers: 
\code{\link{XGBoostArgsCheck}()},
\code{\link{XGBoostClassifierParams}()},
\code{\link{XGBoostFinalParams}()},
\code{\link{XGBoostMultiClassParams}()},
\code{\link{XGBoostRegressionMetrics}()},
\code{\link{XGBoostRegressionParams}()}
}
\author{
Adrian Antico
}
\concept{XGBoost Helpers}
