% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyFunctions.R
\name{XGBoostEvalMetricOptions}
\alias{XGBoostEvalMetricOptions}
\title{XGBoostEvalMetricOptions}
\usage{
XGBoostEvalMetricOptions(TT)
}
\arguments{
\item{TT}{Target Type: 'regression', 'classification', 'multiclass', all lower case}
}
\description{
XGBoostEvalMetricOptions
}
\keyword{internal}
