% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XGBoostHelpers.R
\name{XGBoostFinalParams}
\alias{XGBoostFinalParams}
\title{XGBoostFinalParams}
\usage{
XGBoostFinalParams(
  GridTune. = GridTune,
  PassInGrid. = PassInGrid,
  TrainOnFull. = TrainOnFull,
  LossFunction. = LossFunction,
  eval_metric. = eval_metric,
  NThreads. = NThreads,
  TreeMethod. = TreeMethod,
  BestGrid. = BestGrid,
  Trees. = Trees,
  NumLevels. = NumLevels
)
}
\arguments{
\item{PassInGrid.}{Passthrough}

\item{LossFunction.}{Passthrough}

\item{eval_metric.}{Passthrough}

\item{TreeMethod.}{Passthrough}

\item{BestGrid.}{Passthrough}

\item{Trees.}{Passthrough}

\item{NumLevels.}{Passthrough}

\item{NThreads}{Passthrough}
}
\description{
Parameters for xgboost fitting
}
\seealso{
Other XGBoost Helpers: 
\code{\link{XGBoostArgsCheck}()},
\code{\link{XGBoostDataPrep}()},
\code{\link{XGBoostGridParams}()},
\code{\link{XGBoostRegressionMetrics}()}
}
\author{
Adrian Antico
}
\concept{XGBoost Helpers}
