% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XGBoostHelpers.R
\name{XGBoostGridParams}
\alias{XGBoostGridParams}
\title{XGBoostGridParams}
\usage{
XGBoostGridParams(
  N. = N,
  counter. = NULL,
  NThreads. = -1L,
  Objective. = "reg:logistic",
  BanditArmsN. = NULL,
  EvalMetric. = NULL,
  TreeMethod. = NULL,
  model_path. = NULL,
  NewGrid. = NULL,
  Grid. = NULL,
  GridClusters. = NULL
)
}
\arguments{
\item{N.}{Passthrough}

\item{counter.}{Passthrough}

\item{NThreads.}{= -1L,}

\item{Objective.}{Passthrough}

\item{BanditArmsN.}{Passthrough}

\item{EvalMetric.}{Passthrough}

\item{TreeMethod.}{Passthrough}

\item{model_path.}{Passthrough}

\item{NewGrid.}{Passthrough}

\item{Grid.}{Passthrough}

\item{GridClusters.}{Passthrough}
}
\description{
XGBoostGridParams
}
\seealso{
Other XGBoost Helpers: 
\code{\link{XGBoostArgsCheck}()},
\code{\link{XGBoostDataPrep}()},
\code{\link{XGBoostFinalParams}()},
\code{\link{XGBoostRegressionMetrics}()}
}
\author{
Adrian Antico
}
\concept{XGBoost Helpers}
