% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GridTuning.R
\name{XGBoostGridTuner}
\alias{XGBoostGridTuner}
\title{XGBoostGridTuner}
\usage{
XGBoostGridTuner(
  ModelType = "classification",
  TrainOnFull. = TrainOnFull,
  DebugMode. = DebugMode.,
  TreeMethod. = TreeMethod,
  Trees. = Trees,
  Depth. = max_depth,
  LearningRate. = eta,
  min_child_weight. = min_child_weight,
  subsample. = subsample,
  colsample_bytree. = colsample_bytree,
  LossFunction = LossFunction,
  EvalMetric = eval_metric,
  grid_eval_metric. = grid_eval_metric,
  CostMatrixWeights = CostMatrixWeights,
  TargetColumnName. = TargetColumnName,
  datatrain. = datatrain,
  datavalidate. = datavalidate,
  datatest. = datatest,
  EvalSets. = EvalSets,
  TestTarget. = TestTarget,
  FinalTestTarget. = FinalTestTarget,
  TargetLevels. = TargetLevels,
  MaxRunsWithoutNewWinner = MaxRunsWithoutNewWinner,
  MaxModelsInGrid = MaxModelsInGrid,
  MaxRunMinutes = MaxRunMinutes,
  BaselineComparison. = BaselineComparison,
  SaveModelObjects = SaveModelObjects,
  metadata_path = metadata_path,
  model_path = model_path,
  ModelID = ModelID,
  Verbose. = Verbose,
  NumLevels. = NumLevels
)
}
\arguments{
\item{ModelType}{"classification"}

\item{TrainOnFull.}{TrainOnFull}

\item{DebugMode.}{DebugMode}

\item{TreeMethod.}{TreeMethod}

\item{Trees.}{Trees}

\item{Depth.}{max_depth}

\item{LearningRate.}{eta}

\item{min_child_weight.}{min_child_weight}

\item{subsample.}{subsample}

\item{colsample_bytree.}{colsample_bytree}

\item{LossFunction}{LossFunction}

\item{EvalMetric}{EvalMetric}

\item{grid_eval_metric.}{MultiClass}

\item{CostMatrixWeights}{CostMatrixWeights}

\item{TargetColumnName.}{TargetColumnName}

\item{datatrain.}{datatrain}

\item{datavalidate.}{datavalidate}

\item{datatest.}{datatest}

\item{EvalSets.}{EvalSets}

\item{TestTarget.}{TestTarget}

\item{FinalTestTarget.}{FinalTestTarget}

\item{TargetLevels.}{TargetLevels}

\item{MaxRunsWithoutNewWinner}{MaxRunsWithoutNewWinner}

\item{MaxModelsInGrid}{MaxModelsInGrid}

\item{MaxRunMinutes}{MaxRunMinutes}

\item{BaselineComparison.}{BaselineComparison}

\item{SaveModelObjects}{SaveModelObjects}

\item{metadata_path}{metadata_path}

\item{model_path}{model_path}

\item{ModelID}{ModelID}

\item{Verbose.}{Verbose}

\item{NumLevels.}{NumLevels}

\item{MetricPeriods}{MetricPeriods}

\item{ClassWeights}{ClassWeights}
}
\description{
Execute grid tuning for xgboost
}
\seealso{
Other Reinforcement Learning: 
\code{\link{CatBoostGridTuner}()},
\code{\link{RL_Initialize}()},
\code{\link{RL_ML_Update}()},
\code{\link{RL_Update}()},
\code{\link{RPM_Binomial_Bandit}()}
}
\author{
Adrian Antico
}
\concept{Reinforcement Learning}
