% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyFunctions.R
\name{XGBoostLossFunctionOptions}
\alias{XGBoostLossFunctionOptions}
\title{XGBoostLossFunctionOptions}
\usage{
XGBoostLossFunctionOptions(TT)
}
\arguments{
\item{TT}{Target Type: 'regression', 'classification', 'multiclass', all lower case}
}
\description{
XGBoostLossFunctionOptions
}
\keyword{internal}
