% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XGBoostHelpers.R
\name{XGBoostMultiClassParams}
\alias{XGBoostMultiClassParams}
\title{XGBoostMultiClassParams}
\usage{
XGBoostMultiClassParams(
  counter = NULL,
  num_class = NULL,
  Objective = "multi:softmax",
  NThreads = -1L,
  BanditArmsN = NULL,
  eval_metric = NULL,
  task_type = NULL,
  model_path = NULL,
  NewGrid = NULL,
  Grid = NULL,
  ExperimentalGrid = NULL,
  GridClusters = NULL
)
}
\arguments{
\item{counter}{Passthrough}

\item{num_class}{NULL}

\item{Objective}{Passthrough}

\item{NThreads}{= -1L,}

\item{BanditArmsN}{Passthrough}

\item{eval_metric}{Passthrough}

\item{task_type}{Passthrough}

\item{model_path}{Passthrough}

\item{NewGrid}{Passthrough}

\item{Grid}{Passthrough}

\item{ExperimentalGrid}{Passthrough}

\item{GridClusters}{Passthrough}
}
\description{
XGBoostMultiClassParams
}
\seealso{
Other XGBoost Helpers: 
\code{\link{XGBoostArgsCheck}()},
\code{\link{XGBoostClassifierParams}()},
\code{\link{XGBoostDataPrep}()},
\code{\link{XGBoostFinalParams}()},
\code{\link{XGBoostRegressionMetrics}()},
\code{\link{XGBoostRegressionParams}()}
}
\author{
Adrian Antico
}
\concept{XGBoost Helpers}
