% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReinforcementLearningFunctions.R
\name{XGBoostParameterGrids}
\alias{XGBoostParameterGrids}
\title{XGBoostParameterGrids}
\usage{
XGBoostParameterGrids(
  TaskType = "CPU",
  Shuffles = 1L,
  NTrees = seq(500L, 5000L, 500L),
  Depth = seq(4L, 16L, 2L),
  LearningRate = seq(0.05, 0.4, 0.05),
  MinChildWeight = seq(1, 10, 1),
  SubSample = seq(0.55, 1, 0.05),
  ColSampleByTree = seq(0.55, 1, 0.05)
)
}
\arguments{
\item{TaskType}{"GPU" or "CPU"}

\item{Shuffles}{The number of shuffles you want to apply to each grid}

\item{NTrees}{seq(500L, 5000L, 500L)}

\item{Depth}{seq(4L, 16L, 2L)}

\item{LearningRate}{seq(0.05,0.40,0.05)}

\item{MinChildWeight}{seq(1.0, 10.0, 1.0)}

\item{SubSample}{seq(0.55, 1.0, 0.05)}

\item{ColSampleByTree}{seq(0.55, 1.0, 0.05)}
}
\value{
A list containing data.table's with the parameters shuffled and ready to test in the bandit framework
}
\description{
XGBoostParameterGrids
}
\seealso{
Other Supervised Learning: 
\code{\link{AutoH2OScoring}()},
\code{\link{XGBoostClassifierParams}()},
\code{\link{XGBoostMultiClassParams}()},
\code{\link{XGBoostRegressionMetrics}()},
\code{\link{XGBoostRegressionParams}()}
}
\author{
Adrian Antico
}
\concept{Supervised Learning}
