% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XGBoostHelpers.R
\name{XGBoostRegressionMetrics}
\alias{XGBoostRegressionMetrics}
\title{XGBoostRegressionMetrics}
\usage{
XGBoostRegressionMetrics(grid_eval_metric, MinVal, calibEval)
}
\arguments{
\item{grid_eval_metric}{Passthrough}

\item{MinVal}{= -1L,}

\item{calibEval}{Passthrough}
}
\description{
XGBoostRegressionMetrics
}
\seealso{
Other XGBoost Helpers: 
\code{\link{XGBoostArgsCheck}()},
\code{\link{XGBoostDataPrep}()},
\code{\link{XGBoostFinalParams}()},
\code{\link{XGBoostGridParams}()}
}
\author{
Adrian Antico
}
\concept{XGBoost Helpers}
