% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReinforcementLearningFunctions.R
\name{XGBoostRegressionParams}
\alias{XGBoostRegressionParams}
\title{XGBoostRegressionParams}
\usage{
XGBoostRegressionParams(
  counter = NULL,
  NThreads = -1L,
  BanditArmsN = NULL,
  objective = NULL,
  eval_metric = NULL,
  task_type = NULL,
  model_path = NULL,
  NewGrid = NULL,
  Grid = NULL,
  ExperimentalGrid = NULL,
  GridClusters = NULL
)
}
\arguments{
\item{counter}{Passthrough}

\item{NThreads}{= -1L,}

\item{BanditArmsN}{Passthrough}

\item{objective}{Passthrough}

\item{eval_metric}{Passthrough}

\item{task_type}{Passthrough}

\item{model_path}{Passthrough}

\item{NewGrid}{Passthrough}

\item{Grid}{Passthrough}

\item{ExperimentalGrid}{Passthrough}

\item{GridClusters}{Passthrough}
}
\description{
XGBoostRegressionParams
}
\seealso{
Other Supervised Learning: 
\code{\link{AutoH2OScoring}()},
\code{\link{CatBoostClassifierParams}()},
\code{\link{CatBoostMultiClassParams}()},
\code{\link{CatBoostParameterGrids}()},
\code{\link{CatBoostRegressionParams}()},
\code{\link{XGBoostClassifierParams}()},
\code{\link{XGBoostMultiClassParams}()},
\code{\link{XGBoostParameterGrids}()},
\code{\link{XGBoostRegressionMetrics}()}
}
\author{
Adrian Antico
}
\concept{Supervised Learning}
