% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimpleApps.R
\name{App_TimeSeriesPlots}
\alias{App_TimeSeriesPlots}
\title{App_TimeSeriesPlots}
\usage{
App_TimeSeriesPlots(
  data,
  FeatureNames = NULL,
  GroupVariables = NULL,
  FilterVariable = NULL,
  DateName = NULL,
  AppWidth = 10L,
  LogoWidth = "1200px",
  LogoHeight = "300px",
  Debug = FALSE
)
}
\arguments{
\item{data}{Source data.table}

\item{FeatureNames}{Character vector of feature column names}

\item{GroupVariables}{Character vector of group column names}

\item{FilterVariable}{Variable to use to filter data before plotting}

\item{DateName}{Character scalar of the date column name}

\item{AppWidth}{Width of boxes}

\item{Debug}{FALSE}
}
\description{
Simple shiny app for viewing time series plots with box plots or line plots. You can use up to 3 categorical variables to filter by and one additional variable to filter as a bonus. You'll need to have shiny, shinyWidgets, htmltools
}
\examples{
\dontrun{
# Pull Data
data <- data.table::fread(system.file('tests/QA_DataSets/ThreeGroup-FC-Walmart-XREG3.csv', package = "RemixAutoML"))
data[, Date := as.Date(Date)]

# Run App
RemixAutoML::App_TimeSeriesPlots(
  data,
  FeatureNames = names(data)[5L:ncol(data)],
  GroupVariables = names(data)[seq_len(3L)],
  FilterVariable = 'XREG1',
  DateName = 'Date',
  AppWidth = 9L,
  LogoWidth = '1100px',
  LogoHeight = '259px',
  Debug = TRUE)

# FeatureNames = names(data)[5L:ncol(data)]
# GroupVariables = names(data)[seq_len(3L)]
# FilterVariable = 'XREG1'
# DateName = 'Date'
# Debug = TRUE
# AppWidth = 9L
# LogoWidth = '1100px',
# LogoHeight = '259px',
# Debug = TRUE
}

}
\seealso{
Other Shiny: 
\code{\link{ArgNullCheck2}()},
\code{\link{ArgNullCheck}()},
\code{\link{BlankRow}()},
\code{\link{DateInput}()},
\code{\link{GenerateEvaluationMetrics}()},
\code{\link{NumericInput}()},
\code{\link{PickerInput_GetLevels}()},
\code{\link{PickerInput}()},
\code{\link{PreparePlotData}()},
\code{\link{ReactiveLoadCSV}()},
\code{\link{ReturnParam}()},
\code{\link{StoreArgs}()},
\code{\link{TextInput}()},
\code{\link{observeEventLoad}()}
}
\author{
Adrian Antico
}
\concept{Shiny}
