% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotFunctions.R
\name{AppModelInsights}
\alias{AppModelInsights}
\title{AppModelInsights}
\usage{
AppModelInsights(
  ModelOutputList,
  dt = NULL,
  PlotType = NULL,
  TargetVar = NULL,
  PredictVar = NULL,
  PDPVar = NULL,
  DateVar = NULL,
  FacetVar1 = NULL,
  FacetVar2 = NULL,
  GamFit = FALSE,
  Buckets = 20,
  ShapAgg = "mean",
  Rebuild = FALSE,
  Check2 = FALSE,
  Debug = FALSE
)
}
\arguments{
\item{ModelOutputList}{output list}

\item{dt}{= NULL,}

\item{PlotType}{= NULL,}

\item{TargetVar}{= isolate(YVar()),}

\item{PredictVar}{= isolate(ScoreVar()),}

\item{PDPVar}{= NULL,}

\item{DateVar}{= isolate(DateVar()),}

\item{FacetVar1}{= NULL}

\item{FacetVar2}{= NULL}

\item{GamFit}{= FALSE,}

\item{Buckets}{= 20,}

\item{ShapAgg}{A string for aggregating shapely values for importances. Choices include, 'mean', 'absmean', 'meanabs', 'geomean', 'harmmean', 'sd', 'median', 'absmedian', 'medianabs'}

\item{Rebuild}{= FALSE,}

\item{Check2}{= FALSE,}

\item{Debug}{= FALSE}
}
\description{
Plot model insights in app
}
\seealso{
Other Graphics: 
\code{\link{AddFacet}()},
\code{\link{AutoPlotter}()},
\code{\link{BarPlot}()},
\code{\link{BoxPlot}()},
\code{\link{ChartTheme}()},
\code{\link{HistPlot}()},
\code{\link{ViolinPlot}()},
\code{\link{multiplot}()}
}
\author{
Adrian Antico
}
\concept{Graphics}
