% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimpleApps.R
\name{AppsPlotting}
\alias{AppsPlotting}
\title{AppsPlotting}
\usage{
AppsPlotting(
  BlobStorageURL = NULL,
  IFrameLocation = NULL,
  PlotObjectHome = NULL,
  HeaderColor = "black",
  AppWidth = 8L,
  LogoWidth = "750px",
  LogoHeight = "100px",
  GroupVarsBoxColor = "navy",
  VarsBoxColor = "purple",
  FilterBoxColor = "blue",
  PlotBoxColor = "aqua",
  CreatePlotButtonColor = "royal",
  UpdatePlotButtonColor = "default",
  ResetPlotButtonColor = "default",
  H3Color = "darkblue",
  H4Color = "darkblue",
  AppTextColor = "blue",
  Browser = FALSE,
  Docker = FALSE,
  UserName_Password_DT = NULL,
  Debug = FALSE
)
}
\arguments{
\item{BlobStorageURL}{is the URL path that leads up to the file name. E.g. 'http://localhost:5000/BlobFiles/ViewFile?name='}

\item{IFrameLocation}{Browser location}

\item{PlotObjectHome}{NULL}

\item{HeaderColor}{'black', 'blue', 'purple', 'green', 'red', 'yellow'}

\item{AppWidth}{Width of boxes}

\item{GroupVarsBoxColor}{Choose from 'red', 'yellow', 'aqua', 'blue', 'light-blue', 'green', 'navy', 'teal', 'olive', 'lime', 'orange', 'fuchsia', 'purple', 'maroon', 'black'}

\item{VarsBoxColor}{Choose from 'red', 'yellow', 'aqua', 'blue', 'light-blue', 'green', 'navy', 'teal', 'olive', 'lime', 'orange', 'fuchsia', 'purple', 'maroon', 'black'}

\item{FilterBoxColor}{Choose from 'red', 'yellow', 'aqua', 'blue', 'light-blue', 'green', 'navy', 'teal', 'olive', 'lime', 'orange', 'fuchsia', 'purple', 'maroon', 'black'}

\item{PlotBoxColor}{Choose from 'red', 'yellow', 'aqua', 'blue', 'light-blue', 'green', 'navy', 'teal', 'olive', 'lime', 'orange', 'fuchsia', 'purple', 'maroon', 'black'}

\item{CreatePlotButtonColor}{Choose from 'default', 'primary', 'warning', 'danger', 'success', 'royal'}

\item{UpdatePlotButtonColor}{Choose from 'default', 'primary', 'warning', 'danger', 'success', 'royal'}

\item{ResetPlotButtonColor}{Choose from 'default', 'primary', 'warning', 'danger', 'success', 'royal'}

\item{H3Color}{Header colors 'darkblue'}

\item{H4Color}{Text below headers but not the input cells; 'blue2'}

\item{AppTextColor}{'white'}

\item{Browser}{FALSE}

\item{Docker}{FALSE}

\item{UserName_Password_DT}{NULL. In order to enforce authentication, supply a data.table with columns 'UserName' which contains the names of your users and 'Password' which contains the acceptable passwords. E.g. data.table::data.table(UserName = c('Adrian Antico', 'Guest'), Password = c('Password1', 'Password2')). Case sensitivity applies.}

\item{Debug}{FALSE}
}
\description{
Simple shiny app for viewing time series plots with box plots or line plots. You can use up to 3 categorical variables to filter by and one additional variable to filter as a bonus. You'll need to have shiny, shinyWidgets, htmltools
}
\examples{
\dontrun{
# Pull Data
# data <- data.table::fread(system.file('tests/QA_DataSets/ThreeGroup-FC-Walmart-XREG3.csv', package = "RemixAutoML"))
# data[, Date := as.Date(Date)]

# Run App
# RemixAutoML::AppsPlotting(
#   BlobStorageURL = NULL,
#   IFrameLocation = NULL,
#   PlotObjectHome = NULL,
#   HeaderColor = 'black',
#   AppWidth = 12L,
#   LogoWidth = '1000px',
#   LogoHeight = '100px',
#   GroupVarsBoxColor = 'navy',
#   VarsBoxColor = 'purple',
#   FilterBoxColor = 'blue',
#   PlotBoxColor = 'aqua',
#   CreatePlotButtonColor = 'default',
#   UpdatePlotButtonColor = 'default',
#   ResetPlotButtonColor = 'default',
#   H3Color = 'darkblue',
#   H5Color = 'blue2',
#   AppTextColor = 'white',
#   Docker = FALSE,
#   Browser = FALSE,
#   UserName_Password_DT = NULL,
#   Debug = FALSE)
#
# Debug = TRUE
# HeaderColor = 'black'
# AppWidth = 12L
# LogoWidth = '1000px'
# LogoHeight = '100px'
# GroupVarsBoxColor = 'navy'
# VarsBoxColor = 'purple'
# FilterBoxColor = 'blue'
# PlotBoxColor = 'aqua'
# CreatePlotButtonColor = 'default'
# UpdatePlotButtonColor = 'default'
# ResetPlotButtonColor = 'default'
# H3Color = 'darkblue'
# H4Color = 'blue2'
# AppTextColor = 'white'
# Docker = FALSE
# Browser = FALSE
# UserName_Password_DT = NULL
# Debug = TRUE
}

}
\seealso{
Other GUI: 
\code{\link{RunRemixAutoML}()}
}
\author{
Adrian Antico
}
\concept{GUI}
