% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimpleApps.R
\name{AppsPlotting}
\alias{AppsPlotting}
\title{AppsPlotting}
\usage{
AppsPlotting(
  data,
  XVariable = NULL,
  YVariable = NULL,
  DateName = NULL,
  GroupVariables = NULL,
  FilterVariable = NULL,
  ModelOutputList = NULL,
  HeaderColor = "black",
  AppWidth = 12L,
  LogoWidth = "1000px",
  LogoHeight = "100px",
  GroupVarsBoxColor = "navy",
  VarsBoxColor = "purple",
  FilterBoxColor = "blue",
  PlotBoxColor = "aqua",
  CreatePlotButtonColor = "primary",
  UpdatePlotButtonColor = "primary",
  ResetPlotButtonColor = "primary",
  Browser = FALSE,
  Debug = FALSE
)
}
\arguments{
\item{data}{Source data.table}

\item{XVariable}{Starter column name for x-variable. Not needed if x-axis is the DateName column}

\item{YVariable}{Starter column name for y-variable}

\item{DateName}{Starter column name for date-variable}

\item{GroupVariables}{Starter column name for group-variables}

\item{FilterVariable}{Starter column name for filter-variable}

\item{ModelOutputList}{ML output from RemixAutoML Auto___() SL functions}

\item{HeaderColor}{'black', 'blue', 'purple', 'green', 'red', 'yellow'}

\item{AppWidth}{Width of boxes}

\item{GroupVarsBoxColor}{Choose from 'red', 'yellow', 'aqua', 'blue', 'light-blue', 'green', 'navy', 'teal', 'olive', 'lime', 'orange', 'fuchsia', 'purple', 'maroon', 'black'}

\item{VarsBoxColor}{Choose from 'red', 'yellow', 'aqua', 'blue', 'light-blue', 'green', 'navy', 'teal', 'olive', 'lime', 'orange', 'fuchsia', 'purple', 'maroon', 'black'}

\item{FilterBoxColor}{Choose from 'red', 'yellow', 'aqua', 'blue', 'light-blue', 'green', 'navy', 'teal', 'olive', 'lime', 'orange', 'fuchsia', 'purple', 'maroon', 'black'}

\item{PlotBoxColor}{Choose from 'red', 'yellow', 'aqua', 'blue', 'light-blue', 'green', 'navy', 'teal', 'olive', 'lime', 'orange', 'fuchsia', 'purple', 'maroon', 'black'}

\item{CreatePlotButtonColor}{Choose from 'default', 'primary', 'warning', 'danger', 'success', 'royal'}

\item{UpdatePlotButtonColor}{Choose from 'default', 'primary', 'warning', 'danger', 'success', 'royal'}

\item{ResetPlotButtonColor}{Choose from 'default', 'primary', 'warning', 'danger', 'success', 'royal'}

\item{Browser}{FALSE}

\item{Debug}{FALSE}
}
\description{
Simple shiny app for viewing time series plots with box plots or line plots. You can use up to 3 categorical variables to filter by and one additional variable to filter as a bonus. You'll need to have shiny, shinyWidgets, htmltools
}
\examples{
\dontrun{
# Pull Data
data <- data.table::fread(system.file('tests/QA_DataSets/ThreeGroup-FC-Walmart-XREG3.csv', package = "RemixAutoML"))
data[, Date := as.Date(Date)]

# Run App
RemixAutoML::AppsPlotting(
  data,
  XVariable = NULL,
  YVariable = 'XREG1',
  DateName = 'Date',
  GroupVariables = names(data)[seq_len(3L)],
  FilterVariable = 'XREG1',
  ModelOutputList = NULL,
  HeaderColor = 'black',
  AppWidth = 12L,
  LogoWidth = '1000px',
  LogoHeight = '100px',
  GroupVarsBoxColor = 'navy',
  VarsBoxColor = 'purple',
  FilterBoxColor = 'blue',
  PlotBoxColor = 'aqua',
  CreatePlotButtonColor = 'default',
  UpdatePlotButtonColor = 'default',
  ResetPlotButtonColor = 'default',
  Debug = FALSE)

# XVariable = 'Date'
# YVariable = 'XREG1'
# DateName = 'Date'
# GroupVariables = names(data)[seq_len(3L)]
# FilterVariable = 'XREG1'
# ModelOutputList = NULL
# Debug = TRUE
# HeaderColor = 'black'
# AppWidth = 12L,
# LogoWidth = '1000px',
# LogoHeight = '100px',
# GroupVarsBoxColor = 'navy',
# VarsBoxColor = 'purple',
# FilterBoxColor = 'blue'
# PlotBoxColor = 'aqua',
# CreatePlotButtonColor = 'default',
# UpdatePlotButtonColor = 'default',
# ResetPlotButtonColor = 'default',
# Debug = TRUE
}

}
\seealso{
Other GUI: 
\code{\link{RunRemixAutoML}()}
}
\author{
Adrian Antico
}
\concept{GUI}
