% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoCatBoostHurdleModel.R
\name{AutoCatBoostHurdleModel}
\alias{AutoCatBoostHurdleModel}
\title{AutoCatBoostHurdleModel for generalized hurdle modeling}
\usage{
AutoCatBoostHurdleModel(data, ValidationData = NULL, TestData = NULL,
  Buckets = 0, TargetColumnName = NULL, FeatureColNames = NULL,
  PrimaryDateColumn = NULL, IDcols = NULL,
  TransformNumericColumns = NULL, ClassWeights = NULL,
  SplitRatios = c(0.7, 0.2, 0.1), task_type = "GPU",
  ModelID = "ModelTest", Paths = NULL, MetaDataPaths = NULL,
  SaveModelObjects = TRUE, Trees = 1000, GridTune = TRUE,
  MaxModelsInGrid = 1, NumOfParDepPlots = 10, PassInGrid = NULL)
}
\arguments{
\item{data}{Source training data. Do not include a column that has the class labels for the buckets as they are created internally.}

\item{ValidationData}{Source validation data. Do not include a column that has the class labels for the buckets as they are created internally.}

\item{TestData}{Souce test data. Do not include a column that has the class labels for the buckets as they are created internally.}

\item{Buckets}{A numeric vector of the buckets used for subsetting the data. NOTE: the final Bucket value will first create a subset of data that is less than the value and a second one thereafter for data greater than the bucket value.}

\item{TargetColumnName}{Supply the column name or number for the target variable}

\item{FeatureColNames}{Supply the column names or number of the features (not included the PrimaryDateColumn)}

\item{PrimaryDateColumn}{Supply a date column if the data is functionally related to it}

\item{IDcols}{Includes PrimaryDateColumn and any other columns you want returned in the validation data with predictions}

\item{TransformNumericColumns}{Transform numeric column inside the AutoCatBoostRegression() function}

\item{ClassWeights}{Utilize these for the classifier model}

\item{SplitRatios}{Supply vector of partition ratios. For example, c(0.70,0.20,0,10).}

\item{task_type}{Set to "GPU" or "CPU"}

\item{ModelID}{Define a character name for your models}

\item{Paths}{The path to your folder where you want your model information saved}

\item{MetaDataPaths}{TA character string of your path file to where you want your model evaluation output saved. If left NULL, all output will be saved to Paths.}

\item{SaveModelObjects}{Set to TRUE to save the model objects to file in the folders listed in Paths}

\item{Trees}{Default 15000}

\item{GridTune}{Set to TRUE if you want to grid tune the models}

\item{MaxModelsInGrid}{Set to a numeric value for the number of models to try in grid tune}

\item{NumOfParDepPlots}{Set to pull back N number of partial dependence calibration plots.}

\item{PassInGrid}{Pass in a grid for changing up the parameter settings for catboost}
}
\value{
Returns AutoCatBoostRegression() model objects: VariableImportance.csv, Model, ValidationData.csv, EvalutionPlot.png, EvalutionBoxPlot.png, EvaluationMetrics.csv, ParDepPlots.R a named list of features with partial dependence calibration plots, ParDepBoxPlots.R, GridCollect, and catboostgrid
}
\description{
AutoCatBoostHurdleModel for generalized hurdle modeling. Check out the Readme.Rd on github for more background.
}
\examples{
\donttest{
Output <- RemixAutoML::AutoCatBoostHurdleModel(
  data,
  ValidationData = NULL,
  TestData = NULL,
  Buckets = 1,
  TargetColumnName = "Target_Variable",
  FeatureColNames = 4:ncol(data),
  PrimaryDateColumn = "Date",
  IDcols = 1:3,
  TransformNumericColumns = NULL,
  ClassWeights = NULL,
  SplitRatios = c(0.7, 0.2, 0.1),
  task_type = "GPU",
  ModelID = "ModelID",
  Paths = NULL,
  MetaDataPaths = NULL,
  SaveModelObjects = TRUE,
  Trees = 1000,
  GridTune = FALSE,
  MaxModelsInGrid = 1,
  NumOfParDepPlots = 10,
  PassInGrid = NULL)
}
}
\seealso{
Other Automated Regression: \code{\link{AutoCatBoostRegression}},
  \code{\link{AutoH2oDRFHurdleModel}},
  \code{\link{AutoH2oDRFRegression}},
  \code{\link{AutoH2oGBMHurdleModel}},
  \code{\link{AutoH2oGBMRegression}},
  \code{\link{AutoH2oGLMRegression}},
  \code{\link{AutoH2oMLRegression}}, \code{\link{AutoNLS}},
  \code{\link{AutoXGBoostHurdleModel}},
  \code{\link{AutoXGBoostRegression}}
}
\author{
Adrian Antico
}
\concept{Automated Regression}
