% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoCatBoostHurdleModel.R
\name{AutoCatBoostHurdleModel}
\alias{AutoCatBoostHurdleModel}
\title{AutoCatBoostHurdleModel for generalized hurdle modeling}
\usage{
AutoCatBoostHurdleModel(
  data = NULL,
  TrainOnFull = FALSE,
  ValidationData = NULL,
  TestData = NULL,
  Buckets = 0L,
  TargetColumnName = NULL,
  FeatureColNames = NULL,
  PrimaryDateColumn = NULL,
  IDcols = NULL,
  TransformNumericColumns = NULL,
  Methods = c("BoxCox", "Asinh", "Asin", "Log", "LogPlus1", "Logit", "YeoJohnson"),
  ClassWeights = NULL,
  SplitRatios = c(0.7, 0.2, 0.1),
  task_type = "GPU",
  ModelID = "ModelTest",
  Paths = NULL,
  MetaDataPaths = NULL,
  SaveModelObjects = FALSE,
  ReturnModelObjects = TRUE,
  NumOfParDepPlots = 10L,
  PassInGrid = NULL,
  GridTune = FALSE,
  BaselineComparison = "default",
  MaxModelsInGrid = 1L,
  MaxRunsWithoutNewWinner = 20L,
  MaxRunMinutes = 60L * 60L,
  Shuffles = 2L,
  MetricPeriods = 25L,
  Trees = seq(1000L, 5000L, 500L),
  Depth = 6,
  RandomStrength = 1,
  BorderCount = 128,
  LearningRate = seq(0.01, 0.1, 0.01),
  L2_Leaf_Reg = 3,
  RSM = c(0.8, 0.85, 0.9, 0.95, 1),
  BootStrapType = c("Bayesian", "Bernoulli", "Poisson", "MVS", "No"),
  GrowPolicy = c("SymmetricTree", "Depthwise", "Lossguide")
)
}
\arguments{
\item{data}{Source training data. Do not include a column that has the class labels for the buckets as they are created internally.}

\item{TrainOnFull}{Set to TRUE to use all data}

\item{ValidationData}{Source validation data. Do not include a column that has the class labels for the buckets as they are created internally.}

\item{TestData}{Souce test data. Do not include a column that has the class labels for the buckets as they are created internally.}

\item{Buckets}{A numeric vector of the buckets used for subsetting the data. NOTE: the final Bucket value will first create a subset of data that is less than the value and a second one thereafter for data greater than the bucket value.}

\item{TargetColumnName}{Supply the column name or number for the target variable}

\item{FeatureColNames}{Supply the column names or number of the features (not included the PrimaryDateColumn)}

\item{PrimaryDateColumn}{Supply a date column if the data is functionally related to it}

\item{IDcols}{Includes PrimaryDateColumn and any other columns you want returned in the validation data with predictions}

\item{TransformNumericColumns}{Transform numeric column inside the AutoCatBoostRegression() function}

\item{Methods}{Choose transformation methods}

\item{ClassWeights}{Utilize these for the classifier model}

\item{SplitRatios}{Supply vector of partition ratios. For example, c(0.70,0.20,0,10).}

\item{task_type}{Set to "GPU" or "CPU"}

\item{ModelID}{Define a character name for your models}

\item{Paths}{The path to your folder where you want your model information saved}

\item{MetaDataPaths}{TA character string of your path file to where you want your model evaluation output saved. If left NULL, all output will be saved to Paths.}

\item{SaveModelObjects}{Set to TRUE to save the model objects to file in the folders listed in Paths}

\item{ReturnModelObjects}{TRUE to return the models}

\item{NumOfParDepPlots}{Set to pull back N number of partial dependence calibration plots.}

\item{PassInGrid}{Pass in a grid for changing up the parameter settings for catboost}

\item{GridTune}{Set to TRUE if you want to grid tune the models}

\item{BaselineComparison}{= "default",}

\item{MaxModelsInGrid}{= 1L,}

\item{MaxRunsWithoutNewWinner}{= 20L,}

\item{MaxRunMinutes}{= 60L*60L,}

\item{Shuffles}{= 2L,}

\item{MetricPeriods}{= 25L,}

\item{Trees}{= seq(1000L, 5000L, 500L),}

\item{Depth}{= seq(4L, 8L, 1L),}

\item{RandomStrength}{1}

\item{BorderCount}{128}

\item{LearningRate}{= seq(0.01,0.10,0.01),}

\item{L2_Leaf_Reg}{= seq(1.0, 10.0, 1.0),}

\item{RSM}{= c(0.80, 0.85, 0.90, 0.95, 1.0),}

\item{BootStrapType}{= c("Bayesian", "Bernoulli", "Poisson", "MVS", "No"),}

\item{GrowPolicy}{= c("SymmetricTree", "Depthwise", "Lossguide")}
}
\value{
Returns AutoCatBoostRegression() model objects: VariableImportance.csv, Model, ValidationData.csv, EvalutionPlot.png, EvalutionBoxPlot.png, EvaluationMetrics.csv, ParDepPlots.R a named list of features with partial dependence calibration plots, ParDepBoxPlots.R, GridCollect, and catboostgrid
}
\description{
AutoCatBoostHurdleModel for generalized hurdle modeling. Check out the Readme.Rd on github for more background.
}
\examples{
\dontrun{
Output <- RemixAutoML::AutoCatBoostHurdleModel(

  # Operationalization
  task_type = "GPU",
  ModelID = "ModelTest",
  SaveModelObjects = FALSE,
  ReturnModelObjects = TRUE,

  # Data related args
  data = data,
  TrainOnFull = FALSE,
  ValidationData = NULL,
  TestData = NULL,
  Buckets = 0L,
  TargetColumnName = NULL,
  FeatureColNames = NULL,
  PrimaryDateColumn = NULL,
  IDcols = NULL,

  # Metadata args
  Paths = normalizePath("./"),
  MetaDataPaths = NULL,
  TransformNumericColumns = NULL,
  Methods =
     c("BoxCox", "Asinh", "Asin", "Log",
       "LogPlus1", "Logit", "YeoJohnson"),
  ClassWeights = NULL,
  SplitRatios = c(0.70, 0.20, 0.10),
  NumOfParDepPlots = 10L,

  # Grid tuning setup
  PassInGrid = NULL,
  GridTune = FALSE,
  BaselineComparison = "default",
  MaxModelsInGrid = 1L,
  MaxRunsWithoutNewWinner = 20L,
  MaxRunMinutes = 60L*60L,
  Shuffles = 2L,
  MetricPeriods = 25L,

  # Bandit grid args
  Trees = seq(1000L, 5000L, 500L),
  Depth = seq(4L, 8L, 1L),
  LearningRate = seq(0.01,0.10,0.01),
  L2_Leaf_Reg = seq(1.0, 10.0, 1.0),
  RandomStrength = 1,
  BorderCount = 128,
  RSM = c(0.80, 0.85, 0.90, 0.95, 1.0),
  BootStrapType = c("Bayesian", "Bernoulli",
     "Poisson", "MVS", "No"),
  GrowPolicy = c("SymmetricTree", "Depthwise", "Lossguide"))
}
}
\seealso{
Other Automated Regression: 
\code{\link{AutoH2oDRFHurdleModel}()},
\code{\link{AutoH2oGBMHurdleModel}()},
\code{\link{AutoNLS}()},
\code{\link{AutoXGBoostHurdleModel}()}
}
\author{
Adrian Antico
}
\concept{Automated Regression}
