% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RemixAutoML.R
\name{AutoCatBoostScoring}
\alias{AutoCatBoostScoring}
\title{AutoCatBoostScoring is an automated scoring function that compliments the AutoCatBoost model training functions.}
\usage{
AutoCatBoostScoring(TargetType = NULL, ScoringData = NULL,
  FeatureColNames = NULL, CatFeatures = NULL, IDcols = NULL,
  ModelPath = NULL, ModelID = NULL, ReturnFeatures = TRUE,
  MDP_Impute = TRUE, MDP_CharToFactor = TRUE, MDP_RemoveDates = TRUE,
  MDP_MissFactor = "0", MDP_MissNum = -1)
}
\arguments{
\item{TargetType}{Set this value to "regression", "classification", or "multiclass" to score models built using AutoCatBoostRegression(), AutoCatBoostClassify() or AutoCatBoostMultiClass().}

\item{ScoringData}{This is your data.table of features for scoring. Can be a single row or batch.}

\item{CatFeatures}{Supply either column names or number for your categorical columns used in the AutoCatBoostRegression() function}

\item{IDcols}{Supply ID column numbers for any metadata you want returned with your predicted values}

\item{ModelPath}{Supply your path file used in the AutoCatBoost__() function}

\item{ModelID}{Supply the model ID used in the AutoCatBoost__() function}

\item{ReturnFeatures}{Set to TRUE to return your features with the predicted values.}

\item{MDP_Impute}{Set to TRUE if you did so for modeling and didn't do so before supplying ScoringData in this function}

\item{MDP_CharToFactor}{Set to TRUE to turn your character columns to factors if you didn't do so to your ScoringData that you are supplying to this function}

\item{MDP_RemoveDates}{Set to TRUE if you have date of timestamp columns in your ScoringData}

\item{MDP_MissFactor}{If you set MDP_Impute to TRUE, supply the character values to replace missing values with}

\item{MDP_MissNum}{If you set MDP_Impute to TRUE, supply a numeric value to replace missing values with}

\item{FeatureColumnNames}{Supply either column names or column numbers used in the AutoCatBoostRegression() function}
}
\value{
A data.table of predicted values with the option to return model features as well.
}
\description{
AutoCatBoostScoring is an automated scoring function that compliments the AutoCatBoost model training functions. This function requires you to supply features for scoring. It will run ModelDataPrep() to prepare your features for catboost data conversion and scoring.
}
\examples{
\donttest{
Preds <- AutoCatBoostRegressionScoring(TargetType = "regression",
                                       ScoringData = data,
                                       FeatureColNames = 2:12,
                                       CatFeatures = 12,
                                       IDcols = NULL,
                                       ModelPath = "home",
                                       ModelID = "ModelTest",
                                       ReturnFeatures = TRUE,
                                       MDP_Impute = TRUE,
                                       MDP_CharToFactor = TRUE,
                                       MDP_RemoveDates = TRUE,
                                       MDP_MissFactor = "0",
                                       MDP_MissNum = -1)
}
}
\seealso{
Other Supervised Learning: \code{\link{AutoCatBoostClassifier}},
  \code{\link{AutoCatBoostMultiClass}},
  \code{\link{AutoCatBoostRegression}},
  \code{\link{AutoH2OModeler}},
  \code{\link{AutoH2OScoring}},
  \code{\link{AutoH2oDRFClassifier}},
  \code{\link{AutoH2oDRFMultiClass}},
  \code{\link{AutoH2oDRFRegression}},
  \code{\link{AutoH2oGBMClassifier}},
  \code{\link{AutoH2oGBMMultiClass}},
  \code{\link{AutoH2oGBMRegression}},
  \code{\link{AutoNLS}},
  \code{\link{AutoRecommenderScoring}},
  \code{\link{AutoRecommender}}, \code{\link{AutoTS}},
  \code{\link{AutoXGBoostClassifier}},
  \code{\link{AutoXGBoostMultiClass}},
  \code{\link{AutoXGBoostRegression}},
  \code{\link{AutoXGBoostScoring}}
}
\concept{Supervised Learning}
