% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoCatBoostScoring.R
\name{AutoCatBoostScoring}
\alias{AutoCatBoostScoring}
\title{AutoCatBoostScoring}
\usage{
AutoCatBoostScoring(
  TargetType = NULL,
  ScoringData = NULL,
  FeatureColumnNames = NULL,
  FactorLevelsList = NULL,
  IDcols = NULL,
  OneHot = FALSE,
  ReturnShapValues = FALSE,
  ModelObject = NULL,
  ModelPath = NULL,
  ModelID = NULL,
  ReturnFeatures = TRUE,
  MultiClassTargetLevels = NULL,
  TransformNumeric = FALSE,
  BackTransNumeric = FALSE,
  TargetColumnName = NULL,
  TransformationObject = NULL,
  TransID = NULL,
  TransPath = NULL,
  MDP_Impute = TRUE,
  MDP_CharToFactor = TRUE,
  MDP_RemoveDates = TRUE,
  MDP_MissFactor = "0",
  MDP_MissNum = -1,
  RemoveModel = FALSE
)
}
\arguments{
\item{TargetType}{Set this value to "regression", "classification", "multiclass", or "multiregression" to score models built using AutoCatBoostRegression(), AutoCatBoostClassify() or AutoCatBoostMultiClass().}

\item{ScoringData}{This is your data.table of features for scoring. Can be a single row or batch.}

\item{FeatureColumnNames}{Supply either column names or column numbers used in the AutoCatBoostRegression() function}

\item{FactorLevelsList}{List of factors levels to DummifyDT()}

\item{IDcols}{Supply ID column numbers for any metadata you want returned with your predicted values}

\item{OneHot}{Passsed to DummifyD}

\item{ReturnShapValues}{Set to TRUE to return a data.table of feature contributions to all predicted values generated}

\item{ModelObject}{Supply the model object directly for scoring instead of loading it from file. If you supply this, ModelID and ModelPath will be ignored.}

\item{ModelPath}{Supply your path file used in the AutoCatBoost__() function}

\item{ModelID}{Supply the model ID used in the AutoCatBoost__() function}

\item{ReturnFeatures}{Set to TRUE to return your features with the predicted values.}

\item{MultiClassTargetLevels}{For use with AutoCatBoostMultiClass(). If you saved model objects then this scoring function will locate the target levels file. If you did not save model objects, you can supply the target levels returned from AutoCatBoostMultiClass().}

\item{TransformNumeric}{Set to TRUE if you have features that were transformed automatically from an Auto__Regression() model AND you haven't already transformed them.}

\item{BackTransNumeric}{Set to TRUE to generate back-transformed predicted values. Also, if you return features, those will also be back-transformed.}

\item{TargetColumnName}{Input your target column name used in training if you are utilizing the transformation service}

\item{TransformationObject}{Set to NULL if you didn't use transformations or if you want the function to pull from the file output from the Auto__Regression() function. You can also supply the transformation data.table object with the transformation details versus having it pulled from file.}

\item{TransID}{Set to the ID used for saving the transformation data.table object or set it to the ModelID if you are pulling from file from a build with Auto__Regression().}

\item{TransPath}{Set the path file to the folder where your transformation data.table detail object is stored. If you used the Auto__Regression() to build, set it to the same path as ModelPath.}

\item{MDP_Impute}{Set to TRUE if you did so for modeling and didn't do so before supplying ScoringData in this function}

\item{MDP_CharToFactor}{Set to TRUE to turn your character columns to factors if you didn't do so to your ScoringData that you are supplying to this function}

\item{MDP_RemoveDates}{Set to TRUE if you have date of timestamp columns in your ScoringData}

\item{MDP_MissFactor}{If you set MDP_Impute to TRUE, supply the character values to replace missing values with}

\item{MDP_MissNum}{If you set MDP_Impute to TRUE, supply a numeric value to replace missing values with}

\item{RemoveModel}{Set to TRUE if you want the model removed immediately after scoring}
}
\value{
A data.table of predicted values with the option to return model features as well.
}
\description{
AutoCatBoostScoring is an automated scoring function that compliments the AutoCatBoost model training functions. This function requires you to supply features for scoring. It will run ModelDataPrep() to prepare your features for catboost data conversion and scoring.
}
\examples{
\dontrun{

# Create some dummy correlated data
data <- RemixAutoML::FakeDataGenerator(
  Correlation = 0.85,
  N = 10000,
  ID = 2,
  ZIP = 0,
  AddDate = FALSE,
  Classification = FALSE,
  MultiClass = FALSE)

# Train a Multiple Regression Model (two target variables)
TestModel <- RemixAutoML::AutoCatBoostRegression(

  # GPU or CPU and the number of available GPUs
  task_type = "GPU",
  NumGPUs = 1,

  # Metadata arguments
  ModelID = "Test_Model_1",
  model_path = normalizePath("./"),
  metadata_path = NULL,
  SaveModelObjects = FALSE,
  ReturnModelObjects = TRUE,

  # Data arguments
  data = data,
  TrainOnFull = FALSE,
  ValidationData = NULL,
  TestData = NULL,
  Weights = NULL,
  DummifyCols = FALSE,
  TargetColumnName = c("Adrian","Independent_Variable1"),
  FeatureColNames = names(data)[!names(data) \%in\%
    c("IDcol_1","IDcol_2","Adrian")],
  PrimaryDateColumn = NULL,
  IDcols = c("IDcol_1","IDcol_2"),
  TransformNumericColumns = NULL,
  Methods = c("BoxCox","Asinh","Asin","Log","LogPlus1",
    "Logit","YeoJohnson"),

  # Model evaluation
  eval_metric = "MultiRMSE",
  eval_metric_value = 1.5,
  loss_function = "MultiRMSE",
  loss_function_value = 1.5,
  MetricPeriods = 10L,
  NumOfParDepPlots = ncol(data)-1L-2L,
  EvalPlots = TRUE,

  # Grid tuning
  PassInGrid = NULL,
  GridTune = FALSE,
  MaxModelsInGrid = 100L,
  MaxRunsWithoutNewWinner = 100L,
  MaxRunMinutes = 60*60,
  Shuffles = 4L,
  BaselineComparison = "default",

  # ML Args
  langevin = TRUE,
  diffusion_temperature = 10000,
  Trees = 250,
  Depth = 6,
  L2_Leaf_Reg = 3.0,
  RandomStrength = 1,
  BorderCount = 128,
  LearningRate = seq(0.01,0.10,0.01),
  RSM = c(0.80, 0.85, 0.90, 0.95, 1.0),
  BootStrapType = c("Bayesian","Bernoulli","Poisson","MVS","No"),
  GrowPolicy = c("SymmetricTree", "Depthwise", "Lossguide"))

# Output
TestModel$Model
TestModel$ValidationData
TestModel$EvaluationPlot
TestModel$EvaluationBoxPlot
TestModel$EvaluationMetrics
TestModel$VariableImportance
TestModel$InteractionImportance
TestModel$ShapValuesDT
TestModel$VI_Plot
TestModel$PartialDependencePlots
TestModel$PartialDependenceBoxPlots
TestModel$GridList
TestModel$ColNames
TestModel$TransformationResults

# Score a multiple regression model
Preds <- RemixAutoML::AutoCatBoostScoring(
  TargetType = "multiregression",
  ScoringData = data,
  FeatureColumnNames = names(data)[!names(data) \%in\%
    c("IDcol_1", "IDcol_2","Adrian")],
  FactorLevelsList = TestModel$FactorLevelsList,
  IDcols = c("IDcol_1","IDcol_2"),
  OneHot = FALSE,
  ReturnShapValues = TRUE,
  ModelObject = TestModel$Model,
  ModelPath = NULL, #normalizePath("./"),
  ModelID = "Test_Model_1",
  ReturnFeatures = TRUE,
  MultiClassTargetLevels = NULL,
  TransformNumeric = FALSE,
  BackTransNumeric = FALSE,
  TargetColumnName = NULL,
  TransformationObject = NULL,
  TransID = NULL,
  TransPath = NULL,
  MDP_Impute = TRUE,
  MDP_CharToFactor = TRUE,
  MDP_RemoveDates = TRUE,
  MDP_MissFactor = "0",
  MDP_MissNum = -1,
  RemoveModel = FALSE)
}
}
\seealso{
Other Automated Model Scoring: 
\code{\link{AutoH2OMLScoring}()},
\code{\link{AutoH2OModeler}()},
\code{\link{AutoHurdleScoring}()},
\code{\link{AutoXGBoostScoring}()},
\code{\link{IntermittentDemandScoringDataGenerator}()}
}
\author{
Adrian Antico
}
\concept{Automated Model Scoring}
