% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataWranglingFunctions.R
\name{AutoDataDictionaries}
\alias{AutoDataDictionaries}
\title{AutoDataDictionaries}
\usage{
AutoDataDictionaries(Type = "sqlserver", DBConnection, DDType = 1L,
  Query = NULL, CloseChannel = TRUE)
}
\arguments{
\item{Type}{= "sqlserver" is currently the only system supported}

\item{DBConnection}{This is a RODBC connection object for sql server}

\item{DDType}{Select from 1 - 6 based on this article}

\item{Query}{Supply a query}

\item{CloseChannel}{Set to TRUE to disconnect}
}
\description{
AutoDataDictionaries is a function to return data dictionary data in table form
}
\seealso{
Other Data Wrangling: \code{\link{AutoHierarchicalFourier}},
  \code{\link{ColumnSubsetDataTable}},
  \code{\link{DataDisplayMeta}},
  \code{\link{FakeDataGenerator}},
  \code{\link{FullFactorialCatFeatures}},
  \code{\link{SQL_Server_DBConnection}},
  \code{\link{TimeSeriesMelt}}
}
\author{
Adrian Antico
}
\concept{Data Wrangling}
