% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SQL_Functions.R
\name{AutoDataDictionaries}
\alias{AutoDataDictionaries}
\title{AutoDataDictionaries}
\usage{
AutoDataDictionaries(
  Type = "sqlserver",
  DBConnection,
  DDType = 1L,
  Query = NULL,
  ASIS = FALSE,
  CloseChannel = TRUE
)
}
\arguments{
\item{Type}{= "sqlserver" is currently the only system supported}

\item{DBConnection}{This is a RODBC connection object for sql server}

\item{DDType}{Select from 1 - 6 based on this article}

\item{Query}{Supply a query}

\item{ASIS}{Set to TRUE to pull in values without coercing types}

\item{CloseChannel}{Set to TRUE to disconnect}
}
\description{
AutoDataDictionaries is a function to return data dictionary data in table form
}
\seealso{
Other Database: 
\code{\link{SQL_ClearTable}()},
\code{\link{SQL_DropTable}()},
\code{\link{SQL_Query_Push}()},
\code{\link{SQL_Query}()},
\code{\link{SQL_SaveTable}()},
\code{\link{SQL_Server_BulkPull}()},
\code{\link{SQL_Server_BulkPush}()},
\code{\link{SQL_Server_DBConnection}()},
\code{\link{SQL_UpdateTable}()}
}
\author{
Adrian Antico
}
\concept{Database}
