% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoDataPartition.R
\name{AutoDataPartition}
\alias{AutoDataPartition}
\title{The AutoDataPartition function}
\usage{
AutoDataPartition(data, NumDataSets = 3, Ratios = c(0.7, 0.2, 0.1),
  PartitionType = "random", StratifyColumnNames = NULL,
  StratifyNumericTarget = NULL, StratTargetPrecision = 3,
  TimeColumnName = NULL)
}
\arguments{
\item{data}{Source data to do your partitioning on}

\item{NumDataSets}{The number of total data sets you want built}

\item{Ratios}{A vector of values for how much data each data set should get in each split. E.g. c(0.70, 0.20, 0.10)}

\item{PartitionType}{Set to either "random", "timeseries", or "time". With "random", your data will be paritioned randomly (with stratified sampling if column names are supplied). With "timeseries", you can partition by time with a stratify option (so long as you have an equal number of records for each strata). With "time" you will have data sets generated so that the training data contains the earliest records in time, validation data the second earliest, test data the third earliest, etc.}

\item{StratifyColumnNames}{Supply column names of categorical features to use in a stratified sampling procedure for partitioning the data. Partition type must be "random" to use this option}

\item{StratifyNumericTarget}{Supply a column name that is numeric. Use for "random" PartitionType, you can stratify your numeric variable by splitting up based on percRank to ensure a proper allocation of extreme values in your created data sets.}

\item{StratTargetPrecision}{For "random" PartitionType and when StratifyNumericTarget is not null, precision will be the number of decimals used in the percentile calculation. If you supply a value of 1, deciles will be used. For a value of 2, percentiles will be used. Larger values are supported.}

\item{TimeColumnName}{Supply a date column name or a name of a column with an ID for sorting by time such that the smallest number is the earliest in time.}
}
\value{
Returns a list of data.tables
}
\description{
This function will take your ratings matrix and model and score your data in parallel.
}
\examples{
\donttest{
dataSets <- AutoDataPartition(data,
                              NumDataSets = 3,
                              Ratios = c(0.70,0.20,0.10),
                              PartitionType = "random",
                              StratifyColumnNames = NULL,
                              StratifyNumericTarget = NULL,
                              StratTargetPrecision = 1,
                              TimeColumnName = NULL)
}
}
\seealso{
Other Feature Engineering: \code{\link{AutoTransformationCreate}},
  \code{\link{AutoTransformationScore}},
  \code{\link{AutoWord2VecModeler}},
  \code{\link{CreateCalendarVariables}},
  \code{\link{CreateHolidayVariables}},
  \code{\link{DT_GDL_Feature_Engineering}},
  \code{\link{DummifyDT}},
  \code{\link{GDL_Feature_Engineering}},
  \code{\link{ModelDataPrep}},
  \code{\link{Partial_DT_GDL_Feature_Engineering}},
  \code{\link{Scoring_GDL_Feature_Engineering}}
}
\author{
Adrian Antico and Douglas Pestana
}
\concept{Feature Engineering}
