% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EconometricsFunctions.R
\name{AutoFourierFeatures}
\alias{AutoFourierFeatures}
\title{AutoFourierFeatures}
\usage{
AutoFourierFeatures(data, FourierPairs = NULL, FCPeriods = NULL,
  Time_Unit = NULL, TargetColumn = NULL, DateColumn = NULL,
  GroupVariable = NULL, xregs = NonGroupDateNames)
}
\arguments{
\item{data}{The source data}

\item{FourierPairs}{A number indicating the max number of fourier pairs that will be built}

\item{TargetColumn}{The name of your target column}

\item{DateColumn}{The name of your date column}

\item{GroupVariable}{The name of your group variable}

\item{xregs}{Extra data to merge in}
}
\description{
#' AutoFourierFeatures
}
\seealso{
Other Feature Engineering: \code{\link{AutoDataPartition}},
  \code{\link{AutoLagRollStatsScoring}},
  \code{\link{AutoLagRollStats}},
  \code{\link{AutoTransformationCreate}},
  \code{\link{AutoTransformationScore}},
  \code{\link{AutoWord2VecModeler}},
  \code{\link{CreateCalendarVariables}},
  \code{\link{CreateHolidayVariables}},
  \code{\link{DT_GDL_Feature_Engineering}},
  \code{\link{DummifyDT}},
  \code{\link{IntermittentDemandDataGenerator}},
  \code{\link{ModelDataPrep}},
  \code{\link{Partial_DT_GDL_Feature_Engineering}},
  \code{\link{TimeSeriesFill}}
}
\author{
Adrian Antico
}
\concept{Feature Engineering}
