% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RemixAutoML.R
\name{AutoH20TextPrepScoring}
\alias{AutoH20TextPrepScoring}
\title{AutoH20TextPrepScoring is for NLP scoring}
\usage{
AutoH20TextPrepScoring(data, string)
}
\arguments{
\item{data}{The text data}
}
\description{
This function returns prepared tokenized data for H20 Word2VecModeler scoring
}
\seealso{
Other Misc: \code{\link{AutoWordFreq}},
  \code{\link{ChartTheme}}, \code{\link{PrintObjectsSize}},
  \code{\link{RemixTheme}}, \code{\link{SimpleCap}},
  \code{\link{multiplot}}, \code{\link{percRank}},
  \code{\link{tempDatesFun}}, \code{\link{tokenizeH20}}
}
\author{
Adrian Antico
}
\concept{Misc}
