% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoH2oMLScoring.R
\name{AutoH2OMLScoring}
\alias{AutoH2OMLScoring}
\title{AutoH2OMLScoring is an automated scoring function that compliments the AutoH2o model training functions.}
\usage{
AutoH2OMLScoring(
  ScoringData = NULL,
  ModelObject = NULL,
  ModelType = "mojo",
  H2OShutdown = TRUE,
  MaxMem = "28G",
  NThreads = max(1, parallel::detectCores() - 2),
  JavaOptions = "-Xmx1g -XX:ReservedCodeCacheSize=256m",
  ModelPath = NULL,
  ModelID = NULL,
  ReturnFeatures = TRUE,
  TransformNumeric = FALSE,
  BackTransNumeric = FALSE,
  TargetColumnName = NULL,
  TransformationObject = NULL,
  TransID = NULL,
  TransPath = NULL,
  MDP_Impute = TRUE,
  MDP_CharToFactor = TRUE,
  MDP_RemoveDates = TRUE,
  MDP_MissFactor = "0",
  MDP_MissNum = -1
)
}
\arguments{
\item{ScoringData}{This is your data.table of features for scoring. Can be a single row or batch.}

\item{ModelObject}{Supply a model object from AutoH2oDRF__()}

\item{ModelType}{Set to either "mojo" or "standard" depending on which version you saved}

\item{H2OShutdown}{Set to TRUE is you are scoring a "standard" model file and you aren't planning on continuing to score.}

\item{MaxMem}{Set to you dedicated amount of memory. E.g. "28G"}

\item{NThreads}{Default set to max(1, parallel::detectCores()-2)}

\item{JavaOptions}{Change the default to your machines specification if needed. Default is '-Xmx1g -XX:ReservedCodeCacheSize=256m',}

\item{ModelPath}{Supply your path file used in the AutoH2o__() function}

\item{ModelID}{Supply the model ID used in the AutoH2o__() function}

\item{ReturnFeatures}{Set to TRUE to return your features with the predicted values.}

\item{TransformNumeric}{Set to TRUE if you have features that were transformed automatically from an Auto__Regression() model AND you haven't already transformed them.}

\item{BackTransNumeric}{Set to TRUE to generate back-transformed predicted values. Also, if you return features, those will also be back-transformed.}

\item{TargetColumnName}{Input your target column name used in training if you are utilizing the transformation service}

\item{TransformationObject}{Set to NULL if you didn't use transformations or if you want the function to pull from the file output from the Auto__Regression() function. You can also supply the transformation data.table object with the transformation details versus having it pulled from file.}

\item{TransID}{Set to the ID used for saving the transformation data.table object or set it to the ModelID if you are pulling from file from a build with Auto__Regression().}

\item{TransPath}{Set the path file to the folder where your transformation data.table detail object is stored. If you used the Auto__Regression() to build, set it to the same path as ModelPath.}

\item{MDP_Impute}{Set to TRUE if you did so for modeling and didn't do so before supplying ScoringData in this function}

\item{MDP_CharToFactor}{Set to TRUE to turn your character columns to factors if you didn't do so to your ScoringData that you are supplying to this function}

\item{MDP_RemoveDates}{Set to TRUE if you have date of timestamp columns in your ScoringData}

\item{MDP_MissFactor}{If you set MDP_Impute to TRUE, supply the character values to replace missing values with}

\item{MDP_MissNum}{If you set MDP_Impute to TRUE, supply a numeric value to replace missing values with}
}
\value{
A data.table of predicted values with the option to return model features as well.
}
\description{
AutoH2OMLScoring is an automated scoring function that compliments the AutoH2oGBM__() and AutoH2oDRF__() models training functions. This function requires you to supply features for scoring. It will run ModelDataPrep()to prepare your features for H2O data conversion and scoring.
}
\examples{
\dontrun{
Preds <- AutoH2OMLScoring(
  ScoringData = data,
  ModelObject = NULL,
  ModelType = "mojo",
  H2OShutdown = TRUE,
  MaxMem = "28G",
  NThreads = max(1, parallel::detectCores()-2),
  JavaOptions = '-Xmx1g -XX:ReservedCodeCacheSize=256m',
  ModelPath = normalizePath("./"),
  ModelID = "ModelTest",
  ReturnFeatures = TRUE,
  TransformNumeric = FALSE,
  BackTransNumeric = FALSE,
  TargetColumnName = NULL,
  TransformationObject = NULL,
  TransID = NULL,
  TransPath = NULL,
  MDP_Impute = TRUE,
  MDP_CharToFactor = TRUE,
  MDP_RemoveDates = TRUE,
  MDP_MissFactor = "0",
  MDP_MissNum = -1)
}
}
\seealso{
Other Automated Model Scoring: 
\code{\link{AutoCatBoostScoring}()},
\code{\link{AutoH2OModeler}()},
\code{\link{AutoHurdleScoring}()},
\code{\link{AutoXGBoostScoring}()}
}
\author{
Adrian Antico
}
\concept{Automated Model Scoring}
