% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TextPrepFunctions.R
\name{AutoH2OTextPrepScoring}
\alias{AutoH2OTextPrepScoring}
\title{AutoH2OTextPrepScoring is for NLP scoring}
\usage{
AutoH2OTextPrepScoring(
  data,
  string = NULL,
  MaxMem = NULL,
  NThreads = NULL,
  StartH2O = TRUE
)
}
\arguments{
\item{data}{The text data}

\item{string}{The name of the string column to prepare}

\item{MaxMem}{Amount of memory you want to let H2O utilize}

\item{NThreads}{The number of threads you want to let H2O utilize}

\item{StartH2O}{Set to TRUE to have H2O start inside this function}
}
\description{
This function returns prepared tokenized data for H2O Word2VecModeler scoring
}
\examples{
\dontrun{
data <- AutoH2OTextPrepScoring(data = x,
                               string = "text_column",
                               MaxMem = "28G",
                               NThreads = 8,
                               StartH2O = TRUE)
}
}
\seealso{
Other Misc: 
\code{\link{DeleteFile}()},
\code{\link{PrintToPDF}()},
\code{\link{tokenizeH2O}()}
}
\author{
Adrian Antico
}
\concept{Misc}
