% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EconometricsFunctions.R
\name{AutoHierarchicalFourier}
\alias{AutoHierarchicalFourier}
\title{AutoHierarchicalFourier}
\usage{
AutoHierarchicalFourier(datax = data, xRegs = names(XREGS),
  FourierTermS = FourierTerms, TimeUniT = TimeUnit,
  FC_PeriodS = FC_Periods, TargetColumN = TargetColumn,
  DateColumN = DateColumnName, HierarchGroups = NULL,
  IndependentGroups = NULL)
}
\arguments{
\item{datax}{data}

\item{xRegs}{The XREGS}

\item{FourierTermS}{Number of fourier pairs}

\item{TimeUniT}{Time unit}

\item{FC_PeriodS}{Number of forecast periods}

\item{TargetColumN}{Target column name}

\item{DateColumN}{Date column name}

\item{HierarchGroups}{Character vector of categorical columns to fully interact}

\item{IndependentGroups}{Character vector of categorical columns to run independently}
}
\description{
AutoHierarchicalFourier reverses the difference
}
\seealso{
Other Data Wrangling: \code{\link{AutoDataDictionaries}},
  \code{\link{ColumnSubsetDataTable}},
  \code{\link{DataDisplayMeta}},
  \code{\link{FullFactorialCatFeatures}},
  \code{\link{SQL_Server_DBConnection}},
  \code{\link{TimeSeriesMelt}}
}
\author{
Adrian Antico
}
\concept{Data Wrangling}
