% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoHurdleScoring.R
\name{AutoHurdleScoring}
\alias{AutoHurdleScoring}
\title{AutoHurdleScoring()}
\usage{
AutoHurdleScoring(
  TestData = NULL,
  Path = NULL,
  ModelID = NULL,
  ModelClass = "catboost",
  ArgList = NULL,
  ModelList = NULL
)
}
\arguments{
\item{TestData}{scoring data.table}

\item{Path}{Supply if ArgList is NULL or ModelList is null.}

\item{ModelID}{Supply if ArgList is NULL or ModelList is null. Same as used in model training.}

\item{ModelClass}{Name of model type. "catboost" is currently the only available option}

\item{ArgList}{Output from the hurdle model}

\item{ModelList}{Output from the hurdle model}
}
\value{
A data.table with the final predicted value, the intermediate model predictions, and your source data
}
\description{
AutoHurdleScoring()
}
\examples{
\donttest{

# Pull in some data:
#   N - Change the number of records pulled in
# data <- RemixAutoML::FakeDataGenerator(Correlation = 0.85, N = 25000, ID = 0, ZIP = 1, AddDate = FALSE)
# Path <- "C:/Users/aantico/Documents/Package/GUI_Package"

Output <- RemixAutoML::AutoXGBoostHurdleModel(

   # Operationalization args
   TreeMethod = "hist",
   TrainOnFull = FALSE,
   PassInGrid = NULL,

   # Metadata args
   NThreads = max(1L, parallel::detectCores()-2L),
   ModelID = "ModelTest",
   Paths = normalizePath(Path),
   MetaDataPaths = NULL,

   # data args
   data,
   ValidationData = NULL,
   TestData = NULL,
   Buckets = 0L,
   TargetColumnName = "Adrian",
   FeatureColNames = names(data)[2:ncol(data)],
   IDcols = NULL,

   TransformNumericColumns = NULL,
   SplitRatios = c(0.70, 0.20, 0.10),
   SaveModelObjects = TRUE,
   NumOfParDepPlots = 10L,

   GridTune = FALSE,
   grid_eval_metric = "accuracy",
   MaxModelsInGrid = 1L,
   BaselineComparison = "default",
   MaxRunsWithoutNewWinner = 10L,
   MaxRunMinutes = 60L,

   # bandit hyperparameters
   Trees = 100L,
   eta = seq(0.05,0.40,0.05),
   max_depth = seq(4L, 16L, 2L),

   # random hyperparameters
   min_child_weight = seq(1.0, 10.0, 1.0),
   subsample = seq(0.55, 1.0, 0.05),
   colsample_bytree = seq(0.55, 1.0, 0.05))

# Score models by calling them from file
HurdleScores <- RemixAutoML::AutoHurdleScoring(
   TestData = data,
   Path = Path,
   ModelID = "ModelTest",
   ModelClass = "catboost",
   ModelList = NULL,
   ArgList = NULL)

# Score models by using output object from training
HurdleScores <- RemixAutoML::AutoHurdleScoring(
   TestData = data,
   Path = Path,
   ModelID = "ModelTest",
   ModelClass = "catboost",
   ModelList = Output$ModelList,
   ArgList = Output$ArgsList)
}
}
\seealso{
Other Automated Model Scoring: 
\code{\link{AutoCatBoostScoring}()},
\code{\link{AutoH2OMLScoring}()},
\code{\link{AutoH2OModeler}()},
\code{\link{AutoXGBoostScoring}()}
}
\author{
Adrian Antico
}
\concept{Automated Model Scoring}
