% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoHurdleScoring.R
\name{AutoHurdleScoring}
\alias{AutoHurdleScoring}
\title{AutoHurdleScoring()}
\usage{
AutoHurdleScoring(
  TestData = NULL,
  Path = NULL,
  ModelID = NULL,
  ModelClass = "catboost",
  ArgList = NULL,
  ModelList = NULL
)
}
\arguments{
\item{TestData}{scoring data.table}

\item{Path}{Supply if ArgList is NULL or ModelList is null.}

\item{ModelID}{Supply if ArgList is NULL or ModelList is null. Same as used in model training.}

\item{ModelClass}{Name of model type. "catboost" is currently the only available option}

\item{ArgList}{Output from the hurdle model}

\item{ModelList}{Output from the hurdle model}
}
\value{

}
\description{
AutoHurdleScoring()
}
\examples{
\donttest{

# Score models by calling them from file
HurdleScores <- RemixAutoML::AutoHurdleScoring(
   TestData = data,
   Path = Path,
   ModelID = "ModelTest",
   ModelClass = "catboost",
   ModelList = NULL,
   ArgList = NULL)

# Score models by using output object from training
HurdleScores <- RemixAutoML::AutoHurdleScoring(
   TestData = data,
   Path = Path,
   ModelID = "ModelTest",
   ModelClass = "catboost",
   ModelList = Output$ModelList,
   ArgList = Output$ArgsList)
}
}
\seealso{
Other Automated Model Scoring: 
\code{\link{AutoCatBoostScoring}()},
\code{\link{AutoH2OMLScoring}()},
\code{\link{AutoH2OModeler}()},
\code{\link{AutoXGBoostScoring}()}
}
\author{
Adrian Antico
}
\concept{Automated Model Scoring}
