% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoLightGBMFunnel.R
\name{AutoLightGBMFunnelCARMAScoring}
\alias{AutoLightGBMFunnelCARMAScoring}
\title{AutoLightGBMFunnelCARMAScoring}
\usage{
AutoLightGBMFunnelCARMAScoring(
  TrainData,
  ForwardLookingData = NULL,
  TrainEndDate = NULL,
  ForecastEndDate = NULL,
  ArgsList = NULL,
  TrainOutput = NULL,
  ModelPath = NULL,
  MaxCohortPeriod = NULL,
  DebugMode = FALSE
)
}
\arguments{
\item{TrainData}{Data utilized in training. Do not put the BaseFunnelMeasure in this data set. Put it in the ForwardLookingData object}

\item{ForwardLookingData}{Base funnel measure data. Needs to cover the span of the forecast horizon}

\item{TrainEndDate}{Max date from the training data}

\item{ForecastEndDate}{Max date to forecast out to}

\item{ArgsList}{Output list from AutoCatBoostFunnelCARMA}

\item{TrainOutput}{Pass in the model object to speed up forecasting}

\item{ModelPath}{Path to model location}

\item{MaxCohortPeriod}{Max cohort periods to utilize when forecasting}

\item{DebugMode}{For debugging issues}
}
\description{
AutoLightGBMFunnelCARMAScoring for generating forecasts
}
\seealso{
Other Automated Funnel Data Forecasting: 
\code{\link{AutoCatBoostFunnelCARMAScoring}()},
\code{\link{AutoCatBoostFunnelCARMA}()},
\code{\link{AutoLightGBMFunnelCARMA}()},
\code{\link{AutoXGBoostFunnelCARMAScoring}()},
\code{\link{AutoXGBoostFunnelCARMA}()}
}
\author{
Adrian Antico
}
\concept{Automated Funnel Data Forecasting}
