% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RemixAutoML.R
\name{AutoNLS}
\alias{AutoNLS}
\title{AutoNLS is a function for automatically building nls models}
\usage{
AutoNLS(data, y, x, monotonic = TRUE)
}
\arguments{
\item{data}{Data is the data table you are building the modeling on}

\item{y}{Y is the target variable name in quotes}

\item{x}{X is the independent variable name in quotes}

\item{monotonic}{This is a TRUE/FALSE indicator - choose TRUE if you want monotonic regression over polynomial regression}
}
\value{
A list containing "PredictionData" which is a data table with your original column replaced by the nls model predictions; "ModelName" the model name; "ModelObject" The winning model to later use; "EvaluationMetrics" Model metrics for models with ability to build.
}
\description{
This function will build models for 9 different nls models, along with a non-parametric monotonic regression and a polynomial regression. The models are evaluated, a winner is picked, and the predicted values are stored in your data table.
}
\examples{
# Create Growth Data
data <-
  data.table::data.table(Target = seq(1, 500, 1),
                         Variable = rep(1, 500))
for (i in as.integer(1:500)) {
  if (i == 1) {
    var <- data[i, "Target"][[1]]
    data.table::set(data,
                    i = i,
                    j = 2L,
                    value = var * (1 + runif(1) / 100))
  } else {
    var <- data[i - 1, "Variable"][[1]]
    data.table::set(data,
                    i = i,
                    j = 2L,
                    value = var * (1 + runif(1) / 100))
  }
}

# Add jitter to Target
data[, Target := jitter(Target,
                        factor = 0.25)]

# To keep original values
data1 <- data.table::copy(data)

# Merge and Model data
data11 <- AutoNLS(
  data = data,
  y = "Target",
  x = "Variable",
  monotonic = TRUE
)

# Join predictions to source data
data2 <- merge(
  data1,
  data11$PredictionData,
  by = "Variable",
  all = FALSE
)

# Plot output
ggplot2::ggplot(data2, ggplot2::aes(x = Variable)) +
  ggplot2::geom_line(ggplot2::aes(y = data2[["Target.x"]],
                                  color = "Target")) +
  ggplot2::geom_line(ggplot2::aes(y = data2[["Target.y"]],
                                  color = "Predicted")) +
 RemixAutoML::ChartTheme(Size = 12) +
  ggplot2::ggtitle(paste0("Growth Models AutoNLS: ",
                          data11$ModelName)) +
  ggplot2::ylab("Target Variable") +
  ggplot2::xlab("Independent Variable") +
  ggplot2::scale_colour_manual("Values",
                               breaks = c("Target",
                                          "Predicted"),
                               values = c("red",
                                          "blue"))
summary(data11$ModelObject)
data11$EvaluationMetrics
}
\seealso{
Other Supervised Learning: \code{\link{AutoCatBoostClassifier}},
  \code{\link{AutoCatBoostMultiClass}},
  \code{\link{AutoCatBoostRegression}},
  \code{\link{AutoH2OModeler}},
  \code{\link{AutoH2OScoring}},
  \code{\link{AutoH2oDRFRegression}},
  \code{\link{AutoH2oGBMRegression}},
  \code{\link{AutoRecommenderScoring}},
  \code{\link{AutoRecommender}}, \code{\link{AutoTS}}
}
\author{
Adrian Antico
}
\concept{Supervised Learning}
