% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotFunctions.R
\name{AutoPlotter}
\alias{AutoPlotter}
\title{AutoPlotter}
\usage{
AutoPlotter(
  dt = NULL,
  PlotType = input[["PlotType"]],
  YVar = shiny::isolate(YVar()),
  XVar = shiny::isolate(DateVar()),
  YMin = input[["YMin"]],
  YMax = input[["YMax"]],
  XMin = input[["XMin"]],
  XMax = input[["XMax"]],
  ColorVariables = shiny::isolate(SelectedGroups()),
  SizeVar1 = input[["SizeVar1"]],
  FacetVar1 = shiny::isolate(input[["FacetVar1"]]),
  FacetVar2 = shiny::isolate(input[["FacetVar2"]]),
  YTicks = input[["YTicks"]],
  XTicks = input[["XTicks"]],
  OutlierSize = input[["OutlierSize"]],
  OutlierColor = input[["OutlierColor"]],
  BoxPlotFill = input[["BoxPlotFill"]],
  GamFitScatter = input[["GamFitScatter"]],
  TextSize = input[["TextSize"]],
  AngleX = input[["AngleX"]],
  AngleY = input[["AngleY"]],
  ChartColor = input[["ChartColor"]],
  BorderColor = input[["BorderColor"]],
  TextColor = input[["TextColor"]],
  GridColor = input[["GridColor"]],
  BackGroundColor = input[["BackGroundColor"]],
  Debug = FALSE
)
}
\arguments{
\item{dt}{= NULL}

\item{PlotType}{= input[['PlotType']]}

\item{YVar}{= shiny::isolate(YVar())}

\item{XVar}{= shiny::isolate(DateVar())}

\item{YMin}{= input[['YMin']]}

\item{YMax}{= input[['YMax']]}

\item{XMin}{= input[['XMin']]}

\item{XMax}{= input[['XMax']]}

\item{ColorVariables}{= shiny::isolate(SelectedGroups())}

\item{SizeVar1}{= input[['SizeVar1']]}

\item{FacetVar1}{= shiny::isolate(input[['FacetVar1']])}

\item{FacetVar2}{= shiny::isolate(input[['FacetVar2']])}

\item{YTicks}{= input[['YTicks']]}

\item{XTicks}{= input[['XTicks']]}

\item{OutlierSize}{= input[['OutlierSize']]}

\item{OutlierColor}{= input[['OutlierColor']]}

\item{BoxPlotFill}{= input[['BoxPlotFill']]}

\item{GamFitScatter}{= input[['GamFitScatter']]}

\item{TextSize}{= input[['TextSize']]}

\item{AngleX}{= input[['AngleX']]}

\item{AngleY}{= input[['AngleY']]}

\item{ChartColor}{= input[['ChartColor']]}

\item{BorderColor}{= input[['BorderColor']]}

\item{TextColor}{= input[['TextColor']]}

\item{GridColor}{= input[['GridColor']]}

\item{BackGroundColor}{= input[['BackGroundColor']]}

\item{Debug}{= FALSE}
}
\description{
Create plots
}
\seealso{
Other Graphics: 
\code{\link{ChartTheme}()},
\code{\link{multiplot}()}
}
\author{
Adrian Antico
}
\concept{Graphics}
