% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoRecommender.R
\name{AutoRecommender}
\alias{AutoRecommender}
\title{Automatically build the best recommender model among models available.}
\usage{
AutoRecommender(data, Partition = "Split", KFolds = 2, Ratio = 0.75,
  ProductEvaluation = 1, RatingType = "TopN", RatingsKeep = 20,
  SkipModels = "AssociationRules", ModelMetric = "TPR")
}
\arguments{
\item{data}{This is your BinaryRatingsMatrix. See function RecomDataCreate}

\item{Partition}{Choose from "split", "cross-validation", "bootstrap". See evaluationScheme in recommenderlab for details.}

\item{KFolds}{Choose 2 for traditional train and test. Choose greater than 2 for the number of cross validations}

\item{Ratio}{The ratio for train and test. E.g. 0.75 for 75 percent data allocated to training}

\item{ProductEvaluation}{The number of products you would like to evaluate. Negative values implement all-but schemes.}

\item{RatingType}{Choose from "TopN", "ratings", "ratingMatrix"}

\item{RatingsKeep}{The total ratings you wish to return. Default is 20.}

\item{SkipModels}{AssociationRules runs the slowest and may crash your system. Choose from: "AssociationRules","ItemBasedCF","UserBasedCF","PopularItems","RandomItems"}

\item{ModelMetric}{Choose from "Precision", "Recall", "TPR", or "FPR"}
}
\value{
The winning model used for scoring in the AutoRecommenderScoring function
}
\description{
This function returns the winning model that you pass onto AutoRecommenderScoring
}
\examples{
\donttest{
WinningModel <- AutoRecommender(RatingsMatrix,
                                Partition = "Split",
                                KFolds = 2,
                                Ratio = 0.75,
                                ProductEvaluation = 1,
                                RatingType = "TopN",
                                RatingsKeep = 20,
                                SkipModels = "AssociationRules",
                                ModelMetric = "TPR")
}
}
\seealso{
Other Marketing Modeling: \code{\link{AutoMarketBasketModel}},
  \code{\link{AutoRecomDataCreate}},
  \code{\link{AutoRecommenderScoring}}
}
\author{
Adrian Antico and Douglas Pestana
}
\concept{Marketing Modeling}
