% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RemixAutoML.R
\name{AutoRecommenderScoring}
\alias{AutoRecommenderScoring}
\title{The AutoRecomScoring function scores recommender models from AutoRecommender()}
\usage{
AutoRecommenderScoring(data, WinningModel, EntityColName = "CustomerID",
  ProductColName = "StockCode")
}
\arguments{
\item{data}{The binary ratings matrix from RecomDataCreate()}

\item{WinningModel}{The winning model returned from AutoRecommender()}

\item{EntityColName}{Typically your customer ID}

\item{ProductColName}{Something like "StockCode"}
}
\value{
Returns the prediction data
}
\description{
This function will take your ratings matrix and model and score your data in parallel.
}
\examples{
\donttest{
# F(G(Z(x))): AutoRecommenderScoring(AutoRecommender(RecomDataCreate(TransactionData)))
Results <- AutoRecommenderScoring(
  data = RecomDataCreate(
      data,
      EntityColName = "CustomerID",
      ProductColName = "StockCode",
      MetricColName = "TotalSales"),
  WinningModel = AutoRecommender(
      RecomDataCreate(
        data,
        EntityColName = "CustomerID",
        ProductColName = "StockCode",
        MetricColName = "TotalSales"),
      Partition = "Split",
      KFolds = 2,
      Ratio = 0.75,
      RatingType = "TopN",
      RatingsKeep = 20,
      SkipModels = "AssociationRules",
      ModelMetric = "TPR"),
  EntityColName = "CustomerID",
  ProductColName = "StockCode")
}
}
\seealso{
Other Supervised Learning: \code{\link{AutoCatBoostClassifier}},
  \code{\link{AutoCatBoostRegression}},
  \code{\link{AutoH2OModeler}},
  \code{\link{AutoH2OScoring}}, \code{\link{AutoNLS}},
  \code{\link{AutoRecommender}}, \code{\link{AutoTS}}
}
\author{
Adrian Antico and Douglas Pestana
}
\concept{Supervised Learning}
