% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoRecommenderScoring.R
\name{AutoRecommenderScoring}
\alias{AutoRecommenderScoring}
\title{The AutoRecomScoring function scores recommender models from AutoRecommender()}
\usage{
AutoRecommenderScoring(data, WinningModel, EntityColName = "CustomerID",
  ProductColName = "StockCode", NumItemsReturn = 1)
}
\arguments{
\item{data}{The binary ratings matrix from RecomDataCreate()}

\item{WinningModel}{The winning model returned from AutoRecommender()}

\item{EntityColName}{Typically your customer ID}

\item{ProductColName}{Something like "StockCode"}

\item{NumItemsReturn}{Number of items to return on scoring}
}
\value{
Returns the prediction data
}
\description{
This function will take your ratings matrix and model and score your data in parallel.
}
\examples{
\donttest{
# F(G(Z(x))): AutoRecommenderScoring(AutoRecommender(AutoRecomDataCreate(TransactionData)))
Results <- AutoRecommenderScoring(
  data = AutoRecomDataCreate(
      data,
      EntityColName = "CustomerID",
      ProductColName = "StockCode",
      MetricColName = "TotalSales"),
  WinningModel = AutoRecommender(
      AutoRecomDataCreate(
        data,
        EntityColName = "CustomerID",
        ProductColName = "StockCode",
        MetricColName = "TotalSales"),
      Partition = "Split",
      KFolds = 2,
      Ratio = 0.75,
      RatingType = "TopN",
      RatingsKeep = 20,
      SkipModels = "AssociationRules",
      ModelMetric = "TPR"),
  EntityColName = "CustomerID",
  ProductColName = "StockCode")
}
}
\seealso{
Other Marketing Modeling: \code{\link{AutoMarketBasketModel}},
  \code{\link{AutoRecomDataCreate}},
  \code{\link{AutoRecommender}}
}
\author{
Adrian Antico and Douglas Pestana
}
\concept{Marketing Modeling}
