% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransformationFunctions.R
\name{AutoTransformationCreate}
\alias{AutoTransformationCreate}
\title{AutoTransformationCreate is a function for automatically identifying the optimal transformations for numeric features and transforming them once identified.}
\usage{
AutoTransformationCreate(data, ColumnNames = NULL,
  Methods = c("BoxCox", "YeoJohnson", "Asinh", "Asin", "Logit"),
  Path = NULL, TransID = "ModelID", SaveOutput = FALSE)
}
\arguments{
\item{data}{This is your source data}

\item{ColumnNames}{List your columns names in a vector, for example, c("Target", "IV1")}

\item{Methods}{Choose from "YeoJohnson", "BoxCox", "Asinh", "Asin", and "Logit". "Identity" is always run.}

\item{Path}{Set to the directly where you want to save all of your modeling files}

\item{TransID}{Set to a character value that corresponds with your modeling project}

\item{SaveOutput}{Set to TRUE to save necessary file to run AutoTransformationScore()}
}
\value{
data with transformed columns and the transformation object for back-transforming later
}
\description{
AutoTransformationCreate is a function for automatically identifying the optimal transformations for numeric features and transforming them once identified. This function will loop through your selected transformation options (YeoJohnson, BoxCox, Asinh, Asin, and Logit) and find the one that produces data that is the closest to normally distributed data. It then makes the transformation and collects the metadata information for use in the AutoTransformationScore() function, either by returning the objects (always) or saving them to file (optional).
}
\examples{
Correl <- 0.85
N <- 1000
data <- data.table::data.table(Adrian = runif(N))
data[, x1 := qnorm(Adrian)]
data[, x2 := runif(N)]
data[, Adrian1 := log(pnorm(Correl * x1 +
                            sqrt(1-Correl^2) * qnorm(x2)))]
data <- AutoTransformationCreate(data,
                                 ColumnNames = "Sample",
                                 Methods = c("BoxCox","Asin"),
                                 Path = NULL,
                                 TransID = "Trans",
                                 SaveOutput = FALSE)
}
\seealso{
Other Feature Engineering: \code{\link{AutoDataPartition}},
  \code{\link{AutoTransformationScore}},
  \code{\link{AutoWord2VecModeler}},
  \code{\link{CreateCalendarVariables}},
  \code{\link{DT_GDL_Feature_Engineering}},
  \code{\link{DummifyDT}},
  \code{\link{GDL_Feature_Engineering}},
  \code{\link{ModelDataPrep}},
  \code{\link{Partial_DT_GDL_Feature_Engineering2}},
  \code{\link{Partial_DT_GDL_Feature_Engineering}},
  \code{\link{Scoring_GDL_Feature_Engineering}}
}
\author{
Adrian Antico
}
\concept{Feature Engineering}
