% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RemixAutoML.R
\name{AutoTransformationScore}
\alias{AutoTransformationScore}
\title{Test Identity Transformation}
\usage{
AutoTransformationScore(ScoringData, FinalResults, Type = "Inverse",
  TransID = "TestModel", Path = NULL)
}
\arguments{
\item{ScoringData}{This is your source data}

\item{FinalResults}{This is the FinalResults output object from AutoTransformationCreate().}

\item{Type}{Set to "Inverse" to back-transfrom or "Apply" for applying the transformation.}

\item{TransID}{Set to a character value that corresponds with your modeling project}

\item{Path}{Set to the directly where you want to save all of your modeling files}
}
\value{
Returns AutoCatBoostRegression() model objects: VariableImportance.csv, Model, ValidationData.csv, EvalutionPlot.png, EvalutionBoxPlot.png, EvaluationMetrics.csv, ParDepPlots.R a named list of features with partial dependence calibration plots, ParDepBoxPlots.R, GridCollect, and catboostgrid
}
\description{
Test Identity Transformation
}
\examples{
Correl <- 0.85
N <- 1000
data <- data.table::data.table(Adrian = runif(N))
data[, x1 := qnorm(Adrian)]
data[, x2 := runif(N)]
data[, Adrian1 := log(pnorm(Correl * x1 +
                            sqrt(1-Correl^2) * qnorm(x2)))]
data <- AutoTransformationScore(ScoringData, 
                                FinalResults,
                                Path = NULL,
                                TransID = "Trans")
}
\seealso{
Other Feature Engineering: \code{\link{Apply_Asinh}},
  \code{\link{Apply_Asin}}, \code{\link{Apply_BoxCox}},
  \code{\link{Apply_Logit}},
  \code{\link{Apply_YeoJohnson}},
  \code{\link{AutoDataPartition}},
  \code{\link{AutoTransformationCreate}},
  \code{\link{AutoWord2VecModeler}},
  \code{\link{CreateCalendarVariables}},
  \code{\link{DT_GDL_Feature_Engineering}},
  \code{\link{DummifyDT}},
  \code{\link{FAST_GDL_Feature_Engineering}},
  \code{\link{GDL_Feature_Engineering}},
  \code{\link{InvApply_Asinh}},
  \code{\link{InvApply_Asin}},
  \code{\link{InvApply_BoxCox}},
  \code{\link{InvApply_Logit}},
  \code{\link{InvApply_YeoJohnson}},
  \code{\link{ModelDataPrep}},
  \code{\link{Scoring_GDL_Feature_Engineering}}
}
\author{
Adrian Antico
}
\concept{Feature Engineering}
