% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransformationFunctions.R
\name{AutoTransformationScore}
\alias{AutoTransformationScore}
\title{AutoTransformationScore() is a the complimentary function to AutoTransformationCreate()}
\usage{
AutoTransformationScore(ScoringData, FinalResults, Type = "Inverse",
  TransID = "TestModel", Path = NULL)
}
\arguments{
\item{ScoringData}{This is your source data}

\item{FinalResults}{This is the FinalResults output object from AutoTransformationCreate().}

\item{Type}{Set to "Inverse" to back-transfrom or "Apply" for applying the transformation.}

\item{TransID}{Set to a character value that corresponds with your modeling project}

\item{Path}{Set to the directly where you want to save all of your modeling files}
}
\value{
data with transformed columns
}
\description{
AutoTransformationScore() is a the compliment function to AutoTransformationCreate(). Automatically apply or inverse the transformations you identified in AutoTransformationCreate() to other data sets. This is useful for applying transformations to your validation and test data sets for modeling. It's also useful for back-transforming your target and prediction columns after you have build and score your models so you can obtain statistics on the original features.
}
\examples{
Correl <- 0.85
N <- 1000
data <- data.table::data.table(Adrian = runif(N))
data[, x1 := qnorm(Adrian)]
data[, x2 := runif(N)]
data[, Adrian1 := log(pnorm(Correl * x1 +
                            sqrt(1-Correl^2) * qnorm(x2)))]
data <- AutoTransformationScore(ScoringData,
                                FinalResults,
                                Path = NULL,
                                TransID = "Trans")
}
\seealso{
Other Feature Engineering: \code{\link{AutoDataPartition}},
  \code{\link{AutoTransformationCreate}},
  \code{\link{AutoWord2VecModeler}},
  \code{\link{CreateCalendarVariables}},
  \code{\link{DT_GDL_Feature_Engineering}},
  \code{\link{DummifyDT}},
  \code{\link{GDL_Feature_Engineering}},
  \code{\link{ModelDataPrep}},
  \code{\link{Partial_DT_GDL_Feature_Engineering2}},
  \code{\link{Partial_DT_GDL_Feature_Engineering}},
  \code{\link{Scoring_GDL_Feature_Engineering}}
}
\author{
Adrian Antico
}
\concept{Feature Engineering}
