% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RemixAutoML.R
\name{AutoTS}
\alias{AutoTS}
\title{AutoTS is an automated time series modeling function}
\usage{
AutoTS(data, TargetName = "Targets", DateName = "DateTime",
  FCPeriods = 30, HoldOutPeriods = 30, TimeUnit = "day", Lags = 25,
  SLags = 2, NumCores = 4, SkipModels = NULL, StepWise = TRUE)
}
\arguments{
\item{data}{is the source time series data.table}

\item{TargetName}{is the name of the dependent variable in your data.table}

\item{DateName}{is the name of the date column in your data.table}

\item{FCPeriods}{is the number of periods into the future you wish to forecast}

\item{HoldOutPeriods}{is the number of periods to use for validation testing}

\item{TimeUnit}{is the level of aggregation your dataset comes in}

\item{Lags}{is the number of lags you wish to test in various models (same with moving averages)}

\item{SLags}{is the number of seasonal lags you wish to test in various models (same with moving averages)}

\item{NumCores}{is the number of cores available on your computer}

\item{SkipModels}{Don't run specified models - e.g. exclude all models "ARFIMA" "ARIMA" "ETS" "NNET" "TBATS" "TSLM" "PROPHET"}

\item{StepWise}{Set to TRUE to have ARIMA and ARFIMA run a stepwise selection process. Otherwise, all models will be generated in parallel execution, but still run much slower.}
}
\value{
Returns a list containing 1: A data.table object with a date column and the forecasted values; 2: The model evaluation results; 3: The winning model for later use if desired.
}
\description{
AutoTS builds the best time series models for each type, compares all types, selects the winner, and generates a forecast.
}
\examples{
data <- data.table::data.table(DateTime = as.Date(Sys.time()),
  Target = stats::filter(rnorm(1000,
                               mean = 50,
                               sd = 20),
                         filter=rep(1,10),
                         circular=TRUE))
data[, temp := seq(1:1000)][, DateTime := DateTime - temp][, temp := NULL]
data <- data[order(DateTime)]
output <-   AutoTS(data,
                   TargetName     = "Target",
                   DateName       = "DateTime",
                   FCPeriods      = 30,
                   HoldOutPeriods = 30,
                   TimeUnit       = "day",
                   Lags           = 5,
                   SLags          = 1,
                   NumCores       = 4,
                   SkipModels     = NULL,
                   StepWise       = TRUE)
ForecastData <- output[[1]]
ModelEval    <- output[[2]]
WinningModel <- output[[3]]
}
\seealso{
Other Supervised Learning: \code{\link{AutoH20Modeler}},
  \code{\link{AutoH20Scoring}}, \code{\link{AutoNLS}}
}
\author{
Adrian Antico
}
\concept{Supervised Learning}
