% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoWord2VecModeler.R
\name{AutoWord2VecModeler}
\alias{AutoWord2VecModeler}
\title{Automated word2vec data generation via H2O}
\usage{
AutoWord2VecModeler(data, BuildType = "Combined",
  stringCol = c("Text_Col1", "Text_Col2"),
  KeepStringCol = FALSE, model_path = NULL, vects = 100,
  SaveStopWords = FALSE, MinWords = 1, WindowSize = 12,
  Epochs = 25, StopWords = NULL, SaveModel = "standard",
  Threads = max(1, parallel::detectCores() - 2), MaxMemory = "28G",
  SaveOutput = FALSE)
}
\arguments{
\item{data}{Source data table to merge vects onto}
  
\item{BuildType}{Select "combined" to build a single model for all columns. Otherwise a single model will be build for each text columnn}

\item{stringCol}{A string name for the column to convert via word2vec}

\item{KeepStringCol}{Set to TRUE if you want to keep the original string column that you convert via word2vec}

\item{model_path}{A string path to the location where you want the model and metadata stored}

\item{vects}{The number of vectors to retain from the word2vec model}

\item{SaveStopWords}{Set to TRUE to save the stop words used}

\item{MinWords}{For H2O word2vec model}

\item{WindowSize}{For H2O word2vec model}

\item{Epochs}{For H2O word2vec model}

\item{StopWords}{For H2O word2vec model}

\item{SaveModel}{Set to "standard" to save normally; set to "mojo" to save as mojo. NOTE: while you can save a mojo, I haven't figured out how to score it in the AutoH20Scoring function.}

\item{Threads}{Number of available threads you want to dedicate to model building}

\item{MaxMemory}{Amount of memory you want to dedicate to model building}

\item{SaveOutput}{Set to TRUE to save your models to file}
}
\description{
This function allows you to automatically build a word2vec model and merge the data onto your supplied dataset
}
\examples{
\donttest{
data <- AutoWord2VecModeler(data,
                            stringCol = c("Text_Col1",
                                          "Text_Col2"),
                            KeepStringCol = FALSE,
                            model_path = NULL,
                            vects = 100,
                            SaveStopWords = FALSE,
                            MinWords = 1,
                            WindowSize = 1,
                            Epochs = 25,
                            StopWords = NULL,
                            SaveModel = "standard",
                            Threads = max(1,parallel::detectCores()-2),
                            MaxMemory = "28G",
                            SaveOutput = TRUE)
}
}
\seealso{
Other Feature Engineering: \code{\link{AutoDataPartition}},
  \code{\link{AutoTransformationCreate}},
  \code{\link{AutoTransformationScore}},
  \code{\link{CreateCalendarVariables}},
  \code{\link{DT_GDL_Feature_Engineering}},
  \code{\link{DummifyDT}},
  \code{\link{FAST_GDL_Feature_Engineering}},
  \code{\link{GDL_Feature_Engineering}},
  \code{\link{ModelDataPrep}},
  \code{\link{Scoring_GDL_Feature_Engineering}}
}
\author{
Adrian Antico
}
\concept{Feature Engineering}
