% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TextPrepFunctions.R
\name{AutoWordFreq}
\alias{AutoWordFreq}
\title{Automated Word Frequency and Word Cloud Creation}
\usage{
AutoWordFreq(
  data,
  TextColName = "DESCR",
  GroupColName = "ClusterAllNoTarget",
  GroupLevel = 0,
  RemoveEnglishStopwords = TRUE,
  Stemming = TRUE,
  StopWords = c("bla", "bla2")
)
}
\arguments{
\item{data}{Source data table}

\item{TextColName}{A string name for the column}

\item{GroupColName}{Set to NULL to ignore, otherwise set to Cluster column name (or factor column name)}

\item{GroupLevel}{Must be set if GroupColName is defined. Set to cluster ID (or factor level)}

\item{RemoveEnglishStopwords}{Set to TRUE to remove English stop words, FALSE to ignore}

\item{Stemming}{Set to TRUE to run stemming on your text data}

\item{StopWords}{Add your own stopwords, in vector format}
}
\description{
This function builds a word frequency table and a word cloud. It prepares data, cleans text, and generates output.
}
\examples{
\dontrun{
data <- data.table::data.table(
DESCR = c(
  "Gru", "Gru", "Gru", "Gru", "Gru", "Gru", "Gru",
  "Gru", "Gru", "Gru", "Gru", "Gru", "Gru", "Urkle",
  "Urkle", "Urkle", "Urkle", "Urkle", "Urkle", "Urkle",
  "Gru", "Gru", "Gru", "bears", "bears", "bears",
  "bears", "bears", "bears", "smug", "smug", "smug", "smug",
  "smug", "smug", "smug", "smug", "smug", "smug",
  "smug", "smug", "smug", "smug", "smug", "eats", "eats",
  "eats", "eats", "eats", "eats", "beats", "beats", "beats", "beats",
  "beats", "beats", "beats", "beats", "beats", "beats",
  "beats", "science", "science", "Dwigt", "Dwigt", "Dwigt", "Dwigt",
  "Dwigt", "Dwigt", "Dwigt", "Dwigt", "Dwigt", "Dwigt",
  "Schrute", "Schrute", "Schrute", "Schrute", "Schrute",
  "Schrute", "Schrute", "James", "James", "James", "James",
  "James", "James", "James", "James", "James", "James",
  "Halpert", "Halpert", "Halpert", "Halpert",
  "Halpert", "Halpert", "Halpert", "Halpert"))
data <- AutoWordFreq(
  data,
  TextColName = "DESCR",
  GroupColName = NULL,
  GroupLevel = NULL,
  RemoveEnglishStopwords = FALSE,
  Stemming = FALSE,
  StopWords = c("Bla"))
}
}
\seealso{
Other EDA: 
\code{\link{ScatterCopula}()}
}
\author{
Adrian Antico
}
\concept{EDA}
