% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RemixAutoML.R
\name{AutoXGBoostRegression}
\alias{AutoXGBoostRegression}
\title{AutoXGBoostRegression is an automated XGBoost modeling framework with grid-tuning and model evaluation}
\usage{
AutoXGBoostRegression(data, ValidationData = NULL, TestData = NULL,
  TargetColumnName = NULL, FeatureColNames = NULL,
  CatFeatures = NULL, IDcols = NULL, eval_metric = "RMSE",
  Trees = 50, GridTune = FALSE, grid_eval_metric = "mae",
  TreeMethod = "hist", MaxModelsInGrid = 10, NThreads = 8,
  model_path = NULL, ModelID = "FirstModel", NumOfParDepPlots = 3,
  Verbose = 0, ReturnModelObjects = TRUE, SaveModelObjects = FALSE,
  PassInGrid = NULL)
}
\arguments{
\item{data}{This is your data set for training and testing your model}

\item{ValidationData}{This is your holdout data set used in modeling either refine your hyperparameters.}

\item{TestData}{This is your holdout data set. Catboost using both training and validation data in the training process so you should evaluate out of sample performance with this data set.}

\item{TargetColumnName}{Either supply the target column name OR the column number where the target is located (but not mixed types).}

\item{FeatureColNames}{Either supply the feature column names OR the column number where the target is located (but not mixed types)}

\item{CatFeatures}{Supply a vector of character names of the columns. Internally, the function will convert these columns to dummary variables.}

\item{IDcols}{A vector of column names or column numbers to keep in your data but not include in the modeling.}

\item{eval_metric}{This is the metric used to identify best grid tuned model. Choose from "r2", "RMSE", "MSE", "MAE"}

\item{Trees}{The maximum number of trees you want in your models}

\item{GridTune}{Set to TRUE to run a grid tuning procedure. Set a number in MaxModelsInGrid to tell the procedure how many models you want to test.}

\item{grid_eval_metric}{Choose from "poisson","mae","mape","mse","msle","kl","cs","r2"}

\item{TreeMethod}{Choose from "hist", "gpu_hist"}

\item{MaxModelsInGrid}{Number of models to test from grid options (243 total possible options)}

\item{NThreads}{Set the maximum number of threads you'd like to dedicate to the model run. E.g. 8}

\item{model_path}{A character string of your path file to where you want your output saved}

\item{ModelID}{A character string to name your model and output}

\item{NumOfParDepPlots}{Tell the function the number of partial dependence calibration plots you want to create.}

\item{Verbose}{Set to 0 if you want to suppress model evaluation updates in training}

\item{ReturnModelObjects}{Set to TRUE to output all modeling objects (E.g. plots and evaluation metrics)}

\item{SaveModelObjects}{Set to TRUE to return all modeling objects to your environment}

\item{PassInGrid}{Default is NULL. Provide a data.table of grid options from a previous run.}
}
\value{
Saves to file and returned in list: VariableImportance.csv, Model, ValidationData.csv, EvalutionPlot.png, EvalutionBoxPlot.png, EvaluationMetrics.csv, ParDepPlots.R a named list of features with partial dependence calibration plots, ParDepBoxPlots.R, GridCollect, and GridList
}
\description{
AutoXGBoostRegression is an automated XGBoost modeling framework with grid-tuning and model evaluation that runs a variety of steps. First, the function will run a random grid tune over N number of models and find which model is the best (a default model is always included in that set). Once the model is identified and built, several other outputs are generated: validation data with predictions, evaluation plot, evaluation boxplot, evaluation metrics, variable importance, partial dependence calibration plots, partial dependence calibration box plots, and column names used in model fitting.
}
\examples{
\donttest{
Correl <- 0.85
N <- 10000
data <- data.table::data.table(Target = runif(N))
data[, x1 := qnorm(Target)]
data[, x2 := runif(N)]
data[, Independent_Variable1 := log(pnorm(Correl * x1 +
                                            sqrt(1-Correl^2) * qnorm(x2)))]
data[, Independent_Variable2 := (pnorm(Correl * x1 +
                                         sqrt(1-Correl^2) * qnorm(x2)))]
data[, Independent_Variable3 := exp(pnorm(Correl * x1 +
                                            sqrt(1-Correl^2) * qnorm(x2)))]
data[, Independent_Variable4 := exp(exp(pnorm(Correl * x1 +
                                                sqrt(1-Correl^2) * qnorm(x2))))]
data[, Independent_Variable5 := sqrt(pnorm(Correl * x1 +
                                             sqrt(1-Correl^2) * qnorm(x2)))]
data[, Independent_Variable6 := (pnorm(Correl * x1 +
                                         sqrt(1-Correl^2) * qnorm(x2)))^0.10]
data[, Independent_Variable7 := (pnorm(Correl * x1 +
                                         sqrt(1-Correl^2) * qnorm(x2)))^0.25]
data[, Independent_Variable8 := (pnorm(Correl * x1 +
                                         sqrt(1-Correl^2) * qnorm(x2)))^0.75]
data[, Independent_Variable9 := (pnorm(Correl * x1 +
                                         sqrt(1-Correl^2) * qnorm(x2)))^2]
data[, Independent_Variable10 := (pnorm(Correl * x1 +
                                          sqrt(1-Correl^2) * qnorm(x2)))^4]
data[, Independent_Variable11 := as.factor(
  ifelse(Independent_Variable2 < 0.20, "A",
         ifelse(Independent_Variable2 < 0.40, "B",
                ifelse(Independent_Variable2 < 0.6,  "C",
                       ifelse(Independent_Variable2 < 0.8,  "D", "E")))))]
data[, ':=' (x1 = NULL, x2 = NULL)]
TestModel <- AutoXGBoostRegression(data,
                                   ValidationData = NULL,
                                   TestData = NULL,
                                   TargetColumnName = 1,
                                   FeatureColNames = 2:12,
                                   CatFeatures = 12,
                                   IDcols = NULL,
                                   eval_metric = "RMSE",
                                   Trees = 50,
                                   GridTune = TRUE,
                                   grid_eval_metric = "mae",
                                   MaxModelsInGrid = 10,
                                   NThreads = 8,
                                   TreeMethod = "hist",
                                   model_path = getwd(),
                                   ModelID = "FirstModel",
                                   NumOfParDepPlots = 3,
                                   ReturnModelObjects = TRUE,
                                   SaveModelObjects = FALSE,
                                   PassInGrid = NULL)
}
}
\seealso{
Other Supervised Learning: \code{\link{AutoCatBoostClassifier}},
  \code{\link{AutoCatBoostMultiClass}},
  \code{\link{AutoCatBoostRegression}},
  \code{\link{AutoCatBoostScoring}},
  \code{\link{AutoH2OMLScoring}},
  \code{\link{AutoH2OModeler}},
  \code{\link{AutoH2OScoring}},
  \code{\link{AutoH2oDRFClassifier}},
  \code{\link{AutoH2oDRFMultiClass}},
  \code{\link{AutoH2oDRFRegression}},
  \code{\link{AutoH2oGBMClassifier}},
  \code{\link{AutoH2oGBMMultiClass}},
  \code{\link{AutoH2oGBMRegression}},
  \code{\link{AutoNLS}},
  \code{\link{AutoRecommenderScoring}},
  \code{\link{AutoRecommender}}, \code{\link{AutoTS}},
  \code{\link{AutoXGBoostClassifier}},
  \code{\link{AutoXGBoostMultiClass}},
  \code{\link{AutoXGBoostScoring}}
}
\author{
Adrian Antico
}
\concept{Supervised Learning}
