% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyOutputs.R
\name{BoxPlotModals}
\alias{BoxPlotModals}
\title{BoxPlotModals}
\usage{
BoxPlotModals(id, NumberPlots, AppWidth = 12L)
}
\arguments{
\item{id}{= 'PlotVariablesModals'}

\item{AppWidth}{= 12L}

\item{PlotNumber}{1}
}
\description{
Plot Variables Modals
}
\seealso{
Other Shiny: 
\code{\link{DataTable2}()},
\code{\link{DataTable}()},
\code{\link{DateInput}()},
\code{\link{GenerateEvaluationMetrics}()},
\code{\link{NumericInput}()},
\code{\link{PickerInput_GetLevels2}()},
\code{\link{PickerInput_GetLevels}()},
\code{\link{PickerInput}()},
\code{\link{PreparePlotData}()},
\code{\link{ReturnParam}()},
\code{\link{SelectizeInput}()},
\code{\link{StandardPlotsModal1}()},
\code{\link{StandardPlotsModal2}()},
\code{\link{StandardPlotsModal3}()},
\code{\link{StandardPlotsModal4}()},
\code{\link{TextInput}()}
}
\author{
Adrian Antico
}
\concept{Shiny}
\keyword{internal}
