% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CARMA-HelperFunctions.R
\name{CARMA_Get_IndepentVariablesPass}
\alias{CARMA_Get_IndepentVariablesPass}
\title{CARMA_Get_IndepentVariablesPass}
\usage{
CARMA_Get_IndepentVariablesPass(HierarchGroups)
}
\arguments{
\item{HierarchGroups}{Supply HierarchGroups}
}
\description{
CARMA_Get_IndepentVariablesPass is to help manage carma code
}
\seealso{
Other Carma Helper: 
\code{\link{CARMA_Define_Args}()},
\code{\link{CarmaCatBoostKeepVarsGDL}()},
\code{\link{CarmaH2OKeepVarsGDL}()},
\code{\link{CarmaXGBoostKeepVarsGDL}()}
}
\author{
Adrian Antico
}
\concept{Carma Helper}
