% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelMetrics.R
\name{CarmaHoldoutMetrics}
\alias{CarmaHoldoutMetrics}
\title{CarmaHoldoutMetrics}
\usage{
CarmaHoldoutMetrics(
  DATA = TestDataEval,
  TARGETCOLUMNNAME = TargetColumnName,
  GROUPVARIABLES = GroupingVariables
)
}
\arguments{
\item{DATA}{TestDataEval}

\item{TARGETCOLUMNNAME}{TargetColumnName}

\item{GROUPVARIABLES}{GroupVariables}
}
\description{
CarmaHoldoutMetrics
}
\author{
Adrian Antico
}
\concept{Time Series}
