% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CatBoostHelpers.R
\name{CatBoostParameterGrids}
\alias{CatBoostParameterGrids}
\title{CatBoostParameterGrids}
\usage{
CatBoostParameterGrids(
  TaskType = "CPU",
  Shuffles = 1L,
  NTrees = seq(1000L, 10000L, 1000L),
  Depth = seq(4L, 16L, 2L),
  LearningRate = c(0.01, 0.02, 0.03, 0.04),
  L2_Leaf_Reg = seq(1, 10, 1),
  RandomStrength = seq(1, 2, 0.1),
  BorderCount = seq(32, 256, 32),
  RSM = c(0.8, 0.85, 0.9, 0.95, 1),
  BootStrapType = c("Bayesian", "Bernoulli", "Poisson", "MVS", "No"),
  GrowPolicy = c("SymmetricTree", "Depthwise", "Lossguide")
)
}
\arguments{
\item{TaskType}{"GPU" or "CPU"}

\item{Shuffles}{The number of shuffles you want to apply to each grid}

\item{NTrees}{seq(1000L, 10000L, 1000L)}

\item{Depth}{seq(4L, 16L, 2L)}

\item{LearningRate}{seq(0.01,.10,0.01)}

\item{L2_Leaf_Reg}{c(1.0:10.0)}

\item{RandomStrength}{seq(1, 2, 0.1)}

\item{BorderCount}{seq(32,256,32)}

\item{RSM}{CPU ONLY, Random subspace method.c(0.80, 0.85, 0.90, 0.95, 1.0)}

\item{BootStrapType}{c("Bayesian", "Bernoulli", "Poisson", "MVS", "No")}

\item{GrowPolicy}{c("SymmetricTree", "Depthwise", "Lossguide")}
}
\value{
A list containing data.table's with the parameters shuffled and ready to test in the bandit framework
}
\description{
CatBoostParameterGrids https://catboost.ai/docs/concepts/r-training-parameters.html
}
\seealso{
Other CatBoost Helpers: 
\code{\link{CatBoostArgsCheck}()},
\code{\link{CatBoostDataConversion}()},
\code{\link{CatBoostDataPrep}()},
\code{\link{CatBoostFinalParams}()},
\code{\link{CatBoostGridParams}()},
\code{\link{CatBoostImportances}()},
\code{\link{CatBoostPDF}()},
\code{\link{CatBoostRemoveFiles}()},
\code{\link{CatBoostValidationData}()}
}
\author{
Adrian Antico
}
\concept{CatBoost Helpers}
