% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReinforcementLearningFunctions.R
\name{CatBoostRegressionParams}
\alias{CatBoostRegressionParams}
\title{CatBoostRegressionParams}
\usage{
CatBoostRegressionParams(
  counter = NULL,
  BanditArmsN = NULL,
  HasTime = NULL,
  MetricPeriods = NULL,
  eval_metric = NULL,
  LossFunction = NULL,
  task_type = NULL,
  NumGPUs = NULL,
  model_path = NULL,
  NewGrid = NULL,
  Grid = NULL,
  ExperimentalGrid = NULL,
  GridClusters = NULL
)
}
\arguments{
\item{counter}{Passthrough}

\item{BanditArmsN}{Passthrough}

\item{HasTime}{Passthrough}

\item{MetricPeriods}{Passthrough}

\item{eval_metric}{Passthrough}

\item{LossFunction}{Passthrough}

\item{task_type}{Passthrough}

\item{NumGPUs}{Passthrough}

\item{model_path}{Passthrough}

\item{NewGrid}{Passthrough}

\item{Grid}{Passthrough}

\item{ExperimentalGrid}{Passthrough}

\item{GridClusters}{Passthrough}
}
\description{
CatBoostRegressionParams
}
\seealso{
Other Supervised Learning: 
\code{\link{AutoH2OScoring}()},
\code{\link{CatBoostClassifierParams}()},
\code{\link{CatBoostMultiClassParams}()},
\code{\link{CatBoostParameterGrids}()},
\code{\link{XGBoostClassifierParams}()},
\code{\link{XGBoostMultiClassParams}()},
\code{\link{XGBoostParameterGrids}()},
\code{\link{XGBoostRegressionMetrics}()},
\code{\link{XGBoostRegressionParams}()}
}
\author{
Adrian Antico
}
\concept{Supervised Learning}
