% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscFunctions.R
\name{ChartTheme}
\alias{ChartTheme}
\title{ChartTheme function is a ggplot theme generator for ggplots}
\usage{
ChartTheme(Size = 12, AngleX = 35, AngleY = 0,
  ChartColor = "lightsteelblue1", BorderColor = "darkblue",
  TextColor = "darkblue", GridColor = "white",
  BackGroundColor = "gray95", LegendPosition = "bottom")
}
\arguments{
\item{Size}{The size of the axis labels and title}

\item{AngleX}{The angle of the x axis labels}

\item{AngleY}{The angle of the Y axis labels}

\item{ChartColor}{"lightsteelblue1",}

\item{BorderColor}{"darkblue",}

\item{TextColor}{"darkblue",}

\item{GridColor}{"white",}

\item{BackGroundColor}{"gray95",}

\item{LegendPosition}{Where to place legend}
}
\value{
An object to pass along to ggplot objects following the "+" sign
}
\description{
This function helps your ggplots look professional with the choice of the two main colors that will dominate the theme
}
\examples{
data <- data.table::data.table(DateTime = as.Date(Sys.time()),
  Target = stats::filter(rnorm(1000,
                               mean = 50,
                               sd = 20),
                         filter=rep(1,10),
                         circular=TRUE))
data[, temp := seq(1:1000)][, DateTime := DateTime - temp][, temp := NULL]
data <- data[order(DateTime)]
p <- ggplot2::ggplot(data, ggplot2::aes(x = DateTime, y = Target)) + ggplot2::geom_line()
p <- p + ChartTheme(Size = 12)
}
\seealso{
Other Misc: \code{\link{AutoH2OTextPrepScoring}},
  \code{\link{PrintObjectsSize}}, \code{\link{RemixTheme}},
  \code{\link{SimpleCap}}, \code{\link{multiplot}},
  \code{\link{percRank}}, \code{\link{tempDatesFun}},
  \code{\link{tokenizeH2O}}
}
\author{
Adrian Antico
}
\concept{Misc}
