% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelMetrics.R
\name{ClassificationMetrics}
\alias{ClassificationMetrics}
\title{ClassificationMetrics}
\usage{
ClassificationMetrics(TestData, Target, Predict)
}
\arguments{
\item{TestData}{Test data from your modeling}

\item{Target}{Name of your target variable}

\item{Predict}{Name of your predicted value variable}
}
\description{
ClassificationMetrics
}
\seealso{
Other Model Evaluation: \code{\link{DT_BinaryConfusionMatrix}},
  \code{\link{RemixClassificationMetrics}}
}
\author{
Adrian Antico
}
\concept{Model Evaluation}
