% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelMetrics.R
\name{ClassificationMetrics}
\alias{ClassificationMetrics}
\title{ClassificationMetrics}
\usage{
ClassificationMetrics(TestData, Thresholds, Target, Predict,
  PositiveOutcome, NegativeOutcome, CostMatrix = c(1, 0, 0, 1))
}
\arguments{
\item{TestData}{Test data from your modeling}

\item{Target}{Name of your target variable}

\item{Predict}{Name of your predicted value variable}

\item{PositiveOutcome}{The value of the positive outcome level}

\item{NegativeOutcome}{The value of the negative outcome level}

\item{CostMatrix}{c(True Positive Cost, False Negative Cost, False Positive Cost, True Negative Cost)}
}
\description{
ClassificationMetrics
}
\seealso{
Other Model Evaluation: \code{\link{DT_BinaryConfusionMatrix}},
  \code{\link{RemixClassificationMetrics}}
}
\author{
Adrian Antico
}
\concept{Model Evaluation}
