% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CL.R
\name{CLForecast}
\alias{CLForecast}
\title{CLForecast()}
\usage{
CLForecast(
  data,
  OutputFilePath = NULL,
  FC_BaseFunnelMeasure = NULL,
  SegmentName = NULL,
  MaxDateForecasted = NULL,
  MaxCalendarDate = NULL,
  ArgsList = NULL,
  MaxCohortPeriods = NULL
)
}
\arguments{
\item{data}{Name of your data.table}

\item{OutputFilePath}{Pathfile for where you want to store you forecast data}

\item{FC_BaseFunnelMeasure}{data.table containing the forward looking FC_BaseFunnelMeasure}

\item{MaxDateForecasted}{Supply a value if you want to backtest}

\item{MaxCalendarDate}{Supply a value if you want to backtest}

\item{ArgsList}{Argument list returned from AutoCatBoostChainLadder}

\item{MaxCohortPeriods}{The maximum amount of ArgsList$CohortPeriodsVariable to utilize for forecasting}
}
\value{
Saves metadata and models to files of your choice. Also returns metadata and models from the function. User specifies both options.
}
\description{
CLForecast() for generating forecasts
}
\seealso{
Other Automated Model Scoring: 
\code{\link{AutoCatBoostScoring}()},
\code{\link{AutoH2OMLScoring}()},
\code{\link{AutoH2OModeler}()},
\code{\link{AutoHurdleScoring}()},
\code{\link{AutoXGBoostScoring}()}
}
\author{
Adrian Antico
}
\concept{Automated Model Scoring}
