% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CL.R
\name{CLForecast}
\alias{CLForecast}
\title{CLForecast}
\usage{
CLForecast(
  data,
  OutputFilePath = NULL,
  FC_BaseFunnelMeasure = NULL,
  SegmentName = NULL,
  MaxDateForecasted = NULL,
  MaxCalendarDate = NULL,
  ArgsList = NULL,
  MaxCohortPeriods = NULL
)
}
\arguments{
\item{data}{N}

\item{OutputFilePath}{P}

\item{FC_BaseFunnelMeasure}{d}

\item{SegmentName}{a}

\item{MaxDateForecasted}{S}

\item{MaxCalendarDate}{S}

\item{ArgsList}{A}

\item{MaxCohortPeriods}{T}
}
\value{
S
}
\description{
CLForecast for generating forecasts
}
\seealso{
Other Automated Model Scoring: 
\code{\link{AutoCatBoostScoring}()},
\code{\link{AutoH2OMLScoring}()},
\code{\link{AutoH2OModeler}()},
\code{\link{AutoHurdleScoring}()},
\code{\link{AutoXGBoostScoring}()}
}
\author{
Adrian Antico
}
\concept{Automated Model Scoring}
