% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataWranglingFunctions.R
\name{ColumnSubsetDataTable}
\alias{ColumnSubsetDataTable}
\title{ColumnSubsetDataTable}
\usage{
ColumnSubsetDataTable(
  data,
  TargetColumnName = NULL,
  DateColumnName = NULL,
  GroupVars = NULL
)
}
\arguments{
\item{data}{data.table}

\item{GroupVars}{Group variables}

\item{Target}{Target variable}

\item{Date}{Date variable}
}
\description{
ColumnSubsetDataTable will subset data tables by column
}
\seealso{
Other Data Wrangling: 
\code{\link{AutoDataDictionaries}()},
\code{\link{AutoHierarchicalFourier}()},
\code{\link{DataDisplayMeta}()},
\code{\link{FakeDataGenerator}()},
\code{\link{FullFactorialCatFeatures}()},
\code{\link{SQL_ClearTable}()},
\code{\link{SQL_DropTable}()},
\code{\link{SQL_Query}()},
\code{\link{SQL_SaveTable}()},
\code{\link{SQL_Server_DBConnection}()},
\code{\link{SQL_UpdateTable}()},
\code{\link{TimeSeriesMelt}()}
}
\author{
Adrian Antico
}
\concept{Data Wrangling}
