% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateHolidayVariables.R
\name{CreateHolidayVariables}
\alias{CreateHolidayVariables}
\title{CreateHolidayVariables Create Holiday Count Columns}
\usage{
CreateHolidayVariables(
  data,
  DateCols = NULL,
  HolidayGroups = c("USPublicHolidays", "EasterGroup", "ChristmasGroup",
    "OtherEcclesticalFeasts"),
  Holidays = NULL,
  GroupingVars = NULL,
  Print = FALSE
)
}
\arguments{
\item{data}{This is your data}

\item{DateCols}{Supply either column names or column numbers of your date columns you want to use for creating calendar variables}

\item{HolidayGroups}{Pick groups}

\item{Holidays}{Pick holidays}

\item{GroupingVars}{Grouping variable names}

\item{Print}{Set to TRUE to print iteration number to console}
}
\value{
Returns your data.table with the added holiday indicator variable
}
\description{
CreateHolidayVariables Rapidly creates holiday count variables based on the date columns you provide
}
\examples{
\donttest{
data <- data.table::data.table(Date = as.Date(rep('2018-01-01 00:00:00', 100L)))
data1 <- data.table::data.table(Date = as.Date(rep('2018-01-01 00:00:00', 100L)))
data[, ID := 1L:.N][, Date := Date + lubridate::days(ID)][, ID := NULL]
data1[, ID := 1L:.N][, Date := Date + lubridate::days(ID)][, ID := NULL]
data[, GroupVar := "A"]
data1[, GroupVar := "B"]
data <- data.table::rbindlist(list(data,data1))
data <- CreateHolidayVariables(
   data,
   DateCols = "Date",
   HolidayGroups = c("USPublicHolidays","EasterGroup","ChristmasGroup","OtherEcclesticalFeasts"),
   Holidays = NULL,
   GroupingVars = "GroupVar",
   Print = FALSE)
}
}
\seealso{
Other Feature Engineering: 
\code{\link{AutoDataPartition}()},
\code{\link{AutoFourierFeatures}()},
\code{\link{AutoLagRollStatsScoring}()},
\code{\link{AutoLagRollStats}()},
\code{\link{AutoTransformationCreate}()},
\code{\link{AutoTransformationScore}()},
\code{\link{AutoWord2VecModeler}()},
\code{\link{ContinuousTimeDataGenerator}()},
\code{\link{CreateCalendarVariables}()},
\code{\link{DT_GDL_Feature_Engineering}()},
\code{\link{DummifyDT}()},
\code{\link{H2oAutoencoder}()},
\code{\link{ModelDataPrep}()},
\code{\link{Partial_DT_GDL_Feature_Engineering}()},
\code{\link{TimeSeriesFill}()}
}
\author{
Adrian Antico
}
\concept{Feature Engineering}
