% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SystemFunctions.R
\name{CreateProjectFolders}
\alias{CreateProjectFolders}
\title{PathNormalizer Converts path files to proper path files}
\usage{
CreateProjectFolders(ProjectName = input$ID_NewProjectName,
  RootPath = input$ID_Root_Folder, ExistsButNoProjectList = FALSE,
  Local = FALSE)
}
\arguments{
\item{ProjectName}{This is the name of a project which will be the name of the file created in the root folder}

\item{RootPath}{This is the path file to the root folder}

\item{ExistsButNoProjectList}{Set to TRUE if the folder exists but not the ProjectList file}
}
\value{
Returns a proper path file string
}
\description{
PathNormalizer Converts path files to proper path files
}
\seealso{
Other System Functions: \code{\link{PathNormalizer}}
}
\author{
Adrian Antico
}
\concept{System Functions}
