% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataWranglingFunctions.R
\name{DataDisplayMeta}
\alias{DataDisplayMeta}
\title{DataDisplayMeta}
\usage{
DataDisplayMeta(data)
}
\arguments{
\item{data}{Source data}
}
\description{
DataDisplayMeta
}
\seealso{
Other Data Wrangling: \code{\link{AutoHierarchicalFourier}},
  \code{\link{ColumnSubsetDataTable}},
  \code{\link{FullFactorialCatFeatures}},
  \code{\link{TimeSeriesMelt}}
}
\author{
Adrian Antico
}
\concept{Data Wrangling}
