% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyFunctions.R
\name{DataTable}
\alias{DataTable}
\title{DataTable}
\usage{
DataTable(data)
}
\arguments{
\item{data}{source data.table}
}
\description{
Fully loaded DT::datatable() with args prefilled
}
\examples{
\dontrun{
# Rmarkdown example of DataTable inside a <details> </Details> section

```{r Get Dependencies For DT::datatable(), echo=FALSE,include = FALSE}
# You need this code to conduct the magic dependences attaching...
DT::datatable(matrix())
```

```{js Nest All DT::datatable() inside a details drop down, echo=FALSE}
setTimeout(function() {
  var codes = document.querySelectorAll('.dataTables_wrapper');
  var code, i, d, s, p;
  for (i = 0; i < codes.length; i++) {
    code = codes[i];
    p = code.parentNode;
    d = document.createElement('details');
    s = document.createElement('summary');
    s.innerText = 'Details';
    // <details><summary>Details</summary></details>
      d.appendChild(s);
    // move the code into <details>
      p.replaceChild(d, code);
    d.appendChild(code);
  }
});
```

```{r Example, echo = FALSE}
RemixAutoML::DataTable(data)
````

# Shiny Usage
output$Table <- shiny::renderUI({RemixAutoML::DataTable(data)})

}

}
\seealso{
Other Shiny: 
\code{\link{PreparePlotData}()}
}
\author{
Adrian Antico
}
\concept{Shiny}
