% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataWranglingFunctions.R
\name{DifferenceData}
\alias{DifferenceData}
\title{DifferenceData}
\usage{
DifferenceData(data, ColumnsToDiff = c(names(data)[2:ncol(data)]),
  CARMA = FALSE, TargetVariable = NULL, GroupingVariable = NULL)
}
\arguments{
\item{data}{Source data}

\item{ColumnsToDiff}{The column numbers you want differenced}

\item{CARMA}{Set to TRUE for CARMA functions}

\item{TargetVariable}{The target variable name}

\item{GroupingVariable}{Difference data by group}
}
\description{
DifferenceData differences your data set
}
\seealso{
Other Time Series: \code{\link{AutoBanditNNet}},
  \code{\link{AutoBanditSarima}}, \code{\link{AutoTBATS}},
  \code{\link{CarmaHoldoutMetrics}},
  \code{\link{DifferenceDataReverse}}
}
\author{
Adrian Antico
}
\concept{Time Series}
