% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataWranglingFunctions.R
\name{DifferenceDataReverse}
\alias{DifferenceDataReverse}
\title{DifferenceDataReverse}
\usage{
DifferenceDataReverse(data, ScoreData = Forecasts$Predictions,
  LastRow = DiffTrainOutput$LastRow$Weekly_Sales, CARMA = FALSE,
  TargetCol = TargetColumnName, FirstRow = DiffTrainOutput$FirstRow,
  GroupingVariables = NULL)
}
\arguments{
\item{data}{Pre differenced scoring data}

\item{ScoreData}{Predicted values from ML model}

\item{LastRow}{The last row from training data target variables}

\item{CARMA}{Set to TRUE for CARMA utilization}

\item{TargetCol}{Target column name}

\item{FirstRow}{The first row of the target variable}

\item{GroupingVariables}{Group columns}
}
\description{
DifferenceDataReverse reverses the difference
}
\seealso{
Other Time Series: \code{\link{AutoBanditNNet}},
  \code{\link{AutoBanditSarima}}, \code{\link{AutoTBATS}},
  \code{\link{CarmaHoldoutMetrics}},
  \code{\link{DifferenceData}}
}
\author{
Adrian Antico
}
\concept{Time Series}
