% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Adrian_Antico_Modeling_Tools_Library.R
\name{DummifyDT}
\alias{DummifyDT}
\title{DummifyDT creates dummy variables for the selected columns}
\usage{
DummifyDT(data, cols, KeepBaseCols = TRUE, OneHot = TRUE)
}
\arguments{
\item{data}{the data set to run the micro auc on}

\item{cols}{a vector with the names of the columns you wish to dichotomize}

\item{KeepBaseCols}{set to TRUE to keep the original columns used in the dichotomization process}

\item{OneHot}{Set to TRUE to run one hot encoding, FALSE to generate N columns for N levels}
}
\value{
data table with new dummy variables columns and optionally removes base columns
}
\description{
DummifyDT creates dummy variables for the selected columns
}
\examples{
test <- data.table::data.table(Value = runif(100000),
                   FactorCol = sample(x = c(letters,
                                            LETTERS,
                                            paste0(letters,letters),
                                            paste0(LETTERS,LETTERS),
                                            paste0(letters,LETTERS),
                                            paste0(LETTERS,letters)),
                                      size = 100000,
                                      replace = TRUE))
test <- DummifyDT(data = test,
                  cols = "FactorCol",
                  KeepBaseCols = FALSE)
ncol(test)
test[, sum(FactorCol_gg)]
}
\author{
Adrian Antico
}
