% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DummifyDT.R
\name{DummifyDT}
\alias{DummifyDT}
\title{DummifyDT creates dummy variables for the selected columns.}
\usage{
DummifyDT(data, cols, KeepFactorCols = FALSE, OneHot = FALSE,
  SaveFactorLevels = FALSE, SavePath = NULL,
  ImportFactorLevels = FALSE, FactorLevelsList = NULL,
  ClustScore = FALSE, ReturnFactorLevels = FALSE)
}
\arguments{
\item{data}{The data set to run the micro auc on}

\item{cols}{A vector with the names of the columns you wish to dichotomize}

\item{KeepFactorCols}{Set to TRUE to keep the original columns used in the dichotomization process}

\item{OneHot}{Set to TRUE to run one hot encoding, FALSE to generate N columns for N levels}

\item{SaveFactorLevels}{Set to TRUE to save unique levels of each factor column to file as a csv}

\item{SavePath}{Provide a file path to save your factor levels. Use this for models that you have to create dummy variables for.}

\item{ImportFactorLevels}{Instead of using the data you provide, import the factor levels csv to ensure you build out all of the columns you trained with in modeling.}

\item{FactorLevelsList}{Supply a list of factor variable levels}

\item{ClustScore}{This is for scoring AutoKMeans. Set to FALSE for all other applications.}

\item{ReturnFactorLevels}{If you want a named list of all the factor levels returned, set this to TRUE. Doing so will cause the function to return a list with the source data.table and the list of factor variables' levels}
}
\value{
Either a data table with new dummy variables columns and optionally removes base columns (if ReturnFactorLevels is FALSE), otherwise a list with the data.table and a list of the factor levels.
}
\description{
DummifyDT creates dummy variables for the selected columns. Either one-hot encoding, N+1 columns for N levels, or N columns for N levels.
}
\examples{
test <- data.table::data.table(Value = runif(100000),
                   FactorCol = sample(x = c(letters,
                                            LETTERS,
                                            paste0(letters,letters),
                                            paste0(LETTERS,LETTERS),
                                            paste0(letters,LETTERS),
                                            paste0(LETTERS,letters)),
                                      size = 100000,
                                      replace = TRUE))
test <- DummifyDT(data = test,
                  cols = "FactorCol",
                  KeepFactorCols = FALSE,
                  OneHot = FALSE,
                  SaveFactorLevels = FALSE,
                  SavePath = normalizePath("./"),
                  ImportFactorLevels = FALSE,
                  FactorLevelsList = NULL,
                  ClustScore = FALSE,
                  ReturnFactorLevels = FALSE)
ncol(test)
test[, sum(FactorCol_gg)]
}
\seealso{
Other Feature Engineering: \code{\link{AutoDataPartition}},
  \code{\link{AutoFourierFeatures}},
  \code{\link{AutoLagRollStatsScoring}},
  \code{\link{AutoLagRollStats}},
  \code{\link{AutoTransformationCreate}},
  \code{\link{AutoTransformationScore}},
  \code{\link{AutoWord2VecModeler}},
  \code{\link{ContinuousTimeDataGenerator}},
  \code{\link{CreateCalendarVariables}},
  \code{\link{CreateHolidayVariables}},
  \code{\link{DT_GDL_Feature_Engineering}},
  \code{\link{H2oAutoencoder}},
  \code{\link{ModelDataPrep}},
  \code{\link{Partial_DT_GDL_Feature_Engineering}},
  \code{\link{TimeSeriesFill}}
}
\author{
Adrian Antico
}
\concept{Feature Engineering}
