% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RemixAutoML.R
\name{DummifyDT}
\alias{DummifyDT}
\title{DummifyDT creates dummy variables for the selected columns.}
\usage{
DummifyDT(data, cols, KeepBaseCols = FALSE, OneHot = TRUE,
  ClustScore = FALSE)
}
\arguments{
\item{data}{the data set to run the micro auc on}

\item{cols}{a vector with the names of the columns you wish to dichotomize}

\item{KeepBaseCols}{set to TRUE to keep the original columns used in the dichotomization process}

\item{OneHot}{Set to TRUE to run one hot encoding, FALSE to generate N columns for N levels}

\item{ClustScore}{This is for scoring AutoKMeans. Set to FALSE for all other applications.}
}
\value{
data table with new dummy variables columns and optionally removes base columns
}
\description{
DummifyDT creates dummy variables for the selected columns. Either one-hot encoding, N+1 columns for N levels, or N columns for N levels.
}
\examples{
test <- data.table::data.table(Value = runif(100000),
                   FactorCol = sample(x = c(letters,
                                            LETTERS,
                                            paste0(letters,letters),
                                            paste0(LETTERS,LETTERS),
                                            paste0(letters,LETTERS),
                                            paste0(LETTERS,letters)),
                                      size = 100000,
                                      replace = TRUE))
test <- DummifyDT(data = test,
                  cols = "FactorCol",
                  KeepBaseCols = FALSE)
ncol(test)
test[, sum(FactorCol_gg)]
}
\seealso{
Other Feature Engineering: \code{\link{AutoWord2VecModeler}},
  \code{\link{DT_GDL_Feature_Engineering}},
  \code{\link{FAST_GDL_Feature_Engineering}},
  \code{\link{GDL_Feature_Engineering}},
  \code{\link{ModelDataPrep}},
  \code{\link{Scoring_GDL_Feature_Engineering}}
}
\author{
Adrian Antico
}
\concept{Feature Engineering}
