% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelInsights.R
\name{EDAReport}
\alias{EDAReport}
\title{EDAReport}
\usage{
EDAReport(
  data = NULL,
  DataName = NULL,
  UnivariateVars = NULL,
  CorrVars = NULL,
  TrendVars = NULL,
  TrendDateVar = NULL,
  TrendGroupVar = NULL,
  TargetVar = NULL,
  OutputPath = NULL,
  Theme = "dark"
)
}
\arguments{
\item{data}{NULL}

\item{DataName}{NULL}

\item{UnivariateVars}{NULL}

\item{CorrVars}{NULL}

\item{TrendVars}{NULL}

\item{TrendDateVar}{NULL}

\item{TrendGroupVar}{NULL}

\item{TargetVar}{NULL}

\item{OutputPath}{List of output objects}

\item{Theme}{AutoPlots Theme parameter value}
}
\description{
EDAReport is an Rmarkdown report for viewing EDA results
}
\seealso{
Other Reports: 
\code{\link{ModelInsightsReport}()}
}
\author{
Adrian Antico
}
\concept{Reports}
