% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelEvaluationPlots.R
\name{EvalPlot}
\alias{EvalPlot}
\title{EvalPlot}
\usage{
EvalPlot(
  data,
  PredictionColName = c("PredictedValues"),
  TargetColName = c("ActualValues"),
  GraphType = c("calibration"),
  PercentileBucket = 0.05,
  aggrfun = function(x) mean(x, na.rm = TRUE)
)
}
\arguments{
\item{data}{Data containing predicted values and actual values for comparison}

\item{PredictionColName}{String representation of column name with predicted values from model}

\item{TargetColName}{String representation of column name with target values from model}

\item{GraphType}{Calibration or boxplot - calibration aggregated data based on summary statistic; boxplot shows variation}

\item{PercentileBucket}{Number of buckets to partition the space on (0,1) for evaluation}

\item{aggrfun}{The statistics function used in aggregation, listed as a function}
}
\value{
Calibration plot or boxplot
}
\description{
This function automatically builds calibration plots and calibration boxplots for model evaluation using regression, quantile regression, and binary and multinomial classification
}
\examples{
\dontrun{
# Create fake data
data <- RemixAutoML::FakeDataGenerator(
  Correlation = 0.70, N = 10000000, Classification = TRUE)
data.table::setnames(data, "IDcol_1", "Predict")

# Run function
EvalPlot(data,
         PredictionColName = "Predict",
         TargetColName = "Adrian",
         GraphType = "calibration",
         PercentileBucket = 0.05,
         aggrfun = function(x) mean(x, na.rm = TRUE))
}
}
\seealso{
Other Model Evaluation and Interpretation: 
\code{\link{AutoLimeAid}()},
\code{\link{LimeModel}()},
\code{\link{ML_EvalPlots}()},
\code{\link{ParDepCalPlots}()},
\code{\link{ROCPlot}()},
\code{\link{RedYellowGreen}()},
\code{\link{VI_Plot}()},
\code{\link{threshOptim}()}
}
\author{
Adrian Antico
}
\concept{Model Evaluation and Interpretation}
