% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SQL_Functions.R
\name{ExecuteSSIS}
\alias{ExecuteSSIS}
\title{ExecuteSSIS}
\usage{
ExecuteSSIS(PkgPath = NULL, CSVPath = NULL)
}
\arguments{
\item{PkgPath}{Path to SSIS package includin the package name and the package extension .dtsx}

\item{CSVPath}{Path to the csv output data location including the name of the file and the .csv extension}
}
\description{
Run an SSIS package from R. Function will check to make sure you can run an SSIS package and it will remove the output file if it exists so as to not append data on top of it.
}
\seealso{
Other Database: 
\code{\link{AutoDataDictionaries}()},
\code{\link{SQL_ClearTable}()},
\code{\link{SQL_DropTable}()},
\code{\link{SQL_Query_Push}()},
\code{\link{SQL_Query}()},
\code{\link{SQL_SaveTable}()},
\code{\link{SQL_Server_BulkPull}()},
\code{\link{SQL_Server_BulkPush}()},
\code{\link{SQL_Server_DBConnection}()},
\code{\link{SQL_UpdateTable}()}
}
\author{
Adrian Antico
}
\concept{Database}
