% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FakeDataGenerator.R
\name{FakeDataGenerator}
\alias{FakeDataGenerator}
\title{FakeDataGenerator}
\usage{
FakeDataGenerator(N = 1000, ID = 1, ZIP = 1, AddDate = FALSE,
  Classification = FALSE, MultiClass = FALSE)
}
\arguments{
\item{N}{Number of records}

\item{ID}{Number of IDcols to include}

\item{ZIP}{Parameter for zero inflated data generation}

\item{AddDate}{Set to TRUE to include a date column}

\item{Classification}{Set to TRUE to build classification data}

\item{MultiClass}{Set to TRUE to build MultiClass data}
}
\description{
FakeDataGenerator
}
\seealso{
Other Data Wrangling: \code{\link{AutoDataDictionaries}},
  \code{\link{AutoHierarchicalFourier}},
  \code{\link{ColumnSubsetDataTable}},
  \code{\link{DataDisplayMeta}},
  \code{\link{FullFactorialCatFeatures}},
  \code{\link{SQL_Server_DBConnection}},
  \code{\link{TimeSeriesMelt}}
}
\author{
Adrian Antico
}
\concept{Data Wrangling}
