% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FakeDataGenerator.R
\name{FakeDataGenerator}
\alias{FakeDataGenerator}
\title{FakeDataGenerator}
\usage{
FakeDataGenerator(
  Correlation = 0.7,
  N = 25000L,
  ID = 5L,
  ZIP = 5L,
  FactorCount = 2L,
  AddDate = TRUE,
  Classification = FALSE,
  MultiClass = FALSE
)
}
\arguments{
\item{Correlation}{Set the correlation value for simulated data}

\item{N}{Number of records}

\item{ID}{Number of IDcols to include}

\item{ZIP}{Zero Inflation Model target variable creation. Select from 0 to 5 to create that number of distinctly distributed data, stratifed from small to large}

\item{FactorCount}{Number of factor type columns to create}

\item{AddDate}{Set to TRUE to include a date column}

\item{Classification}{Set to TRUE to build classification data}

\item{MultiClass}{Set to TRUE to build MultiClass data}
}
\description{
FakeDataGenerator
}
\examples{
\donttest{
data <- RemixAutoML::FakeDataGenerator(
   Correlation = 0.70,
   N = 25000L,
   ID = 2L,
   ZIP = 2L,
   FactorCount = 2L,
   AddDate = TRUE,
   Classification = FALSE,
   MultiClass = FALSE)
}
}
\seealso{
Other Data Wrangling: 
\code{\link{AutoDataDictionaries}()},
\code{\link{AutoHierarchicalFourier}()},
\code{\link{ColumnSubsetDataTable}()},
\code{\link{DataDisplayMeta}()},
\code{\link{FullFactorialCatFeatures}()},
\code{\link{SQL_Server_DBConnection}()},
\code{\link{TimeSeriesMelt}()}
}
\author{
Adrian Antico
}
\concept{Data Wrangling}
