% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FakeDataGenerator.R
\name{FakeDataGenerator}
\alias{FakeDataGenerator}
\title{FakeDataGenerator}
\usage{
FakeDataGenerator(
  Correlation = 0.7,
  N = 1000L,
  ID = 5L,
  FactorCount = 2L,
  AddDate = TRUE,
  AddComment = FALSE,
  ZIP = 5L,
  TimeSeries = FALSE,
  TimeSeriesTimeAgg = "day",
  ChainLadderData = FALSE,
  Classification = FALSE,
  MultiClass = FALSE
)
}
\arguments{
\item{Correlation}{Set the correlation value for simulated data}

\item{N}{Number of records}

\item{ID}{Number of IDcols to include}

\item{FactorCount}{Number of factor type columns to create}

\item{AddDate}{Set to TRUE to include a date column}

\item{AddComment}{Set to TRUE to add a comment column}

\item{ZIP}{Zero Inflation Model target variable creation. Select from 0 to 5 to create that number of distinctly distributed data, stratifed from small to large}

\item{TimeSeries}{For testing AutoBanditSarima}

\item{TimeSeriesTimeAgg}{Choose from "1min", "5min", "10min", "15min", "30min", "hour", "day", "week", "month", "quarter", "year",}

\item{ChainLadderData}{Set to TRUE to return Chain Ladder Data for using AutoMLChainLadderTrainer}

\item{Classification}{Set to TRUE to build classification data}

\item{MultiClass}{Set to TRUE to build MultiClass data}
}
\description{
Create fake data for examples
}
\examples{
\dontrun{
data <- RemixAutoML::FakeDataGenerator(
   Correlation = 0.70,
   N = 1000L,
   ID = 2L,
   FactorCount = 2L,
   AddDate = TRUE,
   AddComment = FALSE,
   ZIP = 2L,
   TimeSeries = FALSE,
   ChainLadderData = FALSE,
   Classification = FALSE,
   MultiClass = FALSE)
}
}
\author{
Adrian Antico
}
\concept{Data Wrangling}
