% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataWranglingFunctions.R
\name{FullFactorialCatFeatures}
\alias{FullFactorialCatFeatures}
\title{FullFactorialCatFeatures}
\usage{
FullFactorialCatFeatures(GroupVars = GroupVariables, BottomsUp = TRUE)
}
\arguments{
\item{GroupVars}{Character vector of categorical columns to fully interact}

\item{BottomsUp}{TRUE or FALSE. TRUE starts with the most comlex interaction to the main effects}
}
\description{
FullFactorialCatFeatures reverses the difference
}
\seealso{
Other Data Wrangling: 
\code{\link{AutoDataDictionaries}()},
\code{\link{ColumnSubsetDataTable}()},
\code{\link{DataDisplayMeta}()},
\code{\link{FakeDataGenerator}()},
\code{\link{SQL_ClearTable}()},
\code{\link{SQL_DropTable}()},
\code{\link{SQL_Query_Push}()},
\code{\link{SQL_Query}()},
\code{\link{SQL_SaveTable}()},
\code{\link{SQL_Server_DBConnection}()},
\code{\link{SQL_UpdateTable}()},
\code{\link{TimeSeriesMelt}()}
}
\author{
Adrian Antico
}
\concept{Data Wrangling}
