% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RemixAutoML.R
\name{GDL_Feature_Engineering}
\alias{GDL_Feature_Engineering}
\title{An Automated Feature Engineering Function}
\usage{
GDL_Feature_Engineering(data, lags = seq(1, 5, 1), periods = seq(1, 5,
  1), statsFUNs = "mean", statsNames = "mean", targets = NULL,
  groupingVars = NULL, sortDateName = NULL, timeDiffTarget = NULL,
  timeAgg = NULL, WindowingLag = 1, Type = "Lag", Timer = TRUE,
  SkipCols = NULL, SimpleImpute = TRUE)
}
\arguments{
\item{data}{A data.table you want to run the function on}

\item{lags}{A numeric vector of the specific lags you want to have generated. You must include 1 if WindowingLag = 1.}

\item{periods}{A numeric vector of the specific rolling statistics window sizes you want to utilize in the calculations.}

\item{statsFUNs}{Vector that holds functions for your rolling stats, such as function(x) mean(x), function(x) sd(x), or function(x) quantile(x)}

\item{statsNames}{A character vector of the corresponding names to create for the rollings stats variables.}

\item{targets}{A character vector of the column names for the reference column in which you will build your lags and rolling stats}

\item{groupingVars}{A character vector of categorical variable names you will build your lags and rolling stats by}

\item{sortDateName}{The column name of your date column used to sort events over time}

\item{timeDiffTarget}{Specify a desired name for features created for time between events. Set to NULL if you don't want time between events features created.}

\item{timeAgg}{List the time aggregation level for the time between events features, such as "hour", "day", "week", "month", "quarter", or "year"}

\item{WindowingLag}{Set to 0 to build rolling stats off of target columns directly or set to 1 to build the rolling stats off of the lag-1 target}

\item{Type}{List either "Lag" if you want features built on historical values or "Lead" if you want features built on future values}

\item{Timer}{Set to TRUE if you percentage complete tracker printout}

\item{SkipCols}{Defaults to NULL; otherwise supply a character vector of the names of columns to skip}

\item{SimpleImpute}{Set to TRUE for factor level imputation of "0" and numeric imputation of -1}
}
\value{
data.table of original data plus created lags, rolling stats, and time between event lags and rolling stats
}
\description{
Builds autoregressive and rolling stats from target columns and distributed lags and distributed rolling stats for independent features distributed across time. On top of that, you can also create time between instances along with their associated lags and rolling stats. This function works for data with groups and without groups.
}
\examples{
N = 25116
data <- data.table::data.table(DateTime = as.Date(Sys.time()),
  Target = stats::filter(rnorm(N,
                               mean = 50,
                               sd = 20),
                         filter=rep(1,10),
                         circular=TRUE))
data[, temp := seq(1:N)][, DateTime := DateTime - temp][, temp := NULL]
data <- data[order(DateTime)]
data <- GDL_Feature_Engineering(data,
           lags           = c(seq(1,1,1)),
           periods        = c(3),
           statsFUNs      = "mean",
           statsNames     = "mean",
           targets        = c("Target"),
           groupingVars   = NULL,
           sortDateName   = "DateTime",
           timeDiffTarget = NULL,
           timeAgg        = "days",
           WindowingLag   = 1,
           Type           = "Lag",
           Timer          = TRUE,
           SkipCols       = FALSE,
           SimpleImpute   = TRUE)
}
\seealso{
Other Feature Engineering: \code{\link{AutoDataPartition}},
  \code{\link{AutoTransformationCreate}},
  \code{\link{AutoTransformationScore}},
  \code{\link{AutoWord2VecModeler}},
  \code{\link{CreateCalendarVariables}},
  \code{\link{DT_GDL_Feature_Engineering}},
  \code{\link{DummifyDT}},
  \code{\link{FAST_GDL_Feature_Engineering}},
  \code{\link{ModelDataPrep}},
  \code{\link{Scoring_GDL_Feature_Engineering}}
}
\author{
Adrian Antico
}
\concept{Feature Engineering}
