% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RemixAutoML.R
\name{GenTSAnomVars}
\alias{GenTSAnomVars}
\title{GenTSAnomVars is an automated z-score anomaly detection via GLM-like procedure}
\usage{
GenTSAnomVars(data, ValueCol = "Value", GroupVar1 = "SKU",
  GroupVar2 = NULL, DateVar = "DATE", High = 1.96, Low = -1.96,
  KeepAllCols = FALSE, DataScaled = TRUE)
}
\arguments{
\item{data}{the source residuals data.table}

\item{ValueCol}{the numeric column to run anomaly detection over}

\item{GroupVar1}{this is a group by variable}

\item{GroupVar2}{this is another group by variable}

\item{DateVar}{this is a time variable for grouping}

\item{High}{this is the threshold on the high end}

\item{Low}{this is the threshold on the low end}

\item{KeepAllCols}{set to TRUE to remove the intermediate features}

\item{DataScaled}{set to TRUE if you already scaled your data}
}
\value{
The original data.table with the added columns merged in
}
\description{
GenTSAnomVars is an automated z-score anomaly detection via GLM-like procedure. Data is z-scaled and grouped by factors and time periods to determine which points are above and below the control limits in a cumulative time fashion. Then a cumulative rate is created as the final variable. Set KeepAllCols to FALSE to utilize the intermediate features to create rolling stats from them.
}
\examples{
data <- data.table::data.table(DateTime = as.Date(Sys.time()),
  Target = stats::filter(rnorm(1000,
                               mean = 50,
                               sd = 20),
                         filter=rep(1,10),
                         circular=TRUE))
data[, temp := seq(1:10000)][, DateTime := DateTime - temp][, temp := NULL]
data <- data[order(DateTime)]
x <- data.table::as.data.table(sde::GBM(N=10000)*1000)
data[, predicted := x[-1,]]
stuff    <- GenTSAnomVars(data,
                          ValueCol    = "Target",
                          GroupVar1   = NULL,
                          GroupVar2   = NULL,
                          DateVar     = "DateTime",
                          High        = 1.96,
                          Low         = -1.96,
                          KeepAllCols = TRUE,
                          DataScaled  = FALSE)
}
\seealso{
Other Unsupervised Learning: \code{\link{AutoKMeans}},
  \code{\link{ResidualOutliers}}
}
\author{
Adrian Antico
}
\concept{Unsupervised Learning}
