% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Adrian_Antico_Modeling_Tools_Library.R
\name{ModelDataPrep}
\alias{ModelDataPrep}
\title{Final Data Preparation Function}
\usage{
ModelDataPrep(data, Impute = TRUE, CharToFactor = TRUE,
  MissFactor = "0", MissNum = -1)
}
\arguments{
\item{data}{This is your source data you'd like to modify}

\item{Impute}{Defaults to TRUE which tells the function to impute the data}

\item{CharToFactor}{Defaults to TRUE which tells the function to convert characters to factors}

\item{MissFactor}{Supply the value to impute missing factor levels}

\item{MissNum}{Supply  the value to impute missing numeric values}
}
\value{
Returns the original data table with corrected values
}
\description{
This function replaces inf values with NA, converts characters to factors, and imputes with constants
}
\examples{
data <- data.table::data.table(Value = runif(100000),
                               FactorCol = as.character(sample(x = c(letters,
                                                                     LETTERS,
                                                                     paste0(letters,letters),
                                                                     paste0(LETTERS,LETTERS),
                                                                     paste0(letters,LETTERS),
                                                                     paste0(LETTERS,letters)),
                                                               size = 100000,
                                                               replace = TRUE)))
data <- ModelDataPrep(data,
                      Impute = TRUE,
                      CharToFactor = TRUE,
                      MissFactor = "0",
                      MissNum    = -1)
}
\author{
Adrian Antico
}
