% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureEngineering_DataSets.R
\name{ModelDataPrep}
\alias{ModelDataPrep}
\title{ModelDataPrep}
\usage{
ModelDataPrep(
  data,
  Impute = TRUE,
  CharToFactor = TRUE,
  FactorToChar = FALSE,
  IntToNumeric = TRUE,
  LogicalToBinary = FALSE,
  DateToChar = FALSE,
  IDateConversion = FALSE,
  RemoveDates = FALSE,
  MissFactor = "0",
  MissNum = -1,
  IgnoreCols = NULL
)
}
\arguments{
\item{data}{This is your source data you'd like to modify}

\item{Impute}{Defaults to TRUE which tells the function to impute the data}

\item{CharToFactor}{Defaults to TRUE which tells the function to convert characters to factors}

\item{FactorToChar}{Converts to character}

\item{IntToNumeric}{Defaults to TRUE which tells the function to convert integers to numeric}

\item{LogicalToBinary}{Converts logical values to binary numeric values}

\item{DateToChar}{Converts date columns into character columns}

\item{IDateConversion}{Convert IDateTime to POSIXct and IDate to Date types}

\item{RemoveDates}{Defaults to FALSE. Set to TRUE to remove date columns from your data.table}

\item{MissFactor}{Supply the value to impute missing factor levels}

\item{MissNum}{Supply  the value to impute missing numeric values}

\item{IgnoreCols}{Supply column numbers for columns you want the function to ignore}
}
\value{
Returns the original data table with corrected values
}
\description{
This function replaces inf values with NA, converts characters to factors, and imputes with constants
}
\examples{
\dontrun{
# Create fake data
data <- RemixAutoML::FakeDataGenerator(
  Correlation = 0.75,
  N = 250000L,
  ID = 2L,
  ZIP = 0L,
  FactorCount = 6L,
  AddDate = TRUE,
  Classification = FALSE,
  MultiClass = FALSE)

# Check column types
str(data)

# Convert some factors to character
data <- RemixAutoML::ModelDataPrep(
  data,
  Impute       = TRUE,
  CharToFactor = FALSE,
  FactorToChar = TRUE,
  IntToNumeric = TRUE,
  LogicalToBinary = FALSE,
  DateToChar   = FALSE,
  IDateConversion = FALSE,
  RemoveDates  = TRUE,
  MissFactor   = "0",
  MissNum      = -1,
  IgnoreCols   = c("Factor_1"))

# Check column types
str(data)
}
}
\seealso{
Other Feature Engineering: 
\code{\link{AutoDataPartition}()},
\code{\link{AutoDiffLagN}()},
\code{\link{AutoHierarchicalFourier}()},
\code{\link{AutoInteraction}()},
\code{\link{AutoLagRollMode}()},
\code{\link{AutoLagRollStatsScoring}()},
\code{\link{AutoLagRollStats}()},
\code{\link{AutoTransformationCreate}()},
\code{\link{AutoTransformationScore}()},
\code{\link{AutoWord2VecModeler}()},
\code{\link{AutoWord2VecScoring}()},
\code{\link{CategoricalEncoding}()},
\code{\link{CreateCalendarVariables}()},
\code{\link{CreateHolidayVariables}()},
\code{\link{DummifyDT}()},
\code{\link{H2OAutoencoderScoring}()},
\code{\link{H2OAutoencoder}()},
\code{\link{TimeSeriesFill}()}
}
\author{
Adrian Antico
}
\concept{Feature Engineering}
