% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RemixAutoML.R
\name{multiplot}
\alias{multiplot}
\title{Multiplot is a function for combining multiple plots}
\usage{
multiplot(..., plotlist = NULL, cols = 2, layout = NULL)
}
\arguments{
\item{...}{Passthrough arguments}

\item{plotlist}{This is the list of your charts}

\item{cols}{This is the number of columns in your multiplot}

\item{layout}{Leave NULL}
}
\value{
Multiple ggplots on a single image
}
\description{
Sick of copying this one into your code? Well, not anymore.
}
\examples{
\dontrun{
multiplot(plotlist = list(p1,p2,p3,p4), cols = 2)
}
}
\seealso{
Other Misc: \code{\link{AutoH20TextPrepScoring}},
  \code{\link{ChartTheme}}, \code{\link{PrintObjectsSize}},
  \code{\link{RemixTheme}}, \code{\link{SimpleCap}},
  \code{\link{WordFreq}}, \code{\link{percRank}},
  \code{\link{tempDatesFun}}, \code{\link{tokenizeH20}}
}
\author{
Adrian Antico
}
\concept{Misc}
