% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RemixAutoML.R
\name{multiplot}
\alias{multiplot}
\title{Multiplot is a function for combining multiple plots}
\usage{
multiplot(..., plotlist = NULL, cols = 2, layout = NULL)
}
\arguments{
\item{...}{Passthrough arguments}

\item{plotlist}{This is the list of your charts}

\item{cols}{This is the number of columns in your multiplot}

\item{layout}{Leave NULL}
}
\value{
Multiple ggplots on a single image
}
\description{
Sick of copying this one into your code? Well, not anymore.
}
\examples{
Correl <- 0.85
data <- data.table::data.table(Target = runif(100))
data[, x1 := qnorm(Target)]
data[, x2 := runif(100)]
data[, Independent_Variable1 := log(pnorm(Correl * x1 +
                                            sqrt(1-Correl^2) * qnorm(x2)))]
data[, Predict := (pnorm(Correl * x1 +
                           sqrt(1-Correl^2) * qnorm(x2)))]
p1 <- RemixAutoML::ParDepCalPlots(data,
                                  PredictionColName = "Predict",
                                  TargetColName = "Target",
                                  IndepVar = "Independent_Variable1",
                                  GraphType = "calibration",
                                  PercentileBucket = 0.20,
                                  FactLevels = 10,
                                  Function = function(x) mean(x, na.rm = TRUE))
p2 <- RemixAutoML::ParDepCalPlots(data,
                                  PredictionColName = "Predict",
                                  TargetColName = "Target",
                                  IndepVar = "Independent_Variable1",
                                  GraphType = "boxplot",
                                  PercentileBucket = 0.20,
                                  FactLevels = 10,
                                  Function = function(x) mean(x, na.rm = TRUE))
RemixAutoML::multiplot(plotlist = list(p1,p2), cols = 2)
}
\seealso{
Other Misc: \code{\link{AutoH2OTextPrepScoring}},
  \code{\link{AutoRecomDataCreate}},
  \code{\link{AutoWordFreq}}, \code{\link{ChartTheme}},
  \code{\link{PrintObjectsSize}}, \code{\link{RemixTheme}},
  \code{\link{SimpleCap}}, \code{\link{percRank}},
  \code{\link{tempDatesFun}}, \code{\link{tokenizeH2O}}
}
\author{
Adrian Antico
}
\concept{Misc}
