% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EconometricsFunctions.R
\name{ParallelAutoNNET}
\alias{ParallelAutoNNET}
\title{ParallelAutoNNET to run the 4 data sets at once}
\usage{
ParallelAutoNNET(Output, MetricSelection = "MAE", MaxFourierTerms = 1,
  TrainValidateShare = c(0.5, 0.5), MaxNumberModels = 20,
  MaxRunMinutes = 5, MaxRunsWithoutNewWinner = 12)
}
\arguments{
\item{Output}{The output returned from TimeSeriesDataPrepare()}

\item{MetricSelection}{Choose from MAE, MSE, and MAPE}

\item{MaxFourierTerms}{Fourier pairs}

\item{TrainValidateShare}{The value returned from TimeSeriesPrepare()}

\item{MaxNumberModels}{20}

\item{MaxRunMinutes}{5}

\item{MaxRunsWithoutNewWinner}{12}

\item{TrainValidateShare}{c(0.50,0.50)}
}
\value{
Time series data sets to pass onto auto modeling functions
}
\description{
ParallelAutoNNET to run the 4 data sets at once
}
\examples{
ParallelAutoNNET(
  MetricSelection = "MAE",
  Output = NULL, 
  MaxRunsWithoutNewWinner = 20,
  TrainValidateShare = c(0.50,0.50),
  MaxNumberModels = 5,
  MaxRunMinutes = 5)
}
\seealso{
Other Time Series: \code{\link{AutoBanditNNet}},
  \code{\link{AutoBanditSarima}}, \code{\link{AutoTBATS}},
  \code{\link{CarmaHoldoutMetrics}},
  \code{\link{DifferenceDataReverse}},
  \code{\link{DifferenceData}},
  \code{\link{FinalBuildArfima}},
  \code{\link{FinalBuildArima}},
  \code{\link{FinalBuildETS}},
  \code{\link{FinalBuildNNET}},
  \code{\link{FinalBuildTBATS}},
  \code{\link{FinalBuildTSLM}},
  \code{\link{GenerateParameterGrids}},
  \code{\link{OptimizeArfima}},
  \code{\link{OptimizeArima}}, \code{\link{OptimizeETS}},
  \code{\link{OptimizeNNET}}, \code{\link{OptimizeTBATS}},
  \code{\link{OptimizeTSLM}},
  \code{\link{ParallelAutoARIMA}},
  \code{\link{ParallelAutoArfima}},
  \code{\link{ParallelAutoETS}},
  \code{\link{ParallelAutoTBATS}},
  \code{\link{ParallelAutoTSLM}},
  \code{\link{RL_Performance}},
  \code{\link{Regular_Performance}},
  \code{\link{StackedTimeSeriesEnsembleForecast}},
  \code{\link{TimeSeriesDataPrepare}},
  \code{\link{WideTimeSeriesEnsembleForecast}}
}
\author{
Adrian Antico
}
\concept{Time Series}
