% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RemixAutoML.R
\name{ParDepCalPlots}
\alias{ParDepCalPlots}
\title{ParDepCalPlots automatically builds partial dependence calibration plots for model evaluation}
\usage{
ParDepCalPlots(data, PredictionColName = c("PredictedValues"),
  TargetColName = c("ActualValues"),
  IndepVar = c("Independent_Variable_Name"),
  GraphType = c("calibration"), PercentileBucket = 0.05,
  FactLevels = 10, Function = function(x) base::mean(x, na.rm = TRUE))
}
\arguments{
\item{data}{Data containing predicted values and actual values for comparison}

\item{PredictionColName}{Predicted values column names}

\item{TargetColName}{Target value column names}

\item{IndepVar}{Independent variable column names}

\item{GraphType}{calibration or boxplot - calibration aggregated data based on summary statistic; boxplot shows variation}

\item{PercentileBucket}{Number of buckets to partition the space on (0,1) for evaluation}

\item{FactLevels}{The number of levels to show on the chart (1. Levels are chosen based on frequency; 2. all other levels grouped and labeled as "Other")}

\item{Function}{Supply the function you wish to use for aggregation.}
}
\value{
Partial dependence calibration plot or boxplot
}
\description{
This function automatically builds partial dependence calibration plots and partial dependence calibration boxplots for model evaluation using regression, quantile regression, and binary and multinomial classification
}
\examples{
\dontrun{
ParDepCalPlots(data,
               PredictionColName = "predict",
               TargetColName  = "target",
               IndepVar    = "Independent_Variable",
               GraphType        = "boxplot",
               PercentileBucket = 0.05,
               FactLevels  = 10,
               Function    = function(x) mean(x, na.rm = TRUE))
}
}
\seealso{
Other Model Evaluation and Interpretation: \code{\link{EvalPlot}},
  \code{\link{RedYellowGreen}}, \code{\link{threshOptim}}
}
\author{
Adrian Antico
}
\concept{Model Evaluation and Interpretation}
