% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Adrian_Antico_Modeling_Tools_Library.R
\name{percRank}
\alias{percRank}
\title{Percentile rank function}
\usage{
percRank(x)
}
\arguments{
\item{x}{X is your variable of interest}
}
\value{
vector of percentile ranks
}
\description{
This function computes percentile ranks for each row in your data like Excel's PERCENT_RANK
}
\examples{
\dontrun{
percRank(x)
}
}
\seealso{
Other Misc: \code{\link{ChartTheme}},
  \code{\link{PrintObjectsSize}}, \code{\link{RemixTheme}},
  \code{\link{SimpleCap}}, \code{\link{WordFreq}},
  \code{\link{multiplot}}, \code{\link{tempDatesFun}},
  \code{\link{tokenizeH20}}
}
\author{
Adrian Antico
}
\concept{Misc}
