% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Adrian_Antico_Modeling_Tools_Library.R
\name{percRank}
\alias{percRank}
\title{Percentile rank function}
\usage{
percRank(x)
}
\arguments{
\item{x}{X is your variable of interest}
}
\value{
vector of percentile ranks
}
\description{
This function computes percentile ranks for each row in your data like Excel's PERCENT_RANK
}
\examples{
# percRank(x)
}
\author{
Adrian Antico
}
