% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RemixAutoML.R
\name{percRank}
\alias{percRank}
\title{Percentile rank function}
\usage{
percRank(x)
}
\arguments{
\item{x}{X is your variable of interest}
}
\value{
vector of percentile ranks
}
\description{
This function computes percentile ranks for each row in your data like Excel's PERCENT_RANK
}
\examples{
data <- data.table::data.table(A = runif(100))
data[, Rank := percRank(A)]
data <- data.table::data.table(A = runif(100))
data[, Percentile := percRank(A)]
}
\seealso{
Other Misc: \code{\link{AutoH2OTextPrepScoring}},
  \code{\link{AutoRecomDataCreate}},
  \code{\link{AutoWordFreq}}, \code{\link{ChartTheme}},
  \code{\link{PrintObjectsSize}}, \code{\link{RemixTheme}},
  \code{\link{SimpleCap}}, \code{\link{multiplot}},
  \code{\link{tempDatesFun}}, \code{\link{tokenizeH2O}}
}
\author{
Adrian Antico
}
\concept{Misc}
