% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EDAFunctions.R
\name{PlotGUI}
\alias{PlotGUI}
\title{PlotGUI}
\usage{
PlotGUI()
}
\description{
Spin up the esquisse plotting gui
}
\seealso{
Other EDA: 
\code{\link{AutoWordFreq}()},
\code{\link{BarPlot}()},
\code{\link{BoxPlot}()},
\code{\link{EDA_Histograms}()},
\code{\link{HistPlot}()},
\code{\link{Mode}()},
\code{\link{ScatterCopula}()},
\code{\link{UserBaseEvolution}()},
\code{\link{ViolinPlot}()}
}
\concept{EDA}
