% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SQL_Functions.R
\name{PostGRE_CreateTable}
\alias{PostGRE_CreateTable}
\title{PostGRE_CreateTable}
\usage{
PostGRE_CreateTable(
  data = NULL,
  DBName = NULL,
  Schema = NULL,
  TableName = NULL,
  Connection = NULL,
  CloseConnection = FALSE,
  Temporary = FALSE,
  Host = NULL,
  User = NULL,
  Port = NULL,
  Password = NULL
)
}
\arguments{
\item{data}{Source data.table. If you supply a Schema, data will be ignored.}

\item{DBName}{If Connection is NULL then this must be supplied. database name}

\item{Schema}{Optional. Advised to use if type inference is fuzzy}

\item{TableName}{Name of table you want created}

\item{Connection}{NULL. If supplied, use this: Connection <- DBI::dbConnect(RPostgres::Postgres(), host = Host, dbname = DBName, user = User, port = Port, password = Password)}

\item{CloseConnection}{= FALSE}

\item{Temporary}{If Connection is NULL then this must be supplied. FALSE}

\item{Host}{If Connection is NULL then this must be supplied. host name}

\item{User}{If Connection is NULL then this must be supplied. user name}

\item{Port}{If Connection is NULL then this must be supplied. port name}

\item{Password}{user password}
}
\description{
PostGRE_CreateTable get data from a database table
}
\examples{
\dontrun{
RemixAutoML::PostGRE_CreateTable(
  data,
  DBName = 'Testing',
  Schema = NULL,
  TableName = NULL,
  Temporary = FALSE,
  Connection = NULL,
  CloseConnection = FALSE,
  Host = 'localhost',
  User = 'postgres',
  Port = 5432,
  Password = 'Aa...')
}

}
\seealso{
Other Database: 
\code{\link{AutoDataDictionaries}()},
\code{\link{PostGRE_AppendData}()},
\code{\link{PostGRE_GetTableNames}()},
\code{\link{PostGRE_ListTables}()},
\code{\link{PostGRE_Query}()},
\code{\link{PostGRE_RemoveCreateAppend}()},
\code{\link{PostGRE_RemoveTable}()},
\code{\link{PosteGRE_CreateDatabase}()},
\code{\link{PosteGRE_DropDB}()},
\code{\link{PosteGRE_ListDatabases}()},
\code{\link{SQL_ClearTable}()},
\code{\link{SQL_DropTable}()},
\code{\link{SQL_Query_Push}()},
\code{\link{SQL_Query}()},
\code{\link{SQL_SaveTable}()},
\code{\link{SQL_Server_DBConnection}()}
}
\author{
Adrian Antico
}
\concept{Database}
