% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EconometricsFunctions.R
\name{PredictArima}
\alias{PredictArima}
\title{PredictArima}
\usage{
PredictArima(
  object = Results,
  n.ahead = FCPeriods,
  newxreg = NULL,
  se.fit = TRUE
)
}
\arguments{
\item{object}{Object that stores the output from Arima()}

\item{n.ahead}{Number of forecast periods to forecast}

\item{newxreg}{NULL by default. Forward looking independent variables as matrix type}

\item{se.fit}{Set to FALSE to not return prediction intervals with the forecast}
}
\description{
PredictArima is a function to overwrite the s3 generic <code>getS3method('predict','Arima')</code>
}
\seealso{
Other Time Series Helper: 
\code{\link{FinalBuildArfima}()},
\code{\link{FinalBuildArima}()},
\code{\link{FinalBuildETS}()},
\code{\link{FinalBuildNNET}()},
\code{\link{FinalBuildTBATS}()},
\code{\link{FinalBuildTSLM}()},
\code{\link{GenerateParameterGrids}()},
\code{\link{OptimizeArfima}()},
\code{\link{OptimizeArima}()},
\code{\link{OptimizeETS}()},
\code{\link{OptimizeNNET}()},
\code{\link{OptimizeTBATS}()},
\code{\link{OptimizeTSLM}()},
\code{\link{ParallelAutoARIMA}()},
\code{\link{ParallelAutoArfima}()},
\code{\link{ParallelAutoETS}()},
\code{\link{ParallelAutoNNET}()},
\code{\link{ParallelAutoTBATS}()},
\code{\link{ParallelAutoTSLM}()},
\code{\link{RL_Performance}()},
\code{\link{Regular_Performance}()},
\code{\link{StackedTimeSeriesEnsembleForecast}()},
\code{\link{TimeSeriesDataPrepare}()},
\code{\link{WideTimeSeriesEnsembleForecast}()}
}
\author{
Adrian Antico
}
\concept{Time Series Helper}
