% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RemixAutoML.R
\name{PrintObjectsSize}
\alias{PrintObjectsSize}
\title{PrintObjectsSize prints out the top N objects and their associated sizes, sorted by size}
\usage{
PrintObjectsSize(N = 10)
}
\arguments{
\item{N}{The number of objects to display}
}
\value{
A print to your console of the sizes of the objects in your environment
}
\description{
PrintObjectsSize prints out the top N objects and their associated sizes, sorted by size
}
\examples{
\donttest{
PrintObjectsSize(N = 10)
}
}
\seealso{
Other Misc: \code{\link{AutoH2OTextPrepScoring}},
  \code{\link{AutoRecomDataCreate}},
  \code{\link{AutoWordFreq}}, \code{\link{ChartTheme}},
  \code{\link{RemixTheme}}, \code{\link{SimpleCap}},
  \code{\link{multiplot}}, \code{\link{percRank}},
  \code{\link{tempDatesFun}}, \code{\link{tokenizeH2O}}
}
\author{
Adrian Antico
}
\concept{Misc}
