% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RemixAutoML.R
\name{RecomDataCreate}
\alias{RecomDataCreate}
\title{Convert transactional data.table to a binary ratings matrix}
\usage{
RecomDataCreate(data, EntityColName = "CustomerID",
  ProductColName = "StockCode", MetricColName = "TotalSales")
}
\arguments{
\item{data}{This is your transactional data.table. Must include an Entity (typically customer), ProductCode (such as SKU), and a sales metric (such as total sales).}

\item{EntityColName}{This is the column name in quotes that represents the column name for the Entity, such as customer}

\item{ProductColName}{This is the column name in quotes that represents the column name for the product, such as SKU}

\item{MetricColName}{This is the column name in quotes that represents the column name for the metric, such as total sales}
}
\value{
A BinaryRatingsMatrix
}
\description{
Convert transactional data.table to a binary ratings matrix
}
\examples{
\dontrun{
RatingsMatrix <- RecomDataCreate(data,
                                 EntityColName = "CustomerID",
                                 ProductColName = "StockCode",
                                 MetricColName = "TotalSales")
}
}
\seealso{
Other Misc: \code{\link{AutoH2OTextPrepScoring}},
  \code{\link{AutoWordFreq}}, \code{\link{ChartTheme}},
  \code{\link{PrintObjectsSize}}, \code{\link{RemixTheme}},
  \code{\link{SimpleCap}}, \code{\link{multiplot}},
  \code{\link{percRank}}, \code{\link{tempDatesFun}},
  \code{\link{tokenizeH2O}}
}
\author{
Adrian Antico and Douglas Pestana
}
\concept{Misc}
