% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Adrian_Antico_Modeling_Tools_Library.R
\name{RedYellowGreen}
\alias{RedYellowGreen}
\title{RedYellowGreen is for determining the optimal thresholds for binary classification when do-nothing is an option}
\usage{
RedYellowGreen(data, PredictColNumber = 2, ActualColNumber = 1,
  TruePositiveCost = 0, TrueNegativeCost = 0,
  FalsePositiveCost = -10, FalseNegativeCost = -50, MidTierCost = -2,
  Cores = 8, Precision = 0.01)
}
\arguments{
\item{data}{data is the data table with your predicted and actual values from a classification model}

\item{PredictColNumber}{The column number where the actual target variable is located (in binary form)}

\item{ActualColNumber}{The column number where the predicted values are located}

\item{TruePositiveCost}{This is the utility for generating a true positive prediction}

\item{TrueNegativeCost}{This is the utility for generating a true negative prediction}

\item{FalsePositiveCost}{This is the cost of generating a false positive prediction}

\item{FalseNegativeCost}{This is the cost of generating a false negative prediction}

\item{MidTierCost}{This is the cost of doing nothing (or whatever it means to not classify in your case)}

\item{Cores}{Number of cores on your machine}

\item{Precision}{Set the decimal number to increment by between 0 and 1}
}
\value{
A data table with all evaluated strategies, parameters, and utilities, along with a 3d scatterplot of the results
}
\description{
This function will find the optimial thresholds for applying the main label and for finding the optimial range for doing nothing when you can quantity the cost of doing nothing
}
\examples{
# data <- RedYellowGreen(data,
#                         PredictColNumber  = 1,
#                         ActualColNumber   = 2,
#                         TruePositiveCost  = 0,
#                         TrueNegativeCost  = 0,
#                         FalsePositiveCost = -1,
#                         FalseNegativeCost = -2,
#                         MidTierCost       = -0.5)
}
\seealso{
Other Model Evaluation and Interpretation: \code{\link{EvalPlot}},
  \code{\link{ParDepCalPlots}}, \code{\link{threshOptim}}
}
\author{
Adrian Antico
}
\concept{Model Evaluation and Interpretation}
