% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelMetrics.R
\name{RemixClassificationMetrics}
\alias{RemixClassificationMetrics}
\title{RemixClassificationMetrics}
\usage{
RemixClassificationMetrics(MLModels = c("catboost", "h2ogbm", "h2odrf",
  "xgboost"), TargetVariable = "Value", Thresholds = seq(0.01, 0.99,
  0.01), CostMatrix = c(1, 0, 0, 1), ClassLabels = c(1, 0),
  CatBoostTestData = NULL, H2oGBMTestData = NULL,
  H2oDRFTestData = NULL, XGBoostTestData = NULL)
}
\arguments{
\item{MLModels}{A vector of model names from remixautoml}

\item{TargetVariable}{Name of your target variable}

\item{Thresholds}{seq(0.01,0.99,0.01),}

\item{CostMatrix}{c(1,0,0,1),}

\item{ClassLabels}{c(1,0),}

\item{CatBoostTestData}{Test data returned from AutoCatBoostClassifier}

\item{H2oGBMTestData}{Test data returned from AutoCatBoostClassifier}

\item{H2oDRFTestData}{Test data returned from AutoCatBoostClassifier}

\item{XGBoostTestData}{Test data returned from AutoCatBoostClassifier}
}
\description{
RemixClassificationMetrics
}
\seealso{
Other Model Evaluation: \code{\link{ClassificationMetrics}},
  \code{\link{DT_BinaryConfusionMatrix}}
}
\author{
Adrian Antico
}
\concept{Model Evaluation}
