% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelMetrics.R
\name{RemixClassificationMetrics}
\alias{RemixClassificationMetrics}
\title{RemixClassificationMetrics}
\usage{
RemixClassificationMetrics(
  TargetVariable = NULL,
  Thresholds = seq(0.01, 0.99, 0.01),
  CostMatrix = c(1, 0, 0, 1),
  ClassLabels = c(1, 0),
  ValidationData. = NULL
)
}
\arguments{
\item{TargetVariable}{Name of your target variable}

\item{Thresholds}{seq(0.01,0.99,0.01),}

\item{CostMatrix}{c(1,0,0,1) c(TP utility, FN utility, FP utility, TN utility)}

\item{ClassLabels}{c(1,0),}

\item{ValidationData.}{Test data}
}
\description{
RemixClassificationMetrics
}
\examples{
\dontrun{
RemixClassificationMetrics <- function(
  TargetVariable = "Adrian",
  Thresholds = seq(0.01,0.99,0.01),
  CostMatrix = c(1,0,0,1),
  ClassLabels = c(1,0),
  ValidationData. = ValidationData)
}
}
\seealso{
Other Model Evaluation: 
\code{\link{BinaryMetrics}()},
\code{\link{ClassificationMetrics}()},
\code{\link{DT_BinaryConfusionMatrix}()},
\code{\link{MultiClassMetrics}()},
\code{\link{RegressionMetrics}()}
}
\author{
Adrian Antico
}
\concept{Model Evaluation}
