% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyFunctions.R
\name{ReturnParam}
\alias{ReturnParam}
\title{Save VarName values within a project}
\usage{
ReturnParam(
  input = NULL,
  VarName = "bla",
  Type = "numeric",
  Default = 1,
  Switch = TRUE
)
}
\arguments{
\item{input}{This is the input value within a Shiny context}

\item{VarName}{The name of the VarNameument you want to store}

\item{Type}{'character' 'numeric' 'logical' 'date'}

\item{Default}{default value}

\item{Switch}{= FALSE}
}
\value{
Updates ProjectList inside function
}
\description{
Automatically save VarNames to project list
}
\examples{
\dontrun{
Aggregate <- RemixAutoML::ReturnParam(input, VarName = "TS_AggregateFunction", Type = "character", Default = "mean")
}

}
\seealso{
Other Shiny: 
\code{\link{ArgNullCheck2}()},
\code{\link{ArgNullCheck}()},
\code{\link{BlankRow}()},
\code{\link{DateInput}()},
\code{\link{GenerateEvaluationMetrics}()},
\code{\link{NumericInput}()},
\code{\link{PickerInput_GetLevels2}()},
\code{\link{PickerInput_GetLevels}()},
\code{\link{PickerInput}()},
\code{\link{PreparePlotData}()},
\code{\link{ReactiveLoadCSV}()},
\code{\link{SelectizeInput}()},
\code{\link{StoreArgs}()},
\code{\link{TextInput}()},
\code{\link{observeEventLoad}()}
}
\author{
Adrian Antico
}
\concept{Shiny}
