% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EDAFunctions.R
\name{ScatterCopula}
\alias{ScatterCopula}
\title{ScatterCopula}
\usage{
ScatterCopula(
  data = NULL,
  x_var = NULL,
  y_var = NULL,
  GroupVariable = NULL,
  SampleCount = 100000L,
  FitGam = TRUE,
  color = "darkblue",
  point_size = 0.5,
  text_size = 12,
  x_axis_text_angle = 35,
  y_axis_text_angle = 0,
  chart_color = "lightsteelblue1",
  border_color = "darkblue",
  text_color = "darkblue",
  grid_color = "white",
  background_color = "gray95",
  legend_position = "bottom"
)
}
\arguments{
\item{data}{Source data.table}

\item{x_var}{Numeric variable}

\item{y_var}{Numeric variable}

\item{GroupVariable}{Color options}

\item{SampleCount}{Number of randomized rows to utilize. For speedup and memory purposes}

\item{FitGam}{Add gam fit to scatterplot and copula plot}

\item{color}{= "darkblue"}

\item{point_size}{= 0.50}

\item{text_size}{= 12}

\item{x_axis_text_angle}{= 35}

\item{y_axis_text_angle}{= 0}

\item{chart_color}{= "lightsteelblue1"}

\item{border_color}{= "darkblue"}

\item{text_color}{= "darkblue"}

\item{grid_color}{= "white"}

\item{background_color}{= "gray95"}

\item{legend_position}{= "bottom}
}
\description{
Dual plot. One on original scale and one using empirical copula data
}
\seealso{
Other EDA: 
\code{\link{AutoWordFreq}()},
\code{\link{EDA_Histograms}()},
\code{\link{Mode}()},
\code{\link{PlotGUI}()},
\code{\link{UserBaseEvolution}()}
}
\author{
Adrian Antico
}
\concept{EDA}
