% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Adrian_Antico_Modeling_Tools_Library.R
\name{Scoring_GDL_Feature_Engineering}
\alias{Scoring_GDL_Feature_Engineering}
\title{An Automated Scoring Feature Engineering Function}
\usage{
Scoring_GDL_Feature_Engineering(data, lags = c(1:6, 12, seq(24, 168,
  24)), periods = c(6, 12, 24, 72, 168, 720, 4320, 8640),
  statsFUNs = c(function(x) mean(x, na.rm = TRUE), function(x) sd(x,
  na.rm = TRUE), function(x) quantile(x, probs = 0.85, na.rm = TRUE)),
  statsNames = c("mean", "sd", "q85"),
  targets = c("ScaledConsumption"), groupingVars = c("BADGE_NBR"),
  sortDateName = "DAY_DATE", timeDiffTarget = NULL,
  timeAgg = "hours", WindowingLag = 1, Type = "Lag", Timer = FALSE,
  SkipCols = FALSE, SimpleImpute = TRUE,
  AscRowByGroup = "BadgeRowNum", RecordsKeep = 1)
}
\arguments{
\item{data}{Core instruction file for automation}

\item{lags}{The ceiling amount of memory H20 will utilize}

\item{statsNames}{The corresponding names to append to your colnames created associated with statsFuns}

\item{targets}{The column(s) in which you will build your lags and rolling stats}

\item{groupingVars}{Categorical variables you will build your lags and rolling stats by}

\item{sortDateName}{String name of your core date column in your transaction data}

\item{timeDiffTarget}{List a name in order to create time between events with assiciated lags and rolling features}

\item{timeAgg}{Unit of time to aggregate by}

\item{WindowingLag}{Build moving stats off of target column(s) or one of their lags (1+)}

\item{Type}{input "Lag" if you want features built on historical values; use "Lead" if you want features built on future values}

\item{Timer}{Set to TRUE if you want a time run for the operation; useful when there is grouping}

\item{SkipCols}{Defaults to NULL; otherwise name the vector containing the names of columns to skip}

\item{SimpleImpute}{Set to TRUE for factor level imputation of "0" and numeric imputation of -1}

\item{AscRowByGroup}{Required to have a column with a Row Number by group (if grouping) with 1 being the record for scoring (typically the most current in time)}

\item{RecordsKeep}{List the number of records to retain (1 for last record, 2 for last 2 records, etc.)}

\item{statsFuns}{List of functions for your rolling windows, such as mean, sd, min, max, quantile}
}
\value{
data.table of original data plus newly created features
}
\description{
For scoring purposes (brings back a single row by group), this function creates autoregressive and rolling stats from target columns and distributed lags and distributed rolling stats for independent features distributed across time. On top of that, you can also create time between instances along with their associated lags and rolling stats. This function works for data with groups and without groups.
}
\examples{
quick_model <- Scoring_GDL_Feature_Engineering(quick_model,
                                               lags           = c(1:6,12,seq(24,168,24)),
                                               periods        = c(6,12,24,72,168,720,4320,8640),
                                               statsFUNs      = c(function(x) mean(x, na.rm = TRUE),
                                                                  function(x) sd(x, na.rm = TRUE),
                                                                  function(x) quantile(x, probs = 0.85, na.rm = TRUE)),
                                               statsNames     = c("mean","sd","q85"),
                                               targets        = c("ScaledConsumption"),
                                               groupingVars   = c("BADGE_NBR"),
                                               sortDateName   = "DAY_DATE",
                                               timeDiffTarget = NULL,
                                               timeAgg        = "hours",
                                               WindowingLag   = 1,
                                               Type           = "Lag",
                                               Timer          = FALSE,
                                               SkipCols       = FALSE,
                                               SimpleImpute   = TRUE,
                                               AscRowByGroup  = "BadgeRowNum")
}
\author{
Adrian Antico at RemixInstitute.com
}
