% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotFunctions.R
\name{ShapImportancePlot}
\alias{ShapImportancePlot}
\title{ShapImportancePlot}
\usage{
ShapImportancePlot(data, ShapColNames = NULL, AggMethod = "mean", TopN = 25)
}
\arguments{
\item{data}{Source data.table}

\item{ShapColNames}{Names of the columns that contain shap values you want included}

\item{AggMethod}{A string for aggregating shapely values for importances. Choices include, 'mean', 'absmean', 'meanabs', 'geomean', 'harmmean', 'sd', 'median'}

\item{TopN}{The number of variables to plot}
}
\description{
Generate Variable Importance Plots using Shapely Values of given data set
}
\seealso{
Other Model Insights: 
\code{\link{ModelInsightsReport}()}
}
\author{
Adrian Antico
}
\concept{Model Insights}
