% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RemixAutoML.R
\name{SimpleCap}
\alias{SimpleCap}
\title{SimpleCap function is for capitalizing the first letter of words}
\usage{
SimpleCap(x)
}
\arguments{
\item{x}{Column of interest}
}
\value{
An object to pass along to ggplot objects following the "+" sign
}
\description{
SimpleCap function is for capitalizing the first letter of words (need I say more?)
}
\examples{
x <- "adrian"
x <- SimpleCap(x)
}
\seealso{
Other Misc: \code{\link{AutoH2OTextPrepScoring}},
  \code{\link{AutoRecomDataCreate}},
  \code{\link{AutoWordFreq}}, \code{\link{ChartTheme}},
  \code{\link{PrintObjectsSize}}, \code{\link{RemixTheme}},
  \code{\link{multiplot}}, \code{\link{percRank}},
  \code{\link{tempDatesFun}}, \code{\link{tokenizeH2O}}
}
\author{
Adrian Antico
}
\concept{Misc}
