% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelEvaluationPlots.R
\name{SingleRowShapeShap}
\alias{SingleRowShapeShap}
\title{SingleRowShapeShap}
\usage{
SingleRowShapeShap(ShapData = NULL, EntityID = NULL, DateColumnName = NULL)
}
\arguments{
\item{ShapData}{Scoring data from AutoCatBoostScoring with classification or regression}
}
\description{
SingleRowShapeShap will convert a single row of your shap data into a table
}
\seealso{
Other Model Evaluation and Interpretation: 
\code{\link{AutoShapeShap}()},
\code{\link{CumGainsChart}()},
\code{\link{EvalPlot}()},
\code{\link{ParDepCalPlots}()},
\code{\link{ROCPlot}()},
\code{\link{RedYellowGreen}()},
\code{\link{ResidualPlots}()},
\code{\link{ShapPlot}()},
\code{\link{threshOptim}()}
}
\author{
Adrian Antico
}
\concept{Model Evaluation and Interpretation}
