% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataWranglingFunctions.R
\name{SQL_ClearTable}
\alias{SQL_ClearTable}
\title{SQL_ClearTable}
\usage{
SQL_ClearTable(
  DBConnection,
  SQLTableName = "",
  CloseChannel = TRUE,
  Errors = TRUE
)
}
\arguments{
\item{DBConnection}{RemixAutoML::SQL_Server_DBConnection()}

\item{SQLTableName}{The SQL statement you want to run}

\item{CloseChannel}{TRUE to close when done, FALSE to leave the channel open}

\item{Errors}{Set to TRUE to halt, FALSE to return -1 in cases of errors}
}
\description{
SQL_ClearTable get data from a database
}
\seealso{
Other Data Wrangling: 
\code{\link{AutoDataDictionaries}()},
\code{\link{ColumnSubsetDataTable}()},
\code{\link{DataDisplayMeta}()},
\code{\link{FakeDataGenerator}()},
\code{\link{FullFactorialCatFeatures}()},
\code{\link{SQL_DropTable}()},
\code{\link{SQL_Query_Push}()},
\code{\link{SQL_Query}()},
\code{\link{SQL_SaveTable}()},
\code{\link{SQL_Server_DBConnection}()},
\code{\link{SQL_UpdateTable}()},
\code{\link{TimeSeriesMelt}()}
}
\author{
Adrian Antico
}
\concept{Data Wrangling}
