% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataWranglingFunctions.R
\name{SQL_Query}
\alias{SQL_Query}
\title{SQL_Query}
\usage{
SQL_Query(
  DBConnection,
  Query,
  ASIS = FALSE,
  CloseChannel = TRUE,
  RowsPerBatch = 1024
)
}
\arguments{
\item{DBConnection}{RemixAutoML::SQL_Server_DBConnection()}

\item{Query}{The SQL statement you want to run}

\item{ASIS}{Auto column typing}

\item{CloseChannel}{TRUE to close when done, FALSE to leave the channel open}

\item{RowsPerBatch}{Rows default is 1024}
}
\description{
SQL_Query get data from a database
}
\seealso{
Other Data Wrangling: 
\code{\link{AutoDataDictionaries}()},
\code{\link{ColumnSubsetDataTable}()},
\code{\link{DataDisplayMeta}()},
\code{\link{FakeDataGenerator}()},
\code{\link{FullFactorialCatFeatures}()},
\code{\link{SQL_ClearTable}()},
\code{\link{SQL_DropTable}()},
\code{\link{SQL_Query_Push}()},
\code{\link{SQL_SaveTable}()},
\code{\link{SQL_Server_DBConnection}()},
\code{\link{SQL_UpdateTable}()},
\code{\link{TimeSeriesMelt}()}
}
\author{
Adrian Antico
}
\concept{Data Wrangling}
