% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SQL_Functions.R
\name{SQL_SaveTable}
\alias{SQL_SaveTable}
\title{SQL_SaveTable}
\usage{
SQL_SaveTable(
  DataToPush,
  DBConnection,
  SQLTableName = "",
  RowNames = NULL,
  ColNames = TRUE,
  CloseChannel = TRUE,
  AppendData = FALSE,
  AddPK = TRUE,
  Safer = TRUE
)
}
\arguments{
\item{DataToPush}{data to be sent to warehouse}

\item{DBConnection}{RemixAutoML::SQL_Server_DBConnection()}

\item{SQLTableName}{The SQL statement you want to run}

\item{RowNames}{c("Segment","Date")}

\item{ColNames}{Column names in first row}

\item{CloseChannel}{TRUE to close when done, FALSE to leave the channel open}

\item{AppendData}{TRUE or FALSE}

\item{AddPK}{Add a PK column to table}

\item{Safer}{TRUE}
}
\description{
SQL_SaveTable create a database table
}
\seealso{
Other Database: 
\code{\link{AutoDataDictionaries}()},
\code{\link{ExecuteSSIS}()},
\code{\link{SQL_ClearTable}()},
\code{\link{SQL_DropTable}()},
\code{\link{SQL_Query_Push}()},
\code{\link{SQL_Query}()},
\code{\link{SQL_Server_BulkPull}()},
\code{\link{SQL_Server_BulkPush}()},
\code{\link{SQL_Server_DBConnection}()},
\code{\link{SQL_UpdateTable}()}
}
\author{
Adrian Antico
}
\concept{Database}
