% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataWranglingFunctions.R
\name{SQL_Server_DBConnection}
\alias{SQL_Server_DBConnection}
\title{SQL_Server_DBConnection}
\usage{
SQL_Server_DBConnection(DataBaseName = "", Server = "")
}
\arguments{
\item{DataBaseName}{Name of the database}

\item{Server}{Name of the server to use}
}
\description{
SQL_Server_DBConnection is a function to return data dictionary data in table form
}
\seealso{
Other Data Wrangling: 
\code{\link{AutoDataDictionaries}()},
\code{\link{AutoHierarchicalFourier}()},
\code{\link{ColumnSubsetDataTable}()},
\code{\link{DataDisplayMeta}()},
\code{\link{FakeDataGenerator}()},
\code{\link{FullFactorialCatFeatures}()},
\code{\link{SQL_ClearTable}()},
\code{\link{SQL_DropTable}()},
\code{\link{SQL_Query_Push}()},
\code{\link{SQL_Query}()},
\code{\link{SQL_SaveTable}()},
\code{\link{SQL_UpdateTable}()},
\code{\link{TimeSeriesMelt}()}
}
\author{
Adrian Antico
}
\concept{Data Wrangling}
