% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SQL_Functions.R
\name{SQL_Server_DBConnection}
\alias{SQL_Server_DBConnection}
\title{SQL_Server_DBConnection}
\usage{
SQL_Server_DBConnection(DataBaseName = "", Server = "")
}
\arguments{
\item{DataBaseName}{Name of the database}

\item{Server}{Name of the server to use}
}
\description{
SQL_Server_DBConnection makes a connection to a sql server database
}
\seealso{
Other Database: 
\code{\link{AutoDataDictionaries}()},
\code{\link{SQL_ClearTable}()},
\code{\link{SQL_DropTable}()},
\code{\link{SQL_Query_Push}()},
\code{\link{SQL_Query}()},
\code{\link{SQL_SaveTable}()},
\code{\link{SQL_Server_BulkPull}()},
\code{\link{SQL_Server_BulkPush}()},
\code{\link{SQL_UpdateTable}()}
}
\author{
Adrian Antico
}
\concept{Database}
