% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotFunctions.R
\name{StockPlot}
\alias{StockPlot}
\title{StockPlot}
\usage{
StockPlot(
  Type = "candlestick",
  Metric = "Stock Price",
  TimeAgg = "days",
  Symbol = "TSLA",
  CompanyName = NULL,
  StartDate = "2022-01-01",
  EndDate = Sys.Date() - 1L,
  APIKey = NULL
)
}
\arguments{
\item{Type}{'candlestick', 'ohlc'}

\item{Metric}{Stock Price, Percent Returns (use symbol for percent), Percent Log Returns (use symbol for percent), Index, Quadratic Variation}

\item{TimeAgg}{= 'days', 'weeks', 'months'}

\item{Symbol}{ticker symbol string}

\item{CompanyName}{company name if you have it. ends up in title, that is all}

\item{StartDate}{Supply a start date. E.g. '2022-01-01'}

\item{EndDate}{Supply an end date. E.g. `Sys.Date()`}

\item{APIKey}{Supply your polygon API key}
}
\description{
Create a candlestick plot for stocks. See https://plotly.com/r/figure-labels/
}
\seealso{
Other Graphics: 
\code{\link{AddFacet}()},
\code{\link{AppModelInsights}()},
\code{\link{AutoPlotter}()},
\code{\link{BarPlot}()},
\code{\link{BoxPlot}()},
\code{\link{ChartTheme}()},
\code{\link{CorrMatrixPlot}()},
\code{\link{HeatMapPlot}()},
\code{\link{HistPlot}()},
\code{\link{PlotlyConversion}()},
\code{\link{ViolinPlot}()},
\code{\link{multiplot}()}
}
\author{
Adrian Antico
}
\concept{Graphics}
