% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyFunctions.R
\name{TextInput}
\alias{TextInput}
\title{TextInput}
\usage{
TextInput(
  InputID = "TS_CARMA_HolidayMovingAverages",
  Label = "Path to Data",
  Value = NULL,
  Width = "100\%",
  Placeholder = "NULL"
)
}
\arguments{
\item{InputID}{Feeds ProjectList and inputId. Argument saved in ProjectList}

\item{Label}{Feeds label}

\item{Value}{Default}

\item{Width}{width arg}

\item{Format}{Date format}
}
\value{
PickerInput object for server.R to go into renderUI({PickerInput()})
}
\description{
TextInput automatically builds a text input with ProjectList argument usage if it exists
}
\examples{
\dontrun{
output$TS_Date <- renderUI({
  RemixAutoML::TextInput(InputID = "TS_CARMA_HolidayMovingAverages",
                         Label = "Import Data Creation Date",
                         Value = Sys.Date(),
                         Placeholder = "yyyy-mm-dd")})
}
}
\seealso{
Other Shiny: 
\code{\link{ArgNullCheck2}()},
\code{\link{ArgNullCheck}()},
\code{\link{BlankRow}()},
\code{\link{DateInput}()},
\code{\link{GenerateEvaluationMetrics}()},
\code{\link{NumericInput}()},
\code{\link{PickerInput_GetLevels2}()},
\code{\link{PickerInput_GetLevels}()},
\code{\link{PickerInput}()},
\code{\link{PreparePlotData}()},
\code{\link{ReactiveLoadCSV}()},
\code{\link{ReturnParam}()},
\code{\link{SelectizeInput}()},
\code{\link{StoreArgs}()},
\code{\link{observeEventLoad}()}
}
\author{
Adrian Antico
}
\concept{Shiny}
