% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Adrian_Antico_Modeling_Tools_Library.R
\name{threshOptim}
\alias{threshOptim}
\title{Utility maximizing thresholds for binary classification}
\usage{
threshOptim(data, actTar = 1, predTar = 2, tpProfit = 1,
  tnProfit = 5, fpProfit = -1, fnProfit = -1,
  MidTierDoNothing = FALSE, MidTierCost = -100,
  MidTierLowThresh = 0.25, MidTierHighThresh = 0.75)
}
\arguments{
\item{data}{data is the data table you are building the modeling on}

\item{actTar}{The column number where the actual target variable is located (in binary form)}

\item{predTar}{The column number where the predicted values are located}

\item{tpProfit}{This is the utility for generating a true positive prediction}

\item{tnProfit}{This is the utility for generating a true negative prediction}

\item{fpProfit}{This is the cost of generating a false positive prediction}

\item{fnProfit}{This is the cost of generating a false negative prediction}

\item{MidTierDoNothing}{Set to TRUE if there is a mid-range of predicted probabilities such that you do something else with those}

\item{MidTierCost}{Set cost value for these outcomes}

\item{MidTierLowThresh}{The lower bound probability for the do something else range}

\item{MidTeirHighThresh}{The upper bound probability for the do something else range}
}
\value{
Optimal threshold and corresponding utilities for the range of thresholds tested
}
\description{
This function will return the utility maximizing threshold for future predictions along with the data generated to estimate the threshold
}
\examples{
test <- data.table(actual = ifelse(runif(1000) > 0.5,1,0),target = runif(1000))
data <- threshOptim(test,
                    actTar   = 1,
                    predTar  = 2,
                    tpProfit = 1,
                    tnProfit = 5,
                    fpProfit = -1,
                    fnProfit = -1,
                    MidTierDoNothing = FALSE,
                    MidTierCost = -100,
                    MidTierLowThresh = 0.25,
                    MidTierHighThresh = 0.75)
optimalThreshold <- data[[1]]
allResults       <- data[[2]]
}
\author{
Adrian Antico
}
