% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Adrian_Antico_Modeling_Tools_Library.R
\name{threshOptim}
\alias{threshOptim}
\title{Utility maximizing thresholds for binary classification}
\usage{
threshOptim(data, actTar = "target", predTar = "p1", tpProfit = 0,
  tnProfit = 0, fpProfit = -1, fnProfit = -2)
}
\arguments{
\item{data}{data is the data table you are building the modeling on}

\item{actTar}{The column name where the actual target variable is located (in binary form)}

\item{predTar}{The column name where the predicted values are located}

\item{tpProfit}{This is the utility for generating a true positive prediction}

\item{tnProfit}{This is the utility for generating a true negative prediction}

\item{fpProfit}{This is the cost of generating a false positive prediction}

\item{fnProfit}{This is the cost of generating a false negative prediction}
}
\value{
Optimal threshold and corresponding utilities for the range of thresholds tested
}
\description{
This function will return the utility maximizing threshold for future predictions along with the data generated to estimate the threshold
}
\examples{
# data <- threshOptim(data     = data,
#                      actTar   = "target",
#                      predTar  = "p1",
#                      tpProfit = 0,
#                      tnProfit = 0,
#                      fpProfit = -1,
#                      fnProfit = -2)
#  optimalThreshold <- data[[1]]
#  allResults       <- data[[2]]
}
\seealso{
Other Model Evaluation and Interpretation: \code{\link{EvalPlot}},
  \code{\link{ParDepCalPlots}},
  \code{\link{RedYellowGreen}}
}
\author{
Adrian Antico
}
\concept{Model Evaluation and Interpretation}
