% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeSeriesFill.R
\name{TimeSeriesFill}
\alias{TimeSeriesFill}
\title{TimeSeriesFill For Completing Time Series Data}
\usage{
TimeSeriesFill(data = data, DateColumnName = "Date",
  GroupVariables = NULL, TimeUnit = "days", FillType = "all")
}
\arguments{
\item{data}{Supply your full series data set here}

\item{DateColumnName}{Supply the name of your date column}

\item{GroupVariables}{Supply the column names of your group variables. E.g. "Group" or c("Group1","Group2")}

\item{TimeUnit}{Choose from "second", "minute", "hour", "day", "week", "month", "quarter", "year"}

\item{FillType}{Choose from "all" or "inner". Only relevant for when you have GroupVariables. The "all" option will take the max date and the min date of the entire data set and fill according to those. The "inner" option will grab the max and min dates by group levels and fill each group level based on those.}
}
\value{
Returns a data table with missing time series records filled (currently just zeros)
}
\description{
TimeSeriesFill For Completing Time Series Data For Single Series or Time Series by Group
}
\examples{
\donttest{
data <- TimeSeriesFill(data,
                       DateColumnName = "Date",
                       GroupVariables = "GroupVar",
                       TimeUnit = "days",
                       FillType = "inner")
}
}
\seealso{
Other Feature Engineering: \code{\link{AutoDataPartition}},
  \code{\link{AutoTransformationCreate}},
  \code{\link{AutoTransformationScore}},
  \code{\link{AutoWord2VecModeler}},
  \code{\link{CreateCalendarVariables}},
  \code{\link{CreateHolidayVariables}},
  \code{\link{DT_GDL_Feature_Engineering}},
  \code{\link{DummifyDT}},
  \code{\link{GDL_Feature_Engineering}},
  \code{\link{ModelDataPrep}},
  \code{\link{Partial_DT_GDL_Feature_Engineering}},
  \code{\link{Scoring_GDL_Feature_Engineering}}
}
\concept{Feature Engineering}
