% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeSeriesFill.R
\name{TimeSeriesFill}
\alias{TimeSeriesFill}
\title{TimeSeriesFill For Completing Time Series Data}
\usage{
TimeSeriesFill(
  data = data,
  DateColumnName = "Date",
  GroupVariables = c("Store", "Dept"),
  TimeUnit = "weeks",
  FillType = c("maxmax", "minmax", "maxmin", "minmin"),
  MaxMissingPercent = 0.05,
  SimpleImpute = FALSE
)
}
\arguments{
\item{data}{Supply your full series data set here}

\item{DateColumnName}{Supply the name of your date column}

\item{GroupVariables}{Supply the column names of your group variables. E.g. "Group" or c("Group1","Group2")}

\item{TimeUnit}{Choose from "second", "minute", "hour", "day", "week", "month", "quarter", "year"}

\item{FillType}{Choose from maxmax - Fill from the absolute min date to the absolute max date, minmax - Fill from the max date of the min set to the absolute max date, maxmin - Fill from the absolute min date to the min of the max dates, or minmin - Fill from the max date of the min dates to the min date of the max dates}

\item{MaxMissingPercent}{The maximum amount of missing values an individual series can have to remain and be imputed. Otherwise, they are discarded.}

\item{SimpleImpute}{Set to TRUE or FALSE. With TRUE numeric cols will fill NAs with a -1 and non-numeric cols with a "0"}
}
\value{
Returns a data table with missing time series records filled (currently just zeros)
}
\description{
TimeSeriesFill For Completing Time Series Data For Single Series or Time Series by Group
}
\examples{
\dontrun{

# Pull in data
data <- data <- data.table::fread("https://www.dropbox.com/s/2str3ek4f4cheqi/walmart_train.csv?dl=1")

# Run function
data <- TimeSeriesFill(
  data,
  DateColumnName = "Date",
  GroupVariables = c("Store","Dept"),
  TimeUnit = "weeks",
  FillType = "maxmax",
  SimpleImpute = FALSE)
}
}
\seealso{
Other Feature Engineering: 
\code{\link{AutoDataPartition}()},
\code{\link{AutoHierarchicalFourier}()},
\code{\link{AutoLagRollStatsScoring}()},
\code{\link{AutoLagRollStats}()},
\code{\link{AutoTransformationCreate}()},
\code{\link{AutoTransformationScore}()},
\code{\link{AutoWord2VecModeler}()},
\code{\link{ContinuousTimeDataGenerator}()},
\code{\link{CreateCalendarVariables}()},
\code{\link{CreateHolidayVariables}()},
\code{\link{DT_GDL_Feature_Engineering}()},
\code{\link{DummifyDT}()},
\code{\link{H2oAutoencoder}()},
\code{\link{ModelDataPrep}()},
\code{\link{Partial_DT_GDL_Feature_Engineering}()}
}
\concept{Feature Engineering}
