% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RemixAutoML.R
\name{tokenizeH20}
\alias{tokenizeH20}
\title{For NLP work}
\usage{
tokenizeH20(data)
}
\arguments{
\item{data}{The text data}
}
\description{
This function tokenizes data
}
\seealso{
Other Misc: \code{\link{AutoH20TextPrepScoring}},
  \code{\link{ChartTheme}}, \code{\link{PrintObjectsSize}},
  \code{\link{RemixTheme}}, \code{\link{SimpleCap}},
  \code{\link{WordFreq}}, \code{\link{multiplot}},
  \code{\link{percRank}}, \code{\link{tempDatesFun}}
}
\author{
Adrian Antico
}
\concept{Misc}
