% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Adrian_Antico_Modeling_Tools_Library.R
\name{Word2VecModel}
\alias{Word2VecModel}
\title{Automated word2vec data generation via H20}
\usage{
Word2VecModel(datax, stringCol = c("Q65_Ans", "Q66_Ans", "Q67_Ans",
  "Q69_Ans", "Q70_Ans", "Q71_Ans", "Q72_Ans", "Q73_Ans", "Q74_Ans",
  "Q75_Ans", "Q76_Ans", "Q77_Ans", "Q78_Ans", "Q79_Ans", "Q80_Ans",
  "Q81_Ans", "Q82_Ans", "Q83_Ans"), KeepStringCol = FALSE,
  model_path = getwd(), ModelID = c("Q65_Ans", "Q66_Ans", "Q67_Ans",
  "Q69_Ans", "Q70_Ans", "Q71_Ans", "Q72_Ans", "Q73_Ans", "Q74_Ans",
  "Q75_Ans", "Q76_Ans", "Q77_Ans", "Q78_Ans", "Q79_Ans", "Q80_Ans",
  "Q81_Ans", "Q82_Ans", "Q83_Ans"), vects = 5, SaveStopWords = FALSE,
  MinWords = 1, WindowSize = 1, Epochs = 25, StopWords = NULL,
  SaveModel = "standard")
}
\arguments{
\item{stringCol}{A string name for the column to convert via word2vec}

\item{KeepStringCol}{Set to TRUE if you want to keep the original string column that you convert via word2vec}

\item{model_path}{A string path to the location where you want the model and metadata stored}

\item{ModelID}{A vector of your model names}

\item{vects}{The number of vectors to retain from the word2vec model}

\item{SaveStopWords}{Set to TRUE to save the stop words used}

\item{MinWords}{For H20 word2vec model}

\item{WindowSize}{For H20 word2vec model}

\item{Epochs}{For H20 word2vec model}

\item{StopWords}{For H20 word2vec model}

\item{SaveModel}{Set to "standard" to save normally; set to "mojo" to save as mojo}

\item{data}{Source data table to merge vects onto}
}
\description{
This function allows you to automatically build a word2vec model and merge the data onto your supplied dataset
}
\examples{
Word2VecModel(data,
             stringCol     = "Comment",
             KeepStringCol = FALSE,
             model_path    = getwd(),
             ModelID       = "bla",
             vects         = 50,
             SaveStopWords = FALSE,
             MinWords      = 1,
             WindowSize    = 1,
             Epochs        = 25,
             StopWords     = NULL)
}
\author{
Adrian Antico
}
