% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReinforcementLearningFunctions.R
\name{XGBoostRegressionMetrics}
\alias{XGBoostRegressionMetrics}
\title{XGBoostRegressionMetrics}
\usage{
XGBoostRegressionMetrics(grid_eval_metric, MinVal, calibEval)
}
\arguments{
\item{grid_eval_metric}{Passthrough}

\item{MinVal}{= -1L,}

\item{calibEval}{Passthrough}
}
\description{
XGBoostRegressionMetrics
}
\seealso{
Other Supervised Learning: 
\code{\link{AutoH2OScoring}()},
\code{\link{CatBoostClassifierParams}()},
\code{\link{CatBoostMultiClassParams}()},
\code{\link{CatBoostParameterGrids}()},
\code{\link{CatBoostRegressionParams}()},
\code{\link{XGBoostClassifierParams}()},
\code{\link{XGBoostMultiClassParams}()},
\code{\link{XGBoostParameterGrids}()},
\code{\link{XGBoostRegressionParams}()}
}
\author{
Adrian Antico
}
\concept{Supervised Learning}
